Evision.AKAZE (Evision v0.1.9)

cv

compute(self, images, keypoints)

Variant 1:

Positional Arguments

images: [Evision.Mat].

Image set.

Return

keypoints: [vector_KeyPoint].

Input collection of keypoints. Keypoints for which a descriptor cannot be computed are removed. Sometimes new keypoints can be added, for example: SIFT duplicates keypoint with several dominant orientations (for each orientation).

descriptors: [Evision.Mat].

Computed descriptors. In the second variant of the method descriptors[i] are descriptors computed for a keypoints[i]. Row j is the keypoints (or keypoints[i]) is the descriptor for keypoint j-th keypoint.

Has overloading in C++

Python prototype (for reference):

compute(images, keypoints[, descriptors]) -> keypoints, descriptors

Variant 2:

Computes the descriptors for a set of keypoints detected in an image (first variant) or image set (second variant).

Positional Arguments

image: Evision.Mat.

Image.

Return

keypoints: [KeyPoint].

Input collection of keypoints. Keypoints for which a descriptor cannot be computed are removed. Sometimes new keypoints can be added, for example: SIFT duplicates keypoint with several dominant orientations (for each orientation).

descriptors: Evision.Mat.

Computed descriptors. In the second variant of the method descriptors[i] are descriptors computed for a keypoints[i]. Row j is the keypoints (or keypoints[i]) is the descriptor for keypoint j-th keypoint.

Python prototype (for reference):

compute(image, keypoints[, descriptors]) -> keypoints, descriptors

compute(self, images, keypoints, opts)

Variant 1:

Positional Arguments

images: [Evision.Mat].

Image set.

Return

keypoints: [vector_KeyPoint].

Input collection of keypoints. Keypoints for which a descriptor cannot be computed are removed. Sometimes new keypoints can be added, for example: SIFT duplicates keypoint with several dominant orientations (for each orientation).

descriptors: [Evision.Mat].

Computed descriptors. In the second variant of the method descriptors[i] are descriptors computed for a keypoints[i]. Row j is the keypoints (or keypoints[i]) is the descriptor for keypoint j-th keypoint.

Has overloading in C++

Python prototype (for reference):

compute(images, keypoints[, descriptors]) -> keypoints, descriptors

Variant 2:

Computes the descriptors for a set of keypoints detected in an image (first variant) or image set (second variant).

Positional Arguments

image: Evision.Mat.

Image.

Return

keypoints: [KeyPoint].

Input collection of keypoints. Keypoints for which a descriptor cannot be computed are removed. Sometimes new keypoints can be added, for example: SIFT duplicates keypoint with several dominant orientations (for each orientation).

descriptors: Evision.Mat.

Computed descriptors. In the second variant of the method descriptors[i] are descriptors computed for a keypoints[i]. Row j is the keypoints (or keypoints[i]) is the descriptor for keypoint j-th keypoint.

Python prototype (for reference):

compute(image, keypoints[, descriptors]) -> keypoints, descriptors

create()

The AKAZE constructor

Keyword Arguments

descriptor_type: AKAZE_DescriptorType.

Type of the extracted descriptor: DESCRIPTOR_KAZE, DESCRIPTOR_KAZE_UPRIGHT, DESCRIPTOR_MLDB or DESCRIPTOR_MLDB_UPRIGHT.

descriptor_size: int.

Size of the descriptor in bits. 0 -> Full size

descriptor_channels: int.

Number of channels in the descriptor (1, 2, 3)

threshold: float.

Detector response threshold to accept point

nOctaves: int.

Maximum octave evolution of the image

nOctaveLayers: int.

Default number of sublevels per scale level

diffusivity: KAZE_DiffusivityType.

Diffusivity type. DIFF_PM_G1, DIFF_PM_G2, DIFF_WEICKERT or DIFF_CHARBONNIER

Python prototype (for reference):

create([, descriptor_type[, descriptor_size[, descriptor_channels[, threshold[, nOctaves[, nOctaveLayers[, diffusivity]]]]]]]) -> retval

create(opts)

The AKAZE constructor

Keyword Arguments

descriptor_type: AKAZE_DescriptorType.

Type of the extracted descriptor: DESCRIPTOR_KAZE, DESCRIPTOR_KAZE_UPRIGHT, DESCRIPTOR_MLDB or DESCRIPTOR_MLDB_UPRIGHT.

descriptor_size: int.

Size of the descriptor in bits. 0 -> Full size

descriptor_channels: int.

Number of channels in the descriptor (1, 2, 3)

threshold: float.

Detector response threshold to accept point

nOctaves: int.

Maximum octave evolution of the image

nOctaveLayers: int.

Default number of sublevels per scale level

diffusivity: KAZE_DiffusivityType.

Diffusivity type. DIFF_PM_G1, DIFF_PM_G2, DIFF_WEICKERT or DIFF_CHARBONNIER

Python prototype (for reference):

create([, descriptor_type[, descriptor_size[, descriptor_channels[, threshold[, nOctaves[, nOctaveLayers[, diffusivity]]]]]]]) -> retval

defaultNorm(self)

Python prototype (for reference):

defaultNorm() -> retval

descriptorSize(self)

Python prototype (for reference):

descriptorSize() -> retval

descriptorType(self)

Python prototype (for reference):

descriptorType() -> retval

detect(self, images)

Variant 1:

Positional Arguments

images: [Evision.Mat].

Image set.

Keyword Arguments

masks: [Evision.Mat].

Masks for each input image specifying where to look for keypoints (optional). masks[i] is a mask for images[i].

Return

keypoints: [vector_KeyPoint].

The detected keypoints. In the second variant of the method keypoints[i] is a set of keypoints detected in images[i] .

Has overloading in C++

Python prototype (for reference):

detect(images[, masks]) -> keypoints

Variant 2:

Detects keypoints in an image (first variant) or image set (second variant).

Positional Arguments

image: Evision.Mat.

Image.

Keyword Arguments

mask: Evision.Mat.

Mask specifying where to look for keypoints (optional). It must be a 8-bit integer matrix with non-zero values in the region of interest.

Return

keypoints: [KeyPoint].

The detected keypoints. In the second variant of the method keypoints[i] is a set of keypoints detected in images[i] .

Python prototype (for reference):

detect(image[, mask]) -> keypoints

detect(self, images, opts)

Variant 1:

Positional Arguments

images: [Evision.Mat].

Image set.

Keyword Arguments

masks: [Evision.Mat].

Masks for each input image specifying where to look for keypoints (optional). masks[i] is a mask for images[i].

Return

keypoints: [vector_KeyPoint].

The detected keypoints. In the second variant of the method keypoints[i] is a set of keypoints detected in images[i] .

Has overloading in C++

Python prototype (for reference):

detect(images[, masks]) -> keypoints

Variant 2:

Detects keypoints in an image (first variant) or image set (second variant).

Positional Arguments

image: Evision.Mat.

Image.

Keyword Arguments

mask: Evision.Mat.

Mask specifying where to look for keypoints (optional). It must be a 8-bit integer matrix with non-zero values in the region of interest.

Return

keypoints: [KeyPoint].

The detected keypoints. In the second variant of the method keypoints[i] is a set of keypoints detected in images[i] .

Python prototype (for reference):

detect(image[, mask]) -> keypoints

detectAndCompute(self, image, mask)

Positional Arguments

image: Evision.Mat

mask: Evision.Mat

Keyword Arguments

useProvidedKeypoints: bool.

Return

keypoints: [KeyPoint]

descriptors: Evision.Mat.

Detects keypoints and computes the descriptors

Python prototype (for reference):

detectAndCompute(image, mask[, descriptors[, useProvidedKeypoints]]) -> keypoints, descriptors

detectAndCompute(self, image, mask, opts)

Positional Arguments

image: Evision.Mat

mask: Evision.Mat

Keyword Arguments

useProvidedKeypoints: bool.

Return

keypoints: [KeyPoint]

descriptors: Evision.Mat.

Detects keypoints and computes the descriptors

Python prototype (for reference):

detectAndCompute(image, mask[, descriptors[, useProvidedKeypoints]]) -> keypoints, descriptors

empty(self)

Python prototype (for reference):

empty() -> retval

getDefaultName(self)

Python prototype (for reference):

getDefaultName() -> retval

getDescriptorChannels(self)

Python prototype (for reference):

getDescriptorChannels() -> retval

getDescriptorSize(self)

Python prototype (for reference):

getDescriptorSize() -> retval

getDescriptorType(self)

Python prototype (for reference):

getDescriptorType() -> retval

getDiffusivity(self)

Python prototype (for reference):

getDiffusivity() -> retval

getNOctaveLayers(self)

Python prototype (for reference):

getNOctaveLayers() -> retval

getNOctaves(self)

Python prototype (for reference):

getNOctaves() -> retval

getThreshold(self)

Python prototype (for reference):

getThreshold() -> retval

read(self, fileName)

Variant 1:

Positional Arguments

fileName: String

Python prototype (for reference):

read(fileName) -> None

Variant 2:

Positional Arguments

arg1: FileNode

Python prototype (for reference):

read(arg1) -> None

setDescriptorChannels(self, dch)

Positional Arguments

dch: int

Python prototype (for reference):

setDescriptorChannels(dch) -> None

setDescriptorSize(self, dsize)

Positional Arguments

dsize: int

Python prototype (for reference):

setDescriptorSize(dsize) -> None

setDescriptorType(self, dtype)

Positional Arguments

dtype: AKAZE_DescriptorType

Python prototype (for reference):

setDescriptorType(dtype) -> None

setDiffusivity(self, diff)

Positional Arguments

diff: KAZE_DiffusivityType

Python prototype (for reference):

setDiffusivity(diff) -> None

setNOctaveLayers(self, octaveLayers)

Positional Arguments

octaveLayers: int

Python prototype (for reference):

setNOctaveLayers(octaveLayers) -> None

setNOctaves(self, octaves)

Positional Arguments

octaves: int

Python prototype (for reference):

setNOctaves(octaves) -> None

setThreshold(self, threshold)

Positional Arguments

threshold: double

Python prototype (for reference):

setThreshold(threshold) -> None

write(self, fs)

Variant 1:

Positional Arguments

fs: Ptr<FileStorage>

Keyword Arguments

name: String.

Python prototype (for reference):

write(fs[, name]) -> None

Variant 2:

Positional Arguments

fileName: String

Python prototype (for reference):

write(fileName) -> None

write(self, fs, opts)

Positional Arguments

fs: Ptr<FileStorage>

Keyword Arguments

name: String.

Python prototype (for reference):

write(fs[, name]) -> None

Functions

compute!(self, images, keypoints)

Raising version of compute/3.

compute!(self, images, keypoints, opts)

Raising version of compute/4.

create!()

Raising version of create/0.

create!(opts)

Raising version of create/1.

defaultNorm!(self)

Raising version of defaultNorm/1.

descriptorSize!(self)

Raising version of descriptorSize/1.

descriptorType!(self)

Raising version of descriptorType/1.

detect!(self, images)

Raising version of detect/2.

detect!(self, images, opts)

Raising version of detect/3.

detectAndCompute!(self, image, mask)

Raising version of detectAndCompute/3.

detectAndCompute!(self, image, mask, opts)

Raising version of detectAndCompute/4.

empty!(self)

Raising version of empty/1.

getDefaultName!(self)

Raising version of getDefaultName/1.

getDescriptorChannels!(self)

Raising version of getDescriptorChannels/1.

getDescriptorSize!(self)

Raising version of getDescriptorSize/1.

getDescriptorType!(self)

Raising version of getDescriptorType/1.

getDiffusivity!(self)

Raising version of getDiffusivity/1.

getNOctaveLayers!(self)

Raising version of getNOctaveLayers/1.

getNOctaves!(self)

Raising version of getNOctaves/1.

getThreshold!(self)

Raising version of getThreshold/1.

read!(self, fileName)

Raising version of read/2.

setDescriptorChannels!(self, dch)

Raising version of setDescriptorChannels/2.

setDescriptorSize!(self, dsize)

Raising version of setDescriptorSize/2.

setDescriptorType!(self, dtype)

Raising version of setDescriptorType/2.

setDiffusivity!(self, diff)

Raising version of setDiffusivity/2.

setNOctaveLayers!(self, octaveLayers)

Raising version of setNOctaveLayers/2.

setNOctaves!(self, octaves)

Raising version of setNOctaves/2.

setThreshold!(self, threshold)

Raising version of setThreshold/2.

write!(self, fs)

Raising version of write/2.

write!(self, fs, opts)

Raising version of write/3.