Evision.AlignMTB (Evision v0.1.9)

cv

calculateShift(self, img0, img1)

Calculates shift between two images, i. e. how to shift the second image to correspond it with the first.

Positional Arguments

img0: Evision.Mat.

first image

img1: Evision.Mat.

second image

Python prototype (for reference):

calculateShift(img0, img1) -> retval

computeBitmaps(self, img)

Computes median threshold and exclude bitmaps of given image.

Positional Arguments

img: Evision.Mat.

input image

Return

tb: Evision.Mat.

median threshold bitmap

eb: Evision.Mat.

exclude bitmap

Python prototype (for reference):

computeBitmaps(img[, tb[, eb]]) -> tb, eb

computeBitmaps(self, img, opts)

Computes median threshold and exclude bitmaps of given image.

Positional Arguments

img: Evision.Mat.

input image

Return

tb: Evision.Mat.

median threshold bitmap

eb: Evision.Mat.

exclude bitmap

Python prototype (for reference):

computeBitmaps(img[, tb[, eb]]) -> tb, eb

getCut(self)

Python prototype (for reference):

getCut() -> retval

getExcludeRange(self)

Python prototype (for reference):

getExcludeRange() -> retval

getMaxBits(self)

Python prototype (for reference):

getMaxBits() -> retval

process(self, src, dst)

Short version of process, that doesn't take extra arguments.

Positional Arguments

src: [Evision.Mat].

vector of input images

dst: [Evision.Mat].

vector of aligned images

Python prototype (for reference):

process(src, dst) -> None

process(self, src, dst, times, response)

Positional Arguments

src: [Evision.Mat]

dst: [Evision.Mat]

times: Evision.Mat

response: Evision.Mat

Python prototype (for reference):

process(src, dst, times, response) -> None

setCut(self, value)

Positional Arguments

value: bool

Python prototype (for reference):

setCut(value) -> None

setExcludeRange(self, exclude_range)

Positional Arguments

exclude_range: int

Python prototype (for reference):

setExcludeRange(exclude_range) -> None

setMaxBits(self, max_bits)

Positional Arguments

max_bits: int

Python prototype (for reference):

setMaxBits(max_bits) -> None

shiftMat(self, src, shift)

Helper function, that shift Mat filling new regions with zeros.

Positional Arguments

src: Evision.Mat.

input image

shift: Point.

shift value

Return

dst: Evision.Mat.

result image

Python prototype (for reference):

shiftMat(src, shift[, dst]) -> dst

shiftMat(self, src, shift, opts)

Helper function, that shift Mat filling new regions with zeros.

Positional Arguments

src: Evision.Mat.

input image

shift: Point.

shift value

Return

dst: Evision.Mat.

result image

Python prototype (for reference):

shiftMat(src, shift[, dst]) -> dst

Functions

calculateShift!(self, img0, img1)

Raising version of calculateShift/3.

computeBitmaps!(self, img)

Raising version of computeBitmaps/2.

computeBitmaps!(self, img, opts)

Raising version of computeBitmaps/3.

getCut!(self)

Raising version of getCut/1.

getExcludeRange!(self)

Raising version of getExcludeRange/1.

getMaxBits!(self)

Raising version of getMaxBits/1.

process!(self, src, dst)

Raising version of process/3.

process!(self, src, dst, times, response)

Raising version of process/5.

setCut!(self, value)

Raising version of setCut/2.

setExcludeRange!(self, exclude_range)

Raising version of setExcludeRange/2.

setMaxBits!(self, max_bits)

Raising version of setMaxBits/2.

shiftMat!(self, src, shift)

Raising version of shiftMat/3.

shiftMat!(self, src, shift, opts)

Raising version of shiftMat/4.