Evision.AsyncArray (Evision v0.1.7-dev)

cv

asyncArray()

Python prototype (for reference):

AsyncArray() -> <AsyncArray object>

get(self)

Positional Arguments

Keyword Arguments

dst: UMat.

Fetch the result. Waits for result until container has valid result. Throws exception if exception was stored as a result. Throws exception on invalid container state.

Note: Result or stored exception can be fetched only once.

Python prototype (for reference):

get([, dst]) -> dst

get(self, timeoutNs)

Positional Arguments

timeoutNs: double.

Keyword Arguments

dst: UMat.

Retrieving the result with timeout

@returns true if result is ready, false if the timeout has expired Note: Result or stored exception can be fetched only once.

Python prototype (for reference):

get(timeoutNs[, dst]) -> retval, dst

get(self, timeoutNs, opts)

Positional Arguments

timeoutNs: double.

Keyword Arguments

dst: UMat.

Retrieving the result with timeout

@returns true if result is ready, false if the timeout has expired Note: Result or stored exception can be fetched only once.

Python prototype (for reference):

get(timeoutNs[, dst]) -> retval, dst

release(self)

Python prototype (for reference):

release() -> None

valid(self)

Python prototype (for reference):

valid() -> retval

wait_for(self, timeoutNs)

Python prototype (for reference):

wait_for(timeoutNs) -> retval

Functions

asyncArray!()

Raising version of asyncArray/0.

get!(self)

Raising version of get/1.

get!(self, timeoutNs)

Raising version of get/2.

get!(self, timeoutNs, opts)

Raising version of get/3.

release!(self)

Raising version of release/1.

valid!(self)

Raising version of valid/1.

wait_for!(self, timeoutNs)

Raising version of wait_for/2.