Evision.BFMatcher (Evision v0.1.9)

cv

add(self, descriptors)

Adds descriptors to train a CPU(trainDescCollectionis) or GPU(utrainDescCollectionis) descriptor collection.

Positional Arguments

descriptors: [Evision.Mat].

Descriptors to add. Each descriptors[i] is a set of descriptors from the same train image.

If the collection is not empty, the new descriptors are added to existing train descriptors.

Python prototype (for reference):

add(descriptors) -> None

bfMatcher()

Brute-force matcher constructor (obsolete). Please use BFMatcher.create()

Keyword Arguments

normType: int.

crossCheck: bool.

Python prototype (for reference):

BFMatcher([, normType[, crossCheck]]) -> <BFMatcher object>

bfMatcher(opts)

Brute-force matcher constructor (obsolete). Please use BFMatcher.create()

Keyword Arguments

normType: int.

crossCheck: bool.

Python prototype (for reference):

BFMatcher([, normType[, crossCheck]]) -> <BFMatcher object>

clear(self)

Clears the train descriptor collections.

Python prototype (for reference):

clear() -> None

clone(self)

Clones the matcher.

Keyword Arguments

emptyTrainData: bool.

If emptyTrainData is false, the method creates a deep copy of the object, that is, copies both parameters and train data. If emptyTrainData is true, the method creates an object copy with the current parameters but with empty train data.

Python prototype (for reference):

clone([, emptyTrainData]) -> retval

create()

Brute-force matcher create method.

Keyword Arguments

normType: int.

One of NORM_L1, NORM_L2, NORM_HAMMING, NORM_HAMMING2. L1 and L2 norms are preferable choices for SIFT and SURF descriptors, NORM_HAMMING should be used with ORB, BRISK and BRIEF, NORM_HAMMING2 should be used with ORB when WTA_K==3 or 4 (see ORB::ORB constructor description).

crossCheck: bool.

If it is false, this is will be default BFMatcher behaviour when it finds the k nearest neighbors for each query descriptor. If crossCheck==true, then the knnMatch() method with k=1 will only return pairs (i,j) such that for i-th query descriptor the j-th descriptor in the matcher's collection is the nearest and vice versa, i.e. the BFMatcher will only return consistent pairs. Such technique usually produces best results with minimal number of outliers when there are enough matches. This is alternative to the ratio test, used by D. Lowe in SIFT paper.

Python prototype (for reference):

create([, normType[, crossCheck]]) -> retval

create(opts)

Brute-force matcher create method.

Keyword Arguments

normType: int.

One of NORM_L1, NORM_L2, NORM_HAMMING, NORM_HAMMING2. L1 and L2 norms are preferable choices for SIFT and SURF descriptors, NORM_HAMMING should be used with ORB, BRISK and BRIEF, NORM_HAMMING2 should be used with ORB when WTA_K==3 or 4 (see ORB::ORB constructor description).

crossCheck: bool.

If it is false, this is will be default BFMatcher behaviour when it finds the k nearest neighbors for each query descriptor. If crossCheck==true, then the knnMatch() method with k=1 will only return pairs (i,j) such that for i-th query descriptor the j-th descriptor in the matcher's collection is the nearest and vice versa, i.e. the BFMatcher will only return consistent pairs. Such technique usually produces best results with minimal number of outliers when there are enough matches. This is alternative to the ratio test, used by D. Lowe in SIFT paper.

Python prototype (for reference):

create([, normType[, crossCheck]]) -> retval

empty(self)

Returns true if there are no train descriptors in the both collections.

Python prototype (for reference):

empty() -> retval

getTrainDescriptors(self)

Returns a constant link to the train descriptor collection trainDescCollection .

Python prototype (for reference):

getTrainDescriptors() -> retval

isMaskSupported(self)

Returns true if the descriptor matcher supports masking permissible matches.

Python prototype (for reference):

isMaskSupported() -> retval

knnMatch(self, queryDescriptors, k)

Positional Arguments

queryDescriptors: Evision.Mat.

Query set of descriptors.

k: int.

Count of best matches found per each query descriptor or less if a query descriptor has less than k possible matches in total.

Keyword Arguments

masks: [Evision.Mat].

Set of masks. Each masks[i] specifies permissible matches between the input query descriptors and stored train descriptors from the i-th image trainDescCollection[i].

compactResult: bool.

Parameter used when the mask (or masks) is not empty. If compactResult is false, the matches vector has the same size as queryDescriptors rows. If compactResult is true, the matches vector does not contain matches for fully masked-out query descriptors.

Return

matches: [vector_DMatch].

Matches. Each matches[i] is k or less matches for the same query descriptor.

Has overloading in C++

Python prototype (for reference):

knnMatch(queryDescriptors, k[, masks[, compactResult]]) -> matches

knnMatch(self, queryDescriptors, k, opts)

Variant 1:

Finds the k best matches for each descriptor from a query set.

Positional Arguments

queryDescriptors: Evision.Mat.

Query set of descriptors.

trainDescriptors: Evision.Mat.

Train set of descriptors. This set is not added to the train descriptors collection stored in the class object.

k: int.

Count of best matches found per each query descriptor or less if a query descriptor has less than k possible matches in total.

Keyword Arguments

mask: Evision.Mat.

Mask specifying permissible matches between an input query and train matrices of descriptors.

compactResult: bool.

Parameter used when the mask (or masks) is not empty. If compactResult is false, the matches vector has the same size as queryDescriptors rows. If compactResult is true, the matches vector does not contain matches for fully masked-out query descriptors.

Return

matches: [vector_DMatch].

Matches. Each matches[i] is k or less matches for the same query descriptor.

These extended variants of DescriptorMatcher::match methods find several best matches for each query descriptor. The matches are returned in the distance increasing order. See DescriptorMatcher::match for the details about query and train descriptors.

Python prototype (for reference):

knnMatch(queryDescriptors, trainDescriptors, k[, mask[, compactResult]]) -> matches

Variant 2:

Positional Arguments

queryDescriptors: Evision.Mat.

Query set of descriptors.

k: int.

Count of best matches found per each query descriptor or less if a query descriptor has less than k possible matches in total.

Keyword Arguments

masks: [Evision.Mat].

Set of masks. Each masks[i] specifies permissible matches between the input query descriptors and stored train descriptors from the i-th image trainDescCollection[i].

compactResult: bool.

Parameter used when the mask (or masks) is not empty. If compactResult is false, the matches vector has the same size as queryDescriptors rows. If compactResult is true, the matches vector does not contain matches for fully masked-out query descriptors.

Return

matches: [vector_DMatch].

Matches. Each matches[i] is k or less matches for the same query descriptor.

Has overloading in C++

Python prototype (for reference):

knnMatch(queryDescriptors, k[, masks[, compactResult]]) -> matches

knnMatch(self, queryDescriptors, trainDescriptors, k, opts)

Finds the k best matches for each descriptor from a query set.

Positional Arguments

queryDescriptors: Evision.Mat.

Query set of descriptors.

trainDescriptors: Evision.Mat.

Train set of descriptors. This set is not added to the train descriptors collection stored in the class object.

k: int.

Count of best matches found per each query descriptor or less if a query descriptor has less than k possible matches in total.

Keyword Arguments

mask: Evision.Mat.

Mask specifying permissible matches between an input query and train matrices of descriptors.

compactResult: bool.

Parameter used when the mask (or masks) is not empty. If compactResult is false, the matches vector has the same size as queryDescriptors rows. If compactResult is true, the matches vector does not contain matches for fully masked-out query descriptors.

Return

matches: [vector_DMatch].

Matches. Each matches[i] is k or less matches for the same query descriptor.

These extended variants of DescriptorMatcher::match methods find several best matches for each query descriptor. The matches are returned in the distance increasing order. See DescriptorMatcher::match for the details about query and train descriptors.

Python prototype (for reference):

knnMatch(queryDescriptors, trainDescriptors, k[, mask[, compactResult]]) -> matches

match(self, queryDescriptors)

Positional Arguments

queryDescriptors: Evision.Mat.

Query set of descriptors.

Keyword Arguments

masks: [Evision.Mat].

Set of masks. Each masks[i] specifies permissible matches between the input query descriptors and stored train descriptors from the i-th image trainDescCollection[i].

Return

matches: [DMatch].

Matches. If a query descriptor is masked out in mask , no match is added for this descriptor. So, matches size may be smaller than the query descriptors count.

Has overloading in C++

Python prototype (for reference):

match(queryDescriptors[, masks]) -> matches

match(self, queryDescriptors, opts)

Variant 1:

Finds the best match for each descriptor from a query set.

Positional Arguments

queryDescriptors: Evision.Mat.

Query set of descriptors.

trainDescriptors: Evision.Mat.

Train set of descriptors. This set is not added to the train descriptors collection stored in the class object.

Keyword Arguments

mask: Evision.Mat.

Mask specifying permissible matches between an input query and train matrices of descriptors.

Return

matches: [DMatch].

Matches. If a query descriptor is masked out in mask , no match is added for this descriptor. So, matches size may be smaller than the query descriptors count.

In the first variant of this method, the train descriptors are passed as an input argument. In the second variant of the method, train descriptors collection that was set by DescriptorMatcher::add is used. Optional mask (or masks) can be passed to specify which query and training descriptors can be matched. Namely, queryDescriptors[i] can be matched with trainDescriptors[j] only if mask.at\<uchar>(i,j) is non-zero.

Python prototype (for reference):

match(queryDescriptors, trainDescriptors[, mask]) -> matches

Variant 2:

Positional Arguments

queryDescriptors: Evision.Mat.

Query set of descriptors.

Keyword Arguments

masks: [Evision.Mat].

Set of masks. Each masks[i] specifies permissible matches between the input query descriptors and stored train descriptors from the i-th image trainDescCollection[i].

Return

matches: [DMatch].

Matches. If a query descriptor is masked out in mask , no match is added for this descriptor. So, matches size may be smaller than the query descriptors count.

Has overloading in C++

Python prototype (for reference):

match(queryDescriptors[, masks]) -> matches

match(self, queryDescriptors, trainDescriptors, opts)

Finds the best match for each descriptor from a query set.

Positional Arguments

queryDescriptors: Evision.Mat.

Query set of descriptors.

trainDescriptors: Evision.Mat.

Train set of descriptors. This set is not added to the train descriptors collection stored in the class object.

Keyword Arguments

mask: Evision.Mat.

Mask specifying permissible matches between an input query and train matrices of descriptors.

Return

matches: [DMatch].

Matches. If a query descriptor is masked out in mask , no match is added for this descriptor. So, matches size may be smaller than the query descriptors count.

In the first variant of this method, the train descriptors are passed as an input argument. In the second variant of the method, train descriptors collection that was set by DescriptorMatcher::add is used. Optional mask (or masks) can be passed to specify which query and training descriptors can be matched. Namely, queryDescriptors[i] can be matched with trainDescriptors[j] only if mask.at\<uchar>(i,j) is non-zero.

Python prototype (for reference):

match(queryDescriptors, trainDescriptors[, mask]) -> matches

radiusMatch(self, queryDescriptors, maxDistance)

Positional Arguments

queryDescriptors: Evision.Mat.

Query set of descriptors.

maxDistance: float.

Threshold for the distance between matched descriptors. Distance means here metric distance (e.g. Hamming distance), not the distance between coordinates (which is measured in Pixels)!

Keyword Arguments

masks: [Evision.Mat].

Set of masks. Each masks[i] specifies permissible matches between the input query descriptors and stored train descriptors from the i-th image trainDescCollection[i].

compactResult: bool.

Parameter used when the mask (or masks) is not empty. If compactResult is false, the matches vector has the same size as queryDescriptors rows. If compactResult is true, the matches vector does not contain matches for fully masked-out query descriptors.

Return

matches: [vector_DMatch].

Found matches.

Has overloading in C++

Python prototype (for reference):

radiusMatch(queryDescriptors, maxDistance[, masks[, compactResult]]) -> matches

radiusMatch(self, queryDescriptors, maxDistance, opts)

Variant 1:

For each query descriptor, finds the training descriptors not farther than the specified distance.

Positional Arguments

queryDescriptors: Evision.Mat.

Query set of descriptors.

trainDescriptors: Evision.Mat.

Train set of descriptors. This set is not added to the train descriptors collection stored in the class object.

maxDistance: float.

Threshold for the distance between matched descriptors. Distance means here metric distance (e.g. Hamming distance), not the distance between coordinates (which is measured in Pixels)!

Keyword Arguments

mask: Evision.Mat.

Mask specifying permissible matches between an input query and train matrices of descriptors.

compactResult: bool.

Parameter used when the mask (or masks) is not empty. If compactResult is false, the matches vector has the same size as queryDescriptors rows. If compactResult is true, the matches vector does not contain matches for fully masked-out query descriptors.

Return

matches: [vector_DMatch].

Found matches.

For each query descriptor, the methods find such training descriptors that the distance between the query descriptor and the training descriptor is equal or smaller than maxDistance. Found matches are returned in the distance increasing order.

Python prototype (for reference):

radiusMatch(queryDescriptors, trainDescriptors, maxDistance[, mask[, compactResult]]) -> matches

Variant 2:

Positional Arguments

queryDescriptors: Evision.Mat.

Query set of descriptors.

maxDistance: float.

Threshold for the distance between matched descriptors. Distance means here metric distance (e.g. Hamming distance), not the distance between coordinates (which is measured in Pixels)!

Keyword Arguments

masks: [Evision.Mat].

Set of masks. Each masks[i] specifies permissible matches between the input query descriptors and stored train descriptors from the i-th image trainDescCollection[i].

compactResult: bool.

Parameter used when the mask (or masks) is not empty. If compactResult is false, the matches vector has the same size as queryDescriptors rows. If compactResult is true, the matches vector does not contain matches for fully masked-out query descriptors.

Return

matches: [vector_DMatch].

Found matches.

Has overloading in C++

Python prototype (for reference):

radiusMatch(queryDescriptors, maxDistance[, masks[, compactResult]]) -> matches

radiusMatch(self, queryDescriptors, trainDescriptors, maxDistance, opts)

For each query descriptor, finds the training descriptors not farther than the specified distance.

Positional Arguments

queryDescriptors: Evision.Mat.

Query set of descriptors.

trainDescriptors: Evision.Mat.

Train set of descriptors. This set is not added to the train descriptors collection stored in the class object.

maxDistance: float.

Threshold for the distance between matched descriptors. Distance means here metric distance (e.g. Hamming distance), not the distance between coordinates (which is measured in Pixels)!

Keyword Arguments

mask: Evision.Mat.

Mask specifying permissible matches between an input query and train matrices of descriptors.

compactResult: bool.

Parameter used when the mask (or masks) is not empty. If compactResult is false, the matches vector has the same size as queryDescriptors rows. If compactResult is true, the matches vector does not contain matches for fully masked-out query descriptors.

Return

matches: [vector_DMatch].

Found matches.

For each query descriptor, the methods find such training descriptors that the distance between the query descriptor and the training descriptor is equal or smaller than maxDistance. Found matches are returned in the distance increasing order.

Python prototype (for reference):

radiusMatch(queryDescriptors, trainDescriptors, maxDistance[, mask[, compactResult]]) -> matches

read(self, fileName)

Variant 1:

Positional Arguments

fileName: String

Python prototype (for reference):

read(fileName) -> None

Variant 2:

Positional Arguments

arg1: FileNode

Python prototype (for reference):

read(arg1) -> None

train(self)

Trains a descriptor matcher

Trains a descriptor matcher (for example, the flann index). In all methods to match, the method train() is run every time before matching. Some descriptor matchers (for example, BruteForceMatcher) have an empty implementation of this method. Other matchers really train their inner structures (for example, FlannBasedMatcher trains flann::Index ).

Python prototype (for reference):

train() -> None

write(self, fs)

Variant 1:

Positional Arguments

fs: Ptr<FileStorage>

Keyword Arguments

name: String.

Python prototype (for reference):

write(fs[, name]) -> None

Variant 2:

Positional Arguments

fileName: String

Python prototype (for reference):

write(fileName) -> None

write(self, fs, opts)

Positional Arguments

fs: Ptr<FileStorage>

Keyword Arguments

name: String.

Python prototype (for reference):

write(fs[, name]) -> None

Functions

add!(self, descriptors)

Raising version of add/2.

bfMatcher!()

Raising version of bfMatcher/0.

bfMatcher!(opts)

Raising version of bfMatcher/1.

clear!(self)

Raising version of clear/1.

clone!(self)

Raising version of clone/1.

create!()

Raising version of create/0.

create!(opts)

Raising version of create/1.

empty!(self)

Raising version of empty/1.

getTrainDescriptors!(self)

Raising version of getTrainDescriptors/1.

isMaskSupported!(self)

Raising version of isMaskSupported/1.

knnMatch!(self, queryDescriptors, k)

Raising version of knnMatch/3.

knnMatch!(self, queryDescriptors, trainDescriptors, k)

Raising version of knnMatch/4.

knnMatch!(self, queryDescriptors, trainDescriptors, k, opts)

Raising version of knnMatch/5.

match!(self, queryDescriptors)

Raising version of match/2.

match!(self, queryDescriptors, trainDescriptors)

Raising version of match/3.

match!(self, queryDescriptors, trainDescriptors, opts)

Raising version of match/4.

radiusMatch!(self, queryDescriptors, maxDistance)

Raising version of radiusMatch/3.

radiusMatch!(self, queryDescriptors, trainDescriptors, maxDistance)

Raising version of radiusMatch/4.

radiusMatch!(self, queryDescriptors, trainDescriptors, maxDistance, opts)

Raising version of radiusMatch/5.

read!(self, fileName)

Raising version of read/2.

train!(self)

Raising version of train/1.

write!(self, fs)

Raising version of write/2.

write!(self, fs, opts)

Raising version of write/3.