Evision.BOWTrainer (Evision v0.1.7-dev)

cv

add(self, descriptors)

Adds descriptors to a training set.

descriptor.

The training set is clustered using clustermethod to construct the vocabulary.

Python prototype (for reference):

add(descriptors) -> None

clear(self)

Python prototype (for reference):

clear() -> None

cluster(self)

Has overloading in C++

Python prototype (for reference):

cluster() -> retval

cluster(self, descriptors)

Clusters train descriptors.

Descriptors are not added to the inner train descriptor set.

The vocabulary consists of cluster centers. So, this method returns the vocabulary. In the first variant of the method, train descriptors stored in the object are clustered. In the second variant, input descriptors are clustered.

Python prototype (for reference):

cluster(descriptors) -> retval

descriptorsCount(self)

Returns the count of all descriptors stored in the training set.

Python prototype (for reference):

descriptorsCount() -> retval

getDescriptors(self)

Returns a training set of descriptors.

Python prototype (for reference):

getDescriptors() -> retval

Functions

add!(self, descriptors)

Raising version of add/2.

clear!(self)

Raising version of clear/1.

cluster!(self)

Raising version of cluster/1.

cluster!(self, descriptors)

Raising version of cluster/2.

descriptorsCount!(self)

Raising version of descriptorsCount/1.

getDescriptors!(self)

Raising version of getDescriptors/1.