Evision.BackgroundSubtractorKNN (Evision v0.1.7-dev)

cv

getDetectShadows(self)

Returns the shadow detection flag If true, the algorithm detects shadows and marks them. See createBackgroundSubtractorKNN for details.

Python prototype (for reference):

getDetectShadows() -> retval

getDist2Threshold(self)

Returns the threshold on the squared distance between the pixel and the sample The threshold on the squared distance between the pixel and the sample to decide whether a pixel is close to a data sample.

Python prototype (for reference):

getDist2Threshold() -> retval

getHistory(self)

Returns the number of last frames that affect the background model

Python prototype (for reference):

getHistory() -> retval

getkNNSamples(self)

Returns the number of neighbours, the k in the kNN. K is the number of samples that need to be within dist2Threshold in order to decide that that pixel is matching the kNN background model.

Python prototype (for reference):

getkNNSamples() -> retval

getNSamples(self)

Returns the number of data samples in the background model

Python prototype (for reference):

getNSamples() -> retval

getShadowThreshold(self)

Returns the shadow threshold A shadow is detected if pixel is a darker version of the background. The shadow threshold (Tau in the paper) is a threshold defining how much darker the shadow can be. Tau= 0.5 means that if a pixel is more than twice darker then it is not shadow. See Prati, Mikic, Trivedi and Cucchiara, Detecting Moving Shadows...*, IEEE PAMI,2003.

Python prototype (for reference):

getShadowThreshold() -> retval

getShadowValue(self)

Returns the shadow value Shadow value is the value used to mark shadows in the foreground mask. Default value is 127. Value 0 in the mask always means background, 255 means foreground.

Python prototype (for reference):

getShadowValue() -> retval

setDetectShadows(self, detectShadows)

Enables or disables shadow detection

Python prototype (for reference):

setDetectShadows(detectShadows) -> None

setDist2Threshold(self, dist2Threshold)

Sets the threshold on the squared distance

Python prototype (for reference):

setDist2Threshold(_dist2Threshold) -> None

setHistory(self, history)

Sets the number of last frames that affect the background model

Python prototype (for reference):

setHistory(history) -> None

setkNNSamples(self, nkNN)

Sets the k in the kNN. How many nearest neighbours need to match.

Python prototype (for reference):

setkNNSamples(_nkNN) -> None

setNSamples(self, nN)

Sets the number of data samples in the background model. The model needs to be reinitalized to reserve memory.

Python prototype (for reference):

setNSamples(_nN) -> None

setShadowThreshold(self, threshold)

Sets the shadow threshold

Python prototype (for reference):

setShadowThreshold(threshold) -> None

setShadowValue(self, value)

Sets the shadow value

Python prototype (for reference):

setShadowValue(value) -> None

Functions

getDetectShadows!(self)

Raising version of getDetectShadows/1.

getDist2Threshold!(self)

Raising version of getDist2Threshold/1.

getHistory!(self)

Raising version of getHistory/1.

getkNNSamples!(self)

Raising version of getkNNSamples/1.

getNSamples!(self)

Raising version of getNSamples/1.

getShadowThreshold!(self)

Raising version of getShadowThreshold/1.

getShadowValue!(self)

Raising version of getShadowValue/1.

setDetectShadows!(self, detectShadows)

Raising version of setDetectShadows/2.

setDist2Threshold!(self, dist2Threshold)

Raising version of setDist2Threshold/2.

setHistory!(self, history)

Raising version of setHistory/2.

setkNNSamples!(self, nkNN)

Raising version of setkNNSamples/2.

setNSamples!(self, nN)

Raising version of setNSamples/2.

setShadowThreshold!(self, threshold)

Raising version of setShadowThreshold/2.

setShadowValue!(self, value)

Raising version of setShadowValue/2.