Evision.BackgroundSubtractorMOG2 (Evision v0.1.7-dev)

cv

apply(self, image)

Positional Arguments

image: UMat.

Keyword Arguments

fgmask: UMat.

learningRate: double.

Computes a foreground mask.

learnt. Negative parameter value makes the algorithm to use some automatically chosen learning
rate. 0 means that the background model is not updated at all, 1 means that the background model
is completely reinitialized from the last frame.

Python prototype (for reference):

apply(image[, fgmask[, learningRate]]) -> fgmask

apply(self, image, opts)

Positional Arguments

image: UMat.

Keyword Arguments

fgmask: UMat.

learningRate: double.

Computes a foreground mask.

learnt. Negative parameter value makes the algorithm to use some automatically chosen learning
rate. 0 means that the background model is not updated at all, 1 means that the background model
is completely reinitialized from the last frame.

Python prototype (for reference):

apply(image[, fgmask[, learningRate]]) -> fgmask

getBackgroundRatio(self)

Returns the "background ratio" parameter of the algorithm If a foreground pixel keeps semi-constant value for about backgroundRatio*history frames, it's considered background and added to the model as a center of a new component. It corresponds to TB parameter in the paper.

Python prototype (for reference):

getBackgroundRatio() -> retval

getComplexityReductionThreshold(self)

Returns the complexity reduction threshold This parameter defines the number of samples needed to accept to prove the component exists. CT=0.05 is a default value for all the samples. By setting CT=0 you get an algorithm very similar to the standard Stauffer&Grimson algorithm.

Python prototype (for reference):

getComplexityReductionThreshold() -> retval

getDetectShadows(self)

Returns the shadow detection flag If true, the algorithm detects shadows and marks them. See createBackgroundSubtractorMOG2 for details.

Python prototype (for reference):

getDetectShadows() -> retval

getHistory(self)

Returns the number of last frames that affect the background model

Python prototype (for reference):

getHistory() -> retval

getNMixtures(self)

Returns the number of gaussian components in the background model

Python prototype (for reference):

getNMixtures() -> retval

getShadowThreshold(self)

Returns the shadow threshold A shadow is detected if pixel is a darker version of the background. The shadow threshold (Tau in the paper) is a threshold defining how much darker the shadow can be. Tau= 0.5 means that if a pixel is more than twice darker then it is not shadow. See Prati, Mikic, Trivedi and Cucchiara, Detecting Moving Shadows...*, IEEE PAMI,2003.

Python prototype (for reference):

getShadowThreshold() -> retval

getShadowValue(self)

Returns the shadow value Shadow value is the value used to mark shadows in the foreground mask. Default value is 127. Value 0 in the mask always means background, 255 means foreground.

Python prototype (for reference):

getShadowValue() -> retval

getVarInit(self)

Returns the initial variance of each gaussian component

Python prototype (for reference):

getVarInit() -> retval

getVarMax(self)

Python prototype (for reference):

getVarMax() -> retval

getVarMin(self)

Python prototype (for reference):

getVarMin() -> retval

getVarThreshold(self)

Returns the variance threshold for the pixel-model match The main threshold on the squared Mahalanobis distance to decide if the sample is well described by the background model or not. Related to Cthr from the paper.

Python prototype (for reference):

getVarThreshold() -> retval

getVarThresholdGen(self)

Returns the variance threshold for the pixel-model match used for new mixture component generation Threshold for the squared Mahalanobis distance that helps decide when a sample is close to the existing components (corresponds to Tg in the paper). If a pixel is not close to any component, it is considered foreground or added as a new component. 3 sigma => Tg=3*3=9 is default. A smaller Tg value generates more components. A higher Tg value may result in a small number of components but they can grow too large.

Python prototype (for reference):

getVarThresholdGen() -> retval

setBackgroundRatio(self, ratio)

Sets the "background ratio" parameter of the algorithm

Python prototype (for reference):

setBackgroundRatio(ratio) -> None

setComplexityReductionThreshold(self, ct)

Sets the complexity reduction threshold

Python prototype (for reference):

setComplexityReductionThreshold(ct) -> None

setDetectShadows(self, detectShadows)

Enables or disables shadow detection

Python prototype (for reference):

setDetectShadows(detectShadows) -> None

setHistory(self, history)

Sets the number of last frames that affect the background model

Python prototype (for reference):

setHistory(history) -> None

setNMixtures(self, nmixtures)

Sets the number of gaussian components in the background model. The model needs to be reinitalized to reserve memory.

Python prototype (for reference):

setNMixtures(nmixtures) -> None

setShadowThreshold(self, threshold)

Sets the shadow threshold

Python prototype (for reference):

setShadowThreshold(threshold) -> None

setShadowValue(self, value)

Sets the shadow value

Python prototype (for reference):

setShadowValue(value) -> None

setVarInit(self, varInit)

Sets the initial variance of each gaussian component

Python prototype (for reference):

setVarInit(varInit) -> None

setVarMax(self, varMax)

Python prototype (for reference):

setVarMax(varMax) -> None

setVarMin(self, varMin)

Python prototype (for reference):

setVarMin(varMin) -> None

setVarThreshold(self, varThreshold)

Sets the variance threshold for the pixel-model match

Python prototype (for reference):

setVarThreshold(varThreshold) -> None

setVarThresholdGen(self, varThresholdGen)

Sets the variance threshold for the pixel-model match used for new mixture component generation

Python prototype (for reference):

setVarThresholdGen(varThresholdGen) -> None

Functions

apply!(self, image)

Raising version of apply/2.

apply!(self, image, opts)

Raising version of apply/3.

getBackgroundRatio!(self)

Raising version of getBackgroundRatio/1.

getComplexityReductionThreshold!(self)

Raising version of getComplexityReductionThreshold/1.

getDetectShadows!(self)

Raising version of getDetectShadows/1.

getHistory!(self)

Raising version of getHistory/1.

getNMixtures!(self)

Raising version of getNMixtures/1.

getShadowThreshold!(self)

Raising version of getShadowThreshold/1.

getShadowValue!(self)

Raising version of getShadowValue/1.

getVarInit!(self)

Raising version of getVarInit/1.

getVarMax!(self)

Raising version of getVarMax/1.

getVarMin!(self)

Raising version of getVarMin/1.

getVarThreshold!(self)

Raising version of getVarThreshold/1.

getVarThresholdGen!(self)

Raising version of getVarThresholdGen/1.

setBackgroundRatio!(self, ratio)

Raising version of setBackgroundRatio/2.

setComplexityReductionThreshold!(self, ct)

Raising version of setComplexityReductionThreshold/2.

setDetectShadows!(self, detectShadows)

Raising version of setDetectShadows/2.

setHistory!(self, history)

Raising version of setHistory/2.

setNMixtures!(self, nmixtures)

Raising version of setNMixtures/2.

setShadowThreshold!(self, threshold)

Raising version of setShadowThreshold/2.

setShadowValue!(self, value)

Raising version of setShadowValue/2.

setVarInit!(self, varInit)

Raising version of setVarInit/2.

setVarMax!(self, varMax)

Raising version of setVarMax/2.

setVarMin!(self, varMin)

Raising version of setVarMin/2.

setVarThreshold!(self, varThreshold)

Raising version of setVarThreshold/2.

setVarThresholdGen!(self, varThresholdGen)

Raising version of setVarThresholdGen/2.