Evision.CUDA (Evision v0.1.9)

cv.cuda

createContinuous(rows, cols, type)

Creates a continuous matrix.

Positional Arguments

rows: int.

Row count.

cols: int.

Column count.

type: int.

Type of the matrix.

Return

arr: Evision.Mat.

Destination matrix. This parameter changes only if it has a proper type and area ( \f$\texttt{rows} \times \texttt{cols}\f$ ).

Matrix is called continuous if its elements are stored continuously, that is, without gaps at the end of each row.

Python prototype (for reference):

createContinuous(rows, cols, type[, arr]) -> arr

createContinuous(rows, cols, type, opts)

Creates a continuous matrix.

Positional Arguments

rows: int.

Row count.

cols: int.

Column count.

type: int.

Type of the matrix.

Return

arr: Evision.Mat.

Destination matrix. This parameter changes only if it has a proper type and area ( \f$\texttt{rows} \times \texttt{cols}\f$ ).

Matrix is called continuous if its elements are stored continuously, that is, without gaps at the end of each row.

Python prototype (for reference):

createContinuous(rows, cols, type[, arr]) -> arr

ensureSizeIsEnough(rows, cols, type)

Ensures that the size of a matrix is big enough and the matrix has a proper type.

Positional Arguments

rows: int.

Minimum desired number of rows.

cols: int.

Minimum desired number of columns.

type: int.

Desired matrix type.

Return

arr: Evision.Mat.

Destination matrix.

The function does not reallocate memory if the matrix has proper attributes already.

Python prototype (for reference):

ensureSizeIsEnough(rows, cols, type[, arr]) -> arr

ensureSizeIsEnough(rows, cols, type, opts)

Ensures that the size of a matrix is big enough and the matrix has a proper type.

Positional Arguments

rows: int.

Minimum desired number of rows.

cols: int.

Minimum desired number of columns.

type: int.

Desired matrix type.

Return

arr: Evision.Mat.

Destination matrix.

The function does not reallocate memory if the matrix has proper attributes already.

Python prototype (for reference):

ensureSizeIsEnough(rows, cols, type[, arr]) -> arr

getCudaEnabledDeviceCount()

Returns the number of installed CUDA-enabled devices.

Use this function before any other CUDA functions calls. If OpenCV is compiled without CUDA support, this function returns 0. If the CUDA driver is not installed, or is incompatible, this function returns -1.

Python prototype (for reference):

getCudaEnabledDeviceCount() -> retval

getDevice()

Returns the current device index set by cuda::setDevice or initialized by default.

Python prototype (for reference):

getDevice() -> retval

printCudaDeviceInfo(device)

Positional Arguments

device: int

Python prototype (for reference):

printCudaDeviceInfo(device) -> None

printShortCudaDeviceInfo(device)

Positional Arguments

device: int

Python prototype (for reference):

printShortCudaDeviceInfo(device) -> None

registerPageLocked(m)

Page-locks the memory of matrix and maps it for the device(s).

Positional Arguments

m: Evision.Mat.

Input matrix.

Python prototype (for reference):

registerPageLocked(m) -> None

resetDevice()

Explicitly destroys and cleans up all resources associated with the current device in the current process.

Any subsequent API call to this device will reinitialize the device.

Python prototype (for reference):

resetDevice() -> None

setBufferPoolConfig(deviceId, stackSize, stackCount)

Positional Arguments

deviceId: int

stackSize: size_t

stackCount: int

Python prototype (for reference):

setBufferPoolConfig(deviceId, stackSize, stackCount) -> None

setBufferPoolUsage(on)

Positional Arguments

on: bool

Python prototype (for reference):

setBufferPoolUsage(on) -> None

setDevice(device)

Sets a device and initializes it for the current thread.

Positional Arguments

device: int.

System index of a CUDA device starting with 0.

If the call of this function is omitted, a default device is initialized at the fist CUDA usage.

Python prototype (for reference):

setDevice(device) -> None

unregisterPageLocked(m)

Unmaps the memory of matrix and makes it pageable again.

Positional Arguments

m: Evision.Mat.

Input matrix.

Python prototype (for reference):

unregisterPageLocked(m) -> None

Functions

createContinuous!(rows, cols, type)

Raising version of createContinuous/3.

createContinuous!(rows, cols, type, opts)

Raising version of createContinuous/4.

ensureSizeIsEnough!(rows, cols, type)

Raising version of ensureSizeIsEnough/3.

ensureSizeIsEnough!(rows, cols, type, opts)

Raising version of ensureSizeIsEnough/4.

getCudaEnabledDeviceCount!()

Raising version of getCudaEnabledDeviceCount/0.

getDevice!()

Raising version of getDevice/0.

printCudaDeviceInfo!(device)

Raising version of printCudaDeviceInfo/1.

printShortCudaDeviceInfo!(device)

Raising version of printShortCudaDeviceInfo/1.

registerPageLocked!(m)

Raising version of registerPageLocked/1.

resetDevice!()

Raising version of resetDevice/0.

setBufferPoolConfig!(deviceId, stackSize, stackCount)

Raising version of setBufferPoolConfig/3.

setBufferPoolUsage!(on)

Raising version of setBufferPoolUsage/1.

setDevice!(device)

Raising version of setDevice/1.

unregisterPageLocked!(m)

Raising version of unregisterPageLocked/1.