Evision.CUDA.DeviceInfo (Evision v0.1.7-dev)

cv.cuda

asyncEngineCount(self)

Python prototype (for reference):

asyncEngineCount() -> retval

canMapHostMemory(self)

Python prototype (for reference):

canMapHostMemory() -> retval

clockRate(self)

Python prototype (for reference):

clockRate() -> retval

computeMode(self)

Python prototype (for reference):

computeMode() -> retval

concurrentKernels(self)

Python prototype (for reference):

concurrentKernels() -> retval

cuda_DeviceInfo()

Python prototype (for reference):

DeviceInfo() -> <cuda_DeviceInfo object>

cuda_DeviceInfo(device_id)

The constructors. Constructs the DeviceInfo object for the specified device. If device_id parameter is missed, it constructs an object for the current device.

Python prototype (for reference):

DeviceInfo(device_id) -> <cuda_DeviceInfo object>

deviceID(self)

Returns system index of the CUDA device starting with 0.

Python prototype (for reference):

deviceID() -> retval

eccEnabled(self)

Python prototype (for reference):

ECCEnabled() -> retval

freeMemory(self)

Python prototype (for reference):

freeMemory() -> retval

integrated(self)

Python prototype (for reference):

integrated() -> retval

isCompatible(self)

Checks the CUDA module and device compatibility. This function returns true if the CUDA module can be run on the specified device. Otherwise, it returns false .

Python prototype (for reference):

isCompatible() -> retval

kernelExecTimeoutEnabled(self)

Python prototype (for reference):

kernelExecTimeoutEnabled() -> retval

l2CacheSize(self)

Python prototype (for reference):

l2CacheSize() -> retval

majorVersion(self)

Python prototype (for reference):

majorVersion() -> retval

maxGridSize(self)

Python prototype (for reference):

maxGridSize() -> retval

maxSurface1D(self)

Python prototype (for reference):

maxSurface1D() -> retval

maxSurface1DLayered(self)

Python prototype (for reference):

maxSurface1DLayered() -> retval

maxSurface2D(self)

Python prototype (for reference):

maxSurface2D() -> retval

maxSurface2DLayered(self)

Python prototype (for reference):

maxSurface2DLayered() -> retval

maxSurface3D(self)

Python prototype (for reference):

maxSurface3D() -> retval

maxSurfaceCubemap(self)

Python prototype (for reference):

maxSurfaceCubemap() -> retval

maxSurfaceCubemapLayered(self)

Python prototype (for reference):

maxSurfaceCubemapLayered() -> retval

maxTexture1D(self)

Python prototype (for reference):

maxTexture1D() -> retval

maxTexture1DLayered(self)

Python prototype (for reference):

maxTexture1DLayered() -> retval

maxTexture1DLinear(self)

Python prototype (for reference):

maxTexture1DLinear() -> retval

maxTexture1DMipmap(self)

Python prototype (for reference):

maxTexture1DMipmap() -> retval

maxTexture2D(self)

Python prototype (for reference):

maxTexture2D() -> retval

maxTexture2DGather(self)

Python prototype (for reference):

maxTexture2DGather() -> retval

maxTexture2DLayered(self)

Python prototype (for reference):

maxTexture2DLayered() -> retval

maxTexture2DLinear(self)

Python prototype (for reference):

maxTexture2DLinear() -> retval

maxTexture2DMipmap(self)

Python prototype (for reference):

maxTexture2DMipmap() -> retval

maxTexture3D(self)

Python prototype (for reference):

maxTexture3D() -> retval

maxTextureCubemap(self)

Python prototype (for reference):

maxTextureCubemap() -> retval

maxTextureCubemapLayered(self)

Python prototype (for reference):

maxTextureCubemapLayered() -> retval

maxThreadsDim(self)

Python prototype (for reference):

maxThreadsDim() -> retval

maxThreadsPerBlock(self)

Python prototype (for reference):

maxThreadsPerBlock() -> retval

maxThreadsPerMultiProcessor(self)

Python prototype (for reference):

maxThreadsPerMultiProcessor() -> retval

memoryBusWidth(self)

Python prototype (for reference):

memoryBusWidth() -> retval

memoryClockRate(self)

Python prototype (for reference):

memoryClockRate() -> retval

memPitch(self)

Python prototype (for reference):

memPitch() -> retval

minorVersion(self)

Python prototype (for reference):

minorVersion() -> retval

multiProcessorCount(self)

Python prototype (for reference):

multiProcessorCount() -> retval

pciBusID(self)

Python prototype (for reference):

pciBusID() -> retval

pciDeviceID(self)

Python prototype (for reference):

pciDeviceID() -> retval

pciDomainID(self)

Python prototype (for reference):

pciDomainID() -> retval

queryMemory(self, totalMemory, freeMemory)

Python prototype (for reference):

queryMemory(totalMemory, freeMemory) -> None

regsPerBlock(self)

Python prototype (for reference):

regsPerBlock() -> retval

sharedMemPerBlock(self)

Python prototype (for reference):

sharedMemPerBlock() -> retval

surfaceAlignment(self)

Python prototype (for reference):

surfaceAlignment() -> retval

tccDriver(self)

Python prototype (for reference):

tccDriver() -> retval

textureAlignment(self)

Python prototype (for reference):

textureAlignment() -> retval

texturePitchAlignment(self)

Python prototype (for reference):

texturePitchAlignment() -> retval

totalConstMem(self)

Python prototype (for reference):

totalConstMem() -> retval

totalGlobalMem(self)

Python prototype (for reference):

totalGlobalMem() -> retval

totalMemory(self)

Python prototype (for reference):

totalMemory() -> retval

unifiedAddressing(self)

Python prototype (for reference):

unifiedAddressing() -> retval

warpSize(self)

Python prototype (for reference):

warpSize() -> retval

Functions

asyncEngineCount!(self)

Raising version of asyncEngineCount/1.

canMapHostMemory!(self)

Raising version of canMapHostMemory/1.

clockRate!(self)

Raising version of clockRate/1.

computeMode!(self)

Raising version of computeMode/1.

concurrentKernels!(self)

Raising version of concurrentKernels/1.

cuda_DeviceInfo!()

Raising version of cuda_DeviceInfo/0.

cuda_DeviceInfo!(device_id)

Raising version of cuda_DeviceInfo/1.

deviceID!(self)

Raising version of deviceID/1.

eccEnabled!(self)

Raising version of eccEnabled/1.

freeMemory!(self)

Raising version of freeMemory/1.

integrated!(self)

Raising version of integrated/1.

isCompatible!(self)

Raising version of isCompatible/1.

kernelExecTimeoutEnabled!(self)

Raising version of kernelExecTimeoutEnabled/1.

l2CacheSize!(self)

Raising version of l2CacheSize/1.

majorVersion!(self)

Raising version of majorVersion/1.

maxGridSize!(self)

Raising version of maxGridSize/1.

maxSurface1D!(self)

Raising version of maxSurface1D/1.

maxSurface1DLayered!(self)

Raising version of maxSurface1DLayered/1.

maxSurface2D!(self)

Raising version of maxSurface2D/1.

maxSurface2DLayered!(self)

Raising version of maxSurface2DLayered/1.

maxSurface3D!(self)

Raising version of maxSurface3D/1.

maxSurfaceCubemap!(self)

Raising version of maxSurfaceCubemap/1.

maxSurfaceCubemapLayered!(self)

Raising version of maxSurfaceCubemapLayered/1.

maxTexture1D!(self)

Raising version of maxTexture1D/1.

maxTexture1DLayered!(self)

Raising version of maxTexture1DLayered/1.

maxTexture1DLinear!(self)

Raising version of maxTexture1DLinear/1.

maxTexture1DMipmap!(self)

Raising version of maxTexture1DMipmap/1.

maxTexture2D!(self)

Raising version of maxTexture2D/1.

maxTexture2DGather!(self)

Raising version of maxTexture2DGather/1.

maxTexture2DLayered!(self)

Raising version of maxTexture2DLayered/1.

maxTexture2DLinear!(self)

Raising version of maxTexture2DLinear/1.

maxTexture2DMipmap!(self)

Raising version of maxTexture2DMipmap/1.

maxTexture3D!(self)

Raising version of maxTexture3D/1.

maxTextureCubemap!(self)

Raising version of maxTextureCubemap/1.

maxTextureCubemapLayered!(self)

Raising version of maxTextureCubemapLayered/1.

maxThreadsDim!(self)

Raising version of maxThreadsDim/1.

maxThreadsPerBlock!(self)

Raising version of maxThreadsPerBlock/1.

maxThreadsPerMultiProcessor!(self)

Raising version of maxThreadsPerMultiProcessor/1.

memoryBusWidth!(self)

Raising version of memoryBusWidth/1.

memoryClockRate!(self)

Raising version of memoryClockRate/1.

memPitch!(self)

Raising version of memPitch/1.

minorVersion!(self)

Raising version of minorVersion/1.

multiProcessorCount!(self)

Raising version of multiProcessorCount/1.

pciBusID!(self)

Raising version of pciBusID/1.

pciDeviceID!(self)

Raising version of pciDeviceID/1.

pciDomainID!(self)

Raising version of pciDomainID/1.

queryMemory!(self, totalMemory, freeMemory)

Raising version of queryMemory/3.

regsPerBlock!(self)

Raising version of regsPerBlock/1.

sharedMemPerBlock!(self)

Raising version of sharedMemPerBlock/1.

surfaceAlignment!(self)

Raising version of surfaceAlignment/1.

tccDriver!(self)

Raising version of tccDriver/1.

textureAlignment!(self)

Raising version of textureAlignment/1.

texturePitchAlignment!(self)

Raising version of texturePitchAlignment/1.

totalConstMem!(self)

Raising version of totalConstMem/1.

totalGlobalMem!(self)

Raising version of totalGlobalMem/1.

totalMemory!(self)

Raising version of totalMemory/1.

unifiedAddressing!(self)

Raising version of unifiedAddressing/1.

warpSize!(self)

Raising version of warpSize/1.