Evision.CUDA.HostMem (Evision v0.1.7-dev)

cv.cuda

channels(self)

Python prototype (for reference):

channels() -> retval

clone(self)

Python prototype (for reference):

clone() -> retval

create(self, rows, cols, type)

Python prototype (for reference):

create(rows, cols, type) -> None

createMatHeader(self)

Python prototype (for reference):

createMatHeader() -> retval

cuda_HostMem()

Positional Arguments

Keyword Arguments

alloc_type: HostMem_AllocType.

Python prototype (for reference):

HostMem([, alloc_type]) -> <cuda_HostMem object>

cuda_HostMem(opts)

Variaint 1:

Positional Arguments

arr: UMat

Keyword Arguments

alloc_type: HostMem_AllocType.

Python prototype (for reference):

HostMem(arr[, alloc_type]) -> <cuda_HostMem object>

Variaint 2:

Positional Arguments

Keyword Arguments

alloc_type: HostMem_AllocType.

Python prototype (for reference):

HostMem([, alloc_type]) -> <cuda_HostMem object>

cuda_HostMem(arr, opts)

Variaint 1:

Positional Arguments

size: Size

type: int

Keyword Arguments

alloc_type: HostMem_AllocType.

Python prototype (for reference):

HostMem(size, type[, alloc_type]) -> <cuda_HostMem object>

Variaint 2:

Positional Arguments

arr: UMat

Keyword Arguments

alloc_type: HostMem_AllocType.

Python prototype (for reference):

HostMem(arr[, alloc_type]) -> <cuda_HostMem object>

cuda_HostMem(size, type, opts)

Variaint 1:

Positional Arguments

rows: int

cols: int

type: int

Keyword Arguments

alloc_type: HostMem_AllocType.

Python prototype (for reference):

HostMem(rows, cols, type[, alloc_type]) -> <cuda_HostMem object>

Variaint 2:

Positional Arguments

size: Size

type: int

Keyword Arguments

alloc_type: HostMem_AllocType.

Python prototype (for reference):

HostMem(size, type[, alloc_type]) -> <cuda_HostMem object>

cuda_HostMem(rows, cols, type, opts)

Positional Arguments

rows: int

cols: int

type: int

Keyword Arguments

alloc_type: HostMem_AllocType.

Python prototype (for reference):

HostMem(rows, cols, type[, alloc_type]) -> <cuda_HostMem object>

depth(self)

Python prototype (for reference):

depth() -> retval

elemSize1(self)

Python prototype (for reference):

elemSize1() -> retval

elemSize(self)

Python prototype (for reference):

elemSize() -> retval

empty(self)

Python prototype (for reference):

empty() -> retval

isContinuous(self)

Maps CPU memory to GPU address space and creates the cuda::GpuMat header without reference counting for it. This can be done only if memory was allocated with the SHARED flag and if it is supported by the hardware. Laptops often share video and CPU memory, so address spaces can be mapped, which eliminates an extra copy.

Python prototype (for reference):

isContinuous() -> retval

reshape(self, cn)

Positional Arguments

cn: int

Keyword Arguments

rows: int.

Python prototype (for reference):

reshape(cn[, rows]) -> retval

reshape(self, cn, opts)

Positional Arguments

cn: int

Keyword Arguments

rows: int.

Python prototype (for reference):

reshape(cn[, rows]) -> retval

size(self)

Python prototype (for reference):

size() -> retval

step1(self)

Python prototype (for reference):

step1() -> retval

swap(self, b)

Python prototype (for reference):

swap(b) -> None

type(self)

Python prototype (for reference):

type() -> retval

Functions

channels!(self)

Raising version of channels/1.

clone!(self)

Raising version of clone/1.

create!(self, rows, cols, type)

Raising version of create/4.

createMatHeader!(self)

Raising version of createMatHeader/1.

cuda_HostMem!()

Raising version of cuda_HostMem/0.

cuda_HostMem!(arr)

Raising version of cuda_HostMem/1.

cuda_HostMem!(size, type)

Raising version of cuda_HostMem/2.

cuda_HostMem!(rows, cols, type)

Raising version of cuda_HostMem/3.

cuda_HostMem!(rows, cols, type, opts)

Raising version of cuda_HostMem/4.

depth!(self)

Raising version of depth/1.

elemSize1!(self)

Raising version of elemSize1/1.

elemSize!(self)

Raising version of elemSize/1.

empty!(self)

Raising version of empty/1.

get_step!(self)

Raising version of get_step/1.

get_step(self)

isContinuous!(self)

Raising version of isContinuous/1.

reshape!(self, cn)

Raising version of reshape/2.

reshape!(self, cn, opts)

Raising version of reshape/3.

size!(self)

Raising version of size/1.

step1!(self)

Raising version of step1/1.

swap!(self, b)

Raising version of swap/2.

type!(self)

Raising version of type/1.