Evision.CUDA.Stream (Evision v0.1.7-dev)

cv.cuda

cuda_Stream()

Python prototype (for reference):

Stream() -> <cuda_Stream object>

cuda_Stream(cudaFlags)

Variaint 1:

creates a new Stream using the cudaFlags argument to determine the behaviors of the stream Note: The cudaFlags parameter is passed to the underlying api cudaStreamCreateWithFlags() and supports the same parameter values.

// creates an OpenCV cuda::Stream that manages an asynchronous, non-blocking,
// non-default CUDA stream
cv::cuda::Stream cvStream(cudaStreamNonBlocking);

Python prototype (for reference):

Stream(cudaFlags) -> <cuda_Stream object>

Variaint 2:

Python prototype (for reference):

Stream(allocator) -> <cuda_Stream object>

cudaPtr(self)

Python prototype (for reference):

cudaPtr() -> retval

null()

Adds a callback to be called on the host after all currently enqueued items in the stream have completed. Note: Callbacks must not make any CUDA API calls. Callbacks must not perform any synchronization that may depend on outstanding device work or other callbacks that are not mandated to run earlier. Callbacks without a mandated order (in independent streams) execute in undefined order and may be serialized.

Python prototype (for reference):

Null() -> retval

queryIfComplete(self)

Returns true if the current stream queue is finished. Otherwise, it returns false.

Python prototype (for reference):

queryIfComplete() -> retval

waitEvent(self, event)

Makes a compute stream wait on an event.

Python prototype (for reference):

waitEvent(event) -> None

waitForCompletion(self)

Blocks the current CPU thread until all operations in the stream are complete.

Python prototype (for reference):

waitForCompletion() -> None

Functions

cuda_Stream!()

Raising version of cuda_Stream/0.

cuda_Stream!(cudaFlags)

Raising version of cuda_Stream/1.

cudaPtr!(self)

Raising version of cudaPtr/1.

null!()

Raising version of null/0.

queryIfComplete!(self)

Raising version of queryIfComplete/1.

waitEvent!(self, event)

Raising version of waitEvent/2.

waitForCompletion!(self)

Raising version of waitForCompletion/1.