Evision.CascadeClassifier (Evision v0.1.7-dev)

cv

cascadeClassifier()

Python prototype (for reference):

CascadeClassifier() -> <CascadeClassifier object>

cascadeClassifier(filename)

Loads a classifier from a file.

Python prototype (for reference):

CascadeClassifier(filename) -> <CascadeClassifier object>

convert(oldcascade, newcascade)

Python prototype (for reference):

convert(oldcascade, newcascade) -> retval

detectMultiScale2(self, image)

Positional Arguments

image: UMat.

Keyword Arguments

scaleFactor: double.

minNeighbors: int.

flags: int.

minSize: Size.

maxSize: Size.

Has overloading in C++

rectangles may be partially outside the original image.
of detections is the number of neighboring positively classified rectangles that were joined
together to form the object.
to retain it.
cvHaarDetectObjects. It is not used for a new cascade.

Python prototype (for reference):

detectMultiScale2(image[, scaleFactor[, minNeighbors[, flags[, minSize[, maxSize]]]]]) -> objects, numDetections

detectMultiScale2(self, image, opts)

Positional Arguments

image: UMat.

Keyword Arguments

scaleFactor: double.

minNeighbors: int.

flags: int.

minSize: Size.

maxSize: Size.

Has overloading in C++

rectangles may be partially outside the original image.
of detections is the number of neighboring positively classified rectangles that were joined
together to form the object.
to retain it.
cvHaarDetectObjects. It is not used for a new cascade.

Python prototype (for reference):

detectMultiScale2(image[, scaleFactor[, minNeighbors[, flags[, minSize[, maxSize]]]]]) -> objects, numDetections

detectMultiScale3(self, image)

Positional Arguments

image: UMat

Keyword Arguments

scaleFactor: double.

minNeighbors: int.

flags: int.

minSize: Size.

maxSize: Size.

outputRejectLevels: bool.

Has overloading in C++

This function allows you to retrieve the final stage decision certainty of classification. For this, one needs to set outputRejectLevels on true and provide the rejectLevels and levelWeights parameter. For each resulting detection, levelWeights will then contain the certainty of classification at the final stage. This value can then be used to separate strong from weaker classifications. A code sample on how to use it efficiently can be found below:

Mat img;
vector<double> weights;
vector<int> levels;
vector<Rect> detections;
CascadeClassifier model("/path/to/your/model.xml");
model.detectMultiScale(img, detections, levels, weights, 1.1, 3, 0, Size(), Size(), true);
cerr << "Detection " << detections[0] << " with weight " << weights[0] << endl;

Python prototype (for reference):

detectMultiScale3(image[, scaleFactor[, minNeighbors[, flags[, minSize[, maxSize[, outputRejectLevels]]]]]]) -> objects, rejectLevels, levelWeights

detectMultiScale3(self, image, opts)

Positional Arguments

image: UMat

Keyword Arguments

scaleFactor: double.

minNeighbors: int.

flags: int.

minSize: Size.

maxSize: Size.

outputRejectLevels: bool.

Has overloading in C++

This function allows you to retrieve the final stage decision certainty of classification. For this, one needs to set outputRejectLevels on true and provide the rejectLevels and levelWeights parameter. For each resulting detection, levelWeights will then contain the certainty of classification at the final stage. This value can then be used to separate strong from weaker classifications. A code sample on how to use it efficiently can be found below:

Mat img;
vector<double> weights;
vector<int> levels;
vector<Rect> detections;
CascadeClassifier model("/path/to/your/model.xml");
model.detectMultiScale(img, detections, levels, weights, 1.1, 3, 0, Size(), Size(), true);
cerr << "Detection " << detections[0] << " with weight " << weights[0] << endl;

Python prototype (for reference):

detectMultiScale3(image[, scaleFactor[, minNeighbors[, flags[, minSize[, maxSize[, outputRejectLevels]]]]]]) -> objects, rejectLevels, levelWeights

detectMultiScale(self, image)

Positional Arguments

image: UMat.

Keyword Arguments

scaleFactor: double.

minNeighbors: int.

flags: int.

minSize: Size.

maxSize: Size.

Detects objects of different sizes in the input image. The detected objects are returned as a list of rectangles.

rectangles may be partially outside the original image.
to retain it.
cvHaarDetectObjects. It is not used for a new cascade.

Python prototype (for reference):

detectMultiScale(image[, scaleFactor[, minNeighbors[, flags[, minSize[, maxSize]]]]]) -> objects

detectMultiScale(self, image, opts)

Positional Arguments

image: UMat.

Keyword Arguments

scaleFactor: double.

minNeighbors: int.

flags: int.

minSize: Size.

maxSize: Size.

Detects objects of different sizes in the input image. The detected objects are returned as a list of rectangles.

rectangles may be partially outside the original image.
to retain it.
cvHaarDetectObjects. It is not used for a new cascade.

Python prototype (for reference):

detectMultiScale(image[, scaleFactor[, minNeighbors[, flags[, minSize[, maxSize]]]]]) -> objects

empty(self)

Checks whether the classifier has been loaded.

Python prototype (for reference):

empty() -> retval

getFeatureType(self)

Python prototype (for reference):

getFeatureType() -> retval

getOriginalWindowSize(self)

Python prototype (for reference):

getOriginalWindowSize() -> retval

isOldFormatCascade(self)

Python prototype (for reference):

isOldFormatCascade() -> retval

load(self, filename)

Loads a classifier from a file.

HAAR classifier trained by the haartraining application or a new cascade classifier trained by the
traincascade application.

Python prototype (for reference):

load(filename) -> retval

read(self, node)

Reads a classifier from a FileStorage node. Note: The file may contain a new cascade classifier (trained by the traincascade application) only.

Python prototype (for reference):

read(node) -> retval

Functions

cascadeClassifier!()

Raising version of cascadeClassifier/0.

cascadeClassifier!(filename)

Raising version of cascadeClassifier/1.

convert!(oldcascade, newcascade)

Raising version of convert/2.

detectMultiScale2!(self, image)

Raising version of detectMultiScale2/2.

detectMultiScale2!(self, image, opts)

Raising version of detectMultiScale2/3.

detectMultiScale3!(self, image)

Raising version of detectMultiScale3/2.

detectMultiScale3!(self, image, opts)

Raising version of detectMultiScale3/3.

detectMultiScale!(self, image)

Raising version of detectMultiScale/2.

detectMultiScale!(self, image, opts)

Raising version of detectMultiScale/3.

empty!(self)

Raising version of empty/1.

getFeatureType!(self)

Raising version of getFeatureType/1.

getOriginalWindowSize!(self)

Raising version of getOriginalWindowSize/1.

isOldFormatCascade!(self)

Raising version of isOldFormatCascade/1.

load!(self, filename)

Raising version of load/2.

read!(self, node)

Raising version of read/2.