Evision.DISOpticalFlow (Evision v0.1.7-dev)

cv

create()

Positional Arguments

Keyword Arguments

preset: int.

Creates an instance of DISOpticalFlow

Python prototype (for reference):

create([, preset]) -> retval

create(opts)

Positional Arguments

Keyword Arguments

preset: int.

Creates an instance of DISOpticalFlow

Python prototype (for reference):

create([, preset]) -> retval

getFinestScale(self)

Finest level of the Gaussian pyramid on which the flow is computed (zero level corresponds to the original image resolution). The final flow is obtained by bilinear upscaling. @see setFinestScale

Python prototype (for reference):

getFinestScale() -> retval

getGradientDescentIterations(self)

Maximum number of gradient descent iterations in the patch inverse search stage. Higher values may improve quality in some cases. @see setGradientDescentIterations

Python prototype (for reference):

getGradientDescentIterations() -> retval

getPatchSize(self)

Size of an image patch for matching (in pixels). Normally, default 8x8 patches work well enough in most cases. @see setPatchSize

Python prototype (for reference):

getPatchSize() -> retval

getPatchStride(self)

Stride between neighbor patches. Must be less than patch size. Lower values correspond to higher flow quality. @see setPatchStride

Python prototype (for reference):

getPatchStride() -> retval

getUseMeanNormalization(self)

Whether to use mean-normalization of patches when computing patch distance. It is turned on by default as it typically provides a noticeable quality boost because of increased robustness to illumination variations. Turn it off if you are certain that your sequence doesn't contain any changes in illumination. @see setUseMeanNormalization

Python prototype (for reference):

getUseMeanNormalization() -> retval

getUseSpatialPropagation(self)

Whether to use spatial propagation of good optical flow vectors. This option is turned on by default, as it tends to work better on average and can sometimes help recover from major errors introduced by the coarse-to-fine scheme employed by the DIS optical flow algorithm. Turning this option off can make the output flow field a bit smoother, however. @see setUseSpatialPropagation

Python prototype (for reference):

getUseSpatialPropagation() -> retval

getVariationalRefinementAlpha(self)

Weight of the smoothness term @see setVariationalRefinementAlpha

Python prototype (for reference):

getVariationalRefinementAlpha() -> retval

getVariationalRefinementDelta(self)

Weight of the color constancy term @see setVariationalRefinementDelta

Python prototype (for reference):

getVariationalRefinementDelta() -> retval

getVariationalRefinementGamma(self)

Weight of the gradient constancy term @see setVariationalRefinementGamma

Python prototype (for reference):

getVariationalRefinementGamma() -> retval

getVariationalRefinementIterations(self)

Number of fixed point iterations of variational refinement per scale. Set to zero to disable variational refinement completely. Higher values will typically result in more smooth and high-quality flow. @see setGradientDescentIterations

Python prototype (for reference):

getVariationalRefinementIterations() -> retval

setFinestScale(self, val)

@copybrief getFinestScale @see getFinestScale

Python prototype (for reference):

setFinestScale(val) -> None

setGradientDescentIterations(self, val)

@copybrief getGradientDescentIterations @see getGradientDescentIterations

Python prototype (for reference):

setGradientDescentIterations(val) -> None

setPatchSize(self, val)

@copybrief getPatchSize @see getPatchSize

Python prototype (for reference):

setPatchSize(val) -> None

setPatchStride(self, val)

@copybrief getPatchStride @see getPatchStride

Python prototype (for reference):

setPatchStride(val) -> None

setUseMeanNormalization(self, val)

@copybrief getUseMeanNormalization @see getUseMeanNormalization

Python prototype (for reference):

setUseMeanNormalization(val) -> None

setUseSpatialPropagation(self, val)

@copybrief getUseSpatialPropagation @see getUseSpatialPropagation

Python prototype (for reference):

setUseSpatialPropagation(val) -> None

setVariationalRefinementAlpha(self, val)

@copybrief getVariationalRefinementAlpha @see getVariationalRefinementAlpha

Python prototype (for reference):

setVariationalRefinementAlpha(val) -> None

setVariationalRefinementDelta(self, val)

@copybrief getVariationalRefinementDelta @see getVariationalRefinementDelta

Python prototype (for reference):

setVariationalRefinementDelta(val) -> None

setVariationalRefinementGamma(self, val)

@copybrief getVariationalRefinementGamma @see getVariationalRefinementGamma

Python prototype (for reference):

setVariationalRefinementGamma(val) -> None

setVariationalRefinementIterations(self, val)

@copybrief getGradientDescentIterations @see getGradientDescentIterations

Python prototype (for reference):

setVariationalRefinementIterations(val) -> None

Functions

create!()

Raising version of create/0.

create!(opts)

Raising version of create/1.

getFinestScale!(self)

Raising version of getFinestScale/1.

getGradientDescentIterations!(self)

Raising version of getGradientDescentIterations/1.

getPatchSize!(self)

Raising version of getPatchSize/1.

getPatchStride!(self)

Raising version of getPatchStride/1.

getUseMeanNormalization!(self)

Raising version of getUseMeanNormalization/1.

getUseSpatialPropagation!(self)

Raising version of getUseSpatialPropagation/1.

getVariationalRefinementAlpha!(self)

Raising version of getVariationalRefinementAlpha/1.

getVariationalRefinementDelta!(self)

Raising version of getVariationalRefinementDelta/1.

getVariationalRefinementGamma!(self)

Raising version of getVariationalRefinementGamma/1.

getVariationalRefinementIterations!(self)

Raising version of getVariationalRefinementIterations/1.

setFinestScale!(self, val)

Raising version of setFinestScale/2.

setGradientDescentIterations!(self, val)

Raising version of setGradientDescentIterations/2.

setPatchSize!(self, val)

Raising version of setPatchSize/2.

setPatchStride!(self, val)

Raising version of setPatchStride/2.

setUseMeanNormalization!(self, val)

Raising version of setUseMeanNormalization/2.

setUseSpatialPropagation!(self, val)

Raising version of setUseSpatialPropagation/2.

setVariationalRefinementAlpha!(self, val)

Raising version of setVariationalRefinementAlpha/2.

setVariationalRefinementDelta!(self, val)

Raising version of setVariationalRefinementDelta/2.

setVariationalRefinementGamma!(self, val)

Raising version of setVariationalRefinementGamma/2.

setVariationalRefinementIterations!(self, val)

Raising version of setVariationalRefinementIterations/2.