Evision.DNN (Evision v0.1.5-dev)

Functions

blobFromImage!(image)

Raising version of blobFromImage/1.

blobFromImage!(image, opts)

Raising version of blobFromImage/2.

blobFromImage(image)

blobFromImage(image, opts)

Creates 4-dimensional blob from image. Optionally resizes and crops @p image from center,

subtract @p mean values, scales values by @p scalefactor, swap Blue and Red channels.

image.

size.

mean.
to be in (mean-R, mean-G, mean-B) order if @p image has BGR ordering and @p swapRB is true.

scalefactor.

swapRB.
in 3-channel image is necessary.

crop.

ddepth. @details if @p crop is true, input image is resized so one side after resize is equal to corresponding dimension in @p size and another one is equal or larger. Then, crop from the center is performed. If @p crop is false, direct resize without cropping and preserving aspect ratio is performed. @returns 4-dimensional Mat with NCHW dimensions order.

scalefactor : double

size : Size

mean : Scalar

swapRB : bool

crop : bool

ddepth : int

Python prototype (for reference):

blobFromImage(image[, scalefactor[, size[, mean[, swapRB[, crop[, ddepth]]]]]]) -> retval

blobFromImages!(images)

Raising version of blobFromImages/1.

blobFromImages!(images, opts)

Raising version of blobFromImages/2.

blobFromImages(images)

blobFromImages(images, opts)

Creates 4-dimensional blob from series of images. Optionally resizes and

crops @p images from center, subtract @p mean values, scales values by @p scalefactor,
swap Blue and Red channels.

images.

size.

mean.
to be in (mean-R, mean-G, mean-B) order if @p image has BGR ordering and @p swapRB is true.

scalefactor.

swapRB.
in 3-channel image is necessary.

crop.

ddepth. @details if @p crop is true, input image is resized so one side after resize is equal to corresponding dimension in @p size and another one is equal or larger. Then, crop from the center is performed. If @p crop is false, direct resize without cropping and preserving aspect ratio is performed. @returns 4-dimensional Mat with NCHW dimensions order.

scalefactor : double

size : Size

mean : Scalar

swapRB : bool

crop : bool

ddepth : int

Python prototype (for reference):

blobFromImages(images[, scalefactor[, size[, mean[, swapRB[, crop[, ddepth]]]]]]) -> retval

getAvailableTargets!(be)

Raising version of getAvailableTargets/1.

getAvailableTargets(be)

imagesFromBlob!(blob_)

Raising version of imagesFromBlob/1.

imagesFromBlob!(blob_, opts)

Raising version of imagesFromBlob/2.

imagesFromBlob(blob_)

imagesFromBlob(blob_, opts)

Parse a 4D blob and output the images it contains as 2D arrays through a simpler data structure

(std::vector<cv::Mat>).

blob_.
which you would like to extract the images.

images_.
(CV_32F). They are non normalized neither mean added. The number of returned images equals the first dimension
of the blob (batch size). Every image has a number of channels equals to the second dimension of the blob (depth).

images_ : vector_UMat

Python prototype (for reference):

imagesFromBlob(blob_[, images_]) -> images_

nmsBoxes!(bboxes, scores, score_threshold, nms_threshold)

Raising version of nmsBoxes/4.

nmsBoxes!(bboxes, scores, score_threshold, nms_threshold, opts)

Raising version of nmsBoxes/5.

nmsBoxes(bboxes, scores, score_threshold, nms_threshold)

nmsBoxes(bboxes, scores, score_threshold, nms_threshold, opts)

Performs non maximum suppression given boxes and corresponding scores.

bboxes.

scores.

score_threshold.

nms_threshold.

indices.

eta.

top_k.

eta : float

top_k : int

Python prototype (for reference):

NMSBoxes(bboxes, scores, score_threshold, nms_threshold[, eta[, top_k]]) -> indices

nmsBoxesRotated!(bboxes, scores, score_threshold, nms_threshold)

Raising version of nmsBoxesRotated/4.

nmsBoxesRotated!(bboxes, scores, score_threshold, nms_threshold, opts)

Raising version of nmsBoxesRotated/5.

nmsBoxesRotated(bboxes, scores, score_threshold, nms_threshold)

nmsBoxesRotated(bboxes, scores, score_threshold, nms_threshold, opts)

eta : float

top_k : int

Python prototype (for reference):

NMSBoxesRotated(bboxes, scores, score_threshold, nms_threshold[, eta[, top_k]]) -> indices

readNet!(model)

Raising version of readNet/1.

readNet!(framework, bufferModel)

Raising version of readNet/2.

readNet!(framework, bufferModel, opts)

Raising version of readNet/3.

readNet(model)

readNet(model, opts)

readNet(framework, bufferModel, opts)

Read deep learning network represented in one of the supported formats. @details This is an overloaded member function, provided for convenience.

        It differs from the above function only in what argument(s) it accepts.

framework.

bufferModel.

bufferConfig. @returns Net object.

bufferConfig : vector_uchar

Python prototype (for reference):

readNet(framework, bufferModel[, bufferConfig]) -> retval

readNetFromCaffe!(bufferProto)

Raising version of readNetFromCaffe/1.

readNetFromCaffe!(bufferProto, opts)

Raising version of readNetFromCaffe/2.

readNetFromCaffe(bufferProto)

readNetFromCaffe(bufferProto, opts)

Reads a network model stored in Caffe model in memory.

bufferProto.

bufferModel. @returns Net object.

bufferModel : vector_uchar

Python prototype (for reference):

readNetFromCaffe(bufferProto[, bufferModel]) -> retval

readNetFromDarknet!(bufferCfg)

Raising version of readNetFromDarknet/1.

readNetFromDarknet!(bufferCfg, opts)

Raising version of readNetFromDarknet/2.

readNetFromDarknet(bufferCfg)

readNetFromDarknet(bufferCfg, opts)

Reads a network model stored in <a href="https://pjreddie.com/darknet/">Darknet</a> model files.

bufferCfg.

bufferModel. @returns Net object.

bufferModel : vector_uchar

Python prototype (for reference):

readNetFromDarknet(bufferCfg[, bufferModel]) -> retval

readNetFromModelOptimizer!(bufferModelConfig, bufferWeights)

Raising version of readNetFromModelOptimizer/2.

readNetFromModelOptimizer(bufferModelConfig, bufferWeights)

readNetFromONNX!(buffer)

Raising version of readNetFromONNX/1.

readNetFromONNX(buffer)

readNetFromTensorflow!(bufferModel)

Raising version of readNetFromTensorflow/1.

readNetFromTensorflow!(bufferModel, opts)

Raising version of readNetFromTensorflow/2.

readNetFromTensorflow(bufferModel)

readNetFromTensorflow(bufferModel, opts)

Reads a network model stored in <a href="https://www.tensorflow.org/">TensorFlow</a> framework's format.

bufferModel.

bufferConfig. @returns Net object.

bufferConfig : vector_uchar

Python prototype (for reference):

readNetFromTensorflow(bufferModel[, bufferConfig]) -> retval

readNetFromTorch!(model)

Raising version of readNetFromTorch/1.

readNetFromTorch!(model, opts)

Raising version of readNetFromTorch/2.

readNetFromTorch(model)

readNetFromTorch(model, opts)

Reads a network model stored in <a href="http://torch.ch">Torch7</a> framework's format.

model.

isBinary.

evaluate. @returns Net object. Note: Ascii mode of Torch serializer is more preferable, because binary mode extensively use long type of C language, which has various bit-length on different systems. The loading file must contain serialized <a href="https://github.com/torch/nn/blob/master/doc/module.md">nn.Module</a> object with importing network. Try to eliminate a custom objects from serialazing data to avoid importing errors. List of supported layers (i.e. object instances derived from Torch nn.Module class):

nn.Sequential

nn.Parallel

nn.Concat

nn.Linear

nn.SpatialConvolution

nn.SpatialMaxPooling, nn.SpatialAveragePooling

nn.ReLU, nn.TanH, nn.Sigmoid

nn.Reshape

nn.SoftMax, nn.LogSoftMax Also some equivalents of these classes from cunn, cudnn, and fbcunn may be successfully imported.

isBinary : bool

evaluate : bool

Python prototype (for reference):

readNetFromTorch(model[, isBinary[, evaluate]]) -> retval

readTensorFromONNX!(path)

Raising version of readTensorFromONNX/1.

readTensorFromONNX(path)

readTorchBlob!(filename)

Raising version of readTorchBlob/1.

readTorchBlob!(filename, opts)

Raising version of readTorchBlob/2.

readTorchBlob(filename)

readTorchBlob(filename, opts)

Loads blob which was serialized as torch.Tensor object of Torch7 framework. @warning This function has the same limitations as readNetFromTorch().

isBinary : bool

Python prototype (for reference):

readTorchBlob(filename[, isBinary]) -> retval

shrinkCaffeModel!(src, dst)

Raising version of shrinkCaffeModel/2.

shrinkCaffeModel!(src, dst, opts)

Raising version of shrinkCaffeModel/3.

shrinkCaffeModel(src, dst)

shrinkCaffeModel(src, dst, opts)

Convert all weights of Caffe network to half precision floating point.

src.
          precision floating point weights (usually has `.caffemodel` extension).

dst.

layersTypes.
                  By default, converts only Convolutional and Fully-Connected layers'
                  weights.
Note: Shrinked model has no origin float32 weights so it can't be used
   in origin Caffe framework anymore. However the structure of data
   is taken from NVidia's Caffe fork: https://github.com/NVIDIA/caffe.
   So the resulting model may be used there.

layersTypes : vector_String

Python prototype (for reference):

shrinkCaffeModel(src, dst[, layersTypes]) -> None

softNMSBoxes!(bboxes, scores, score_threshold, nms_threshold)

Raising version of softNMSBoxes/4.

softNMSBoxes!(bboxes, scores, score_threshold, nms_threshold, opts)

Raising version of softNMSBoxes/5.

softNMSBoxes(bboxes, scores, score_threshold, nms_threshold)

softNMSBoxes(bboxes, scores, score_threshold, nms_threshold, opts)

Performs soft non maximum suppression given boxes and corresponding scores. Reference: https://arxiv.org/abs/1704.04503

bboxes.

scores.

updated_scores.

score_threshold.

nms_threshold.

indices.

top_k.

sigma.

method. @see SoftNMSMethod

top_k : size_t

sigma : float

method : SoftNMSMethod

Python prototype (for reference):

softNMSBoxes(bboxes, scores, score_threshold, nms_threshold[, top_k[, sigma[, method]]]) -> updated_scores, indices

writeTextGraph!(model, output)

Raising version of writeTextGraph/2.

writeTextGraph(model, output)