Evision.DNN.ClassificationModel (Evision v0.1.9)

Functions

classify!(self, frame)

Raising version of classify/2.

classify(self, frame)

Positional Arguments

frame: Evision.Mat

Return

classId: int

conf: float

Has overloading in C++

Python prototype (for reference):

classify(frame) -> classId, conf

dnn_ClassificationModel!(network)

Raising version of dnn_ClassificationModel/1.

dnn_ClassificationModel!(model, opts)

Raising version of dnn_ClassificationModel/2.

dnn_ClassificationModel(network)

Variant 1:

Create model from deep learning network.

Positional Arguments

network: Net.

Net object.

Python prototype (for reference):

ClassificationModel(network) -> <dnn_ClassificationModel object>

Variant 2:

Create classification model from network represented in one of the supported formats. An order of @p model and @p config arguments does not matter.

Positional Arguments

model: String.

Binary file contains trained weights.

Keyword Arguments

config: String.

Text file contains network configuration.

Python prototype (for reference):

ClassificationModel(model[, config]) -> <dnn_ClassificationModel object>

dnn_ClassificationModel(model, opts)

Create classification model from network represented in one of the supported formats. An order of @p model and @p config arguments does not matter.

Positional Arguments

model: String.

Binary file contains trained weights.

Keyword Arguments

config: String.

Text file contains network configuration.

Python prototype (for reference):

ClassificationModel(model[, config]) -> <dnn_ClassificationModel object>

getEnableSoftmaxPostProcessing!(self)

Raising version of getEnableSoftmaxPostProcessing/1.

getEnableSoftmaxPostProcessing(self)

Get enable/disable softmax post processing option.

This option defaults to false, softmax post processing is not applied within the classify() function.

Python prototype (for reference):

getEnableSoftmaxPostProcessing() -> retval

setEnableSoftmaxPostProcessing!(self, enable)

Raising version of setEnableSoftmaxPostProcessing/2.

setEnableSoftmaxPostProcessing(self, enable)

Set enable/disable softmax post processing option.

Positional Arguments

enable: bool.

Set enable softmax post processing within the classify() function.

If this option is true, softmax is applied after forward inference within the classify() function to convert the confidences range to [0.0-1.0]. This function allows you to toggle this behavior. Please turn true when not contain softmax layer in model.

Python prototype (for reference):

setEnableSoftmaxPostProcessing(enable) -> retval