Evision.DNN.DetectionModel (Evision v0.1.9)

Functions

detect!(self, frame)

Raising version of detect/2.

detect!(self, frame, opts)

Raising version of detect/3.

detect(self, frame)

Given the @p input frame, create input blob, run net and return result detections.

Positional Arguments

frame: Evision.Mat

Keyword Arguments

confThreshold: float.

A threshold used to filter boxes by confidences.

nmsThreshold: float.

A threshold used in non maximum suppression.

Return

classIds: [int].

Class indexes in result detection.

confidences: [float].

A set of corresponding confidences.

boxes: [Rect].

A set of bounding boxes.

Python prototype (for reference):

detect(frame[, confThreshold[, nmsThreshold]]) -> classIds, confidences, boxes

detect(self, frame, opts)

Given the @p input frame, create input blob, run net and return result detections.

Positional Arguments

frame: Evision.Mat

Keyword Arguments

confThreshold: float.

A threshold used to filter boxes by confidences.

nmsThreshold: float.

A threshold used in non maximum suppression.

Return

classIds: [int].

Class indexes in result detection.

confidences: [float].

A set of corresponding confidences.

boxes: [Rect].

A set of bounding boxes.

Python prototype (for reference):

detect(frame[, confThreshold[, nmsThreshold]]) -> classIds, confidences, boxes

dnn_DetectionModel!(network)

Raising version of dnn_DetectionModel/1.

dnn_DetectionModel!(model, opts)

Raising version of dnn_DetectionModel/2.

dnn_DetectionModel(network)

Variant 1:

Create model from deep learning network.

Positional Arguments

network: Net.

Net object.

Python prototype (for reference):

DetectionModel(network) -> <dnn_DetectionModel object>

Variant 2:

Create detection model from network represented in one of the supported formats. An order of @p model and @p config arguments does not matter.

Positional Arguments

model: String.

Binary file contains trained weights.

Keyword Arguments

config: String.

Text file contains network configuration.

Python prototype (for reference):

DetectionModel(model[, config]) -> <dnn_DetectionModel object>

dnn_DetectionModel(model, opts)

Create detection model from network represented in one of the supported formats. An order of @p model and @p config arguments does not matter.

Positional Arguments

model: String.

Binary file contains trained weights.

Keyword Arguments

config: String.

Text file contains network configuration.

Python prototype (for reference):

DetectionModel(model[, config]) -> <dnn_DetectionModel object>

getNmsAcrossClasses!(self)

Raising version of getNmsAcrossClasses/1.

getNmsAcrossClasses(self)

Getter for nmsAcrossClasses. This variable defaults to false, such that when non max suppression is used during the detect() function, it will do so only per-class

Python prototype (for reference):

getNmsAcrossClasses() -> retval

setNmsAcrossClasses!(self, value)

Raising version of setNmsAcrossClasses/2.

setNmsAcrossClasses(self, value)

nmsAcrossClasses defaults to false, such that when non max suppression is used during the detect() function, it will do so per-class. This function allows you to toggle this behaviour.

Positional Arguments

value: bool.

The new value for nmsAcrossClasses

Python prototype (for reference):

setNmsAcrossClasses(value) -> retval