Evision.DNN.Layer (Evision v0.1.7-dev)

Functions

finalize!(self, inputs)

Raising version of finalize/2.

finalize!(self, inputs, opts)

Raising version of finalize/3.

finalize(self, inputs)

Positional Arguments

inputs: vector_UMat

Keyword Arguments

outputs: vector_UMat.

Computes and sets internal parameters according to inputs, outputs and blobs. If this method is called after network has allocated all memory for input and output blobs and before inferencing.

Python prototype (for reference):

finalize(inputs[, outputs]) -> outputs

finalize(self, inputs, opts)

Positional Arguments

inputs: vector_UMat

Keyword Arguments

outputs: vector_UMat.

Computes and sets internal parameters according to inputs, outputs and blobs. If this method is called after network has allocated all memory for input and output blobs and before inferencing.

Python prototype (for reference):

finalize(inputs[, outputs]) -> outputs

get_blobs!(self)

Raising version of get_blobs/1.

get_blobs(self)

get_name!(self)

Raising version of get_name/1.

get_name(self)

get_preferableTarget!(self)

Raising version of get_preferableTarget/1.

get_preferableTarget(self)

get_type!(self)

Raising version of get_type/1.

get_type(self)

outputNameToIndex!(self, outputName)

Raising version of outputNameToIndex/2.

outputNameToIndex(self, outputName)

Returns index of output blob in output array. @see inputNameToIndex()

Python prototype (for reference):

outputNameToIndex(outputName) -> retval

run!(self, inputs, internals)

Raising version of run/3.

run!(self, inputs, internals, opts)

Raising version of run/4.

run(self, inputs, internals)

Positional Arguments

inputs: vector_Mat

internals: vector_Mat

Keyword Arguments

outputs: vector_Mat.

Allocates layer and computes output. @deprecated This method will be removed in the future release.

Python prototype (for reference):

run(inputs, internals[, outputs]) -> outputs, internals

run(self, inputs, internals, opts)

Positional Arguments

inputs: vector_Mat

internals: vector_Mat

Keyword Arguments

outputs: vector_Mat.

Allocates layer and computes output. @deprecated This method will be removed in the future release.

Python prototype (for reference):

run(inputs, internals[, outputs]) -> outputs, internals

set_blobs!(self, prop)

Raising version of set_blobs/2.

set_blobs(self, prop)