Evision.DNN.Model (Evision v0.1.7-dev)

Functions

dnn_Model!(network)

Raising version of dnn_Model/1.

dnn_Model!(model, opts)

Raising version of dnn_Model/2.

dnn_Model(network)

Variaint 1:

Create model from deep learning network.

Python prototype (for reference):

Model(network) -> <dnn_Model object>

Variaint 2:

Positional Arguments

model: String.

Keyword Arguments

config: String.

Create model from deep learning network represented in one of the supported formats. An order of @p model and @p config arguments does not matter.

Python prototype (for reference):

Model(model[, config]) -> <dnn_Model object>

dnn_Model(model, opts)

Positional Arguments

model: String.

Keyword Arguments

config: String.

Create model from deep learning network represented in one of the supported formats. An order of @p model and @p config arguments does not matter.

Python prototype (for reference):

Model(model[, config]) -> <dnn_Model object>

predict!(self, frame)

Raising version of predict/2.

predict!(self, frame, opts)

Raising version of predict/3.

predict(self, frame)

Positional Arguments

frame: UMat

Keyword Arguments

outs: vector_UMat.

Given the @p input frame, create input blob, run net and return the output @p blobs.

Python prototype (for reference):

predict(frame[, outs]) -> outs

predict(self, frame, opts)

Positional Arguments

frame: UMat

Keyword Arguments

outs: vector_UMat.

Given the @p input frame, create input blob, run net and return the output @p blobs.

Python prototype (for reference):

predict(frame[, outs]) -> outs

setInputCrop!(self, crop)

Raising version of setInputCrop/2.

setInputCrop(self, crop)

Set flag crop for frame.

Python prototype (for reference):

setInputCrop(crop) -> retval

setInputMean!(self, mean)

Raising version of setInputMean/2.

setInputMean(self, mean)

Set mean value for frame.

Python prototype (for reference):

setInputMean(mean) -> retval

setInputParams!(self)

Raising version of setInputParams/1.

setInputParams(self)

Positional Arguments

Keyword Arguments

scale: double.

size: Size.

mean: Scalar.

swapRB: bool.

crop: bool.

Set preprocessing parameters for frame.

  blob(n, c, y, x) = scale * resize( frame(y, x, c) ) - mean(c) )

Python prototype (for reference):

setInputParams([, scale[, size[, mean[, swapRB[, crop]]]]]) -> None

setInputScale!(self, scale)

Raising version of setInputScale/2.

setInputScale(self, scale)

Set scalefactor value for frame.

Python prototype (for reference):

setInputScale(scale) -> retval

setInputSize!(self, size)

Raising version of setInputSize/2.

setInputSize!(self, width, height)

Raising version of setInputSize/3.

setInputSize(self, size)

Set input size for frame. Note: If shape of the new blob less than 0, then frame size not change.

Python prototype (for reference):

setInputSize(size) -> retval

setInputSize(self, width, height)

Has overloading in C++

Python prototype (for reference):

setInputSize(width, height) -> retval

setInputSwapRB!(self, swapRB)

Raising version of setInputSwapRB/2.

setInputSwapRB(self, swapRB)

Set flag swapRB for frame.

Python prototype (for reference):

setInputSwapRB(swapRB) -> retval

setPreferableBackend!(self, backendId)

Raising version of setPreferableBackend/2.

setPreferableBackend(self, backendId)

Python prototype (for reference):

setPreferableBackend(backendId) -> retval

setPreferableTarget!(self, targetId)

Raising version of setPreferableTarget/2.

setPreferableTarget(self, targetId)

Python prototype (for reference):

setPreferableTarget(targetId) -> retval