Evision.DNN.Net (Evision v0.1.5-dev)

Functions

connect!(self, outPin, inpPin)

Raising version of connect/3.

connect(self, outPin, inpPin)

dump!(self)

Raising version of dump/1.

dump(self)

dumpToFile!(self, path)

Raising version of dumpToFile/2.

dumpToFile(self, path)

empty!(self)

Raising version of empty/1.

empty(self)

enableFusion!(self, fusion)

Raising version of enableFusion/2.

enableFusion(self, fusion)

forward!(self, opts)

Raising version of forward/2.

forward(self, opts \\ [])

Runs forward pass to compute outputs of layers listed in @p outBlobNames.

outputBlobs.

outBlobNames.

outputBlobs : vector_UMat

Python prototype (for reference):

forward(outBlobNames[, outputBlobs]) -> outputBlobs

forwardAndRetrieve!(self, outBlobNames)

Raising version of forwardAndRetrieve/2.

forwardAndRetrieve(self, outBlobNames)

forwardAsync!(self)

Raising version of forwardAsync/1.

forwardAsync(self)

getFLOPS!(self, netInputShape)

Raising version of getFLOPS/2.

getFLOPS!(self, layerId, netInputShape)

Raising version of getFLOPS/3.

getFLOPS(self, netInputShape)

getFLOPS(self, layerId, netInputShape)

getInputDetails!(self)

Raising version of getInputDetails/1.

getInputDetails(self)

getLayer!(self, layerName)

Raising version of getLayer/2.

getLayer(self, layerName)

getLayerId!(self, layer)

Raising version of getLayerId/2.

getLayerId(self, layer)

getLayerNames!(self)

Raising version of getLayerNames/1.

getLayerNames(self)

getLayerShapes!(self, opts)

Raising version of getLayerShapes/2.

getLayerShapes(self, opts \\ [])

Has overloading in C++

Python prototype (for reference):

getLayerShapes(netInputShapes, layerId) -> inLayerShapes, outLayerShapes

getLayersCount!(self, layerType)

Raising version of getLayersCount/2.

getLayersCount(self, layerType)

getLayersShapes!(self, opts)

Raising version of getLayersShapes/2.

getLayersShapes(self, opts \\ [])

Has overloading in C++

Python prototype (for reference):

getLayersShapes(netInputShape) -> layersIds, inLayersShapes, outLayersShapes

getLayerTypes!(self)

Raising version of getLayerTypes/1.

getLayerTypes(self)

getMemoryConsumption!(self, netInputShape)

Raising version of getMemoryConsumption/2.

getMemoryConsumption!(self, layerId, netInputShape)

Raising version of getMemoryConsumption/3.

getMemoryConsumption(self, netInputShape)

getMemoryConsumption(self, layerId, netInputShape)

getOutputDetails!(self)

Raising version of getOutputDetails/1.

getOutputDetails(self)

getParam!(self, layerName)

Raising version of getParam/2.

getParam!(self, layerName, opts)

Raising version of getParam/3.

getParam(self, layerName)

getParam(self, layerName, opts)

numParam : int

Python prototype (for reference):

getParam(layerName[, numParam]) -> retval

getPerfProfile!(self)

Raising version of getPerfProfile/1.

getPerfProfile(self)

getUnconnectedOutLayers!(self)

Raising version of getUnconnectedOutLayers/1.

getUnconnectedOutLayers(self)

getUnconnectedOutLayersNames!(self)

Raising version of getUnconnectedOutLayersNames/1.

getUnconnectedOutLayersNames(self)

net!()

Raising version of net/0.

net()

quantize!(self, calibData, inputsDtype, outputsDtype)

Raising version of quantize/4.

quantize(self, calibData, inputsDtype, outputsDtype)

readFromModelOptimizer!(bufferModelConfig, bufferWeights)

Raising version of readFromModelOptimizer/2.

readFromModelOptimizer(bufferModelConfig, bufferWeights)

setHalideScheduler!(self, scheduler)

Raising version of setHalideScheduler/2.

setHalideScheduler(self, scheduler)

setInput!(self, blob)

Raising version of setInput/2.

setInput!(self, blob, opts)

Raising version of setInput/3.

setInput(self, blob)

setInput(self, blob, opts)

Sets the new input value for the network

blob.

name.

scalefactor.

mean. @see connect(String, String) to know format of the descriptor. If scale or mean values are specified, a final input blob is computed as: \f[input(n,c,h,w) = scalefactor \times (blob(n,c,h,w) - mean_c)\f]

name : String

scalefactor : double

mean : Scalar

Python prototype (for reference):

setInput(blob[, name[, scalefactor[, mean]]]) -> None

setInputShape!(self, inputName, shape)

Raising version of setInputShape/3.

setInputShape(self, inputName, shape)

setInputsNames!(self, inputBlobNames)

Raising version of setInputsNames/2.

setInputsNames(self, inputBlobNames)

setParam!(self, layerName, numParam, blob)

Raising version of setParam/4.

setParam(self, layerName, numParam, blob)

setPreferableBackend!(self, backendId)

Raising version of setPreferableBackend/2.

setPreferableBackend(self, backendId)

setPreferableTarget!(self, targetId)

Raising version of setPreferableTarget/2.

setPreferableTarget(self, targetId)