Evision.DNN.SegmentationModel (Evision v0.1.7-dev)

Functions

dnn_SegmentationModel!(network)

Raising version of dnn_SegmentationModel/1.

dnn_SegmentationModel!(model, opts)

Raising version of dnn_SegmentationModel/2.

dnn_SegmentationModel(network)

Variaint 1:

Create model from deep learning network.

Python prototype (for reference):

SegmentationModel(network) -> <dnn_SegmentationModel object>

Variaint 2:

Positional Arguments

model: String.

Keyword Arguments

config: String.

Create segmentation model from network represented in one of the supported formats. An order of @p model and @p config arguments does not matter.

Python prototype (for reference):

SegmentationModel(model[, config]) -> <dnn_SegmentationModel object>

dnn_SegmentationModel(model, opts)

Positional Arguments

model: String.

Keyword Arguments

config: String.

Create segmentation model from network represented in one of the supported formats. An order of @p model and @p config arguments does not matter.

Python prototype (for reference):

SegmentationModel(model[, config]) -> <dnn_SegmentationModel object>

segment!(self, frame)

Raising version of segment/2.

segment!(self, frame, opts)

Raising version of segment/3.

segment(self, frame)

Positional Arguments

frame: UMat

Keyword Arguments

mask: UMat.

Given the @p input frame, create input blob, run net

Python prototype (for reference):

segment(frame[, mask]) -> mask

segment(self, frame, opts)

Positional Arguments

frame: UMat

Keyword Arguments

mask: UMat.

Given the @p input frame, create input blob, run net

Python prototype (for reference):

segment(frame[, mask]) -> mask