Evision.DNN.TextDetectionModelDB (Evision v0.1.9)

Functions

dnn_TextDetectionModel_DB!(model)

Raising version of dnn_TextDetectionModel_DB/1.

dnn_TextDetectionModel_DB!(model, opts)

Raising version of dnn_TextDetectionModel_DB/2.

dnn_TextDetectionModel_DB(model)

Create text detection model from network represented in one of the supported formats. An order of @p model and @p config arguments does not matter.

Positional Arguments

model: string.

Binary file contains trained weights.

Keyword Arguments

config: string.

Text file contains network configuration.

Python prototype (for reference):

TextDetectionModel_DB(model[, config]) -> <dnn_TextDetectionModel_DB object>

dnn_TextDetectionModel_DB(model, opts)

Create text detection model from network represented in one of the supported formats. An order of @p model and @p config arguments does not matter.

Positional Arguments

model: string.

Binary file contains trained weights.

Keyword Arguments

config: string.

Text file contains network configuration.

Python prototype (for reference):

TextDetectionModel_DB(model[, config]) -> <dnn_TextDetectionModel_DB object>

getBinaryThreshold!(self)

Raising version of getBinaryThreshold/1.

getBinaryThreshold(self)

Python prototype (for reference):

getBinaryThreshold() -> retval

getMaxCandidates!(self)

Raising version of getMaxCandidates/1.

getMaxCandidates(self)

Python prototype (for reference):

getMaxCandidates() -> retval

getPolygonThreshold!(self)

Raising version of getPolygonThreshold/1.

getPolygonThreshold(self)

Python prototype (for reference):

getPolygonThreshold() -> retval

getUnclipRatio!(self)

Raising version of getUnclipRatio/1.

getUnclipRatio(self)

Python prototype (for reference):

getUnclipRatio() -> retval

setBinaryThreshold!(self, binaryThreshold)

Raising version of setBinaryThreshold/2.

setBinaryThreshold(self, binaryThreshold)

Positional Arguments

binaryThreshold: float

Python prototype (for reference):

setBinaryThreshold(binaryThreshold) -> retval

setMaxCandidates!(self, maxCandidates)

Raising version of setMaxCandidates/2.

setMaxCandidates(self, maxCandidates)

Positional Arguments

maxCandidates: int

Python prototype (for reference):

setMaxCandidates(maxCandidates) -> retval

setPolygonThreshold!(self, polygonThreshold)

Raising version of setPolygonThreshold/2.

setPolygonThreshold(self, polygonThreshold)

Positional Arguments

polygonThreshold: float

Python prototype (for reference):

setPolygonThreshold(polygonThreshold) -> retval

setUnclipRatio!(self, unclipRatio)

Raising version of setUnclipRatio/2.

setUnclipRatio(self, unclipRatio)

Positional Arguments

unclipRatio: double

Python prototype (for reference):

setUnclipRatio(unclipRatio) -> retval