Evision.DNN.TextRecognitionModel (Evision v0.1.9)

Functions

dnn_TextRecognitionModel!(model)

Raising version of dnn_TextRecognitionModel/1.

dnn_TextRecognitionModel!(model, opts)

Raising version of dnn_TextRecognitionModel/2.

dnn_TextRecognitionModel(model)

Create text recognition model from network represented in one of the supported formats Call setDecodeType() and setVocabulary() after constructor to initialize the decoding method

Positional Arguments

model: string.

Binary file contains trained weights

Keyword Arguments

config: string.

Text file contains network configuration

Python prototype (for reference):

TextRecognitionModel(model[, config]) -> <dnn_TextRecognitionModel object>

dnn_TextRecognitionModel(model, opts)

Create text recognition model from network represented in one of the supported formats Call setDecodeType() and setVocabulary() after constructor to initialize the decoding method

Positional Arguments

model: string.

Binary file contains trained weights

Keyword Arguments

config: string.

Text file contains network configuration

Python prototype (for reference):

TextRecognitionModel(model[, config]) -> <dnn_TextRecognitionModel object>

getDecodeType!(self)

Raising version of getDecodeType/1.

getDecodeType(self)

Get the decoding method

@return the decoding method

Python prototype (for reference):

getDecodeType() -> retval

getVocabulary!(self)

Raising version of getVocabulary/1.

getVocabulary(self)

Get the vocabulary for recognition.

@return vocabulary the associated vocabulary

Python prototype (for reference):

getVocabulary() -> retval

recognize!(self, frame)

Raising version of recognize/2.

recognize!(self, frame, roiRects)

Raising version of recognize/3.

recognize(self, frame)

Given the @p input frame, create input blob, run net and return recognition result

Positional Arguments

frame: Evision.Mat.

The input image

@return The text recognition result

Python prototype (for reference):

recognize(frame) -> retval

recognize(self, frame, roiRects)

Given the @p input frame, create input blob, run net and return recognition result

Positional Arguments

frame: Evision.Mat.

The input image

roiRects: [Evision.Mat].

List of text detection regions of interest (cv::Rect, CV_32SC4). ROIs is be cropped as the network inputs

Return

results: [string].

A set of text recognition results.

Python prototype (for reference):

recognize(frame, roiRects) -> results

setDecodeOptsCTCPrefixBeamSearch!(self, beamSize)

Raising version of setDecodeOptsCTCPrefixBeamSearch/2.

setDecodeOptsCTCPrefixBeamSearch!(self, beamSize, opts)

Raising version of setDecodeOptsCTCPrefixBeamSearch/3.

setDecodeOptsCTCPrefixBeamSearch(self, beamSize)

Set the decoding method options for "CTC-prefix-beam-search" decode usage

Positional Arguments

beamSize: int.

Beam size for search

Keyword Arguments

vocPruneSize: int.

Parameter to optimize big vocabulary search, only take top @p vocPruneSize tokens in each search step, @p vocPruneSize <= 0 stands for disable this prune.

Python prototype (for reference):

setDecodeOptsCTCPrefixBeamSearch(beamSize[, vocPruneSize]) -> retval

setDecodeOptsCTCPrefixBeamSearch(self, beamSize, opts)

Set the decoding method options for "CTC-prefix-beam-search" decode usage

Positional Arguments

beamSize: int.

Beam size for search

Keyword Arguments

vocPruneSize: int.

Parameter to optimize big vocabulary search, only take top @p vocPruneSize tokens in each search step, @p vocPruneSize <= 0 stands for disable this prune.

Python prototype (for reference):

setDecodeOptsCTCPrefixBeamSearch(beamSize[, vocPruneSize]) -> retval

setDecodeType!(self, decodeType)

Raising version of setDecodeType/2.

setDecodeType(self, decodeType)

Set the decoding method of translating the network output into string

Positional Arguments

decodeType: string.The decoding method of translating the network output into string, currently supported type:

"CTC-greedy" greedy decoding for the output of CTC-based methods

"CTC-prefix-beam-search" Prefix beam search decoding for the output of CTC-based methods

Python prototype (for reference):

setDecodeType(decodeType) -> retval

setVocabulary!(self, vocabulary)

Raising version of setVocabulary/2.

setVocabulary(self, vocabulary)

Set the vocabulary for recognition.

Positional Arguments

vocabulary: [string].

the associated vocabulary of the network.

Python prototype (for reference):

setVocabulary(vocabulary) -> retval