Evision.Detail (Evision v0.1.7-dev)

cv.detail

calibrateRotatingCamera(hs)

Positional Arguments

hs: vector_Mat

Keyword Arguments

k: Mat.

Estimates focal lengths for each given camera.

Python prototype (for reference):

calibrateRotatingCamera(Hs[, K]) -> retval, K

calibrateRotatingCamera(hs, opts)

Positional Arguments

hs: vector_Mat

Keyword Arguments

k: Mat.

Estimates focal lengths for each given camera.

Python prototype (for reference):

calibrateRotatingCamera(Hs[, K]) -> retval, K

computeImageFeatures2(featuresFinder, image)

Positional Arguments

featuresFinder: Ptr<Feature2D>.

image: UMat.

Keyword Arguments

mask: UMat.

Python prototype (for reference):

computeImageFeatures2(featuresFinder, image[, mask]) -> features

computeImageFeatures2(featuresFinder, image, opts)

Positional Arguments

featuresFinder: Ptr<Feature2D>.

image: UMat.

Keyword Arguments

mask: UMat.

Python prototype (for reference):

computeImageFeatures2(featuresFinder, image[, mask]) -> features

computeImageFeatures(featuresFinder, images)

Positional Arguments

featuresFinder: Ptr<Feature2D>.

images: vector_UMat.

Keyword Arguments

masks: vector_UMat.

Python prototype (for reference):

computeImageFeatures(featuresFinder, images[, masks]) -> features

computeImageFeatures(featuresFinder, images, opts)

Positional Arguments

featuresFinder: Ptr<Feature2D>.

images: vector_UMat.

Keyword Arguments

masks: vector_UMat.

Python prototype (for reference):

computeImageFeatures(featuresFinder, images[, masks]) -> features

createLaplacePyr(img, num_levels, pyr)

Python prototype (for reference):

createLaplacePyr(img, num_levels, pyr) -> pyr

createLaplacePyrGpu(img, num_levels, pyr)

Python prototype (for reference):

createLaplacePyrGpu(img, num_levels, pyr) -> pyr

createWeightMap(mask, sharpness, weight)

Python prototype (for reference):

createWeightMap(mask, sharpness, weight) -> weight

focalsFromHomography(h, f0, f1, f0_ok, f1_ok)

Tries to estimate focal lengths from the given homography under the assumption that the camera undergoes rotations around its centre only. See "Construction of Panoramic Image Mosaics with Global and Local Alignment" by Heung-Yeung Shum and Richard Szeliski.

Python prototype (for reference):

focalsFromHomography(H, f0, f1, f0_ok, f1_ok) -> None

leaveBiggestComponent(features, pairwise_matches, conf_threshold)

Python prototype (for reference):

leaveBiggestComponent(features, pairwise_matches, conf_threshold) -> retval

matchesGraphAsString(pathes, pairwise_matches, conf_threshold)

Python prototype (for reference):

matchesGraphAsString(pathes, pairwise_matches, conf_threshold) -> retval

normalizeUsingWeightMap(weight, src)

Python prototype (for reference):

normalizeUsingWeightMap(weight, src) -> src

overlapRoi(tl1, tl2, sz1, sz2, roi)

Python prototype (for reference):

overlapRoi(tl1, tl2, sz1, sz2, roi) -> retval

restoreImageFromLaplacePyr(pyr)

Python prototype (for reference):

restoreImageFromLaplacePyr(pyr) -> pyr

restoreImageFromLaplacePyrGpu(pyr)

Python prototype (for reference):

restoreImageFromLaplacePyrGpu(pyr) -> pyr

resultRoi(corners, sizes)

Python prototype (for reference):

resultRoi(corners, sizes) -> retval

resultRoiIntersection(corners, sizes)

Python prototype (for reference):

resultRoiIntersection(corners, sizes) -> retval

resultTl(corners)

Python prototype (for reference):

resultTl(corners) -> retval

selectRandomSubset(count, size, subset)

Python prototype (for reference):

selectRandomSubset(count, size, subset) -> None

stitchingLogLevel()

Python prototype (for reference):

stitchingLogLevel() -> retval

waveCorrect(rmats, kind)

Tries to make panorama more horizontal (or vertical).

Python prototype (for reference):

waveCorrect(rmats, kind) -> rmats

Functions

calibrateRotatingCamera!(hs)

Raising version of calibrateRotatingCamera/1.

calibrateRotatingCamera!(hs, opts)

Raising version of calibrateRotatingCamera/2.

computeImageFeatures2!(featuresFinder, image)

Raising version of computeImageFeatures2/2.

computeImageFeatures2!(featuresFinder, image, opts)

Raising version of computeImageFeatures2/3.

computeImageFeatures!(featuresFinder, images)

Raising version of computeImageFeatures/2.

computeImageFeatures!(featuresFinder, images, opts)

Raising version of computeImageFeatures/3.

createLaplacePyr!(img, num_levels, pyr)

Raising version of createLaplacePyr/3.

createLaplacePyrGpu!(img, num_levels, pyr)

Raising version of createLaplacePyrGpu/3.

createWeightMap!(mask, sharpness, weight)

Raising version of createWeightMap/3.

focalsFromHomography!(h, f0, f1, f0_ok, f1_ok)

Raising version of focalsFromHomography/5.

leaveBiggestComponent!(features, pairwise_matches, conf_threshold)

Raising version of leaveBiggestComponent/3.

matchesGraphAsString!(pathes, pairwise_matches, conf_threshold)

Raising version of matchesGraphAsString/3.

normalizeUsingWeightMap!(weight, src)

Raising version of normalizeUsingWeightMap/2.

overlapRoi!(tl1, tl2, sz1, sz2, roi)

Raising version of overlapRoi/5.

restoreImageFromLaplacePyr!(pyr)

Raising version of restoreImageFromLaplacePyr/1.

restoreImageFromLaplacePyrGpu!(pyr)

Raising version of restoreImageFromLaplacePyrGpu/1.

resultRoi!(corners, sizes)

Raising version of resultRoi/2.

resultRoiIntersection!(corners, sizes)

Raising version of resultRoiIntersection/2.

resultTl!(corners)

Raising version of resultTl/1.

selectRandomSubset!(count, size, subset)

Raising version of selectRandomSubset/3.

stitchingLogLevel!()

Raising version of stitchingLogLevel/0.

waveCorrect!(rmats, kind)

Raising version of waveCorrect/2.