Evision.Detail.AffineBestOf2NearestMatcher (Evision v0.1.7-dev)

cv.detail

apply2(self, features)

Positional Arguments

features: vector_ImageFeatures.

Keyword Arguments

mask: UMat.

Performs images matching. The function is parallelized with the TBB library. @sa detail::MatchesInfo

Python prototype (for reference):

apply2(features[, mask]) -> pairwise_matches

apply2(self, features, opts)

Positional Arguments

features: vector_ImageFeatures.

Keyword Arguments

mask: UMat.

Performs images matching. The function is parallelized with the TBB library. @sa detail::MatchesInfo

Python prototype (for reference):

apply2(features[, mask]) -> pairwise_matches

apply(self, features1, features2)

Has overloading in C++

Python prototype (for reference):

apply(features1, features2) -> matches_info

collectGarbage(self)

Python prototype (for reference):

collectGarbage() -> None

create()

Positional Arguments

Keyword Arguments

try_use_gpu: bool.

match_conf: float.

num_matches_thresh1: int.

num_matches_thresh2: int.

Python prototype (for reference):

create([, try_use_gpu[, match_conf[, num_matches_thresh1[, num_matches_thresh2]]]]) -> retval

create(opts)

Positional Arguments

Keyword Arguments

try_use_gpu: bool.

match_conf: float.

num_matches_thresh1: int.

num_matches_thresh2: int.

Python prototype (for reference):

create([, try_use_gpu[, match_conf[, num_matches_thresh1[, num_matches_thresh2]]]]) -> retval

detail_AffineBestOf2NearestMatcher()

Positional Arguments

Keyword Arguments

full_affine: bool.

try_use_gpu: bool.

match_conf: float.

num_matches_thresh1: int.

Constructs a "best of 2 nearest" matcher that expects affine transformation between images

transformation with 4 degrees of freedom using only rotation, translation and uniform scaling
estimation used in the inliers classification step

@sa cv::estimateAffine2D cv::estimateAffinePartial2D

Python prototype (for reference):

AffineBestOf2NearestMatcher([, full_affine[, try_use_gpu[, match_conf[, num_matches_thresh1]]]]) -> <detail_AffineBestOf2NearestMatcher object>

detail_AffineBestOf2NearestMatcher(opts)

Positional Arguments

Keyword Arguments

full_affine: bool.

try_use_gpu: bool.

match_conf: float.

num_matches_thresh1: int.

Constructs a "best of 2 nearest" matcher that expects affine transformation between images

transformation with 4 degrees of freedom using only rotation, translation and uniform scaling
estimation used in the inliers classification step

@sa cv::estimateAffine2D cv::estimateAffinePartial2D

Python prototype (for reference):

AffineBestOf2NearestMatcher([, full_affine[, try_use_gpu[, match_conf[, num_matches_thresh1]]]]) -> <detail_AffineBestOf2NearestMatcher object>

isThreadSafe(self)

@return True, if it's possible to use the same matcher instance in parallel, false otherwise

Python prototype (for reference):

isThreadSafe() -> retval

Functions

apply2!(self, features)

Raising version of apply2/2.

apply2!(self, features, opts)

Raising version of apply2/3.

apply!(self, features1, features2)

Raising version of apply/3.

collectGarbage!(self)

Raising version of collectGarbage/1.

create!()

Raising version of create/0.

create!(opts)

Raising version of create/1.

detail_AffineBestOf2NearestMatcher!()

Raising version of detail_AffineBestOf2NearestMatcher/0.

detail_AffineBestOf2NearestMatcher!(opts)

Raising version of detail_AffineBestOf2NearestMatcher/1.

isThreadSafe!(self)

Raising version of isThreadSafe/1.