Evision.FaceDetectorYN (Evision v0.1.9)

cv

create(model, config, input_size)

Creates an instance of this class with given parameters

Positional Arguments

model: String.

the path to the requested model

config: String.

the path to the config file for compability, which is not requested for ONNX models

input_size: Size.

the size of the input image

Keyword Arguments

score_threshold: float.

the threshold to filter out bounding boxes of score smaller than the given value

nms_threshold: float.

the threshold to suppress bounding boxes of IoU bigger than the given value

top_k: int.

keep top K bboxes before NMS

backend_id: int.

the id of backend

target_id: int.

the id of target device

Python prototype (for reference):

create(model, config, input_size[, score_threshold[, nms_threshold[, top_k[, backend_id[, target_id]]]]]) -> retval

create(model, config, input_size, opts)

Creates an instance of this class with given parameters

Positional Arguments

model: String.

the path to the requested model

config: String.

the path to the config file for compability, which is not requested for ONNX models

input_size: Size.

the size of the input image

Keyword Arguments

score_threshold: float.

the threshold to filter out bounding boxes of score smaller than the given value

nms_threshold: float.

the threshold to suppress bounding boxes of IoU bigger than the given value

top_k: int.

keep top K bboxes before NMS

backend_id: int.

the id of backend

target_id: int.

the id of target device

Python prototype (for reference):

create(model, config, input_size[, score_threshold[, nms_threshold[, top_k[, backend_id[, target_id]]]]]) -> retval

detect(self, image)

A simple interface to detect face from given image

Positional Arguments

image: Evision.Mat.

an image to detect

Return

faces: Evision.Mat.

detection results stored in a cv::Mat

Python prototype (for reference):

detect(image[, faces]) -> retval, faces

detect(self, image, opts)

A simple interface to detect face from given image

Positional Arguments

image: Evision.Mat.

an image to detect

Return

faces: Evision.Mat.

detection results stored in a cv::Mat

Python prototype (for reference):

detect(image[, faces]) -> retval, faces

getInputSize(self)

Python prototype (for reference):

getInputSize() -> retval

getNMSThreshold(self)

Python prototype (for reference):

getNMSThreshold() -> retval

getScoreThreshold(self)

Python prototype (for reference):

getScoreThreshold() -> retval

getTopK(self)

Python prototype (for reference):

getTopK() -> retval

setInputSize(self, input_size)

Set the size for the network input, which overwrites the input size of creating model. Call this method when the size of input image does not match the input size when creating model

Positional Arguments

input_size: Size.

the size of the input image

Python prototype (for reference):

setInputSize(input_size) -> None

setNMSThreshold(self, nms_threshold)

Set the Non-maximum-suppression threshold to suppress bounding boxes that have IoU greater than the given value

Positional Arguments

nms_threshold: float.

threshold for NMS operation

Python prototype (for reference):

setNMSThreshold(nms_threshold) -> None

setScoreThreshold(self, score_threshold)

Set the score threshold to filter out bounding boxes of score less than the given value

Positional Arguments

score_threshold: float.

threshold for filtering out bounding boxes

Python prototype (for reference):

setScoreThreshold(score_threshold) -> None

setTopK(self, top_k)

Set the number of bounding boxes preserved before NMS

Positional Arguments

top_k: int.

the number of bounding boxes to preserve from top rank based on score

Python prototype (for reference):

setTopK(top_k) -> None

Functions

create!(model, config, input_size)

Raising version of create/3.

create!(model, config, input_size, opts)

Raising version of create/4.

detect!(self, image)

Raising version of detect/2.

detect!(self, image, opts)

Raising version of detect/3.

getInputSize!(self)

Raising version of getInputSize/1.

getNMSThreshold!(self)

Raising version of getNMSThreshold/1.

getScoreThreshold!(self)

Raising version of getScoreThreshold/1.

getTopK!(self)

Raising version of getTopK/1.

setInputSize!(self, input_size)

Raising version of setInputSize/2.

setNMSThreshold!(self, nms_threshold)

Raising version of setNMSThreshold/2.

setScoreThreshold!(self, score_threshold)

Raising version of setScoreThreshold/2.

setTopK!(self, top_k)

Raising version of setTopK/2.