Evision.FileNode (Evision v0.1.7-dev)

cv

at(self, i)

Has overloading in C++

Python prototype (for reference):

at(i) -> retval

empty(self)

Python prototype (for reference):

empty() -> retval

fileNode()

The constructors. These constructors are used to create a default file node, construct it from obsolete structures or from the another file node.

Python prototype (for reference):

FileNode() -> <FileNode object>

getNode(self, nodename)

Has overloading in C++

Python prototype (for reference):

getNode(nodename) -> retval

isInt(self)

Python prototype (for reference):

isInt() -> retval

isMap(self)

Python prototype (for reference):

isMap() -> retval

isNamed(self)

Python prototype (for reference):

isNamed() -> retval

isNone(self)

Python prototype (for reference):

isNone() -> retval

isReal(self)

Python prototype (for reference):

isReal() -> retval

isSeq(self)

Python prototype (for reference):

isSeq() -> retval

isString(self)

Python prototype (for reference):

isString() -> retval

keys(self)

Returns keys of a mapping node. @returns Keys of a mapping node.

Python prototype (for reference):

keys() -> retval

mat(self)

Python prototype (for reference):

mat() -> retval

name(self)

Python prototype (for reference):

name() -> retval

rawSize(self)

Python prototype (for reference):

rawSize() -> retval

real(self)

Internal method used when reading FileStorage. Sets the type (int, real or string) and value of the previously created node.

Python prototype (for reference):

real() -> retval

size(self)

Python prototype (for reference):

size() -> retval

string(self)

Python prototype (for reference):

string() -> retval

type(self)

Returns type of the node. @returns Type of the node. See FileNode::Type

Python prototype (for reference):

type() -> retval

Functions

at!(self, i)

Raising version of at/2.

empty!(self)

Raising version of empty/1.

fileNode!()

Raising version of fileNode/0.

getNode!(self, nodename)

Raising version of getNode/2.

isInt!(self)

Raising version of isInt/1.

isMap!(self)

Raising version of isMap/1.

isNamed!(self)

Raising version of isNamed/1.

isNone!(self)

Raising version of isNone/1.

isReal!(self)

Raising version of isReal/1.

isSeq!(self)

Raising version of isSeq/1.

isString!(self)

Raising version of isString/1.

keys!(self)

Raising version of keys/1.

mat!(self)

Raising version of mat/1.

name!(self)

Raising version of name/1.

rawSize!(self)

Raising version of rawSize/1.

real!(self)

Raising version of real/1.

size!(self)

Raising version of size/1.

string!(self)

Raising version of string/1.

type!(self)

Raising version of type/1.