Evision.HighGui (Evision v0.1.9)

OpenCV High-level Graphical User Interface

cv.highgui

destroyAllWindows()

close all windows

destroyWindow(winname)

close a named window

imshow(winname, mat)

Show a mat in a named window

winname. The name of the window.

mat. The image.

Example

{:ok, mat} = Evision.imread("example.jpg")
Evision.imshow("OpenCV", mat)
# the following line may be necessary on your system
# will try to improve this later
Evision.waitkey(0)

waitkey(delay)

wait for user keyboard event for a delay amount of time (ms)

Functions

destroyAllWindows!()

Raising version of destroyAllWindows/0.

destroyWindow!(winname)

Raising version of destroyWindow/1.

imshow!(winname, mat)

Raising version of imshow/2.

waitkey!(delay)

Raising version of waitkey/1.