Evision.KAZE (Evision v0.1.7-dev)

cv

compute(self, images, keypoints)

Variaint 1:

Positional Arguments

images: vector_UMat.

keypoints: vector_vector_KeyPoint.

Keyword Arguments

descriptors: vector_UMat.

Has overloading in C++

computed are removed. Sometimes new keypoints can be added, for example: SIFT duplicates keypoint
with several dominant orientations (for each orientation).
descriptors computed for a keypoints[i]. Row j is the keypoints (or keypoints[i]) is the
descriptor for keypoint j-th keypoint.

Python prototype (for reference):

compute(images, keypoints[, descriptors]) -> keypoints, descriptors

Variaint 2:

Positional Arguments

image: UMat. Image.

keypoints: vector_KeyPoint.

Keyword Arguments

descriptors: UMat.

Computes the descriptors for a set of keypoints detected in an image (first variant) or image set (second variant).

computed are removed. Sometimes new keypoints can be added, for example: SIFT duplicates keypoint
with several dominant orientations (for each orientation).
descriptors computed for a keypoints[i]. Row j is the keypoints (or keypoints[i]) is the
descriptor for keypoint j-th keypoint.

Python prototype (for reference):

compute(image, keypoints[, descriptors]) -> keypoints, descriptors

compute(self, images, keypoints, opts)

Variaint 1:

Positional Arguments

images: vector_UMat.

keypoints: vector_vector_KeyPoint.

Keyword Arguments

descriptors: vector_UMat.

Has overloading in C++

computed are removed. Sometimes new keypoints can be added, for example: SIFT duplicates keypoint
with several dominant orientations (for each orientation).
descriptors computed for a keypoints[i]. Row j is the keypoints (or keypoints[i]) is the
descriptor for keypoint j-th keypoint.

Python prototype (for reference):

compute(images, keypoints[, descriptors]) -> keypoints, descriptors

Variaint 2:

Positional Arguments

image: UMat. Image.

keypoints: vector_KeyPoint.

Keyword Arguments

descriptors: UMat.

Computes the descriptors for a set of keypoints detected in an image (first variant) or image set (second variant).

computed are removed. Sometimes new keypoints can be added, for example: SIFT duplicates keypoint
with several dominant orientations (for each orientation).
descriptors computed for a keypoints[i]. Row j is the keypoints (or keypoints[i]) is the
descriptor for keypoint j-th keypoint.

Python prototype (for reference):

compute(image, keypoints[, descriptors]) -> keypoints, descriptors

create()

Positional Arguments

Keyword Arguments

extended: bool.

upright: bool.

threshold: float.

nOctaves: int.

nOctaveLayers: int.

diffusivity: KAZE_DiffusivityType.

The KAZE constructor

DIFF_CHARBONNIER

Python prototype (for reference):

create([, extended[, upright[, threshold[, nOctaves[, nOctaveLayers[, diffusivity]]]]]]) -> retval

create(opts)

Positional Arguments

Keyword Arguments

extended: bool.

upright: bool.

threshold: float.

nOctaves: int.

nOctaveLayers: int.

diffusivity: KAZE_DiffusivityType.

The KAZE constructor

DIFF_CHARBONNIER

Python prototype (for reference):

create([, extended[, upright[, threshold[, nOctaves[, nOctaveLayers[, diffusivity]]]]]]) -> retval

defaultNorm(self)

Python prototype (for reference):

defaultNorm() -> retval

descriptorSize(self)

Python prototype (for reference):

descriptorSize() -> retval

descriptorType(self)

Python prototype (for reference):

descriptorType() -> retval

detect(self, images)

Variaint 1:

Positional Arguments

images: vector_UMat.

Keyword Arguments

masks: vector_UMat.

Has overloading in C++

of keypoints detected in images[i] .
masks[i] is a mask for images[i].

Python prototype (for reference):

detect(images[, masks]) -> keypoints

Variaint 2:

Positional Arguments

image: UMat. Image.

Keyword Arguments

mask: UMat.

Detects keypoints in an image (first variant) or image set (second variant).

of keypoints detected in images[i] .
matrix with non-zero values in the region of interest.

Python prototype (for reference):

detect(image[, mask]) -> keypoints

detect(self, images, opts)

Variaint 1:

Positional Arguments

images: vector_UMat.

Keyword Arguments

masks: vector_UMat.

Has overloading in C++

of keypoints detected in images[i] .
masks[i] is a mask for images[i].

Python prototype (for reference):

detect(images[, masks]) -> keypoints

Variaint 2:

Positional Arguments

image: UMat. Image.

Keyword Arguments

mask: UMat.

Detects keypoints in an image (first variant) or image set (second variant).

of keypoints detected in images[i] .
matrix with non-zero values in the region of interest.

Python prototype (for reference):

detect(image[, mask]) -> keypoints

detectAndCompute(self, image, mask)

Positional Arguments

image: UMat

mask: UMat

Keyword Arguments

descriptors: UMat.

useProvidedKeypoints: bool.

Detects keypoints and computes the descriptors

Python prototype (for reference):

detectAndCompute(image, mask[, descriptors[, useProvidedKeypoints]]) -> keypoints, descriptors

detectAndCompute(self, image, mask, opts)

Positional Arguments

image: UMat

mask: UMat

Keyword Arguments

descriptors: UMat.

useProvidedKeypoints: bool.

Detects keypoints and computes the descriptors

Python prototype (for reference):

detectAndCompute(image, mask[, descriptors[, useProvidedKeypoints]]) -> keypoints, descriptors

empty(self)

Python prototype (for reference):

empty() -> retval

getDefaultName(self)

Python prototype (for reference):

getDefaultName() -> retval

getDiffusivity(self)

Python prototype (for reference):

getDiffusivity() -> retval

getExtended(self)

Python prototype (for reference):

getExtended() -> retval

getNOctaveLayers(self)

Python prototype (for reference):

getNOctaveLayers() -> retval

getNOctaves(self)

Python prototype (for reference):

getNOctaves() -> retval

getThreshold(self)

Python prototype (for reference):

getThreshold() -> retval

getUpright(self)

Python prototype (for reference):

getUpright() -> retval

read(self, arg1)

Variaint 1:

Python prototype (for reference):

read(arg1) -> None

Variaint 2:

Python prototype (for reference):

read(fileName) -> None

setDiffusivity(self, diff)

Python prototype (for reference):

setDiffusivity(diff) -> None

setExtended(self, extended)

Python prototype (for reference):

setExtended(extended) -> None

setNOctaveLayers(self, octaveLayers)

Python prototype (for reference):

setNOctaveLayers(octaveLayers) -> None

setNOctaves(self, octaves)

Python prototype (for reference):

setNOctaves(octaves) -> None

setThreshold(self, threshold)

Python prototype (for reference):

setThreshold(threshold) -> None

setUpright(self, upright)

Python prototype (for reference):

setUpright(upright) -> None

write(self, fs)

Variaint 1:

Positional Arguments

fs: Ptr<FileStorage>

Keyword Arguments

name: String.

Python prototype (for reference):

write(fs[, name]) -> None

Variaint 2:

Python prototype (for reference):

write(fileName) -> None

write(self, fs, opts)

Positional Arguments

fs: Ptr<FileStorage>

Keyword Arguments

name: String.

Python prototype (for reference):

write(fs[, name]) -> None

Functions

compute!(self, images, keypoints)

Raising version of compute/3.

compute!(self, images, keypoints, opts)

Raising version of compute/4.

create!()

Raising version of create/0.

create!(opts)

Raising version of create/1.

defaultNorm!(self)

Raising version of defaultNorm/1.

descriptorSize!(self)

Raising version of descriptorSize/1.

descriptorType!(self)

Raising version of descriptorType/1.

detect!(self, images)

Raising version of detect/2.

detect!(self, images, opts)

Raising version of detect/3.

detectAndCompute!(self, image, mask)

Raising version of detectAndCompute/3.

detectAndCompute!(self, image, mask, opts)

Raising version of detectAndCompute/4.

empty!(self)

Raising version of empty/1.

getDefaultName!(self)

Raising version of getDefaultName/1.

getDiffusivity!(self)

Raising version of getDiffusivity/1.

getExtended!(self)

Raising version of getExtended/1.

getNOctaveLayers!(self)

Raising version of getNOctaveLayers/1.

getNOctaves!(self)

Raising version of getNOctaves/1.

getThreshold!(self)

Raising version of getThreshold/1.

getUpright!(self)

Raising version of getUpright/1.

read!(self, arg1)

Raising version of read/2.

setDiffusivity!(self, diff)

Raising version of setDiffusivity/2.

setExtended!(self, extended)

Raising version of setExtended/2.

setNOctaveLayers!(self, octaveLayers)

Raising version of setNOctaveLayers/2.

setNOctaves!(self, octaves)

Raising version of setNOctaves/2.

setThreshold!(self, threshold)

Raising version of setThreshold/2.

setUpright!(self, upright)

Raising version of setUpright/2.

write!(self, fs)

Raising version of write/2.

write!(self, fs, opts)

Raising version of write/3.