Evision.KeyPoint (Evision v0.1.9)

cv

convert(points2f)

Positional Arguments

points2f: [Point2f].

Array of (x,y) coordinates of each keypoint

Keyword Arguments

size: float.

keypoint diameter

response: float.

keypoint detector response on the keypoint (that is, strength of the keypoint)

octave: int.

pyramid octave in which the keypoint has been detected

class_id: int.

object id

Return

keypoints: [KeyPoint].

Keypoints obtained from any feature detection algorithm like SIFT/SURF/ORB

Has overloading in C++

Python prototype (for reference):

convert(points2f[, size[, response[, octave[, class_id]]]]) -> keypoints

convert(points2f, opts)

Positional Arguments

points2f: [Point2f].

Array of (x,y) coordinates of each keypoint

Keyword Arguments

size: float.

keypoint diameter

response: float.

keypoint detector response on the keypoint (that is, strength of the keypoint)

octave: int.

pyramid octave in which the keypoint has been detected

class_id: int.

object id

Return

keypoints: [KeyPoint].

Keypoints obtained from any feature detection algorithm like SIFT/SURF/ORB

Has overloading in C++

Python prototype (for reference):

convert(points2f[, size[, response[, octave[, class_id]]]]) -> keypoints

keyPoint()

Python prototype (for reference):

KeyPoint() -> <KeyPoint object>

keyPoint(x, y, size)

Positional Arguments

x: float.

x-coordinate of the keypoint

y: float.

y-coordinate of the keypoint

size: float.

keypoint diameter

Keyword Arguments

angle: float.

keypoint orientation

response: float.

keypoint detector response on the keypoint (that is, strength of the keypoint)

octave: int.

pyramid octave in which the keypoint has been detected

class_id: int.

object id

Python prototype (for reference):

KeyPoint(x, y, size[, angle[, response[, octave[, class_id]]]]) -> <KeyPoint object>

keyPoint(x, y, size, opts)

Positional Arguments

x: float.

x-coordinate of the keypoint

y: float.

y-coordinate of the keypoint

size: float.

keypoint diameter

Keyword Arguments

angle: float.

keypoint orientation

response: float.

keypoint detector response on the keypoint (that is, strength of the keypoint)

octave: int.

pyramid octave in which the keypoint has been detected

class_id: int.

object id

Python prototype (for reference):

KeyPoint(x, y, size[, angle[, response[, octave[, class_id]]]]) -> <KeyPoint object>

overlap(kp1, kp2)

Positional Arguments

kp1: KeyPoint.

First keypoint

kp2: KeyPoint.

Second keypoint

This method computes overlap for pair of keypoints. Overlap is the ratio between area of keypoint regions' intersection and area of keypoint regions' union (considering keypoint region as circle). If they don't overlap, we get zero. If they coincide at same location with same size, we get 1.

Python prototype (for reference):

overlap(kp1, kp2) -> retval

Functions

convert!(points2f)

Raising version of convert/1.

convert!(points2f, opts)

Raising version of convert/2.

get_angle!(self)

Raising version of get_angle/1.

get_angle(self)

get_class_id!(self)

Raising version of get_class_id/1.

get_class_id(self)

get_octave!(self)

Raising version of get_octave/1.

get_octave(self)

get_pt!(self)

Raising version of get_pt/1.

get_pt(self)

get_response!(self)

Raising version of get_response/1.

get_response(self)

get_size!(self)

Raising version of get_size/1.

get_size(self)

keyPoint!()

Raising version of keyPoint/0.

keyPoint!(x, y, size)

Raising version of keyPoint/3.

keyPoint!(x, y, size, opts)

Raising version of keyPoint/4.

overlap!(kp1, kp2)

Raising version of overlap/2.

set_angle!(self, prop)

Raising version of set_angle/2.

set_angle(self, prop)

set_class_id!(self, prop)

Raising version of set_class_id/2.

set_class_id(self, prop)

set_octave!(self, prop)

Raising version of set_octave/2.

set_octave(self, prop)

set_pt!(self, prop)

Raising version of set_pt/2.

set_pt(self, prop)

set_response!(self, prop)

Raising version of set_response/2.

set_response(self, prop)

set_size!(self, prop)

Raising version of set_size/2.

set_size(self, prop)