Evision.ML.ANNMLP (Evision v0.1.5-dev)

cv

clear(self)

getDefaultName(self)

read(self, fn_)

save(self, filename)

write(self, fs)

write(self, fs, opts)

simplified API for language bindings Has overloading in C++

name : String

Python prototype (for reference):

write(fs[, name]) -> None

cv.ml

calcError(self, data, test)

calcError(self, data, test, opts)

Computes error on the training or test dataset

data.

test. computed over the training subset of the data. Please note that if you loaded a completely different dataset to evaluate already trained classifier, you will probably want not to set the test subset at all with TrainData::setTrainTestSplitRatio and specify test=false, so that the error is computed for the whole new set. Yes, this sounds a bit confusing.

resp. The method uses StatModel::predict to compute the error. For regression models the error is computed as RMS, for classifiers - as a percent of missclassified samples (0%-100%).

resp : UMat

Python prototype (for reference):

calcError(data, test[, resp]) -> retval, resp

create()

empty(self)

getAnnealCoolingRatio(self)

getAnnealFinalT(self)

getAnnealInitialT(self)

getAnnealItePerStep(self)

getBackpropMomentumScale(self)

getBackpropWeightScale(self)

getLayerSizes(self)

getRpropDW0(self)

getRpropDWMax(self)

getRpropDWMin(self)

getRpropDWMinus(self)

getRpropDWPlus(self)

getTermCriteria(self)

getTrainMethod(self)

getVarCount(self)

getWeights(self, layerIdx)

isClassifier(self)

isTrained(self)

load(filepath)

predict(self, samples)

predict(self, samples, opts)

Predicts response(s) for the provided sample(s)

samples.

results.

flags.

results : UMat

flags : int

Python prototype (for reference):

predict(samples[, results[, flags]]) -> retval, results

setActivationFunction(self, type)

setActivationFunction(self, type, opts)

Initialize the activation function for each neuron. Currently the default and the only fully supported activation function is ANN_MLP::SIGMOID_SYM.

type.

param1.

param2.

param1 : double

param2 : double

Python prototype (for reference):

setActivationFunction(type[, param1[, param2]]) -> None

setAnnealCoolingRatio(self, val)

setAnnealFinalT(self, val)

setAnnealInitialT(self, val)

setAnnealItePerStep(self, val)

setBackpropMomentumScale(self, val)

setBackpropWeightScale(self, val)

setLayerSizes(self, layer_sizes)

setRpropDW0(self, val)

setRpropDWMax(self, val)

setRpropDWMin(self, val)

setRpropDWMinus(self, val)

setRpropDWPlus(self, val)

setTermCriteria(self, val)

setTrainMethod(self, method)

setTrainMethod(self, method, opts)

Sets training method and common parameters.

method.

param1.

param2.

param1 : double

param2 : double

Python prototype (for reference):

setTrainMethod(method[, param1[, param2]]) -> None

train(self, trainData)

train(self, trainData, opts)

train(self, samples, layout, responses)

Functions

calcError!(self, data, test)

Raising version of calcError/3.

calcError!(self, data, test, opts)

Raising version of calcError/4.

clear!(self)

Raising version of clear/1.

create!()

Raising version of create/0.

empty!(self)

Raising version of empty/1.

getAnnealCoolingRatio!(self)

Raising version of getAnnealCoolingRatio/1.

getAnnealFinalT!(self)

Raising version of getAnnealFinalT/1.

getAnnealInitialT!(self)

Raising version of getAnnealInitialT/1.

getAnnealItePerStep!(self)

Raising version of getAnnealItePerStep/1.

getBackpropMomentumScale!(self)

Raising version of getBackpropMomentumScale/1.

getBackpropWeightScale!(self)

Raising version of getBackpropWeightScale/1.

getDefaultName!(self)

Raising version of getDefaultName/1.

getLayerSizes!(self)

Raising version of getLayerSizes/1.

getRpropDW0!(self)

Raising version of getRpropDW0/1.

getRpropDWMax!(self)

Raising version of getRpropDWMax/1.

getRpropDWMin!(self)

Raising version of getRpropDWMin/1.

getRpropDWMinus!(self)

Raising version of getRpropDWMinus/1.

getRpropDWPlus!(self)

Raising version of getRpropDWPlus/1.

getTermCriteria!(self)

Raising version of getTermCriteria/1.

getTrainMethod!(self)

Raising version of getTrainMethod/1.

getVarCount!(self)

Raising version of getVarCount/1.

getWeights!(self, layerIdx)

Raising version of getWeights/2.

isClassifier!(self)

Raising version of isClassifier/1.

isTrained!(self)

Raising version of isTrained/1.

load!(filepath)

Raising version of load/1.

predict!(self, samples)

Raising version of predict/2.

predict!(self, samples, opts)

Raising version of predict/3.

read!(self, fn_)

Raising version of read/2.

save!(self, filename)

Raising version of save/2.

setActivationFunction!(self, type)

Raising version of setActivationFunction/2.

setActivationFunction!(self, type, opts)

Raising version of setActivationFunction/3.

setAnnealCoolingRatio!(self, val)

Raising version of setAnnealCoolingRatio/2.

setAnnealFinalT!(self, val)

Raising version of setAnnealFinalT/2.

setAnnealInitialT!(self, val)

Raising version of setAnnealInitialT/2.

setAnnealItePerStep!(self, val)

Raising version of setAnnealItePerStep/2.

setBackpropMomentumScale!(self, val)

Raising version of setBackpropMomentumScale/2.

setBackpropWeightScale!(self, val)

Raising version of setBackpropWeightScale/2.

setLayerSizes!(self, layer_sizes)

Raising version of setLayerSizes/2.

setRpropDW0!(self, val)

Raising version of setRpropDW0/2.

setRpropDWMax!(self, val)

Raising version of setRpropDWMax/2.

setRpropDWMin!(self, val)

Raising version of setRpropDWMin/2.

setRpropDWMinus!(self, val)

Raising version of setRpropDWMinus/2.

setRpropDWPlus!(self, val)

Raising version of setRpropDWPlus/2.

setTermCriteria!(self, val)

Raising version of setTermCriteria/2.

setTrainMethod!(self, method)

Raising version of setTrainMethod/2.

setTrainMethod!(self, method, opts)

Raising version of setTrainMethod/3.

train!(self, trainData)

Raising version of train/2.

train!(self, trainData, opts)

Raising version of train/3.

train!(self, samples, layout, responses)

Raising version of train/4.

write!(self, fs)

Raising version of write/2.

write!(self, fs, opts)

Raising version of write/3.