Evision.ML.EM (Evision v0.1.7-dev)

cv

clear(self)

Clears the algorithm state

Python prototype (for reference):

clear() -> None

getDefaultName(self)

Returns the algorithm string identifier. This string is used as top level xml/yml node tag when the object is saved to a file or string.

Python prototype (for reference):

getDefaultName() -> retval

read(self, fn_)

Reads algorithm parameters from a file storage

Python prototype (for reference):

read(fn_) -> None

save(self, filename)

Saves the algorithm to a file. In order to make this method work, the derived class must implement Algorithm::write(FileStorage& fs).

Python prototype (for reference):

save(filename) -> None

write(self, fs)

Positional Arguments

fs: Ptr<FileStorage>

Keyword Arguments

name: String.

simplified API for language bindings Has overloading in C++

Python prototype (for reference):

write(fs[, name]) -> None

write(self, fs, opts)

Positional Arguments

fs: Ptr<FileStorage>

Keyword Arguments

name: String.

simplified API for language bindings Has overloading in C++

Python prototype (for reference):

write(fs[, name]) -> None

cv.ml

calcError(self, data, test)

Positional Arguments

data: Ptr<TrainData>.

test: bool.

Keyword Arguments

resp: UMat.

Computes error on the training or test dataset

computed over the training subset of the data. Please note that if you loaded a completely
different dataset to evaluate already trained classifier, you will probably want not to set
the test subset at all with TrainData::setTrainTestSplitRatio and specify test=false, so
that the error is computed for the whole new set. Yes, this sounds a bit confusing.

The method uses StatModel::predict to compute the error. For regression models the error is computed as RMS, for classifiers - as a percent of missclassified samples (0%-100%).

Python prototype (for reference):

calcError(data, test[, resp]) -> retval, resp

calcError(self, data, test, opts)

Positional Arguments

data: Ptr<TrainData>.

test: bool.

Keyword Arguments

resp: UMat.

Computes error on the training or test dataset

computed over the training subset of the data. Please note that if you loaded a completely
different dataset to evaluate already trained classifier, you will probably want not to set
the test subset at all with TrainData::setTrainTestSplitRatio and specify test=false, so
that the error is computed for the whole new set. Yes, this sounds a bit confusing.

The method uses StatModel::predict to compute the error. For regression models the error is computed as RMS, for classifiers - as a percent of missclassified samples (0%-100%).

Python prototype (for reference):

calcError(data, test[, resp]) -> retval, resp

create()

Creates empty %EM model. The model should be trained then using StatModel::train(traindata, flags) method. Alternatively, you can use one of the EM::train* methods or load it from file using Algorithm::load\<EM>(filename).

Python prototype (for reference):

create() -> retval

empty(self)

Python prototype (for reference):

empty() -> retval

getClustersNumber(self)

@see setClustersNumber

Python prototype (for reference):

getClustersNumber() -> retval

getCovarianceMatrixType(self)

@see setCovarianceMatrixType

Python prototype (for reference):

getCovarianceMatrixType() -> retval

getCovs(self)

Positional Arguments

Keyword Arguments

covs: vector_Mat.

Returns covariation matrices Returns vector of covariation matrices. Number of matrices is the number of gaussian mixtures, each matrix is a square floating-point matrix NxN, where N is the space dimensionality.

Python prototype (for reference):

getCovs([, covs]) -> covs

getMeans(self)

Returns the cluster centers (means of the Gaussian mixture) Returns matrix with the number of rows equal to the number of mixtures and number of columns equal to the space dimensionality.

Python prototype (for reference):

getMeans() -> retval

getTermCriteria(self)

@see setTermCriteria

Python prototype (for reference):

getTermCriteria() -> retval

getVarCount(self)

Returns the number of variables in training samples

Python prototype (for reference):

getVarCount() -> retval

getWeights(self)

Returns weights of the mixtures Returns vector with the number of elements equal to the number of mixtures.

Python prototype (for reference):

getWeights() -> retval

isClassifier(self)

Returns true if the model is classifier

Python prototype (for reference):

isClassifier() -> retval

isTrained(self)

Returns true if the model is trained

Python prototype (for reference):

isTrained() -> retval

load(filepath)

Positional Arguments

filepath: String.

Keyword Arguments

nodeName: String.

Loads and creates a serialized EM from a file Use EM::save to serialize and store an EM to disk. Load the EM from this file again, by calling this function with the path to the file. Optionally specify the node for the file containing the classifier

Python prototype (for reference):

load(filepath[, nodeName]) -> retval

load(filepath, opts)

Positional Arguments

filepath: String.

Keyword Arguments

nodeName: String.

Loads and creates a serialized EM from a file Use EM::save to serialize and store an EM to disk. Load the EM from this file again, by calling this function with the path to the file. Optionally specify the node for the file containing the classifier

Python prototype (for reference):

load(filepath[, nodeName]) -> retval

predict2(self, sample)

Positional Arguments

sample: UMat.

Keyword Arguments

probs: UMat.

Returns a likelihood logarithm value and an index of the most probable mixture component for the given sample.

\f$1 \times dims\f$ or \f$dims \times 1\f$ size.
given the sample. It has \f$1 \times nclusters\f$ size and CV_64FC1 type.

The method returns a two-element double vector. Zero element is a likelihood logarithm value for the sample. First element is an index of the most probable mixture component for the given sample.

Python prototype (for reference):

predict2(sample[, probs]) -> retval, probs

predict2(self, sample, opts)

Positional Arguments

sample: UMat.

Keyword Arguments

probs: UMat.

Returns a likelihood logarithm value and an index of the most probable mixture component for the given sample.

\f$1 \times dims\f$ or \f$dims \times 1\f$ size.
given the sample. It has \f$1 \times nclusters\f$ size and CV_64FC1 type.

The method returns a two-element double vector. Zero element is a likelihood logarithm value for the sample. First element is an index of the most probable mixture component for the given sample.

Python prototype (for reference):

predict2(sample[, probs]) -> retval, probs

predict(self, samples)

Positional Arguments

samples: UMat.

Keyword Arguments

results: UMat.

flags: int.

Returns posterior probabilities for the provided samples

posterior probabilities for each sample from the input

Python prototype (for reference):

predict(samples[, results[, flags]]) -> retval, results

predict(self, samples, opts)

Positional Arguments

samples: UMat.

Keyword Arguments

results: UMat.

flags: int.

Returns posterior probabilities for the provided samples

posterior probabilities for each sample from the input

Python prototype (for reference):

predict(samples[, results[, flags]]) -> retval, results

setClustersNumber(self, val)

@copybrief getClustersNumber @see getClustersNumber

Python prototype (for reference):

setClustersNumber(val) -> None

setCovarianceMatrixType(self, val)

@copybrief getCovarianceMatrixType @see getCovarianceMatrixType

Python prototype (for reference):

setCovarianceMatrixType(val) -> None

setTermCriteria(self, val)

@copybrief getTermCriteria @see getTermCriteria

Python prototype (for reference):

setTermCriteria(val) -> None

train(self, trainData)

Positional Arguments

trainData: Ptr<TrainData>.

Keyword Arguments

flags: int.

Trains the statistical model

created with TrainData::create.
new training samples, not completely overwritten (such as NormalBayesClassifier or ANN_MLP).

Python prototype (for reference):

train(trainData[, flags]) -> retval

train(self, trainData, opts)

Positional Arguments

trainData: Ptr<TrainData>.

Keyword Arguments

flags: int.

Trains the statistical model

created with TrainData::create.
new training samples, not completely overwritten (such as NormalBayesClassifier or ANN_MLP).

Python prototype (for reference):

train(trainData[, flags]) -> retval

train(self, samples, layout, responses)

Trains the statistical model

Python prototype (for reference):

train(samples, layout, responses) -> retval

trainE(self, samples, means0)

Positional Arguments

samples: UMat.

means0: UMat.

Keyword Arguments

covs0: UMat.

weights0: UMat.

logLikelihoods: UMat.

labels: UMat.

probs: UMat.

Estimate the Gaussian mixture parameters from a samples set. This variation starts with Expectation step. You need to provide initial means \f$a_k\f$ of mixture components. Optionally you can pass initial weights \f$\pi_k\f$ and covariance matrices \f$S_k\f$ of mixture components.

one-channel matrix, each row of which is a sample. If the matrix does not have CV_64F type
it will be converted to the inner matrix of such type for the further computing.
\f$nclusters \times dims\f$ size. If the matrix does not have CV_64F type it will be
converted to the inner matrix of such type for the further computing.
covariance matrices is a one-channel matrix of \f$dims \times dims\f$ size. If the matrices
do not have CV_64F type they will be converted to the inner matrices of such type for the
further computing.
floating-point matrix with \f$1 \times nclusters\f$ or \f$nclusters \times 1\f$ size.
each sample. It has \f$nsamples \times 1\f$ size and CV_64FC1 type.
\f$\texttt{labels}\_i=\texttt{arg max}\_k(p\_{i,k}), i=1..N\f$ (indices of the most probable
mixture component for each sample). It has \f$nsamples \times 1\f$ size and CV_32SC1 type.
mixture component given the each sample. It has \f$nsamples \times nclusters\f$ size and
CV_64FC1 type.

Python prototype (for reference):

trainE(samples, means0[, covs0[, weights0[, logLikelihoods[, labels[, probs]]]]]) -> retval, logLikelihoods, labels, probs

trainE(self, samples, means0, opts)

Positional Arguments

samples: UMat.

means0: UMat.

Keyword Arguments

covs0: UMat.

weights0: UMat.

logLikelihoods: UMat.

labels: UMat.

probs: UMat.

Estimate the Gaussian mixture parameters from a samples set. This variation starts with Expectation step. You need to provide initial means \f$a_k\f$ of mixture components. Optionally you can pass initial weights \f$\pi_k\f$ and covariance matrices \f$S_k\f$ of mixture components.

one-channel matrix, each row of which is a sample. If the matrix does not have CV_64F type
it will be converted to the inner matrix of such type for the further computing.
\f$nclusters \times dims\f$ size. If the matrix does not have CV_64F type it will be
converted to the inner matrix of such type for the further computing.
covariance matrices is a one-channel matrix of \f$dims \times dims\f$ size. If the matrices
do not have CV_64F type they will be converted to the inner matrices of such type for the
further computing.
floating-point matrix with \f$1 \times nclusters\f$ or \f$nclusters \times 1\f$ size.
each sample. It has \f$nsamples \times 1\f$ size and CV_64FC1 type.
\f$\texttt{labels}\_i=\texttt{arg max}\_k(p\_{i,k}), i=1..N\f$ (indices of the most probable
mixture component for each sample). It has \f$nsamples \times 1\f$ size and CV_32SC1 type.
mixture component given the each sample. It has \f$nsamples \times nclusters\f$ size and
CV_64FC1 type.

Python prototype (for reference):

trainE(samples, means0[, covs0[, weights0[, logLikelihoods[, labels[, probs]]]]]) -> retval, logLikelihoods, labels, probs

trainEM(self, samples)

Positional Arguments

samples: UMat.

Keyword Arguments

logLikelihoods: UMat.

labels: UMat.

probs: UMat.

Estimate the Gaussian mixture parameters from a samples set. This variation starts with Expectation step. Initial values of the model parameters will be estimated by the k-means algorithm. Unlike many of the ML models, %EM is an unsupervised learning algorithm and it does not take responses (class labels or function values) as input. Instead, it computes the Maximum Likelihood Estimate of the Gaussian mixture parameters from an input sample set, stores all the parameters inside the structure: \f$p_{i,k}\f$ in probs, \f$a_k\f$ in means , \f$S_k\f$ in covs[k], \f$\pi_k\f$ in weights , and optionally computes the output "class label" for each sample: \f$\texttt{labels}_i=\texttt{arg max}_k(p_{i,k}), i=1..N\f$ (indices of the most probable mixture component for each sample). The trained model can be used further for prediction, just like any other classifier. The trained model is similar to the NormalBayesClassifier.

one-channel matrix, each row of which is a sample. If the matrix does not have CV_64F type
it will be converted to the inner matrix of such type for the further computing.
each sample. It has \f$nsamples \times 1\f$ size and CV_64FC1 type.
\f$\texttt{labels}\_i=\texttt{arg max}\_k(p\_{i,k}), i=1..N\f$ (indices of the most probable
mixture component for each sample). It has \f$nsamples \times 1\f$ size and CV_32SC1 type.
mixture component given the each sample. It has \f$nsamples \times nclusters\f$ size and
CV_64FC1 type.

Python prototype (for reference):

trainEM(samples[, logLikelihoods[, labels[, probs]]]) -> retval, logLikelihoods, labels, probs

trainEM(self, samples, opts)

Positional Arguments

samples: UMat.

Keyword Arguments

logLikelihoods: UMat.

labels: UMat.

probs: UMat.

Estimate the Gaussian mixture parameters from a samples set. This variation starts with Expectation step. Initial values of the model parameters will be estimated by the k-means algorithm. Unlike many of the ML models, %EM is an unsupervised learning algorithm and it does not take responses (class labels or function values) as input. Instead, it computes the Maximum Likelihood Estimate of the Gaussian mixture parameters from an input sample set, stores all the parameters inside the structure: \f$p_{i,k}\f$ in probs, \f$a_k\f$ in means , \f$S_k\f$ in covs[k], \f$\pi_k\f$ in weights , and optionally computes the output "class label" for each sample: \f$\texttt{labels}_i=\texttt{arg max}_k(p_{i,k}), i=1..N\f$ (indices of the most probable mixture component for each sample). The trained model can be used further for prediction, just like any other classifier. The trained model is similar to the NormalBayesClassifier.

one-channel matrix, each row of which is a sample. If the matrix does not have CV_64F type
it will be converted to the inner matrix of such type for the further computing.
each sample. It has \f$nsamples \times 1\f$ size and CV_64FC1 type.
\f$\texttt{labels}\_i=\texttt{arg max}\_k(p\_{i,k}), i=1..N\f$ (indices of the most probable
mixture component for each sample). It has \f$nsamples \times 1\f$ size and CV_32SC1 type.
mixture component given the each sample. It has \f$nsamples \times nclusters\f$ size and
CV_64FC1 type.

Python prototype (for reference):

trainEM(samples[, logLikelihoods[, labels[, probs]]]) -> retval, logLikelihoods, labels, probs

trainM(self, samples, probs0)

Positional Arguments

samples: UMat.

probs0: UMat.

Keyword Arguments

logLikelihoods: UMat.

labels: UMat.

probs: UMat.

Estimate the Gaussian mixture parameters from a samples set. This variation starts with Maximization step. You need to provide initial probabilities \f$p_{i,k}\f$ to use this option.

one-channel matrix, each row of which is a sample. If the matrix does not have CV_64F type
it will be converted to the inner matrix of such type for the further computing.
each sample. It has \f$nsamples \times 1\f$ size and CV_64FC1 type.
\f$\texttt{labels}\_i=\texttt{arg max}\_k(p\_{i,k}), i=1..N\f$ (indices of the most probable
mixture component for each sample). It has \f$nsamples \times 1\f$ size and CV_32SC1 type.
mixture component given the each sample. It has \f$nsamples \times nclusters\f$ size and
CV_64FC1 type.

Python prototype (for reference):

trainM(samples, probs0[, logLikelihoods[, labels[, probs]]]) -> retval, logLikelihoods, labels, probs

trainM(self, samples, probs0, opts)

Positional Arguments

samples: UMat.

probs0: UMat.

Keyword Arguments

logLikelihoods: UMat.

labels: UMat.

probs: UMat.

Estimate the Gaussian mixture parameters from a samples set. This variation starts with Maximization step. You need to provide initial probabilities \f$p_{i,k}\f$ to use this option.

one-channel matrix, each row of which is a sample. If the matrix does not have CV_64F type
it will be converted to the inner matrix of such type for the further computing.
each sample. It has \f$nsamples \times 1\f$ size and CV_64FC1 type.
\f$\texttt{labels}\_i=\texttt{arg max}\_k(p\_{i,k}), i=1..N\f$ (indices of the most probable
mixture component for each sample). It has \f$nsamples \times 1\f$ size and CV_32SC1 type.
mixture component given the each sample. It has \f$nsamples \times nclusters\f$ size and
CV_64FC1 type.

Python prototype (for reference):

trainM(samples, probs0[, logLikelihoods[, labels[, probs]]]) -> retval, logLikelihoods, labels, probs

Functions

calcError!(self, data, test)

Raising version of calcError/3.

calcError!(self, data, test, opts)

Raising version of calcError/4.

clear!(self)

Raising version of clear/1.

create!()

Raising version of create/0.

empty!(self)

Raising version of empty/1.

getClustersNumber!(self)

Raising version of getClustersNumber/1.

getCovarianceMatrixType!(self)

Raising version of getCovarianceMatrixType/1.

getCovs!(self)

Raising version of getCovs/1.

getDefaultName!(self)

Raising version of getDefaultName/1.

getMeans!(self)

Raising version of getMeans/1.

getTermCriteria!(self)

Raising version of getTermCriteria/1.

getVarCount!(self)

Raising version of getVarCount/1.

getWeights!(self)

Raising version of getWeights/1.

isClassifier!(self)

Raising version of isClassifier/1.

isTrained!(self)

Raising version of isTrained/1.

load!(filepath)

Raising version of load/1.

load!(filepath, opts)

Raising version of load/2.

predict2!(self, sample)

Raising version of predict2/2.

predict2!(self, sample, opts)

Raising version of predict2/3.

predict!(self, samples)

Raising version of predict/2.

predict!(self, samples, opts)

Raising version of predict/3.

read!(self, fn_)

Raising version of read/2.

save!(self, filename)

Raising version of save/2.

setClustersNumber!(self, val)

Raising version of setClustersNumber/2.

setCovarianceMatrixType!(self, val)

Raising version of setCovarianceMatrixType/2.

setTermCriteria!(self, val)

Raising version of setTermCriteria/2.

train!(self, trainData)

Raising version of train/2.

train!(self, trainData, opts)

Raising version of train/3.

train!(self, samples, layout, responses)

Raising version of train/4.

trainE!(self, samples, means0)

Raising version of trainE/3.

trainE!(self, samples, means0, opts)

Raising version of trainE/4.

trainEM!(self, samples)

Raising version of trainEM/2.

trainEM!(self, samples, opts)

Raising version of trainEM/3.

trainM!(self, samples, probs0)

Raising version of trainM/3.

trainM!(self, samples, probs0, opts)

Raising version of trainM/4.

write!(self, fs)

Raising version of write/2.

write!(self, fs, opts)

Raising version of write/3.