Evision.ML.LogisticRegression (Evision v0.1.7-dev)

cv

clear(self)

Clears the algorithm state

Python prototype (for reference):

clear() -> None

getDefaultName(self)

Returns the algorithm string identifier. This string is used as top level xml/yml node tag when the object is saved to a file or string.

Python prototype (for reference):

getDefaultName() -> retval

read(self, fn_)

Reads algorithm parameters from a file storage

Python prototype (for reference):

read(fn_) -> None

save(self, filename)

Saves the algorithm to a file. In order to make this method work, the derived class must implement Algorithm::write(FileStorage& fs).

Python prototype (for reference):

save(filename) -> None

write(self, fs)

Positional Arguments

fs: Ptr<FileStorage>

Keyword Arguments

name: String.

simplified API for language bindings Has overloading in C++

Python prototype (for reference):

write(fs[, name]) -> None

write(self, fs, opts)

Positional Arguments

fs: Ptr<FileStorage>

Keyword Arguments

name: String.

simplified API for language bindings Has overloading in C++

Python prototype (for reference):

write(fs[, name]) -> None

cv.ml

calcError(self, data, test)

Positional Arguments

data: Ptr<TrainData>.

test: bool.

Keyword Arguments

resp: UMat.

Computes error on the training or test dataset

computed over the training subset of the data. Please note that if you loaded a completely
different dataset to evaluate already trained classifier, you will probably want not to set
the test subset at all with TrainData::setTrainTestSplitRatio and specify test=false, so
that the error is computed for the whole new set. Yes, this sounds a bit confusing.

The method uses StatModel::predict to compute the error. For regression models the error is computed as RMS, for classifiers - as a percent of missclassified samples (0%-100%).

Python prototype (for reference):

calcError(data, test[, resp]) -> retval, resp

calcError(self, data, test, opts)

Positional Arguments

data: Ptr<TrainData>.

test: bool.

Keyword Arguments

resp: UMat.

Computes error on the training or test dataset

computed over the training subset of the data. Please note that if you loaded a completely
different dataset to evaluate already trained classifier, you will probably want not to set
the test subset at all with TrainData::setTrainTestSplitRatio and specify test=false, so
that the error is computed for the whole new set. Yes, this sounds a bit confusing.

The method uses StatModel::predict to compute the error. For regression models the error is computed as RMS, for classifiers - as a percent of missclassified samples (0%-100%).

Python prototype (for reference):

calcError(data, test[, resp]) -> retval, resp

create()

Creates empty model. Creates Logistic Regression model with parameters given.

Python prototype (for reference):

create() -> retval

empty(self)

Python prototype (for reference):

empty() -> retval

get_learnt_thetas(self)

This function returns the trained parameters arranged across rows. For a two class classification problem, it returns a row matrix. It returns learnt parameters of the Logistic Regression as a matrix of type CV_32F.

Python prototype (for reference):

get_learnt_thetas() -> retval

getIterations(self)

@see setIterations

Python prototype (for reference):

getIterations() -> retval

getLearningRate(self)

@see setLearningRate

Python prototype (for reference):

getLearningRate() -> retval

getMiniBatchSize(self)

@see setMiniBatchSize

Python prototype (for reference):

getMiniBatchSize() -> retval

getRegularization(self)

@see setRegularization

Python prototype (for reference):

getRegularization() -> retval

getTermCriteria(self)

@see setTermCriteria

Python prototype (for reference):

getTermCriteria() -> retval

getTrainMethod(self)

@see setTrainMethod

Python prototype (for reference):

getTrainMethod() -> retval

getVarCount(self)

Returns the number of variables in training samples

Python prototype (for reference):

getVarCount() -> retval

isClassifier(self)

Returns true if the model is classifier

Python prototype (for reference):

isClassifier() -> retval

isTrained(self)

Returns true if the model is trained

Python prototype (for reference):

isTrained() -> retval

load(filepath)

Positional Arguments

filepath: String.

Keyword Arguments

nodeName: String.

Loads and creates a serialized LogisticRegression from a file Use LogisticRegression::save to serialize and store an LogisticRegression to disk. Load the LogisticRegression from this file again, by calling this function with the path to the file. Optionally specify the node for the file containing the classifier

Python prototype (for reference):

load(filepath[, nodeName]) -> retval

load(filepath, opts)

Positional Arguments

filepath: String.

Keyword Arguments

nodeName: String.

Loads and creates a serialized LogisticRegression from a file Use LogisticRegression::save to serialize and store an LogisticRegression to disk. Load the LogisticRegression from this file again, by calling this function with the path to the file. Optionally specify the node for the file containing the classifier

Python prototype (for reference):

load(filepath[, nodeName]) -> retval

predict(self, samples)

Positional Arguments

samples: UMat.

Keyword Arguments

results: UMat.

flags: int.

Predicts responses for input samples and returns a float type.

contains variables (features) of one object being classified. Should have data type CV_32F.

Python prototype (for reference):

predict(samples[, results[, flags]]) -> retval, results

predict(self, samples, opts)

Positional Arguments

samples: UMat.

Keyword Arguments

results: UMat.

flags: int.

Predicts responses for input samples and returns a float type.

contains variables (features) of one object being classified. Should have data type CV_32F.

Python prototype (for reference):

predict(samples[, results[, flags]]) -> retval, results

setIterations(self, val)

@copybrief getIterations @see getIterations

Python prototype (for reference):

setIterations(val) -> None

setLearningRate(self, val)

@copybrief getLearningRate @see getLearningRate

Python prototype (for reference):

setLearningRate(val) -> None

setMiniBatchSize(self, val)

@copybrief getMiniBatchSize @see getMiniBatchSize

Python prototype (for reference):

setMiniBatchSize(val) -> None

setRegularization(self, val)

@copybrief getRegularization @see getRegularization

Python prototype (for reference):

setRegularization(val) -> None

setTermCriteria(self, val)

@copybrief getTermCriteria @see getTermCriteria

Python prototype (for reference):

setTermCriteria(val) -> None

setTrainMethod(self, val)

@copybrief getTrainMethod @see getTrainMethod

Python prototype (for reference):

setTrainMethod(val) -> None

train(self, trainData)

Positional Arguments

trainData: Ptr<TrainData>.

Keyword Arguments

flags: int.

Trains the statistical model

created with TrainData::create.
new training samples, not completely overwritten (such as NormalBayesClassifier or ANN_MLP).

Python prototype (for reference):

train(trainData[, flags]) -> retval

train(self, trainData, opts)

Positional Arguments

trainData: Ptr<TrainData>.

Keyword Arguments

flags: int.

Trains the statistical model

created with TrainData::create.
new training samples, not completely overwritten (such as NormalBayesClassifier or ANN_MLP).

Python prototype (for reference):

train(trainData[, flags]) -> retval

train(self, samples, layout, responses)

Trains the statistical model

Python prototype (for reference):

train(samples, layout, responses) -> retval

Functions

calcError!(self, data, test)

Raising version of calcError/3.

calcError!(self, data, test, opts)

Raising version of calcError/4.

clear!(self)

Raising version of clear/1.

create!()

Raising version of create/0.

empty!(self)

Raising version of empty/1.

get_learnt_thetas!(self)

Raising version of get_learnt_thetas/1.

getDefaultName!(self)

Raising version of getDefaultName/1.

getIterations!(self)

Raising version of getIterations/1.

getLearningRate!(self)

Raising version of getLearningRate/1.

getMiniBatchSize!(self)

Raising version of getMiniBatchSize/1.

getRegularization!(self)

Raising version of getRegularization/1.

getTermCriteria!(self)

Raising version of getTermCriteria/1.

getTrainMethod!(self)

Raising version of getTrainMethod/1.

getVarCount!(self)

Raising version of getVarCount/1.

isClassifier!(self)

Raising version of isClassifier/1.

isTrained!(self)

Raising version of isTrained/1.

load!(filepath)

Raising version of load/1.

load!(filepath, opts)

Raising version of load/2.

predict!(self, samples)

Raising version of predict/2.

predict!(self, samples, opts)

Raising version of predict/3.

read!(self, fn_)

Raising version of read/2.

save!(self, filename)

Raising version of save/2.

setIterations!(self, val)

Raising version of setIterations/2.

setLearningRate!(self, val)

Raising version of setLearningRate/2.

setMiniBatchSize!(self, val)

Raising version of setMiniBatchSize/2.

setRegularization!(self, val)

Raising version of setRegularization/2.

setTermCriteria!(self, val)

Raising version of setTermCriteria/2.

setTrainMethod!(self, val)

Raising version of setTrainMethod/2.

train!(self, trainData)

Raising version of train/2.

train!(self, trainData, opts)

Raising version of train/3.

train!(self, samples, layout, responses)

Raising version of train/4.

write!(self, fs)

Raising version of write/2.

write!(self, fs, opts)

Raising version of write/3.