Evision.ML.RTrees (Evision v0.1.5-dev)

cv

clear(self)

getDefaultName(self)

read(self, fn_)

save(self, filename)

write(self, fs)

write(self, fs, opts)

simplified API for language bindings Has overloading in C++

name : String

Python prototype (for reference):

write(fs[, name]) -> None

cv.ml

calcError(self, data, test)

calcError(self, data, test, opts)

Computes error on the training or test dataset

data.

test. computed over the training subset of the data. Please note that if you loaded a completely different dataset to evaluate already trained classifier, you will probably want not to set the test subset at all with TrainData::setTrainTestSplitRatio and specify test=false, so that the error is computed for the whole new set. Yes, this sounds a bit confusing.

resp. The method uses StatModel::predict to compute the error. For regression models the error is computed as RMS, for classifiers - as a percent of missclassified samples (0%-100%).

resp : UMat

Python prototype (for reference):

calcError(data, test[, resp]) -> retval, resp

create()

empty(self)

getActiveVarCount(self)

getCalculateVarImportance(self)

getCVFolds(self)

getMaxCategories(self)

getMaxDepth(self)

getMinSampleCount(self)

getOOBError(self)

getPriors(self)

getRegressionAccuracy(self)

getTermCriteria(self)

getTruncatePrunedTree(self)

getUse1SERule(self)

getUseSurrogates(self)

getVarCount(self)

getVarImportance(self)

getVotes(self, samples, flags)

getVotes(self, samples, flags, opts)

Returns the result of each individual tree in the forest. In case the model is a regression problem, the method will return each of the trees' results for each of the sample cases. If the model is a classifier, it will return a Mat with samples + 1 rows, where the first row gives the class number and the following rows return the votes each class had for each sample.

samples.

results.

flags.

results : UMat

Python prototype (for reference):

getVotes(samples, flags[, results]) -> results

isClassifier(self)

isTrained(self)

load(filepath)

load(filepath, opts)

Loads and creates a serialized RTree from a file Use RTree::save to serialize and store an RTree to disk. Load the RTree from this file again, by calling this function with the path to the file. Optionally specify the node for the file containing the classifier

filepath.

nodeName.

nodeName : String

Python prototype (for reference):

load(filepath[, nodeName]) -> retval

predict(self, samples)

predict(self, samples, opts)

Predicts response(s) for the provided sample(s)

samples.

results.

flags.

results : UMat

flags : int

Python prototype (for reference):

predict(samples[, results[, flags]]) -> retval, results

setActiveVarCount(self, val)

setCalculateVarImportance(self, val)

setCVFolds(self, val)

setMaxCategories(self, val)

setMaxDepth(self, val)

setMinSampleCount(self, val)

setPriors(self, val)

setRegressionAccuracy(self, val)

setTermCriteria(self, val)

setTruncatePrunedTree(self, val)

setUse1SERule(self, val)

setUseSurrogates(self, val)

train(self, trainData)

train(self, trainData, opts)

train(self, samples, layout, responses)

Functions

calcError!(self, data, test)

Raising version of calcError/3.

calcError!(self, data, test, opts)

Raising version of calcError/4.

clear!(self)

Raising version of clear/1.

create!()

Raising version of create/0.

empty!(self)

Raising version of empty/1.

getActiveVarCount!(self)

Raising version of getActiveVarCount/1.

getCalculateVarImportance!(self)

Raising version of getCalculateVarImportance/1.

getCVFolds!(self)

Raising version of getCVFolds/1.

getDefaultName!(self)

Raising version of getDefaultName/1.

getMaxCategories!(self)

Raising version of getMaxCategories/1.

getMaxDepth!(self)

Raising version of getMaxDepth/1.

getMinSampleCount!(self)

Raising version of getMinSampleCount/1.

getOOBError!(self)

Raising version of getOOBError/1.

getPriors!(self)

Raising version of getPriors/1.

getRegressionAccuracy!(self)

Raising version of getRegressionAccuracy/1.

getTermCriteria!(self)

Raising version of getTermCriteria/1.

getTruncatePrunedTree!(self)

Raising version of getTruncatePrunedTree/1.

getUse1SERule!(self)

Raising version of getUse1SERule/1.

getUseSurrogates!(self)

Raising version of getUseSurrogates/1.

getVarCount!(self)

Raising version of getVarCount/1.

getVarImportance!(self)

Raising version of getVarImportance/1.

getVotes!(self, samples, flags)

Raising version of getVotes/3.

getVotes!(self, samples, flags, opts)

Raising version of getVotes/4.

isClassifier!(self)

Raising version of isClassifier/1.

isTrained!(self)

Raising version of isTrained/1.

load!(filepath)

Raising version of load/1.

load!(filepath, opts)

Raising version of load/2.

predict!(self, samples)

Raising version of predict/2.

predict!(self, samples, opts)

Raising version of predict/3.

read!(self, fn_)

Raising version of read/2.

save!(self, filename)

Raising version of save/2.

setActiveVarCount!(self, val)

Raising version of setActiveVarCount/2.

setCalculateVarImportance!(self, val)

Raising version of setCalculateVarImportance/2.

setCVFolds!(self, val)

Raising version of setCVFolds/2.

setMaxCategories!(self, val)

Raising version of setMaxCategories/2.

setMaxDepth!(self, val)

Raising version of setMaxDepth/2.

setMinSampleCount!(self, val)

Raising version of setMinSampleCount/2.

setPriors!(self, val)

Raising version of setPriors/2.

setRegressionAccuracy!(self, val)

Raising version of setRegressionAccuracy/2.

setTermCriteria!(self, val)

Raising version of setTermCriteria/2.

setTruncatePrunedTree!(self, val)

Raising version of setTruncatePrunedTree/2.

setUse1SERule!(self, val)

Raising version of setUse1SERule/2.

setUseSurrogates!(self, val)

Raising version of setUseSurrogates/2.

train!(self, trainData)

Raising version of train/2.

train!(self, trainData, opts)

Raising version of train/3.

train!(self, samples, layout, responses)

Raising version of train/4.

write!(self, fs)

Raising version of write/2.

write!(self, fs, opts)

Raising version of write/3.