Evision.ML.SVM (Evision v0.1.9)

cv

clear(self)

Clears the algorithm state

Python prototype (for reference):

clear() -> None

getDefaultName(self)

Returns the algorithm string identifier. This string is used as top level xml/yml node tag when the object is saved to a file or string.

Python prototype (for reference):

getDefaultName() -> retval

read(self, fn_)

Reads algorithm parameters from a file storage

Positional Arguments

fn_: FileNode

Python prototype (for reference):

read(fn_) -> None

save(self, filename)

Positional Arguments

filename: String

Saves the algorithm to a file. In order to make this method work, the derived class must implement Algorithm::write(FileStorage& fs).

Python prototype (for reference):

save(filename) -> None

write(self, fs)

simplified API for language bindings

Positional Arguments

fs: Ptr<FileStorage>

Keyword Arguments

name: String.

Has overloading in C++

Python prototype (for reference):

write(fs[, name]) -> None

write(self, fs, opts)

simplified API for language bindings

Positional Arguments

fs: Ptr<FileStorage>

Keyword Arguments

name: String.

Has overloading in C++

Python prototype (for reference):

write(fs[, name]) -> None

cv.ml

calcError(self, data, test)

Computes error on the training or test dataset

Positional Arguments

data: Ptr<TrainData>.

the training data

test: bool.

if true, the error is computed over the test subset of the data, otherwise it's computed over the training subset of the data. Please note that if you loaded a completely different dataset to evaluate already trained classifier, you will probably want not to set the test subset at all with TrainData::setTrainTestSplitRatio and specify test=false, so that the error is computed for the whole new set. Yes, this sounds a bit confusing.

Return

resp: Evision.Mat.

the optional output responses.

The method uses StatModel::predict to compute the error. For regression models the error is computed as RMS, for classifiers - as a percent of missclassified samples (0%-100%).

Python prototype (for reference):

calcError(data, test[, resp]) -> retval, resp

calcError(self, data, test, opts)

Computes error on the training or test dataset

Positional Arguments

data: Ptr<TrainData>.

the training data

test: bool.

if true, the error is computed over the test subset of the data, otherwise it's computed over the training subset of the data. Please note that if you loaded a completely different dataset to evaluate already trained classifier, you will probably want not to set the test subset at all with TrainData::setTrainTestSplitRatio and specify test=false, so that the error is computed for the whole new set. Yes, this sounds a bit confusing.

Return

resp: Evision.Mat.

the optional output responses.

The method uses StatModel::predict to compute the error. For regression models the error is computed as RMS, for classifiers - as a percent of missclassified samples (0%-100%).

Python prototype (for reference):

calcError(data, test[, resp]) -> retval, resp

create()

Creates empty model. Use StatModel::train to train the model. Since %SVM has several parameters, you may want to find the best parameters for your problem, it can be done with SVM::trainAuto.

Python prototype (for reference):

create() -> retval

empty(self)

Python prototype (for reference):

empty() -> retval

getC(self)

@see setC

Python prototype (for reference):

getC() -> retval

getClassWeights(self)

@see setClassWeights

Python prototype (for reference):

getClassWeights() -> retval

getCoef0(self)

@see setCoef0

Python prototype (for reference):

getCoef0() -> retval

getDecisionFunction(self, i)

Retrieves the decision function

Positional Arguments

i: int.

the index of the decision function. If the problem solved is regression, 1-class or 2-class classification, then there will be just one decision function and the index should always be 0. Otherwise, in the case of N-class classification, there will be \f$N(N-1)/2\f$ decision functions.

Return

alpha: Evision.Mat.

the optional output vector for weights, corresponding to different support vectors. In the case of linear %SVM all the alpha's will be 1's.

svidx: Evision.Mat.

the optional output vector of indices of support vectors within the matrix of support vectors (which can be retrieved by SVM::getSupportVectors). In the case of linear %SVM each decision function consists of a single "compressed" support vector.

The method returns rho parameter of the decision function, a scalar subtracted from the weighted sum of kernel responses.

Python prototype (for reference):

getDecisionFunction(i[, alpha[, svidx]]) -> retval, alpha, svidx

getDecisionFunction(self, i, opts)

Retrieves the decision function

Positional Arguments

i: int.

the index of the decision function. If the problem solved is regression, 1-class or 2-class classification, then there will be just one decision function and the index should always be 0. Otherwise, in the case of N-class classification, there will be \f$N(N-1)/2\f$ decision functions.

Return

alpha: Evision.Mat.

the optional output vector for weights, corresponding to different support vectors. In the case of linear %SVM all the alpha's will be 1's.

svidx: Evision.Mat.

the optional output vector of indices of support vectors within the matrix of support vectors (which can be retrieved by SVM::getSupportVectors). In the case of linear %SVM each decision function consists of a single "compressed" support vector.

The method returns rho parameter of the decision function, a scalar subtracted from the weighted sum of kernel responses.

Python prototype (for reference):

getDecisionFunction(i[, alpha[, svidx]]) -> retval, alpha, svidx

getDefaultGridPtr(param_id)

Generates a grid for %SVM parameters.

Positional Arguments

param_id: int.

%SVM parameters IDs that must be one of the SVM::ParamTypes. The grid is generated for the parameter with this ID.

The function generates a grid pointer for the specified parameter of the %SVM algorithm. The grid may be passed to the function SVM::trainAuto.

Python prototype (for reference):

getDefaultGridPtr(param_id) -> retval

getDegree(self)

@see setDegree

Python prototype (for reference):

getDegree() -> retval

getGamma(self)

@see setGamma

Python prototype (for reference):

getGamma() -> retval

getKernelType(self)

Type of a %SVM kernel. See SVM::KernelTypes. Default value is SVM::RBF.

Python prototype (for reference):

getKernelType() -> retval

getNu(self)

@see setNu

Python prototype (for reference):

getNu() -> retval

getP(self)

@see setP

Python prototype (for reference):

getP() -> retval

getSupportVectors(self)

Retrieves all the support vectors

The method returns all the support vectors as a floating-point matrix, where support vectors are stored as matrix rows.

Python prototype (for reference):

getSupportVectors() -> retval

getTermCriteria(self)

@see setTermCriteria

Python prototype (for reference):

getTermCriteria() -> retval

getType(self)

@see setType

Python prototype (for reference):

getType() -> retval

getUncompressedSupportVectors(self)

Retrieves all the uncompressed support vectors of a linear %SVM

The method returns all the uncompressed support vectors of a linear %SVM that the compressed support vector, used for prediction, was derived from. They are returned in a floating-point matrix, where the support vectors are stored as matrix rows.

Python prototype (for reference):

getUncompressedSupportVectors() -> retval

getVarCount(self)

Returns the number of variables in training samples

Python prototype (for reference):

getVarCount() -> retval

isClassifier(self)

Returns true if the model is classifier

Python prototype (for reference):

isClassifier() -> retval

isTrained(self)

Returns true if the model is trained

Python prototype (for reference):

isTrained() -> retval

load(filepath)

Loads and creates a serialized svm from a file

Positional Arguments

filepath: String.

path to serialized svm

Use SVM::save to serialize and store an SVM to disk. Load the SVM from this file again, by calling this function with the path to the file.

Python prototype (for reference):

load(filepath) -> retval

predict(self, samples)

Predicts response(s) for the provided sample(s)

Positional Arguments

samples: Evision.Mat.

The input samples, floating-point matrix

Keyword Arguments

flags: int.

The optional flags, model-dependent. See cv::ml::StatModel::Flags.

Return

results: Evision.Mat.

The optional output matrix of results.

Python prototype (for reference):

predict(samples[, results[, flags]]) -> retval, results

predict(self, samples, opts)

Predicts response(s) for the provided sample(s)

Positional Arguments

samples: Evision.Mat.

The input samples, floating-point matrix

Keyword Arguments

flags: int.

The optional flags, model-dependent. See cv::ml::StatModel::Flags.

Return

results: Evision.Mat.

The optional output matrix of results.

Python prototype (for reference):

predict(samples[, results[, flags]]) -> retval, results

setC(self, val)

Positional Arguments

val: double

@copybrief getC @see getC

Python prototype (for reference):

setC(val) -> None

setClassWeights(self, val)

Positional Arguments

val: Evision.Mat

@copybrief getClassWeights @see getClassWeights

Python prototype (for reference):

setClassWeights(val) -> None

setCoef0(self, val)

Positional Arguments

val: double

@copybrief getCoef0 @see getCoef0

Python prototype (for reference):

setCoef0(val) -> None

setDegree(self, val)

Positional Arguments

val: double

@copybrief getDegree @see getDegree

Python prototype (for reference):

setDegree(val) -> None

setGamma(self, val)

Positional Arguments

val: double

@copybrief getGamma @see getGamma

Python prototype (for reference):

setGamma(val) -> None

setKernel(self, kernelType)

Positional Arguments

kernelType: int

Initialize with one of predefined kernels. See SVM::KernelTypes.

Python prototype (for reference):

setKernel(kernelType) -> None

setNu(self, val)

Positional Arguments

val: double

@copybrief getNu @see getNu

Python prototype (for reference):

setNu(val) -> None

setP(self, val)

Positional Arguments

val: double

@copybrief getP @see getP

Python prototype (for reference):

setP(val) -> None

setTermCriteria(self, val)

Positional Arguments

val: TermCriteria

@copybrief getTermCriteria @see getTermCriteria

Python prototype (for reference):

setTermCriteria(val) -> None

setType(self, val)

Positional Arguments

val: int

@copybrief getType @see getType

Python prototype (for reference):

setType(val) -> None

train(self, trainData)

Trains the statistical model

Positional Arguments

trainData: Ptr<TrainData>.

training data that can be loaded from file using TrainData::loadFromCSV or created with TrainData::create.

Keyword Arguments

flags: int.

optional flags, depending on the model. Some of the models can be updated with the new training samples, not completely overwritten (such as NormalBayesClassifier or ANN_MLP).

Python prototype (for reference):

train(trainData[, flags]) -> retval

train(self, trainData, opts)

Trains the statistical model

Positional Arguments

trainData: Ptr<TrainData>.

training data that can be loaded from file using TrainData::loadFromCSV or created with TrainData::create.

Keyword Arguments

flags: int.

optional flags, depending on the model. Some of the models can be updated with the new training samples, not completely overwritten (such as NormalBayesClassifier or ANN_MLP).

Python prototype (for reference):

train(trainData[, flags]) -> retval

train(self, samples, layout, responses)

Trains the statistical model

Positional Arguments

samples: Evision.Mat.

training samples

layout: int.

See ml::SampleTypes.

responses: Evision.Mat.

vector of responses associated with the training samples.

Python prototype (for reference):

train(samples, layout, responses) -> retval

trainAuto(self, samples, layout, responses)

Trains an %SVM with optimal parameters

Positional Arguments

samples: Evision.Mat.

training samples

layout: int.

See ml::SampleTypes.

responses: Evision.Mat.

vector of responses associated with the training samples.

Keyword Arguments

kFold: int.

Cross-validation parameter. The training set is divided into kFold subsets. One subset is used to test the model, the others form the train set. So, the %SVM algorithm is

cgrid: Ptr<ParamGrid>.

gammaGrid: Ptr<ParamGrid>.

grid for gamma

pGrid: Ptr<ParamGrid>.

nuGrid: Ptr<ParamGrid>.

coeffGrid: Ptr<ParamGrid>.

grid for coeff

degreeGrid: Ptr<ParamGrid>.

grid for degree

balanced: bool.

If true and the problem is 2-class classification then the method creates more balanced cross-validation subsets that is proportions between classes in subsets are close to such proportion in the whole train dataset.

The method trains the %SVM model automatically by choosing the optimal parameters C, gamma, p, nu, coef0, degree. Parameters are considered optimal when the cross-validation estimate of the test set error is minimal. This function only makes use of SVM::getDefaultGrid for parameter optimization and thus only offers rudimentary parameter options. This function works for the classification (SVM::C_SVC or SVM::NU_SVC) as well as for the regression (SVM::EPS_SVR or SVM::NU_SVR). If it is SVM::ONE_CLASS, no optimization is made and the usual %SVM with parameters specified in params is executed.

Python prototype (for reference):

trainAuto(samples, layout, responses[, kFold[, Cgrid[, gammaGrid[, pGrid[, nuGrid[, coeffGrid[, degreeGrid[, balanced]]]]]]]]) -> retval

trainAuto(self, samples, layout, responses, opts)

Trains an %SVM with optimal parameters

Positional Arguments

samples: Evision.Mat.

training samples

layout: int.

See ml::SampleTypes.

responses: Evision.Mat.

vector of responses associated with the training samples.

Keyword Arguments

kFold: int.

Cross-validation parameter. The training set is divided into kFold subsets. One subset is used to test the model, the others form the train set. So, the %SVM algorithm is

cgrid: Ptr<ParamGrid>.

gammaGrid: Ptr<ParamGrid>.

grid for gamma

pGrid: Ptr<ParamGrid>.

nuGrid: Ptr<ParamGrid>.

coeffGrid: Ptr<ParamGrid>.

grid for coeff

degreeGrid: Ptr<ParamGrid>.

grid for degree

balanced: bool.

If true and the problem is 2-class classification then the method creates more balanced cross-validation subsets that is proportions between classes in subsets are close to such proportion in the whole train dataset.

The method trains the %SVM model automatically by choosing the optimal parameters C, gamma, p, nu, coef0, degree. Parameters are considered optimal when the cross-validation estimate of the test set error is minimal. This function only makes use of SVM::getDefaultGrid for parameter optimization and thus only offers rudimentary parameter options. This function works for the classification (SVM::C_SVC or SVM::NU_SVC) as well as for the regression (SVM::EPS_SVR or SVM::NU_SVR). If it is SVM::ONE_CLASS, no optimization is made and the usual %SVM with parameters specified in params is executed.

Python prototype (for reference):

trainAuto(samples, layout, responses[, kFold[, Cgrid[, gammaGrid[, pGrid[, nuGrid[, coeffGrid[, degreeGrid[, balanced]]]]]]]]) -> retval

Functions

calcError!(self, data, test)

Raising version of calcError/3.

calcError!(self, data, test, opts)

Raising version of calcError/4.

clear!(self)

Raising version of clear/1.

create!()

Raising version of create/0.

empty!(self)

Raising version of empty/1.

getC!(self)

Raising version of getC/1.

getClassWeights!(self)

Raising version of getClassWeights/1.

getCoef0!(self)

Raising version of getCoef0/1.

getDecisionFunction!(self, i)

Raising version of getDecisionFunction/2.

getDecisionFunction!(self, i, opts)

Raising version of getDecisionFunction/3.

getDefaultGridPtr!(param_id)

Raising version of getDefaultGridPtr/1.

getDefaultName!(self)

Raising version of getDefaultName/1.

getDegree!(self)

Raising version of getDegree/1.

getGamma!(self)

Raising version of getGamma/1.

getKernelType!(self)

Raising version of getKernelType/1.

getNu!(self)

Raising version of getNu/1.

getP!(self)

Raising version of getP/1.

getSupportVectors!(self)

Raising version of getSupportVectors/1.

getTermCriteria!(self)

Raising version of getTermCriteria/1.

getType!(self)

Raising version of getType/1.

getUncompressedSupportVectors!(self)

Raising version of getUncompressedSupportVectors/1.

getVarCount!(self)

Raising version of getVarCount/1.

isClassifier!(self)

Raising version of isClassifier/1.

isTrained!(self)

Raising version of isTrained/1.

load!(filepath)

Raising version of load/1.

predict!(self, samples)

Raising version of predict/2.

predict!(self, samples, opts)

Raising version of predict/3.

read!(self, fn_)

Raising version of read/2.

save!(self, filename)

Raising version of save/2.

setC!(self, val)

Raising version of setC/2.

setClassWeights!(self, val)

Raising version of setClassWeights/2.

setCoef0!(self, val)

Raising version of setCoef0/2.

setDegree!(self, val)

Raising version of setDegree/2.

setGamma!(self, val)

Raising version of setGamma/2.

setKernel!(self, kernelType)

Raising version of setKernel/2.

setNu!(self, val)

Raising version of setNu/2.

setP!(self, val)

Raising version of setP/2.

setTermCriteria!(self, val)

Raising version of setTermCriteria/2.

setType!(self, val)

Raising version of setType/2.

train!(self, trainData)

Raising version of train/2.

train!(self, trainData, opts)

Raising version of train/3.

train!(self, samples, layout, responses)

Raising version of train/4.

trainAuto!(self, samples, layout, responses)

Raising version of trainAuto/4.

trainAuto!(self, samples, layout, responses, opts)

Raising version of trainAuto/5.

write!(self, fs)

Raising version of write/2.

write!(self, fs, opts)

Raising version of write/3.