Evision.ML.SVMSGD (Evision v0.1.9)

cv

clear(self)

Clears the algorithm state

Python prototype (for reference):

clear() -> None

getDefaultName(self)

Returns the algorithm string identifier. This string is used as top level xml/yml node tag when the object is saved to a file or string.

Python prototype (for reference):

getDefaultName() -> retval

read(self, fn_)

Reads algorithm parameters from a file storage

Positional Arguments

fn_: FileNode

Python prototype (for reference):

read(fn_) -> None

save(self, filename)

Positional Arguments

filename: String

Saves the algorithm to a file. In order to make this method work, the derived class must implement Algorithm::write(FileStorage& fs).

Python prototype (for reference):

save(filename) -> None

write(self, fs)

simplified API for language bindings

Positional Arguments

fs: Ptr<FileStorage>

Keyword Arguments

name: String.

Has overloading in C++

Python prototype (for reference):

write(fs[, name]) -> None

write(self, fs, opts)

simplified API for language bindings

Positional Arguments

fs: Ptr<FileStorage>

Keyword Arguments

name: String.

Has overloading in C++

Python prototype (for reference):

write(fs[, name]) -> None

cv.ml

calcError(self, data, test)

Computes error on the training or test dataset

Positional Arguments

data: Ptr<TrainData>.

the training data

test: bool.

if true, the error is computed over the test subset of the data, otherwise it's computed over the training subset of the data. Please note that if you loaded a completely different dataset to evaluate already trained classifier, you will probably want not to set the test subset at all with TrainData::setTrainTestSplitRatio and specify test=false, so that the error is computed for the whole new set. Yes, this sounds a bit confusing.

Return

resp: Evision.Mat.

the optional output responses.

The method uses StatModel::predict to compute the error. For regression models the error is computed as RMS, for classifiers - as a percent of missclassified samples (0%-100%).

Python prototype (for reference):

calcError(data, test[, resp]) -> retval, resp

calcError(self, data, test, opts)

Computes error on the training or test dataset

Positional Arguments

data: Ptr<TrainData>.

the training data

test: bool.

if true, the error is computed over the test subset of the data, otherwise it's computed over the training subset of the data. Please note that if you loaded a completely different dataset to evaluate already trained classifier, you will probably want not to set the test subset at all with TrainData::setTrainTestSplitRatio and specify test=false, so that the error is computed for the whole new set. Yes, this sounds a bit confusing.

Return

resp: Evision.Mat.

the optional output responses.

The method uses StatModel::predict to compute the error. For regression models the error is computed as RMS, for classifiers - as a percent of missclassified samples (0%-100%).

Python prototype (for reference):

calcError(data, test[, resp]) -> retval, resp

create()

Creates empty model. Use StatModel::train to train the model. Since %SVMSGD has several parameters, you may want to find the best parameters for your problem or use setOptimalParameters() to set some default parameters.

Python prototype (for reference):

create() -> retval

empty(self)

Python prototype (for reference):

empty() -> retval

getInitialStepSize(self)

@see setInitialStepSize

Python prototype (for reference):

getInitialStepSize() -> retval

getMarginRegularization(self)

@see setMarginRegularization

Python prototype (for reference):

getMarginRegularization() -> retval

getMarginType(self)

@see setMarginType

Python prototype (for reference):

getMarginType() -> retval

getShift(self)

@return the shift of the trained model (decision function f(x) = weights * x + shift).

Python prototype (for reference):

getShift() -> retval

getStepDecreasingPower(self)

@see setStepDecreasingPower

Python prototype (for reference):

getStepDecreasingPower() -> retval

getSvmsgdType(self)

@see setSvmsgdType

Python prototype (for reference):

getSvmsgdType() -> retval

getTermCriteria(self)

@see setTermCriteria

Python prototype (for reference):

getTermCriteria() -> retval

getVarCount(self)

Returns the number of variables in training samples

Python prototype (for reference):

getVarCount() -> retval

getWeights(self)

@return the weights of the trained model (decision function f(x) = weights * x + shift).

Python prototype (for reference):

getWeights() -> retval

isClassifier(self)

Returns true if the model is classifier

Python prototype (for reference):

isClassifier() -> retval

isTrained(self)

Returns true if the model is trained

Python prototype (for reference):

isTrained() -> retval

load(filepath)

Loads and creates a serialized SVMSGD from a file

Positional Arguments

filepath: String.

path to serialized SVMSGD

Keyword Arguments

nodeName: String.

name of node containing the classifier

Use SVMSGD::save to serialize and store an SVMSGD to disk. Load the SVMSGD from this file again, by calling this function with the path to the file. Optionally specify the node for the file containing the classifier

Python prototype (for reference):

load(filepath[, nodeName]) -> retval

load(filepath, opts)

Loads and creates a serialized SVMSGD from a file

Positional Arguments

filepath: String.

path to serialized SVMSGD

Keyword Arguments

nodeName: String.

name of node containing the classifier

Use SVMSGD::save to serialize and store an SVMSGD to disk. Load the SVMSGD from this file again, by calling this function with the path to the file. Optionally specify the node for the file containing the classifier

Python prototype (for reference):

load(filepath[, nodeName]) -> retval

predict(self, samples)

Predicts response(s) for the provided sample(s)

Positional Arguments

samples: Evision.Mat.

The input samples, floating-point matrix

Keyword Arguments

flags: int.

The optional flags, model-dependent. See cv::ml::StatModel::Flags.

Return

results: Evision.Mat.

The optional output matrix of results.

Python prototype (for reference):

predict(samples[, results[, flags]]) -> retval, results

predict(self, samples, opts)

Predicts response(s) for the provided sample(s)

Positional Arguments

samples: Evision.Mat.

The input samples, floating-point matrix

Keyword Arguments

flags: int.

The optional flags, model-dependent. See cv::ml::StatModel::Flags.

Return

results: Evision.Mat.

The optional output matrix of results.

Python prototype (for reference):

predict(samples[, results[, flags]]) -> retval, results

setInitialStepSize(self, initialStepSize)

Positional Arguments

initialStepSize: float

@copybrief getInitialStepSize @see getInitialStepSize

Python prototype (for reference):

setInitialStepSize(InitialStepSize) -> None

setMarginRegularization(self, marginRegularization)

Positional Arguments

marginRegularization: float

@copybrief getMarginRegularization @see getMarginRegularization

Python prototype (for reference):

setMarginRegularization(marginRegularization) -> None

setMarginType(self, marginType)

Positional Arguments

marginType: int

@copybrief getMarginType @see getMarginType

Python prototype (for reference):

setMarginType(marginType) -> None

setOptimalParameters(self)

Function sets optimal parameters values for chosen SVM SGD model.

Keyword Arguments

svmsgdType: int.

is the type of SVMSGD classifier.

marginType: int.

is the type of margin constraint.

Python prototype (for reference):

setOptimalParameters([, svmsgdType[, marginType]]) -> None

setStepDecreasingPower(self, stepDecreasingPower)

Positional Arguments

stepDecreasingPower: float

@copybrief getStepDecreasingPower @see getStepDecreasingPower

Python prototype (for reference):

setStepDecreasingPower(stepDecreasingPower) -> None

setSvmsgdType(self, svmsgdType)

Positional Arguments

svmsgdType: int

@copybrief getSvmsgdType @see getSvmsgdType

Python prototype (for reference):

setSvmsgdType(svmsgdType) -> None

setTermCriteria(self, val)

Positional Arguments

val: TermCriteria

@copybrief getTermCriteria @see getTermCriteria

Python prototype (for reference):

setTermCriteria(val) -> None

train(self, trainData)

Trains the statistical model

Positional Arguments

trainData: Ptr<TrainData>.

training data that can be loaded from file using TrainData::loadFromCSV or created with TrainData::create.

Keyword Arguments

flags: int.

optional flags, depending on the model. Some of the models can be updated with the new training samples, not completely overwritten (such as NormalBayesClassifier or ANN_MLP).

Python prototype (for reference):

train(trainData[, flags]) -> retval

train(self, trainData, opts)

Trains the statistical model

Positional Arguments

trainData: Ptr<TrainData>.

training data that can be loaded from file using TrainData::loadFromCSV or created with TrainData::create.

Keyword Arguments

flags: int.

optional flags, depending on the model. Some of the models can be updated with the new training samples, not completely overwritten (such as NormalBayesClassifier or ANN_MLP).

Python prototype (for reference):

train(trainData[, flags]) -> retval

train(self, samples, layout, responses)

Trains the statistical model

Positional Arguments

samples: Evision.Mat.

training samples

layout: int.

See ml::SampleTypes.

responses: Evision.Mat.

vector of responses associated with the training samples.

Python prototype (for reference):

train(samples, layout, responses) -> retval

Functions

calcError!(self, data, test)

Raising version of calcError/3.

calcError!(self, data, test, opts)

Raising version of calcError/4.

clear!(self)

Raising version of clear/1.

create!()

Raising version of create/0.

empty!(self)

Raising version of empty/1.

getDefaultName!(self)

Raising version of getDefaultName/1.

getInitialStepSize!(self)

Raising version of getInitialStepSize/1.

getMarginRegularization!(self)

Raising version of getMarginRegularization/1.

getMarginType!(self)

Raising version of getMarginType/1.

getShift!(self)

Raising version of getShift/1.

getStepDecreasingPower!(self)

Raising version of getStepDecreasingPower/1.

getSvmsgdType!(self)

Raising version of getSvmsgdType/1.

getTermCriteria!(self)

Raising version of getTermCriteria/1.

getVarCount!(self)

Raising version of getVarCount/1.

getWeights!(self)

Raising version of getWeights/1.

isClassifier!(self)

Raising version of isClassifier/1.

isTrained!(self)

Raising version of isTrained/1.

load!(filepath)

Raising version of load/1.

load!(filepath, opts)

Raising version of load/2.

predict!(self, samples)

Raising version of predict/2.

predict!(self, samples, opts)

Raising version of predict/3.

read!(self, fn_)

Raising version of read/2.

save!(self, filename)

Raising version of save/2.

setInitialStepSize!(self, initialStepSize)

Raising version of setInitialStepSize/2.

setMarginRegularization!(self, marginRegularization)

Raising version of setMarginRegularization/2.

setMarginType!(self, marginType)

Raising version of setMarginType/2.

setOptimalParameters!(self)

Raising version of setOptimalParameters/1.

setStepDecreasingPower!(self, stepDecreasingPower)

Raising version of setStepDecreasingPower/2.

setSvmsgdType!(self, svmsgdType)

Raising version of setSvmsgdType/2.

setTermCriteria!(self, val)

Raising version of setTermCriteria/2.

train!(self, trainData)

Raising version of train/2.

train!(self, trainData, opts)

Raising version of train/3.

train!(self, samples, layout, responses)

Raising version of train/4.

write!(self, fs)

Raising version of write/2.

write!(self, fs, opts)

Raising version of write/3.