Evision.ML.StatModel (Evision v0.1.5-dev)

cv

clear(self)

getDefaultName(self)

read(self, fn_)

save(self, filename)

write(self, fs)

write(self, fs, opts)

simplified API for language bindings Has overloading in C++

name : String

Python prototype (for reference):

write(fs[, name]) -> None

cv.ml

calcError(self, data, test)

calcError(self, data, test, opts)

Computes error on the training or test dataset

data.

test. computed over the training subset of the data. Please note that if you loaded a completely different dataset to evaluate already trained classifier, you will probably want not to set the test subset at all with TrainData::setTrainTestSplitRatio and specify test=false, so that the error is computed for the whole new set. Yes, this sounds a bit confusing.

resp. The method uses StatModel::predict to compute the error. For regression models the error is computed as RMS, for classifiers - as a percent of missclassified samples (0%-100%).

resp : UMat

Python prototype (for reference):

calcError(data, test[, resp]) -> retval, resp

empty(self)

getVarCount(self)

isClassifier(self)

isTrained(self)

predict(self, samples)

predict(self, samples, opts)

Predicts response(s) for the provided sample(s)

samples.

results.

flags.

results : UMat

flags : int

Python prototype (for reference):

predict(samples[, results[, flags]]) -> retval, results

train(self, trainData)

train(self, trainData, opts)

train(self, samples, layout, responses)

Functions

calcError!(self, data, test)

Raising version of calcError/3.

calcError!(self, data, test, opts)

Raising version of calcError/4.

clear!(self)

Raising version of clear/1.

empty!(self)

Raising version of empty/1.

getDefaultName!(self)

Raising version of getDefaultName/1.

getVarCount!(self)

Raising version of getVarCount/1.

isClassifier!(self)

Raising version of isClassifier/1.

isTrained!(self)

Raising version of isTrained/1.

predict!(self, samples)

Raising version of predict/2.

predict!(self, samples, opts)

Raising version of predict/3.

read!(self, fn_)

Raising version of read/2.

save!(self, filename)

Raising version of save/2.

train!(self, trainData)

Raising version of train/2.

train!(self, trainData, opts)

Raising version of train/3.

train!(self, samples, layout, responses)

Raising version of train/4.

write!(self, fs)

Raising version of write/2.

write!(self, fs, opts)

Raising version of write/3.