Evision.ML.TrainData (Evision v0.1.5-dev)

cv.ml

create(samples, layout, responses)

create(samples, layout, responses, opts)

Creates training data from in-memory arrays.

samples.

layout.

responses. single row or as a single column. The matrix should have type CV_32F or CV_32S (in the former case the responses are considered as ordered by default; in the latter case - as categorical)

varIdx. (CV_32S) containing 0-based variable indices or byte vector (CV_8U) containing a mask of active variables.

sampleIdx. vector (CV_32S) containing 0-based sample indices or byte vector (CV_8U) containing a mask of training samples.

sampleWeights.

varType. <number_of_variables_in_responses>`, containing types of each input and output variable. See ml::VariableTypes. - varIdx : UMat - sampleIdx : UMat - sampleWeights : UMat - varType : UMat Python prototype (for reference): create(samples, layout, responses[, varIdx[, sampleIdx[, sampleWeights[, varType]]]]) -> retval

getCatCount(self, vi)

getCatMap(self)

getCatOfs(self)

getClassLabels(self)

getDefaultSubstValues(self)

getLayout(self)

getMissing(self)

getNAllVars(self)

getNames(self, names)

getNormCatResponses(self)

getNSamples(self)

getNTestSamples(self)

getNTrainSamples(self)

getNVars(self)

getResponses(self)

getResponseType(self)

getSample(self, varIdx, sidx, buf)

getSamples(self)

getSampleWeights(self)

getSubMatrix(matrix, idx, layout)

getSubVector(vec, idx)

getTestNormCatResponses(self)

getTestResponses(self)

getTestSampleIdx(self)

getTestSamples(self)

getTestSampleWeights(self)

getTrainNormCatResponses(self)

getTrainResponses(self)

getTrainSampleIdx(self)

getTrainSamples(self)

getTrainSampleWeights(self)

getValues(self, vi, sidx, values)

getVarIdx(self)

getVarSymbolFlags(self)

getVarType(self)

setTrainTestSplit(self, count)

setTrainTestSplit(self, count, opts)

Splits the training data into the training and test parts @sa TrainData::setTrainTestSplitRatio

shuffle : bool

Python prototype (for reference):

setTrainTestSplit(count[, shuffle]) -> None

setTrainTestSplitRatio(self, ratio)

setTrainTestSplitRatio(self, ratio, opts)

Splits the training data into the training and test parts The function selects a subset of specified relative size and then returns it as the training set. If the function is not called, all the data is used for training. Please, note that for each of TrainData::getTrain* there is corresponding TrainData::getTest*, so that the test subset can be retrieved and processed as well. @sa TrainData::setTrainTestSplit

shuffle : bool

Python prototype (for reference):

setTrainTestSplitRatio(ratio[, shuffle]) -> None

shuffleTrainTest(self)

Functions

create!(samples, layout, responses)

Raising version of create/3.

create!(samples, layout, responses, opts)

Raising version of create/4.

getCatCount!(self, vi)

Raising version of getCatCount/2.

getCatMap!(self)

Raising version of getCatMap/1.

getCatOfs!(self)

Raising version of getCatOfs/1.

getClassLabels!(self)

Raising version of getClassLabels/1.

getDefaultSubstValues!(self)

Raising version of getDefaultSubstValues/1.

getLayout!(self)

Raising version of getLayout/1.

getMissing!(self)

Raising version of getMissing/1.

getNAllVars!(self)

Raising version of getNAllVars/1.

getNames!(self, names)

Raising version of getNames/2.

getNormCatResponses!(self)

Raising version of getNormCatResponses/1.

getNSamples!(self)

Raising version of getNSamples/1.

getNTestSamples!(self)

Raising version of getNTestSamples/1.

getNTrainSamples!(self)

Raising version of getNTrainSamples/1.

getNVars!(self)

Raising version of getNVars/1.

getResponses!(self)

Raising version of getResponses/1.

getResponseType!(self)

Raising version of getResponseType/1.

getSample!(self, varIdx, sidx, buf)

Raising version of getSample/4.

getSamples!(self)

Raising version of getSamples/1.

getSampleWeights!(self)

Raising version of getSampleWeights/1.

getSubMatrix!(matrix, idx, layout)

Raising version of getSubMatrix/3.

getSubVector!(vec, idx)

Raising version of getSubVector/2.

getTestNormCatResponses!(self)

Raising version of getTestNormCatResponses/1.

getTestResponses!(self)

Raising version of getTestResponses/1.

getTestSampleIdx!(self)

Raising version of getTestSampleIdx/1.

getTestSamples!(self)

Raising version of getTestSamples/1.

getTestSampleWeights!(self)

Raising version of getTestSampleWeights/1.

getTrainNormCatResponses!(self)

Raising version of getTrainNormCatResponses/1.

getTrainResponses!(self)

Raising version of getTrainResponses/1.

getTrainSampleIdx!(self)

Raising version of getTrainSampleIdx/1.

getTrainSamples!(self)

Raising version of getTrainSamples/1.

getTrainSampleWeights!(self)

Raising version of getTrainSampleWeights/1.

getValues!(self, vi, sidx, values)

Raising version of getValues/4.

getVarIdx!(self)

Raising version of getVarIdx/1.

getVarSymbolFlags!(self)

Raising version of getVarSymbolFlags/1.

getVarType!(self)

Raising version of getVarType/1.

setTrainTestSplit!(self, count)

Raising version of setTrainTestSplit/2.

setTrainTestSplit!(self, count, opts)

Raising version of setTrainTestSplit/3.

setTrainTestSplitRatio!(self, ratio)

Raising version of setTrainTestSplitRatio/2.

setTrainTestSplitRatio!(self, ratio, opts)

Raising version of setTrainTestSplitRatio/3.

shuffleTrainTest!(self)

Raising version of shuffleTrainTest/1.