Evision.Mat (Evision v0.1.5-dev)

OpenCV Mat

Types

channels_from_binary()

@type channels_from_binary() :: 1 | 3 | 4

mat_type()

@type mat_type() ::
  {:u, 8} | {:u, 16} | {:s, 8} | {:s, 16} | {:s, 32} | {:f, 32} | {:f, 64}

Types for mat

cv.Mat

abs(mat)

@spec abs(reference()) :: {:ok, reference()} | {:error, String.t()}

add(lhs, rhs)

@spec add(reference(), reference()) :: {:ok, reference()} | {:error, String.t()}

add(lhs, rhs, type)

@spec add(reference(), reference(), mat_type()) ::
  {:ok, reference()} | {:error, String.t()}

as_type(mat, type)

@spec as_type(reference(), mat_type()) :: {:ok, reference()} | {:error, String.t()}

at(mat, position)

@spec at(reference(), non_neg_integer()) :: {:ok, number()} | {:error, String.t()}

bitwise_and(lhs, rhs)

@spec bitwise_and(reference(), reference()) ::
  {:ok, reference()} | {:error, String.t()}

bitwise_not(mat)

@spec bitwise_not(reference()) :: {:ok, reference()} | {:error, String.t()}

bitwise_or(lhs, rhs)

@spec bitwise_or(reference(), reference()) ::
  {:ok, reference()} | {:error, String.t()}

bitwise_xor(lhs, rhs)

@spec bitwise_xor(reference(), reference()) ::
  {:ok, reference()} | {:error, String.t()}

ceil(mat)

@spec ceil(reference()) :: {:ok, reference()} | {:error, String.t()}

clip(mat, lower, upper)

@spec clip(reference(), number(), number()) ::
  {:ok, reference()} | {:error, String.t()}

clone(mat)

@spec clone(reference()) :: {:ok, reference()} | {:error, String.t()}

cmp(lhs, rhs, op)

@spec cmp(reference(), reference(), atom()) ::
  {:ok, reference()} | {:error, String.t()}

divide(lhs, rhs)

@spec divide(reference(), reference()) :: {:ok, reference()} | {:error, String.t()}

divide(lhs, rhs, type)

@spec divide(reference(), reference(), mat_type()) ::
  {:ok, reference()} | {:error, String.t()}

dot(mat_a, mat_b)

@spec dot(reference(), reference()) :: {:ok, reference()} | {:error, String.t()}

empty()

@spec empty() :: :ok

expm1(mat)

@spec expm1(reference()) :: {:ok, reference()} | {:error, String.t()}

eye(n, type)

floor(mat)

@spec floor(reference()) :: {:ok, reference()} | {:error, String.t()}

from_binary(binary, type, rows, cols, channels)

@spec from_binary(
  binary(),
  mat_type(),
  pos_integer(),
  pos_integer(),
  channels_from_binary()
) ::
  {:ok, reference()} | {:error, String.t()}

Create Mat from binary (pixel) data

binary. The binary pixel data

type. type={t, l} is one of [{:u, 8}, {:s, 8}, {:u, 16}, {:s, 16}, {:s, 32}, {:f, 32}, {:f, 64}]

rows. Number of rows (i.e., the height of the image)

cols. Number of cols (i.e., the width of the image)

channels. Number of channels, only valid if in [1, 3, 4]

from_binary_by_shape(binary, type, shape)

logical_and(lhs, rhs)

@spec logical_and(reference(), reference()) ::
  {:ok, reference()} | {:error, String.t()}

logical_or(lhs, rhs)

@spec logical_or(reference(), reference()) ::
  {:ok, reference()} | {:error, String.t()}

logical_xor(lhs, rhs)

@spec logical_xor(reference(), reference()) ::
  {:ok, reference()} | {:error, String.t()}

matrix_multiply(lhs, rhs, out_type \\ nil)

@spec matrix_multiply(reference(), reference(), mat_type() | nil) ::
  {:ok, reference()} | {:error, String.t()}

multiply(lhs, rhs)

@spec multiply(reference(), reference()) :: {:ok, reference()} | {:error, String.t()}

multiply(lhs, rhs, type)

@spec multiply(reference(), reference(), mat_type()) ::
  {:ok, reference()} | {:error, String.t()}

negate(mat)

@spec negate(reference()) :: {:ok, reference()} | {:error, String.t()}

number(number, type)

ones(shape, type)

reshape(mat, shape)

round(mat)

@spec round(reference()) :: {:ok, reference()} | {:error, String.t()}

setTo(mat, value, mask)

@spec setTo(reference(), number(), reference()) ::
  {:ok, reference()} | {:error, String.t()}

shape(mat)

sign(mat)

@spec sign(reference()) :: {:ok, reference()} | {:error, String.t()}

subtract(lhs, rhs)

@spec subtract(reference(), reference()) :: {:ok, reference()} | {:error, String.t()}

subtract(lhs, rhs, type)

@spec subtract(reference(), reference(), mat_type()) ::
  {:ok, reference()} | {:error, String.t()}

to_binary(mat, limit \\ 0)

@spec to_binary(reference(), non_neg_integer()) ::
  {:ok, binary()} | {:error, String.t()}

type(mat)

@spec type(reference()) :: {:ok, mat_type()} | {:error, String.t()}

zeros(shape, type)

Functions

abs!(mat)

Raising version of abs/1.

add!(lhs, rhs)

Raising version of add/2.

add!(lhs, rhs, type)

Raising version of add/3.

arange!(from, to, step, type)

Raising version of arange/4.

arange!(from, to, step, type, shape)

Raising version of arange/5.

arange(from, to, step, type)

arange(from, to, step, type, shape)

as_type!(mat, type)

Raising version of as_type/2.

at!(mat, position)

Raising version of at/2.

bitwise_and!(lhs, rhs)

Raising version of bitwise_and/2.

bitwise_not!(mat)

Raising version of bitwise_not/1.

bitwise_or!(lhs, rhs)

Raising version of bitwise_or/2.

bitwise_xor!(lhs, rhs)

Raising version of bitwise_xor/2.

broadcast_to!(mat, to_shape)

Raising version of broadcast_to/2.

broadcast_to!(mat, to_shape, force_src_shape)

Raising version of broadcast_to/3.

broadcast_to(mat, to_shape)

broadcast_to(mat, to_shape, force_src_shape)

ceil!(mat)

Raising version of ceil/1.

clip!(mat, lower, upper)

Raising version of clip/3.

clone!(mat)

Raising version of clone/1.

cmp!(lhs, rhs, op)

Raising version of cmp/3.

divide!(lhs, rhs)

Raising version of divide/2.

divide!(lhs, rhs, type)

Raising version of divide/3.

dot!(mat_a, mat_b)

Raising version of dot/2.

empty!()

Raising version of empty/0.

expm1!(mat)

Raising version of expm1/1.

eye!(n, type)

Raising version of eye/2.

floor!(mat)

Raising version of floor/1.

from_binary!(binary, type, rows, cols, channels)

Raising version of from_binary/5.

from_binary_by_shape!(binary, type, shape)

Raising version of from_binary_by_shape/3.

full!(shape, number, type)

Raising version of full/3.

full(shape, number, type)

logical_and!(lhs, rhs)

Raising version of logical_and/2.

logical_or!(lhs, rhs)

Raising version of logical_or/2.

logical_xor!(lhs, rhs)

Raising version of logical_xor/2.

matrix_multiply!(lhs, rhs)

Raising version of matrix_multiply/2.

matrix_multiply!(lhs, rhs, out_type)

Raising version of matrix_multiply/3.

multiply!(lhs, rhs)

Raising version of multiply/2.

multiply!(lhs, rhs, type)

Raising version of multiply/3.

negate!(mat)

Raising version of negate/1.

number!(number, type)

Raising version of number/2.

ones!(shape, type)

Raising version of ones/2.

reshape!(mat, shape)

Raising version of reshape/2.

round!(mat)

Raising version of round/1.

setTo!(mat, value, mask)

Raising version of setTo/3.

shape!(mat)

Raising version of shape/1.

sign!(mat)

Raising version of sign/1.

squeeze!(mat)

Raising version of squeeze/1.

squeeze(mat)

subtract!(lhs, rhs)

Raising version of subtract/2.

subtract!(lhs, rhs, type)

Raising version of subtract/3.

to_batched!(mat, batch_size, opts)

Raising version of to_batched/3.

to_batched!(mat, batch_size, as_shape, opts)

Raising version of to_batched/4.

to_batched(mat, batch_size, opts)

@spec to_batched(reference(), pos_integer(), keyword()) ::
  {:ok, [reference()]} | {:error, String.t()}

@spec to_batched(reference(), pos_integer(), keyword()) ::
  {:ok, [reference()]} | {:error, String.t()}

to_batched(mat, batch_size, as_shape, opts)

to_binary!(mat)

Raising version of to_binary/1.

to_binary!(mat, limit)

Raising version of to_binary/2.

transpose!(mat)

Raising version of transpose/1.

transpose!(mat, axes, opts)

Raising version of transpose/3.

transpose(mat)

Transpose a matrix

Parameters

mat. The matrix. by default it reverses the order of the axes.

transpose(mat, axes, opts \\ [])

@spec transpose(
  reference(),
  [non_neg_integer()],
  keyword()
) :: {:ok, reference()} | {:error, String.t()}

Transpose a matrix

Parameters

mat. The matrix.

axes. list of ints. It must be a list which contains a permutation of [0,1,..,N-1] where N is the number of axes of mat. The i’th axis of the returned array will correspond to the axis numbered axes[i] of the input.

opts. Keyword options.

as_shape. A tuple or list which overwrites the shape of the matrix (the total number of elements must be equal to the one as in its original shape). For example, a 4x4 matrix can be treated as a 2x2x2x2 matrix and transposed with axes=[2,1,3,0] in a single call.

When specified, it combines the reshape and transpose operation in a single NIF call.

type!(mat)

Raising version of type/1.

zeros!(shape, type)

Raising version of zeros/2.