Evision.Nx (Evision v0.1.7-dev)

OpenCV's cv::mat to Nx tensor.

external

to_mat(t)

@spec to_mat(Nx.t()) :: {:ok, reference()} | {:error, String.t()}

Converts a tensor of Nx to Mat of evision (OpenCV).

If the tensor has three dimensions, it is expected to have shape{height, width, channels}.

to_mat(t, as_shape)

to_mat(binary, type, rows, cols, channels)

to_mat_2d(t)

to_nx(mat, backend \\ Evision.Backend)

Transform an Evision.Mat reference to Nx.tensor.

The resulting tensor is in the shape {height, width, channels}.

Example

iex> {:ok, mat} = Evision.imread("/path/to/exist/img.png")
iex> nx_tensor = Evision.Nx.to_nx(mat)
...> #Nx.Tensor<
...>    u8[1080][1920][3]
...>    [[ ... pixel data ... ]]
...> >

Functions

to_mat!(t)

Raising version of to_mat/1.

to_mat!(t, as_shape)

Raising version of to_mat/2.

to_mat!(binary, type, rows, cols, channels)

Raising version of to_mat/5.

to_mat_2d!(t)

Raising version of to_mat_2d/1.

to_nx!(mat)

Raising version of to_nx/1.

to_nx!(mat, backend)

Raising version of to_nx/2.