Evision.ORB (Evision v0.1.9)

cv

compute(self, images, keypoints)

Variant 1:

Positional Arguments

images: [Evision.Mat].

Image set.

Return

keypoints: [vector_KeyPoint].

Input collection of keypoints. Keypoints for which a descriptor cannot be computed are removed. Sometimes new keypoints can be added, for example: SIFT duplicates keypoint with several dominant orientations (for each orientation).

descriptors: [Evision.Mat].

Computed descriptors. In the second variant of the method descriptors[i] are descriptors computed for a keypoints[i]. Row j is the keypoints (or keypoints[i]) is the descriptor for keypoint j-th keypoint.

Has overloading in C++

Python prototype (for reference):

compute(images, keypoints[, descriptors]) -> keypoints, descriptors

Variant 2:

Computes the descriptors for a set of keypoints detected in an image (first variant) or image set (second variant).

Positional Arguments

image: Evision.Mat.

Image.

Return

keypoints: [KeyPoint].

Input collection of keypoints. Keypoints for which a descriptor cannot be computed are removed. Sometimes new keypoints can be added, for example: SIFT duplicates keypoint with several dominant orientations (for each orientation).

descriptors: Evision.Mat.

Computed descriptors. In the second variant of the method descriptors[i] are descriptors computed for a keypoints[i]. Row j is the keypoints (or keypoints[i]) is the descriptor for keypoint j-th keypoint.

Python prototype (for reference):

compute(image, keypoints[, descriptors]) -> keypoints, descriptors

compute(self, images, keypoints, opts)

Variant 1:

Positional Arguments

images: [Evision.Mat].

Image set.

Return

keypoints: [vector_KeyPoint].

Input collection of keypoints. Keypoints for which a descriptor cannot be computed are removed. Sometimes new keypoints can be added, for example: SIFT duplicates keypoint with several dominant orientations (for each orientation).

descriptors: [Evision.Mat].

Computed descriptors. In the second variant of the method descriptors[i] are descriptors computed for a keypoints[i]. Row j is the keypoints (or keypoints[i]) is the descriptor for keypoint j-th keypoint.

Has overloading in C++

Python prototype (for reference):

compute(images, keypoints[, descriptors]) -> keypoints, descriptors

Variant 2:

Computes the descriptors for a set of keypoints detected in an image (first variant) or image set (second variant).

Positional Arguments

image: Evision.Mat.

Image.

Return

keypoints: [KeyPoint].

Input collection of keypoints. Keypoints for which a descriptor cannot be computed are removed. Sometimes new keypoints can be added, for example: SIFT duplicates keypoint with several dominant orientations (for each orientation).

descriptors: Evision.Mat.

Computed descriptors. In the second variant of the method descriptors[i] are descriptors computed for a keypoints[i]. Row j is the keypoints (or keypoints[i]) is the descriptor for keypoint j-th keypoint.

Python prototype (for reference):

compute(image, keypoints[, descriptors]) -> keypoints, descriptors

create()

The ORB constructor

Keyword Arguments

nfeatures: int.

The maximum number of features to retain.

scaleFactor: float.

Pyramid decimation ratio, greater than 1. scaleFactor==2 means the classical pyramid, where each next level has 4x less pixels than the previous, but such a big scale factor will degrade feature matching scores dramatically. On the other hand, too close to 1 scale factor will mean that to cover certain scale range you will need more pyramid levels and so the speed will suffer.

nlevels: int.

The number of pyramid levels. The smallest level will have linear size equal to input_image_linear_size/pow(scaleFactor, nlevels - firstLevel).

edgeThreshold: int.

This is size of the border where the features are not detected. It should roughly match the patchSize parameter.

firstLevel: int.

The level of pyramid to put source image to. Previous layers are filled with upscaled source image.

wTA_K: int.

The number of points that produce each element of the oriented BRIEF descriptor. The default value 2 means the BRIEF where we take a random point pair and compare their brightnesses, so we get 0/1 response. Other possible values are 3 and 4. For example, 3 means that we take 3 random points (of course, those point coordinates are random, but they are generated from the pre-defined seed, so each element of BRIEF descriptor is computed deterministically from the pixel rectangle), find point of maximum brightness and output index of the winner (0, 1 or 2). Such output will occupy 2 bits, and therefore it will need a special variant of Hamming distance, denoted as NORM_HAMMING2 (2 bits per bin). When WTA_K=4, we take 4 random points to compute each bin (that will also occupy 2 bits with possible values 0, 1, 2 or 3).

scoreType: ORB_ScoreType.

The default HARRIS_SCORE means that Harris algorithm is used to rank features (the score is written to KeyPoint::score and is used to retain best nfeatures features); FAST_SCORE is alternative value of the parameter that produces slightly less stable keypoints, but it is a little faster to compute.

patchSize: int.

size of the patch used by the oriented BRIEF descriptor. Of course, on smaller pyramid layers the perceived image area covered by a feature will be larger.

fastThreshold: int.

the fast threshold

Python prototype (for reference):

create([, nfeatures[, scaleFactor[, nlevels[, edgeThreshold[, firstLevel[, WTA_K[, scoreType[, patchSize[, fastThreshold]]]]]]]]]) -> retval

create(opts)

The ORB constructor

Keyword Arguments

nfeatures: int.

The maximum number of features to retain.

scaleFactor: float.

Pyramid decimation ratio, greater than 1. scaleFactor==2 means the classical pyramid, where each next level has 4x less pixels than the previous, but such a big scale factor will degrade feature matching scores dramatically. On the other hand, too close to 1 scale factor will mean that to cover certain scale range you will need more pyramid levels and so the speed will suffer.

nlevels: int.

The number of pyramid levels. The smallest level will have linear size equal to input_image_linear_size/pow(scaleFactor, nlevels - firstLevel).

edgeThreshold: int.

This is size of the border where the features are not detected. It should roughly match the patchSize parameter.

firstLevel: int.

The level of pyramid to put source image to. Previous layers are filled with upscaled source image.

wTA_K: int.

The number of points that produce each element of the oriented BRIEF descriptor. The default value 2 means the BRIEF where we take a random point pair and compare their brightnesses, so we get 0/1 response. Other possible values are 3 and 4. For example, 3 means that we take 3 random points (of course, those point coordinates are random, but they are generated from the pre-defined seed, so each element of BRIEF descriptor is computed deterministically from the pixel rectangle), find point of maximum brightness and output index of the winner (0, 1 or 2). Such output will occupy 2 bits, and therefore it will need a special variant of Hamming distance, denoted as NORM_HAMMING2 (2 bits per bin). When WTA_K=4, we take 4 random points to compute each bin (that will also occupy 2 bits with possible values 0, 1, 2 or 3).

scoreType: ORB_ScoreType.

The default HARRIS_SCORE means that Harris algorithm is used to rank features (the score is written to KeyPoint::score and is used to retain best nfeatures features); FAST_SCORE is alternative value of the parameter that produces slightly less stable keypoints, but it is a little faster to compute.

patchSize: int.

size of the patch used by the oriented BRIEF descriptor. Of course, on smaller pyramid layers the perceived image area covered by a feature will be larger.

fastThreshold: int.

the fast threshold

Python prototype (for reference):

create([, nfeatures[, scaleFactor[, nlevels[, edgeThreshold[, firstLevel[, WTA_K[, scoreType[, patchSize[, fastThreshold]]]]]]]]]) -> retval

defaultNorm(self)

Python prototype (for reference):

defaultNorm() -> retval

descriptorSize(self)

Python prototype (for reference):

descriptorSize() -> retval

descriptorType(self)

Python prototype (for reference):

descriptorType() -> retval

detect(self, images)

Variant 1:

Positional Arguments

images: [Evision.Mat].

Image set.

Keyword Arguments

masks: [Evision.Mat].

Masks for each input image specifying where to look for keypoints (optional). masks[i] is a mask for images[i].

Return

keypoints: [vector_KeyPoint].

The detected keypoints. In the second variant of the method keypoints[i] is a set of keypoints detected in images[i] .

Has overloading in C++

Python prototype (for reference):

detect(images[, masks]) -> keypoints

Variant 2:

Detects keypoints in an image (first variant) or image set (second variant).

Positional Arguments

image: Evision.Mat.

Image.

Keyword Arguments

mask: Evision.Mat.

Mask specifying where to look for keypoints (optional). It must be a 8-bit integer matrix with non-zero values in the region of interest.

Return

keypoints: [KeyPoint].

The detected keypoints. In the second variant of the method keypoints[i] is a set of keypoints detected in images[i] .

Python prototype (for reference):

detect(image[, mask]) -> keypoints

detect(self, images, opts)

Variant 1:

Positional Arguments

images: [Evision.Mat].

Image set.

Keyword Arguments

masks: [Evision.Mat].

Masks for each input image specifying where to look for keypoints (optional). masks[i] is a mask for images[i].

Return

keypoints: [vector_KeyPoint].

The detected keypoints. In the second variant of the method keypoints[i] is a set of keypoints detected in images[i] .

Has overloading in C++

Python prototype (for reference):

detect(images[, masks]) -> keypoints

Variant 2:

Detects keypoints in an image (first variant) or image set (second variant).

Positional Arguments

image: Evision.Mat.

Image.

Keyword Arguments

mask: Evision.Mat.

Mask specifying where to look for keypoints (optional). It must be a 8-bit integer matrix with non-zero values in the region of interest.

Return

keypoints: [KeyPoint].

The detected keypoints. In the second variant of the method keypoints[i] is a set of keypoints detected in images[i] .

Python prototype (for reference):

detect(image[, mask]) -> keypoints

detectAndCompute(self, image, mask)

Positional Arguments

image: Evision.Mat

mask: Evision.Mat

Keyword Arguments

useProvidedKeypoints: bool.

Return

keypoints: [KeyPoint]

descriptors: Evision.Mat.

Detects keypoints and computes the descriptors

Python prototype (for reference):

detectAndCompute(image, mask[, descriptors[, useProvidedKeypoints]]) -> keypoints, descriptors

detectAndCompute(self, image, mask, opts)

Positional Arguments

image: Evision.Mat

mask: Evision.Mat

Keyword Arguments

useProvidedKeypoints: bool.

Return

keypoints: [KeyPoint]

descriptors: Evision.Mat.

Detects keypoints and computes the descriptors

Python prototype (for reference):

detectAndCompute(image, mask[, descriptors[, useProvidedKeypoints]]) -> keypoints, descriptors

empty(self)

Python prototype (for reference):

empty() -> retval

getDefaultName(self)

Python prototype (for reference):

getDefaultName() -> retval

getEdgeThreshold(self)

Python prototype (for reference):

getEdgeThreshold() -> retval

getFastThreshold(self)

Python prototype (for reference):

getFastThreshold() -> retval

getFirstLevel(self)

Python prototype (for reference):

getFirstLevel() -> retval

getMaxFeatures(self)

Python prototype (for reference):

getMaxFeatures() -> retval

getNLevels(self)

Python prototype (for reference):

getNLevels() -> retval

getPatchSize(self)

Python prototype (for reference):

getPatchSize() -> retval

getScaleFactor(self)

Python prototype (for reference):

getScaleFactor() -> retval

getScoreType(self)

Python prototype (for reference):

getScoreType() -> retval

getWTA_K(self)

Python prototype (for reference):

getWTA_K() -> retval

read(self, fileName)

Variant 1:

Positional Arguments

fileName: String

Python prototype (for reference):

read(fileName) -> None

Variant 2:

Positional Arguments

arg1: FileNode

Python prototype (for reference):

read(arg1) -> None

setEdgeThreshold(self, edgeThreshold)

Positional Arguments

edgeThreshold: int

Python prototype (for reference):

setEdgeThreshold(edgeThreshold) -> None

setFastThreshold(self, fastThreshold)

Positional Arguments

fastThreshold: int

Python prototype (for reference):

setFastThreshold(fastThreshold) -> None

setFirstLevel(self, firstLevel)

Positional Arguments

firstLevel: int

Python prototype (for reference):

setFirstLevel(firstLevel) -> None

setMaxFeatures(self, maxFeatures)

Positional Arguments

maxFeatures: int

Python prototype (for reference):

setMaxFeatures(maxFeatures) -> None

setNLevels(self, nlevels)

Positional Arguments

nlevels: int

Python prototype (for reference):

setNLevels(nlevels) -> None

setPatchSize(self, patchSize)

Positional Arguments

patchSize: int

Python prototype (for reference):

setPatchSize(patchSize) -> None

setScaleFactor(self, scaleFactor)

Positional Arguments

scaleFactor: double

Python prototype (for reference):

setScaleFactor(scaleFactor) -> None

setScoreType(self, scoreType)

Positional Arguments

scoreType: ORB_ScoreType

Python prototype (for reference):

setScoreType(scoreType) -> None

setWTA_K(self, wta_k)

Positional Arguments

wta_k: int

Python prototype (for reference):

setWTA_K(wta_k) -> None

write(self, fs)

Variant 1:

Positional Arguments

fs: Ptr<FileStorage>

Keyword Arguments

name: String.

Python prototype (for reference):

write(fs[, name]) -> None

Variant 2:

Positional Arguments

fileName: String

Python prototype (for reference):

write(fileName) -> None

write(self, fs, opts)

Positional Arguments

fs: Ptr<FileStorage>

Keyword Arguments

name: String.

Python prototype (for reference):

write(fs[, name]) -> None

Functions

compute!(self, images, keypoints)

Raising version of compute/3.

compute!(self, images, keypoints, opts)

Raising version of compute/4.

create!()

Raising version of create/0.

create!(opts)

Raising version of create/1.

defaultNorm!(self)

Raising version of defaultNorm/1.

descriptorSize!(self)

Raising version of descriptorSize/1.

descriptorType!(self)

Raising version of descriptorType/1.

detect!(self, images)

Raising version of detect/2.

detect!(self, images, opts)

Raising version of detect/3.

detectAndCompute!(self, image, mask)

Raising version of detectAndCompute/3.

detectAndCompute!(self, image, mask, opts)

Raising version of detectAndCompute/4.

empty!(self)

Raising version of empty/1.

getDefaultName!(self)

Raising version of getDefaultName/1.

getEdgeThreshold!(self)

Raising version of getEdgeThreshold/1.

getFastThreshold!(self)

Raising version of getFastThreshold/1.

getFirstLevel!(self)

Raising version of getFirstLevel/1.

getMaxFeatures!(self)

Raising version of getMaxFeatures/1.

getNLevels!(self)

Raising version of getNLevels/1.

getPatchSize!(self)

Raising version of getPatchSize/1.

getScaleFactor!(self)

Raising version of getScaleFactor/1.

getScoreType!(self)

Raising version of getScoreType/1.

getWTA_K!(self)

Raising version of getWTA_K/1.

read!(self, fileName)

Raising version of read/2.

setEdgeThreshold!(self, edgeThreshold)

Raising version of setEdgeThreshold/2.

setFastThreshold!(self, fastThreshold)

Raising version of setFastThreshold/2.

setFirstLevel!(self, firstLevel)

Raising version of setFirstLevel/2.

setMaxFeatures!(self, maxFeatures)

Raising version of setMaxFeatures/2.

setNLevels!(self, nlevels)

Raising version of setNLevels/2.

setPatchSize!(self, patchSize)

Raising version of setPatchSize/2.

setScaleFactor!(self, scaleFactor)

Raising version of setScaleFactor/2.

setScoreType!(self, scoreType)

Raising version of setScoreType/2.

setWTA_K!(self, wta_k)

Raising version of setWTA_K/2.

write!(self, fs)

Raising version of write/2.

write!(self, fs, opts)

Raising version of write/3.