Evision.PyRotationWarper (Evision v0.1.7-dev)

cv

buildMaps(self, src_size, k, r)

Positional Arguments

src_size: Size.

k: UMat.

r: UMat.

Keyword Arguments

xmap: UMat.

ymap: UMat.

Builds the projection maps according to the given camera data. @return Projected image minimum bounding box

Python prototype (for reference):

buildMaps(src_size, K, R[, xmap[, ymap]]) -> retval, xmap, ymap

buildMaps(self, src_size, k, r, opts)

Positional Arguments

src_size: Size.

k: UMat.

r: UMat.

Keyword Arguments

xmap: UMat.

ymap: UMat.

Builds the projection maps according to the given camera data. @return Projected image minimum bounding box

Python prototype (for reference):

buildMaps(src_size, K, R[, xmap[, ymap]]) -> retval, xmap, ymap

getScale(self)

Python prototype (for reference):

getScale() -> retval

pyRotationWarper()

Python prototype (for reference):

PyRotationWarper() -> <PyRotationWarper object>

pyRotationWarper(type, scale)

Python prototype (for reference):

PyRotationWarper(type, scale) -> <PyRotationWarper object>

setScale(self, arg1)

Python prototype (for reference):

setScale(arg1) -> None

warp(self, src, k, r, interp_mode, border_mode)

Positional Arguments

src: UMat.

k: UMat.

r: UMat.

interp_mode: int.

border_mode: int.

Keyword Arguments

dst: UMat.

Projects the image. @return Project image top-left corner

Python prototype (for reference):

warp(src, K, R, interp_mode, border_mode[, dst]) -> retval, dst

warp(self, src, k, r, interp_mode, border_mode, opts)

Positional Arguments

src: UMat.

k: UMat.

r: UMat.

interp_mode: int.

border_mode: int.

Keyword Arguments

dst: UMat.

Projects the image. @return Project image top-left corner

Python prototype (for reference):

warp(src, K, R, interp_mode, border_mode[, dst]) -> retval, dst

warpBackward(self, src, k, r, interp_mode, border_mode, dst_size)

Positional Arguments

src: UMat.

k: UMat.

r: UMat.

interp_mode: int.

border_mode: int.

dst_size: Size.

Keyword Arguments

dst: UMat.

Projects the image backward.

Python prototype (for reference):

warpBackward(src, K, R, interp_mode, border_mode, dst_size[, dst]) -> dst

warpBackward(self, src, k, r, interp_mode, border_mode, dst_size, opts)

Positional Arguments

src: UMat.

k: UMat.

r: UMat.

interp_mode: int.

border_mode: int.

dst_size: Size.

Keyword Arguments

dst: UMat.

Projects the image backward.

Python prototype (for reference):

warpBackward(src, K, R, interp_mode, border_mode, dst_size[, dst]) -> dst

warpPoint(self, pt, k, r)

Projects the image point. @return Projected point

Python prototype (for reference):

warpPoint(pt, K, R) -> retval

warpPointBackward(self, pt, k, r)

Python prototype (for reference):

warpPointBackward(pt, K, R) -> retval

warpRoi(self, src_size, k, r)

@return Projected image minimum bounding box

Python prototype (for reference):

warpRoi(src_size, K, R) -> retval

Functions

buildMaps!(self, src_size, k, r)

Raising version of buildMaps/4.

buildMaps!(self, src_size, k, r, opts)

Raising version of buildMaps/5.

getScale!(self)

Raising version of getScale/1.

pyRotationWarper!()

Raising version of pyRotationWarper/0.

pyRotationWarper!(type, scale)

Raising version of pyRotationWarper/2.

setScale!(self, arg1)

Raising version of setScale/2.

warp!(self, src, k, r, interp_mode, border_mode)

Raising version of warp/6.

warp!(self, src, k, r, interp_mode, border_mode, opts)

Raising version of warp/7.

warpBackward!(self, src, k, r, interp_mode, border_mode, dst_size)

Raising version of warpBackward/7.

warpBackward!(self, src, k, r, interp_mode, border_mode, dst_size, opts)

Raising version of warpBackward/8.

warpPoint!(self, pt, k, r)

Raising version of warpPoint/4.

warpPointBackward!(self, pt, k, r)

Raising version of warpPointBackward/4.

warpRoi!(self, src_size, k, r)

Raising version of warpRoi/4.