Evision.QRCodeDetector (Evision v0.1.9)

cv

decode(self, img, points)

Decodes QR code in image once it's found by the detect() method.

Positional Arguments

img: Evision.Mat.

grayscale or color (BGR) image containing QR code.

points: Evision.Mat.

Quadrangle vertices found by detect() method (or some other algorithm).

Return

straight_qrcode: Evision.Mat.

The optional output image containing rectified and binarized QR code

Returns UTF8-encoded output string or empty string if the code cannot be decoded.

Python prototype (for reference):

decode(img, points[, straight_qrcode]) -> retval, straight_qrcode

decode(self, img, points, opts)

Decodes QR code in image once it's found by the detect() method.

Positional Arguments

img: Evision.Mat.

grayscale or color (BGR) image containing QR code.

points: Evision.Mat.

Quadrangle vertices found by detect() method (or some other algorithm).

Return

straight_qrcode: Evision.Mat.

The optional output image containing rectified and binarized QR code

Returns UTF8-encoded output string or empty string if the code cannot be decoded.

Python prototype (for reference):

decode(img, points[, straight_qrcode]) -> retval, straight_qrcode

decodeCurved(self, img, points)

Decodes QR code on a curved surface in image once it's found by the detect() method.

Positional Arguments

img: Evision.Mat.

grayscale or color (BGR) image containing QR code.

points: Evision.Mat.

Quadrangle vertices found by detect() method (or some other algorithm).

Return

straight_qrcode: Evision.Mat.

The optional output image containing rectified and binarized QR code

Returns UTF8-encoded output string or empty string if the code cannot be decoded.

Python prototype (for reference):

decodeCurved(img, points[, straight_qrcode]) -> retval, straight_qrcode

decodeCurved(self, img, points, opts)

Decodes QR code on a curved surface in image once it's found by the detect() method.

Positional Arguments

img: Evision.Mat.

grayscale or color (BGR) image containing QR code.

points: Evision.Mat.

Quadrangle vertices found by detect() method (or some other algorithm).

Return

straight_qrcode: Evision.Mat.

The optional output image containing rectified and binarized QR code

Returns UTF8-encoded output string or empty string if the code cannot be decoded.

Python prototype (for reference):

decodeCurved(img, points[, straight_qrcode]) -> retval, straight_qrcode

decodeMulti(self, img, points)

Decodes QR codes in image once it's found by the detect() method.

Positional Arguments

img: Evision.Mat.

grayscale or color (BGR) image containing QR codes.

points: Evision.Mat.

vector of Quadrangle vertices found by detect() method (or some other algorithm).

Return

decoded_info: [string].

UTF8-encoded output vector of string or empty vector of string if the codes cannot be decoded.

straight_qrcode: [Evision.Mat].

The optional output vector of images containing rectified and binarized QR codes

Python prototype (for reference):

decodeMulti(img, points[, straight_qrcode]) -> retval, decoded_info, straight_qrcode

decodeMulti(self, img, points, opts)

Decodes QR codes in image once it's found by the detect() method.

Positional Arguments

img: Evision.Mat.

grayscale or color (BGR) image containing QR codes.

points: Evision.Mat.

vector of Quadrangle vertices found by detect() method (or some other algorithm).

Return

decoded_info: [string].

UTF8-encoded output vector of string or empty vector of string if the codes cannot be decoded.

straight_qrcode: [Evision.Mat].

The optional output vector of images containing rectified and binarized QR codes

Python prototype (for reference):

decodeMulti(img, points[, straight_qrcode]) -> retval, decoded_info, straight_qrcode

detect(self, img)

Detects QR code in image and returns the quadrangle containing the code.

Positional Arguments

img: Evision.Mat.

grayscale or color (BGR) image containing (or not) QR code.

Return

points: Evision.Mat.

Output vector of vertices of the minimum-area quadrangle containing the code.

Python prototype (for reference):

detect(img[, points]) -> retval, points

detect(self, img, opts)

Detects QR code in image and returns the quadrangle containing the code.

Positional Arguments

img: Evision.Mat.

grayscale or color (BGR) image containing (or not) QR code.

Return

points: Evision.Mat.

Output vector of vertices of the minimum-area quadrangle containing the code.

Python prototype (for reference):

detect(img[, points]) -> retval, points

detectAndDecode(self, img)

Both detects and decodes QR code

Positional Arguments

img: Evision.Mat.

grayscale or color (BGR) image containing QR code.

Return

points: Evision.Mat.

optional output array of vertices of the found QR code quadrangle. Will be empty if not found.

straight_qrcode: Evision.Mat.

The optional output image containing rectified and binarized QR code

Python prototype (for reference):

detectAndDecode(img[, points[, straight_qrcode]]) -> retval, points, straight_qrcode

detectAndDecode(self, img, opts)

Both detects and decodes QR code

Positional Arguments

img: Evision.Mat.

grayscale or color (BGR) image containing QR code.

Return

points: Evision.Mat.

optional output array of vertices of the found QR code quadrangle. Will be empty if not found.

straight_qrcode: Evision.Mat.

The optional output image containing rectified and binarized QR code

Python prototype (for reference):

detectAndDecode(img[, points[, straight_qrcode]]) -> retval, points, straight_qrcode

detectAndDecodeCurved(self, img)

Both detects and decodes QR code on a curved surface

Positional Arguments

img: Evision.Mat.

grayscale or color (BGR) image containing QR code.

Return

points: Evision.Mat.

optional output array of vertices of the found QR code quadrangle. Will be empty if not found.

straight_qrcode: Evision.Mat.

The optional output image containing rectified and binarized QR code

Python prototype (for reference):

detectAndDecodeCurved(img[, points[, straight_qrcode]]) -> retval, points, straight_qrcode

detectAndDecodeCurved(self, img, opts)

Both detects and decodes QR code on a curved surface

Positional Arguments

img: Evision.Mat.

grayscale or color (BGR) image containing QR code.

Return

points: Evision.Mat.

optional output array of vertices of the found QR code quadrangle. Will be empty if not found.

straight_qrcode: Evision.Mat.

The optional output image containing rectified and binarized QR code

Python prototype (for reference):

detectAndDecodeCurved(img[, points[, straight_qrcode]]) -> retval, points, straight_qrcode

detectAndDecodeMulti(self, img)

Both detects and decodes QR codes

Positional Arguments

img: Evision.Mat.

grayscale or color (BGR) image containing QR codes.

Return

decoded_info: [string].

UTF8-encoded output vector of string or empty vector of string if the codes cannot be decoded.

points: Evision.Mat.

optional output vector of vertices of the found QR code quadrangles. Will be empty if not found.

straight_qrcode: [Evision.Mat].

The optional output vector of images containing rectified and binarized QR codes

Python prototype (for reference):

detectAndDecodeMulti(img[, points[, straight_qrcode]]) -> retval, decoded_info, points, straight_qrcode

detectAndDecodeMulti(self, img, opts)

Both detects and decodes QR codes

Positional Arguments

img: Evision.Mat.

grayscale or color (BGR) image containing QR codes.

Return

decoded_info: [string].

UTF8-encoded output vector of string or empty vector of string if the codes cannot be decoded.

points: Evision.Mat.

optional output vector of vertices of the found QR code quadrangles. Will be empty if not found.

straight_qrcode: [Evision.Mat].

The optional output vector of images containing rectified and binarized QR codes

Python prototype (for reference):

detectAndDecodeMulti(img[, points[, straight_qrcode]]) -> retval, decoded_info, points, straight_qrcode

detectMulti(self, img)

Detects QR codes in image and returns the vector of the quadrangles containing the codes.

Positional Arguments

img: Evision.Mat.

grayscale or color (BGR) image containing (or not) QR codes.

Return

points: Evision.Mat.

Output vector of vector of vertices of the minimum-area quadrangle containing the codes.

Python prototype (for reference):

detectMulti(img[, points]) -> retval, points

detectMulti(self, img, opts)

Detects QR codes in image and returns the vector of the quadrangles containing the codes.

Positional Arguments

img: Evision.Mat.

grayscale or color (BGR) image containing (or not) QR codes.

Return

points: Evision.Mat.

Output vector of vector of vertices of the minimum-area quadrangle containing the codes.

Python prototype (for reference):

detectMulti(img[, points]) -> retval, points

qrCodeDetector()

Python prototype (for reference):

QRCodeDetector() -> <QRCodeDetector object>

setEpsX(self, epsX)

sets the epsilon used during the horizontal scan of QR code stop marker detection.

Positional Arguments

epsX: double.

Epsilon neighborhood, which allows you to determine the horizontal pattern of the scheme 1:1:3:1:1 according to QR code standard.

Python prototype (for reference):

setEpsX(epsX) -> None

setEpsY(self, epsY)

sets the epsilon used during the vertical scan of QR code stop marker detection.

Positional Arguments

epsY: double.

Epsilon neighborhood, which allows you to determine the vertical pattern of the scheme 1:1:3:1:1 according to QR code standard.

Python prototype (for reference):

setEpsY(epsY) -> None

Functions

decode!(self, img, points)

Raising version of decode/3.

decode!(self, img, points, opts)

Raising version of decode/4.

decodeCurved!(self, img, points)

Raising version of decodeCurved/3.

decodeCurved!(self, img, points, opts)

Raising version of decodeCurved/4.

decodeMulti!(self, img, points)

Raising version of decodeMulti/3.

decodeMulti!(self, img, points, opts)

Raising version of decodeMulti/4.

detect!(self, img)

Raising version of detect/2.

detect!(self, img, opts)

Raising version of detect/3.

detectAndDecode!(self, img)

Raising version of detectAndDecode/2.

detectAndDecode!(self, img, opts)

Raising version of detectAndDecode/3.

detectAndDecodeCurved!(self, img)

Raising version of detectAndDecodeCurved/2.

detectAndDecodeCurved!(self, img, opts)

Raising version of detectAndDecodeCurved/3.

detectAndDecodeMulti!(self, img)

Raising version of detectAndDecodeMulti/2.

detectAndDecodeMulti!(self, img, opts)

Raising version of detectAndDecodeMulti/3.

detectMulti!(self, img)

Raising version of detectMulti/2.

detectMulti!(self, img, opts)

Raising version of detectMulti/3.

qrCodeDetector!()

Raising version of qrCodeDetector/0.

setEpsX!(self, epsX)

Raising version of setEpsX/2.

setEpsY!(self, epsY)

Raising version of setEpsY/2.