Evision.SIFT (Evision v0.1.7-dev)

cv

compute(self, images, keypoints)

Variaint 1:

Positional Arguments

images: vector_UMat.

keypoints: vector_vector_KeyPoint.

Keyword Arguments

descriptors: vector_UMat.

Has overloading in C++

computed are removed. Sometimes new keypoints can be added, for example: SIFT duplicates keypoint
with several dominant orientations (for each orientation).
descriptors computed for a keypoints[i]. Row j is the keypoints (or keypoints[i]) is the
descriptor for keypoint j-th keypoint.

Python prototype (for reference):

compute(images, keypoints[, descriptors]) -> keypoints, descriptors

Variaint 2:

Positional Arguments

image: UMat. Image.

keypoints: vector_KeyPoint.

Keyword Arguments

descriptors: UMat.

Computes the descriptors for a set of keypoints detected in an image (first variant) or image set (second variant).

computed are removed. Sometimes new keypoints can be added, for example: SIFT duplicates keypoint
with several dominant orientations (for each orientation).
descriptors computed for a keypoints[i]. Row j is the keypoints (or keypoints[i]) is the
descriptor for keypoint j-th keypoint.

Python prototype (for reference):

compute(image, keypoints[, descriptors]) -> keypoints, descriptors

compute(self, images, keypoints, opts)

Variaint 1:

Positional Arguments

images: vector_UMat.

keypoints: vector_vector_KeyPoint.

Keyword Arguments

descriptors: vector_UMat.

Has overloading in C++

computed are removed. Sometimes new keypoints can be added, for example: SIFT duplicates keypoint
with several dominant orientations (for each orientation).
descriptors computed for a keypoints[i]. Row j is the keypoints (or keypoints[i]) is the
descriptor for keypoint j-th keypoint.

Python prototype (for reference):

compute(images, keypoints[, descriptors]) -> keypoints, descriptors

Variaint 2:

Positional Arguments

image: UMat. Image.

keypoints: vector_KeyPoint.

Keyword Arguments

descriptors: UMat.

Computes the descriptors for a set of keypoints detected in an image (first variant) or image set (second variant).

computed are removed. Sometimes new keypoints can be added, for example: SIFT duplicates keypoint
with several dominant orientations (for each orientation).
descriptors computed for a keypoints[i]. Row j is the keypoints (or keypoints[i]) is the
descriptor for keypoint j-th keypoint.

Python prototype (for reference):

compute(image, keypoints[, descriptors]) -> keypoints, descriptors

create()

Positional Arguments

Keyword Arguments

nfeatures: int.

nOctaveLayers: int.

contrastThreshold: double.

edgeThreshold: double.

sigma: double.

(measured in SIFT algorithm as the local contrast) number of octaves is computed automatically from the image resolution. (low-contrast) regions. The larger the threshold, the less features are produced by the detector.

Note: The contrast threshold will be divided by nOctaveLayers when the filtering is applied. When nOctaveLayers is set to default and if you want to use the value used in D. Lowe paper, 0.03, set this argument to 0.09.

is different from the contrastThreshold, i.e. the larger the edgeThreshold, the less features are
filtered out (more features are retained).
is captured with a weak camera with soft lenses, you might want to reduce the number.

Python prototype (for reference):

create([, nfeatures[, nOctaveLayers[, contrastThreshold[, edgeThreshold[, sigma]]]]]) -> retval

create(opts)

Positional Arguments

Keyword Arguments

nfeatures: int.

nOctaveLayers: int.

contrastThreshold: double.

edgeThreshold: double.

sigma: double.

(measured in SIFT algorithm as the local contrast) number of octaves is computed automatically from the image resolution. (low-contrast) regions. The larger the threshold, the less features are produced by the detector.

Note: The contrast threshold will be divided by nOctaveLayers when the filtering is applied. When nOctaveLayers is set to default and if you want to use the value used in D. Lowe paper, 0.03, set this argument to 0.09.

is different from the contrastThreshold, i.e. the larger the edgeThreshold, the less features are
filtered out (more features are retained).
is captured with a weak camera with soft lenses, you might want to reduce the number.

Python prototype (for reference):

create([, nfeatures[, nOctaveLayers[, contrastThreshold[, edgeThreshold[, sigma]]]]]) -> retval

create(nfeatures, nOctaveLayers, contrastThreshold, edgeThreshold, sigma, descriptorType)

Create SIFT with specified descriptorType.

(measured in SIFT algorithm as the local contrast)
number of octaves is computed automatically from the image resolution.
(low-contrast) regions. The larger the threshold, the less features are produced by the detector.

Note: The contrast threshold will be divided by nOctaveLayers when the filtering is applied. When nOctaveLayers is set to default and if you want to use the value used in D. Lowe paper, 0.03, set this argument to 0.09.

is different from the contrastThreshold, i.e. the larger the edgeThreshold, the less features are
filtered out (more features are retained).
is captured with a weak camera with soft lenses, you might want to reduce the number.

Python prototype (for reference):

create(nfeatures, nOctaveLayers, contrastThreshold, edgeThreshold, sigma, descriptorType) -> retval

defaultNorm(self)

Python prototype (for reference):

defaultNorm() -> retval

descriptorSize(self)

Python prototype (for reference):

descriptorSize() -> retval

descriptorType(self)

Python prototype (for reference):

descriptorType() -> retval

detect(self, images)

Variaint 1:

Positional Arguments

images: vector_UMat.

Keyword Arguments

masks: vector_UMat.

Has overloading in C++

of keypoints detected in images[i] .
masks[i] is a mask for images[i].

Python prototype (for reference):

detect(images[, masks]) -> keypoints

Variaint 2:

Positional Arguments

image: UMat. Image.

Keyword Arguments

mask: UMat.

Detects keypoints in an image (first variant) or image set (second variant).

of keypoints detected in images[i] .
matrix with non-zero values in the region of interest.

Python prototype (for reference):

detect(image[, mask]) -> keypoints

detect(self, images, opts)

Variaint 1:

Positional Arguments

images: vector_UMat.

Keyword Arguments

masks: vector_UMat.

Has overloading in C++

of keypoints detected in images[i] .
masks[i] is a mask for images[i].

Python prototype (for reference):

detect(images[, masks]) -> keypoints

Variaint 2:

Positional Arguments

image: UMat. Image.

Keyword Arguments

mask: UMat.

Detects keypoints in an image (first variant) or image set (second variant).

of keypoints detected in images[i] .
matrix with non-zero values in the region of interest.

Python prototype (for reference):

detect(image[, mask]) -> keypoints

detectAndCompute(self, image, mask)

Positional Arguments

image: UMat

mask: UMat

Keyword Arguments

descriptors: UMat.

useProvidedKeypoints: bool.

Detects keypoints and computes the descriptors

Python prototype (for reference):

detectAndCompute(image, mask[, descriptors[, useProvidedKeypoints]]) -> keypoints, descriptors

detectAndCompute(self, image, mask, opts)

Positional Arguments

image: UMat

mask: UMat

Keyword Arguments

descriptors: UMat.

useProvidedKeypoints: bool.

Detects keypoints and computes the descriptors

Python prototype (for reference):

detectAndCompute(image, mask[, descriptors[, useProvidedKeypoints]]) -> keypoints, descriptors

empty(self)

Python prototype (for reference):

empty() -> retval

getDefaultName(self)

Python prototype (for reference):

getDefaultName() -> retval

read(self, arg1)

Variaint 1:

Python prototype (for reference):

read(arg1) -> None

Variaint 2:

Python prototype (for reference):

read(fileName) -> None

write(self, fs)

Variaint 1:

Positional Arguments

fs: Ptr<FileStorage>

Keyword Arguments

name: String.

Python prototype (for reference):

write(fs[, name]) -> None

Variaint 2:

Python prototype (for reference):

write(fileName) -> None

write(self, fs, opts)

Positional Arguments

fs: Ptr<FileStorage>

Keyword Arguments

name: String.

Python prototype (for reference):

write(fs[, name]) -> None

Functions

compute!(self, images, keypoints)

Raising version of compute/3.

compute!(self, images, keypoints, opts)

Raising version of compute/4.

create!()

Raising version of create/0.

create!(opts)

Raising version of create/1.

create!(nfeatures, nOctaveLayers, contrastThreshold, edgeThreshold, sigma, descriptorType)

Raising version of create/6.

defaultNorm!(self)

Raising version of defaultNorm/1.

descriptorSize!(self)

Raising version of descriptorSize/1.

descriptorType!(self)

Raising version of descriptorType/1.

detect!(self, images)

Raising version of detect/2.

detect!(self, images, opts)

Raising version of detect/3.

detectAndCompute!(self, image, mask)

Raising version of detectAndCompute/3.

detectAndCompute!(self, image, mask, opts)

Raising version of detectAndCompute/4.

empty!(self)

Raising version of empty/1.

getDefaultName!(self)

Raising version of getDefaultName/1.

read!(self, arg1)

Raising version of read/2.

write!(self, fs)

Raising version of write/2.

write!(self, fs, opts)

Raising version of write/3.