Evision.Samples (Evision v0.1.9)

cv.samples

addSamplesDataSearchPath(path)

Override search data path by adding new search location

Positional Arguments

path: String.

Path to used samples data

Use this only to override default behavior Passed paths are used in LIFO order.

Python prototype (for reference):

addSamplesDataSearchPath(path) -> None

addSamplesDataSearchSubDirectory(subdir)

Append samples search data sub directory

Positional Arguments

subdir: String.

samples data sub directory

General usage is to add OpenCV modules name (<opencv_contrib>/modules/<name>/samples/data -> <name>/samples/data + modules/<name>/samples/data). Passed subdirectories are used in LIFO order.

Python prototype (for reference):

addSamplesDataSearchSubDirectory(subdir) -> None

findFile(relative_path)

Try to find requested data file

Positional Arguments

relative_path: String.

Relative path to data file

Keyword Arguments

required: bool.

Specify "file not found" handling. If true, function prints information message and raises cv::Exception. If false, function returns empty result

silentMode: bool.

Disables messages

Search directories:

Directories passed via addSamplesDataSearchPath()

OPENCV_SAMPLES_DATA_PATH_HINT environment variable

OPENCV_SAMPLES_DATA_PATH environment variable If parameter value is not empty and nothing is found then stop searching.

Detects build/install path based on: a. current working directory (CWD) b. and/or binary module location (opencv_core/opencv_world, doesn't work with static linkage)

Scan <source>/{,data,samples/data} directories if build directory is detected or the current directory is in source tree.

Scan <install>/share/OpenCV directory if install directory is detected. @see cv::utils::findDataFile @return Returns path (absolute or relative to the current directory) or empty string if file is not found

Python prototype (for reference):

findFile(relative_path[, required[, silentMode]]) -> retval

findFile(relative_path, opts)

Try to find requested data file

Positional Arguments

relative_path: String.

Relative path to data file

Keyword Arguments

required: bool.

Specify "file not found" handling. If true, function prints information message and raises cv::Exception. If false, function returns empty result

silentMode: bool.

Disables messages

Search directories:

Directories passed via addSamplesDataSearchPath()

OPENCV_SAMPLES_DATA_PATH_HINT environment variable

OPENCV_SAMPLES_DATA_PATH environment variable If parameter value is not empty and nothing is found then stop searching.

Detects build/install path based on: a. current working directory (CWD) b. and/or binary module location (opencv_core/opencv_world, doesn't work with static linkage)

Scan <source>/{,data,samples/data} directories if build directory is detected or the current directory is in source tree.

Scan <install>/share/OpenCV directory if install directory is detected. @see cv::utils::findDataFile @return Returns path (absolute or relative to the current directory) or empty string if file is not found

Python prototype (for reference):

findFile(relative_path[, required[, silentMode]]) -> retval

findFileOrKeep(relative_path)

Positional Arguments

relative_path: String

Keyword Arguments

silentMode: bool.

Python prototype (for reference):

findFileOrKeep(relative_path[, silentMode]) -> retval

findFileOrKeep(relative_path, opts)

Positional Arguments

relative_path: String

Keyword Arguments

silentMode: bool.

Python prototype (for reference):

findFileOrKeep(relative_path[, silentMode]) -> retval

Functions

addSamplesDataSearchPath!(path)

Raising version of addSamplesDataSearchPath/1.

addSamplesDataSearchSubDirectory!(subdir)

Raising version of addSamplesDataSearchSubDirectory/1.

findFile!(relative_path)

Raising version of findFile/1.

findFile!(relative_path, opts)

Raising version of findFile/2.

findFileOrKeep!(relative_path)

Raising version of findFileOrKeep/1.

findFileOrKeep!(relative_path, opts)

Raising version of findFileOrKeep/2.