Evision.Segmentation.IntelligentScissorsMB (Evision v0.1.9)

cv.segmentation

applyImage(self, image)

Specify input image and extract image features

Positional Arguments

image: Evision.Mat.

input image. Type is #CV_8UC1 / #CV_8UC3

Python prototype (for reference):

applyImage(image) -> retval

applyImageFeatures(self, non_edge, gradient_direction, gradient_magnitude)

Specify custom features of input image

Positional Arguments

non_edge: Evision.Mat.

Specify cost of non-edge pixels. Type is CV_8UC1. Expected values are {0, 1}.

gradient_direction: Evision.Mat.

Specify gradient direction feature. Type is CV_32FC2. Values are expected to be normalized: x^2 + y^2 == 1

gradient_magnitude: Evision.Mat.

Specify cost of gradient magnitude function: Type is CV_32FC1. Values should be in range [0, 1].

Keyword Arguments

image: Evision.Mat.

Optional parameter. Must be specified if subset of features is specified (non-specified features are calculated internally)

Customized advanced variant of applyImage() call.

Python prototype (for reference):

applyImageFeatures(non_edge, gradient_direction, gradient_magnitude[, image]) -> retval

applyImageFeatures(self, non_edge, gradient_direction, gradient_magnitude, opts)

Specify custom features of input image

Positional Arguments

non_edge: Evision.Mat.

Specify cost of non-edge pixels. Type is CV_8UC1. Expected values are {0, 1}.

gradient_direction: Evision.Mat.

Specify gradient direction feature. Type is CV_32FC2. Values are expected to be normalized: x^2 + y^2 == 1

gradient_magnitude: Evision.Mat.

Specify cost of gradient magnitude function: Type is CV_32FC1. Values should be in range [0, 1].

Keyword Arguments

image: Evision.Mat.

Optional parameter. Must be specified if subset of features is specified (non-specified features are calculated internally)

Customized advanced variant of applyImage() call.

Python prototype (for reference):

applyImageFeatures(non_edge, gradient_direction, gradient_magnitude[, image]) -> retval

buildMap(self, sourcePt)

Prepares a map of optimal paths for the given source point on the image

Positional Arguments

sourcePt: Point.

The source point used to find the paths

Note: applyImage() / applyImageFeatures() must be called before this call

Python prototype (for reference):

buildMap(sourcePt) -> None

getContour(self, targetPt)

Extracts optimal contour for the given target point on the image

Positional Arguments

targetPt: Point.

The target point

Keyword Arguments

backward: bool.

Flag to indicate reverse order of retrived pixels (use "true" value to fetch points from the target to the source point)

Return

contour: Evision.Mat.

The list of pixels which contains optimal path between the source and the target points of the image. Type is CV_32SC2 (compatible with std::vector<Point>)

Note: buildMap() must be called before this call

Python prototype (for reference):

getContour(targetPt[, contour[, backward]]) -> contour

getContour(self, targetPt, opts)

Extracts optimal contour for the given target point on the image

Positional Arguments

targetPt: Point.

The target point

Keyword Arguments

backward: bool.

Flag to indicate reverse order of retrived pixels (use "true" value to fetch points from the target to the source point)

Return

contour: Evision.Mat.

The list of pixels which contains optimal path between the source and the target points of the image. Type is CV_32SC2 (compatible with std::vector<Point>)

Note: buildMap() must be called before this call

Python prototype (for reference):

getContour(targetPt[, contour[, backward]]) -> contour

segmentation_IntelligentScissorsMB()

Python prototype (for reference):

IntelligentScissorsMB() -> <segmentation_IntelligentScissorsMB object>

setEdgeFeatureCannyParameters(self, threshold1, threshold2)

Switch edge feature extractor to use Canny edge detector

Positional Arguments

threshold1: double

threshold2: double

Keyword Arguments

apertureSize: int.

l2gradient: bool.

Note: "Laplacian Zero-Crossing" feature extractor is used by default (following to original article) @sa Canny

Python prototype (for reference):

setEdgeFeatureCannyParameters(threshold1, threshold2[, apertureSize[, L2gradient]]) -> retval

setEdgeFeatureCannyParameters(self, threshold1, threshold2, opts)

Switch edge feature extractor to use Canny edge detector

Positional Arguments

threshold1: double

threshold2: double

Keyword Arguments

apertureSize: int.

l2gradient: bool.

Note: "Laplacian Zero-Crossing" feature extractor is used by default (following to original article) @sa Canny

Python prototype (for reference):

setEdgeFeatureCannyParameters(threshold1, threshold2[, apertureSize[, L2gradient]]) -> retval

setEdgeFeatureZeroCrossingParameters(self)

Switch to "Laplacian Zero-Crossing" edge feature extractor and specify its parameters

Keyword Arguments

gradient_magnitude_min_value: float.

Minimal gradient magnitude value for edge pixels (default: 0, check is disabled)

This feature extractor is used by default according to article. Implementation has additional filtering for regions with low-amplitude noise. This filtering is enabled through parameter of minimal gradient amplitude (use some small value 4, 8, 16). Note: Current implementation of this feature extractor is based on processing of grayscale images (color image is converted to grayscale image first). Note: Canny edge detector is a bit slower, but provides better results (especially on color images): use setEdgeFeatureCannyParameters().

Python prototype (for reference):

setEdgeFeatureZeroCrossingParameters([, gradient_magnitude_min_value]) -> retval

setGradientMagnitudeMaxLimit(self)

Specify gradient magnitude max value threshold

Keyword Arguments

gradient_magnitude_threshold_max: float.

Specify gradient magnitude max value threshold (default: 0, disabled)

Zero limit value is used to disable gradient magnitude thresholding (default behavior, as described in original article). Otherwize pixels with gradient magnitude >= threshold have zero cost. Note: Thresholding should be used for images with irregular regions (to avoid stuck on parameters from high-contract areas, like embedded logos).

Python prototype (for reference):

setGradientMagnitudeMaxLimit([, gradient_magnitude_threshold_max]) -> retval

setWeights(self, weight_non_edge, weight_gradient_direction, weight_gradient_magnitude)

Specify weights of feature functions

Positional Arguments

weight_non_edge: float.

Specify cost of non-edge pixels (default: 0.43f)

weight_gradient_direction: float.

Specify cost of gradient direction function (default: 0.43f)

weight_gradient_magnitude: float.

Specify cost of gradient magnitude function (default: 0.14f)

Consider keeping weights normalized (sum of weights equals to 1.0) Discrete dynamic programming (DP) goal is minimization of costs between pixels.

Python prototype (for reference):

setWeights(weight_non_edge, weight_gradient_direction, weight_gradient_magnitude) -> retval

Functions

applyImage!(self, image)

Raising version of applyImage/2.

applyImageFeatures!(self, non_edge, gradient_direction, gradient_magnitude)

Raising version of applyImageFeatures/4.

applyImageFeatures!(self, non_edge, gradient_direction, gradient_magnitude, opts)

Raising version of applyImageFeatures/5.

buildMap!(self, sourcePt)

Raising version of buildMap/2.

getContour!(self, targetPt)

Raising version of getContour/2.

getContour!(self, targetPt, opts)

Raising version of getContour/3.

segmentation_IntelligentScissorsMB!()

Raising version of segmentation_IntelligentScissorsMB/0.

setEdgeFeatureCannyParameters!(self, threshold1, threshold2)

Raising version of setEdgeFeatureCannyParameters/3.

setEdgeFeatureCannyParameters!(self, threshold1, threshold2, opts)

Raising version of setEdgeFeatureCannyParameters/4.

setEdgeFeatureZeroCrossingParameters!(self)

Raising version of setEdgeFeatureZeroCrossingParameters/1.

setGradientMagnitudeMaxLimit!(self)

Raising version of setGradientMagnitudeMaxLimit/1.

setWeights!(self, weight_non_edge, weight_gradient_direction, weight_gradient_magnitude)

Raising version of setWeights/4.