Evision.SparseOpticalFlow (Evision v0.1.9)

cv

calc(self, prevImg, nextImg, prevPts, nextPts)

Calculates a sparse optical flow.

Positional Arguments

prevImg: Evision.Mat.

First input image.

nextImg: Evision.Mat.

Second input image of the same size and the same type as prevImg.

prevPts: Evision.Mat.

Vector of 2D points for which the flow needs to be found.

Return

nextPts: Evision.Mat.

Output vector of 2D points containing the calculated new positions of input features in the second image.

status: Evision.Mat.

Output status vector. Each element of the vector is set to 1 if the flow for the corresponding features has been found. Otherwise, it is set to 0.

err: Evision.Mat.

Optional output vector that contains error response for each point (inverse confidence).

Python prototype (for reference):

calc(prevImg, nextImg, prevPts, nextPts[, status[, err]]) -> nextPts, status, err

calc(self, prevImg, nextImg, prevPts, nextPts, opts)

Calculates a sparse optical flow.

Positional Arguments

prevImg: Evision.Mat.

First input image.

nextImg: Evision.Mat.

Second input image of the same size and the same type as prevImg.

prevPts: Evision.Mat.

Vector of 2D points for which the flow needs to be found.

Return

nextPts: Evision.Mat.

Output vector of 2D points containing the calculated new positions of input features in the second image.

status: Evision.Mat.

Output status vector. Each element of the vector is set to 1 if the flow for the corresponding features has been found. Otherwise, it is set to 0.

err: Evision.Mat.

Optional output vector that contains error response for each point (inverse confidence).

Python prototype (for reference):

calc(prevImg, nextImg, prevPts, nextPts[, status[, err]]) -> nextPts, status, err

Functions

calc!(self, prevImg, nextImg, prevPts, nextPts)

Raising version of calc/5.

calc!(self, prevImg, nextImg, prevPts, nextPts, opts)

Raising version of calc/6.