Evision.StereoBM (Evision v0.1.9)

cv

compute(self, left, right)

Computes disparity map for the specified stereo pair

Positional Arguments

left: Evision.Mat.

Left 8-bit single-channel image.

right: Evision.Mat.

Right image of the same size and the same type as the left one.

Return

disparity: Evision.Mat.

Output disparity map. It has the same size as the input images. Some algorithms, like StereoBM or StereoSGBM compute 16-bit fixed-point disparity map (where each disparity value has 4 fractional bits), whereas other algorithms output 32-bit floating-point disparity map.

Python prototype (for reference):

compute(left, right[, disparity]) -> disparity

compute(self, left, right, opts)

Computes disparity map for the specified stereo pair

Positional Arguments

left: Evision.Mat.

Left 8-bit single-channel image.

right: Evision.Mat.

Right image of the same size and the same type as the left one.

Return

disparity: Evision.Mat.

Output disparity map. It has the same size as the input images. Some algorithms, like StereoBM or StereoSGBM compute 16-bit fixed-point disparity map (where each disparity value has 4 fractional bits), whereas other algorithms output 32-bit floating-point disparity map.

Python prototype (for reference):

compute(left, right[, disparity]) -> disparity

create()

Creates StereoBM object

Keyword Arguments

numDisparities: int.

the disparity search range. For each pixel algorithm will find the best disparity from 0 (default minimum disparity) to numDisparities. The search range can then be shifted by changing the minimum disparity.

blockSize: int.

the linear size of the blocks compared by the algorithm. The size should be odd (as the block is centered at the current pixel). Larger block size implies smoother, though less accurate disparity map. Smaller block size gives more detailed disparity map, but there is higher chance for algorithm to find a wrong correspondence.

The function create StereoBM object. You can then call StereoBM::compute() to compute disparity for a specific stereo pair.

Python prototype (for reference):

create([, numDisparities[, blockSize]]) -> retval

create(opts)

Creates StereoBM object

Keyword Arguments

numDisparities: int.

the disparity search range. For each pixel algorithm will find the best disparity from 0 (default minimum disparity) to numDisparities. The search range can then be shifted by changing the minimum disparity.

blockSize: int.

the linear size of the blocks compared by the algorithm. The size should be odd (as the block is centered at the current pixel). Larger block size implies smoother, though less accurate disparity map. Smaller block size gives more detailed disparity map, but there is higher chance for algorithm to find a wrong correspondence.

The function create StereoBM object. You can then call StereoBM::compute() to compute disparity for a specific stereo pair.

Python prototype (for reference):

create([, numDisparities[, blockSize]]) -> retval

getBlockSize(self)

Python prototype (for reference):

getBlockSize() -> retval

getDisp12MaxDiff(self)

Python prototype (for reference):

getDisp12MaxDiff() -> retval

getMinDisparity(self)

Python prototype (for reference):

getMinDisparity() -> retval

getNumDisparities(self)

Python prototype (for reference):

getNumDisparities() -> retval

getPreFilterCap(self)

Python prototype (for reference):

getPreFilterCap() -> retval

getPreFilterSize(self)

Python prototype (for reference):

getPreFilterSize() -> retval

getPreFilterType(self)

Python prototype (for reference):

getPreFilterType() -> retval

getROI1(self)

Python prototype (for reference):

getROI1() -> retval

getROI2(self)

Python prototype (for reference):

getROI2() -> retval

getSmallerBlockSize(self)

Python prototype (for reference):

getSmallerBlockSize() -> retval

getSpeckleRange(self)

Python prototype (for reference):

getSpeckleRange() -> retval

getSpeckleWindowSize(self)

Python prototype (for reference):

getSpeckleWindowSize() -> retval

getTextureThreshold(self)

Python prototype (for reference):

getTextureThreshold() -> retval

getUniquenessRatio(self)

Python prototype (for reference):

getUniquenessRatio() -> retval

setBlockSize(self, blockSize)

Positional Arguments

blockSize: int

Python prototype (for reference):

setBlockSize(blockSize) -> None

setDisp12MaxDiff(self, disp12MaxDiff)

Positional Arguments

disp12MaxDiff: int

Python prototype (for reference):

setDisp12MaxDiff(disp12MaxDiff) -> None

setMinDisparity(self, minDisparity)

Positional Arguments

minDisparity: int

Python prototype (for reference):

setMinDisparity(minDisparity) -> None

setNumDisparities(self, numDisparities)

Positional Arguments

numDisparities: int

Python prototype (for reference):

setNumDisparities(numDisparities) -> None

setPreFilterCap(self, preFilterCap)

Positional Arguments

preFilterCap: int

Python prototype (for reference):

setPreFilterCap(preFilterCap) -> None

setPreFilterSize(self, preFilterSize)

Positional Arguments

preFilterSize: int

Python prototype (for reference):

setPreFilterSize(preFilterSize) -> None

setPreFilterType(self, preFilterType)

Positional Arguments

preFilterType: int

Python prototype (for reference):

setPreFilterType(preFilterType) -> None

setROI1(self, roi1)

Positional Arguments

roi1: Rect

Python prototype (for reference):

setROI1(roi1) -> None

setROI2(self, roi2)

Positional Arguments

roi2: Rect

Python prototype (for reference):

setROI2(roi2) -> None

setSmallerBlockSize(self, blockSize)

Positional Arguments

blockSize: int

Python prototype (for reference):

setSmallerBlockSize(blockSize) -> None

setSpeckleRange(self, speckleRange)

Positional Arguments

speckleRange: int

Python prototype (for reference):

setSpeckleRange(speckleRange) -> None

setSpeckleWindowSize(self, speckleWindowSize)

Positional Arguments

speckleWindowSize: int

Python prototype (for reference):

setSpeckleWindowSize(speckleWindowSize) -> None

setTextureThreshold(self, textureThreshold)

Positional Arguments

textureThreshold: int

Python prototype (for reference):

setTextureThreshold(textureThreshold) -> None

setUniquenessRatio(self, uniquenessRatio)

Positional Arguments

uniquenessRatio: int

Python prototype (for reference):

setUniquenessRatio(uniquenessRatio) -> None

Functions

compute!(self, left, right)

Raising version of compute/3.

compute!(self, left, right, opts)

Raising version of compute/4.

create!()

Raising version of create/0.

create!(opts)

Raising version of create/1.

getBlockSize!(self)

Raising version of getBlockSize/1.

getDisp12MaxDiff!(self)

Raising version of getDisp12MaxDiff/1.

getMinDisparity!(self)

Raising version of getMinDisparity/1.

getNumDisparities!(self)

Raising version of getNumDisparities/1.

getPreFilterCap!(self)

Raising version of getPreFilterCap/1.

getPreFilterSize!(self)

Raising version of getPreFilterSize/1.

getPreFilterType!(self)

Raising version of getPreFilterType/1.

getROI1!(self)

Raising version of getROI1/1.

getROI2!(self)

Raising version of getROI2/1.

getSmallerBlockSize!(self)

Raising version of getSmallerBlockSize/1.

getSpeckleRange!(self)

Raising version of getSpeckleRange/1.

getSpeckleWindowSize!(self)

Raising version of getSpeckleWindowSize/1.

getTextureThreshold!(self)

Raising version of getTextureThreshold/1.

getUniquenessRatio!(self)

Raising version of getUniquenessRatio/1.

setBlockSize!(self, blockSize)

Raising version of setBlockSize/2.

setDisp12MaxDiff!(self, disp12MaxDiff)

Raising version of setDisp12MaxDiff/2.

setMinDisparity!(self, minDisparity)

Raising version of setMinDisparity/2.

setNumDisparities!(self, numDisparities)

Raising version of setNumDisparities/2.

setPreFilterCap!(self, preFilterCap)

Raising version of setPreFilterCap/2.

setPreFilterSize!(self, preFilterSize)

Raising version of setPreFilterSize/2.

setPreFilterType!(self, preFilterType)

Raising version of setPreFilterType/2.

setROI1!(self, roi1)

Raising version of setROI1/2.

setROI2!(self, roi2)

Raising version of setROI2/2.

setSmallerBlockSize!(self, blockSize)

Raising version of setSmallerBlockSize/2.

setSpeckleRange!(self, speckleRange)

Raising version of setSpeckleRange/2.

setSpeckleWindowSize!(self, speckleWindowSize)

Raising version of setSpeckleWindowSize/2.

setTextureThreshold!(self, textureThreshold)

Raising version of setTextureThreshold/2.

setUniquenessRatio!(self, uniquenessRatio)

Raising version of setUniquenessRatio/2.