Evision.StereoMatcher (Evision v0.1.9)

cv

compute(self, left, right)

Computes disparity map for the specified stereo pair

Positional Arguments

left: Evision.Mat.

Left 8-bit single-channel image.

right: Evision.Mat.

Right image of the same size and the same type as the left one.

Return

disparity: Evision.Mat.

Output disparity map. It has the same size as the input images. Some algorithms, like StereoBM or StereoSGBM compute 16-bit fixed-point disparity map (where each disparity value has 4 fractional bits), whereas other algorithms output 32-bit floating-point disparity map.

Python prototype (for reference):

compute(left, right[, disparity]) -> disparity

compute(self, left, right, opts)

Computes disparity map for the specified stereo pair

Positional Arguments

left: Evision.Mat.

Left 8-bit single-channel image.

right: Evision.Mat.

Right image of the same size and the same type as the left one.

Return

disparity: Evision.Mat.

Output disparity map. It has the same size as the input images. Some algorithms, like StereoBM or StereoSGBM compute 16-bit fixed-point disparity map (where each disparity value has 4 fractional bits), whereas other algorithms output 32-bit floating-point disparity map.

Python prototype (for reference):

compute(left, right[, disparity]) -> disparity

getBlockSize(self)

Python prototype (for reference):

getBlockSize() -> retval

getDisp12MaxDiff(self)

Python prototype (for reference):

getDisp12MaxDiff() -> retval

getMinDisparity(self)

Python prototype (for reference):

getMinDisparity() -> retval

getNumDisparities(self)

Python prototype (for reference):

getNumDisparities() -> retval

getSpeckleRange(self)

Python prototype (for reference):

getSpeckleRange() -> retval

getSpeckleWindowSize(self)

Python prototype (for reference):

getSpeckleWindowSize() -> retval

setBlockSize(self, blockSize)

Positional Arguments

blockSize: int

Python prototype (for reference):

setBlockSize(blockSize) -> None

setDisp12MaxDiff(self, disp12MaxDiff)

Positional Arguments

disp12MaxDiff: int

Python prototype (for reference):

setDisp12MaxDiff(disp12MaxDiff) -> None

setMinDisparity(self, minDisparity)

Positional Arguments

minDisparity: int

Python prototype (for reference):

setMinDisparity(minDisparity) -> None

setNumDisparities(self, numDisparities)

Positional Arguments

numDisparities: int

Python prototype (for reference):

setNumDisparities(numDisparities) -> None

setSpeckleRange(self, speckleRange)

Positional Arguments

speckleRange: int

Python prototype (for reference):

setSpeckleRange(speckleRange) -> None

setSpeckleWindowSize(self, speckleWindowSize)

Positional Arguments

speckleWindowSize: int

Python prototype (for reference):

setSpeckleWindowSize(speckleWindowSize) -> None

Functions

compute!(self, left, right)

Raising version of compute/3.

compute!(self, left, right, opts)

Raising version of compute/4.

getBlockSize!(self)

Raising version of getBlockSize/1.

getDisp12MaxDiff!(self)

Raising version of getDisp12MaxDiff/1.

getMinDisparity!(self)

Raising version of getMinDisparity/1.

getNumDisparities!(self)

Raising version of getNumDisparities/1.

getSpeckleRange!(self)

Raising version of getSpeckleRange/1.

getSpeckleWindowSize!(self)

Raising version of getSpeckleWindowSize/1.

setBlockSize!(self, blockSize)

Raising version of setBlockSize/2.

setDisp12MaxDiff!(self, disp12MaxDiff)

Raising version of setDisp12MaxDiff/2.

setMinDisparity!(self, minDisparity)

Raising version of setMinDisparity/2.

setNumDisparities!(self, numDisparities)

Raising version of setNumDisparities/2.

setSpeckleRange!(self, speckleRange)

Raising version of setSpeckleRange/2.

setSpeckleWindowSize!(self, speckleWindowSize)

Raising version of setSpeckleWindowSize/2.