Evision.StereoSGBM (Evision v0.1.5-dev)

cv

compute(self, left, right)

compute(self, left, right, opts)

Computes disparity map for the specified stereo pair

left.

right.

disparity. like StereoBM or StereoSGBM compute 16-bit fixed-point disparity map (where each disparity value has 4 fractional bits), whereas other algorithms output 32-bit floating-point disparity map.

disparity : UMat

Python prototype (for reference):

compute(left, right[, disparity]) -> disparity

create()

create(opts)

Creates StereoSGBM object

minDisparity. rectification algorithms can shift images, so this parameter needs to be adjusted accordingly.

numDisparities. zero. In the current implementation, this parameter must be divisible by 16.

blockSize. somewhere in the 3..11 range.

p1.

p2. the smoother the disparity is. P1 is the penalty on the disparity change by plus or minus 1 between neighbor pixels. P2 is the penalty on the disparity change by more than 1 between neighbor pixels. The algorithm requires P2 > P1 . See stereo_match.cpp sample where some reasonably good P1 and P2 values are shown (like 8*number_of_image_channels*blockSize*blockSize and 32*number_of_image_channels*blockSize*blockSize , respectively).

disp12MaxDiff. disparity check. Set it to a non-positive value to disable the check.

preFilterCap. computes x-derivative at each pixel and clips its value by [-preFilterCap, preFilterCap] interval. The result values are passed to the Birchfield-Tomasi pixel cost function.

uniquenessRatio. value should "win" the second best value to consider the found match correct. Normally, a value within the 5-15 range is good enough.

speckleWindowSize. and invalidate. Set it to 0 to disable speckle filtering. Otherwise, set it somewhere in the 50-200 range.

speckleRange. filtering, set the parameter to a positive value, it will be implicitly multiplied by 16. Normally, 1 or 2 is good enough.

mode. algorithm. It will consume O(W*H*numDisparities) bytes, which is large for 640x480 stereo and huge for HD-size pictures. By default, it is set to false . The first constructor initializes StereoSGBM with all the default parameters. So, you only have to set StereoSGBM::numDisparities at minimum. The second constructor enables you to set each parameter to a custom value.

minDisparity : int

numDisparities : int

blockSize : int

p1 : int

p2 : int

disp12MaxDiff : int

preFilterCap : int

uniquenessRatio : int

speckleWindowSize : int

speckleRange : int

mode : int

Python prototype (for reference):

create([, minDisparity[, numDisparities[, blockSize[, P1[, P2[, disp12MaxDiff[, preFilterCap[, uniquenessRatio[, speckleWindowSize[, speckleRange[, mode]]]]]]]]]]]) -> retval

getBlockSize(self)

getDisp12MaxDiff(self)

getMinDisparity(self)

getMode(self)

getNumDisparities(self)

getP1(self)

getP2(self)

getPreFilterCap(self)

getSpeckleRange(self)

getSpeckleWindowSize(self)

getUniquenessRatio(self)

setBlockSize(self, blockSize)

setDisp12MaxDiff(self, disp12MaxDiff)

setMinDisparity(self, minDisparity)

setMode(self, mode)

setNumDisparities(self, numDisparities)

setP1(self, p1)

setP2(self, p2)

setPreFilterCap(self, preFilterCap)

setSpeckleRange(self, speckleRange)

setSpeckleWindowSize(self, speckleWindowSize)

setUniquenessRatio(self, uniquenessRatio)

Functions

compute!(self, left, right)

Raising version of compute/3.

compute!(self, left, right, opts)

Raising version of compute/4.

create!()

Raising version of create/0.

create!(opts)

Raising version of create/1.

getBlockSize!(self)

Raising version of getBlockSize/1.

getDisp12MaxDiff!(self)

Raising version of getDisp12MaxDiff/1.

getMinDisparity!(self)

Raising version of getMinDisparity/1.

getMode!(self)

Raising version of getMode/1.

getNumDisparities!(self)

Raising version of getNumDisparities/1.

getP1!(self)

Raising version of getP1/1.

getP2!(self)

Raising version of getP2/1.

getPreFilterCap!(self)

Raising version of getPreFilterCap/1.

getSpeckleRange!(self)

Raising version of getSpeckleRange/1.

getSpeckleWindowSize!(self)

Raising version of getSpeckleWindowSize/1.

getUniquenessRatio!(self)

Raising version of getUniquenessRatio/1.

setBlockSize!(self, blockSize)

Raising version of setBlockSize/2.

setDisp12MaxDiff!(self, disp12MaxDiff)

Raising version of setDisp12MaxDiff/2.

setMinDisparity!(self, minDisparity)

Raising version of setMinDisparity/2.

setMode!(self, mode)

Raising version of setMode/2.

setNumDisparities!(self, numDisparities)

Raising version of setNumDisparities/2.

setP1!(self, p1)

Raising version of setP1/2.

setP2!(self, p2)

Raising version of setP2/2.

setPreFilterCap!(self, preFilterCap)

Raising version of setPreFilterCap/2.

setSpeckleRange!(self, speckleRange)

Raising version of setSpeckleRange/2.

setSpeckleWindowSize!(self, speckleWindowSize)

Raising version of setSpeckleWindowSize/2.

setUniquenessRatio!(self, uniquenessRatio)

Raising version of setUniquenessRatio/2.