Evision.Stitcher (Evision v0.1.9)

cv

composePanorama(self)

##### Return

pano: Evision.Mat.

Has overloading in C++

Python prototype (for reference):

composePanorama([, pano]) -> retval, pano

composePanorama(self, images)

These functions try to compose the given images (or images stored internally from the other function calls) into the final pano under the assumption that the image transformations were estimated before.

Positional Arguments

images: [Evision.Mat].

Input images.

Return

pano: Evision.Mat.

Final pano.

Note: Use the functions only if you're aware of the stitching pipeline, otherwise use Stitcher::stitch. @return Status code.

Python prototype (for reference):

composePanorama(images[, pano]) -> retval, pano

composePanorama(self, images, opts)

These functions try to compose the given images (or images stored internally from the other function calls) into the final pano under the assumption that the image transformations were estimated before.

Positional Arguments

images: [Evision.Mat].

Input images.

Return

pano: Evision.Mat.

Final pano.

Note: Use the functions only if you're aware of the stitching pipeline, otherwise use Stitcher::stitch. @return Status code.

Python prototype (for reference):

composePanorama(images[, pano]) -> retval, pano

compositingResol(self)

Python prototype (for reference):

compositingResol() -> retval

create()

Creates a Stitcher configured in one of the stitching modes.

Keyword Arguments

mode: Mode.

Scenario for stitcher operation. This is usually determined by source of images to stitch and their transformation. Default parameters will be chosen for operation in given scenario.

@return Stitcher class instance.

Python prototype (for reference):

create([, mode]) -> retval

create(opts)

Creates a Stitcher configured in one of the stitching modes.

Keyword Arguments

mode: Mode.

Scenario for stitcher operation. This is usually determined by source of images to stitch and their transformation. Default parameters will be chosen for operation in given scenario.

@return Stitcher class instance.

Python prototype (for reference):

create([, mode]) -> retval

estimateTransform(self, images)

These functions try to match the given images and to estimate rotations of each camera.

Positional Arguments

images: [Evision.Mat].

Input images.

Keyword Arguments

masks: [Evision.Mat].

Masks for each input image specifying where to look for keypoints (optional).

Note: Use the functions only if you're aware of the stitching pipeline, otherwise use Stitcher::stitch. @return Status code.

Python prototype (for reference):

estimateTransform(images[, masks]) -> retval

estimateTransform(self, images, opts)

These functions try to match the given images and to estimate rotations of each camera.

Positional Arguments

images: [Evision.Mat].

Input images.

Keyword Arguments

masks: [Evision.Mat].

Masks for each input image specifying where to look for keypoints (optional).

Note: Use the functions only if you're aware of the stitching pipeline, otherwise use Stitcher::stitch. @return Status code.

Python prototype (for reference):

estimateTransform(images[, masks]) -> retval

interpolationFlags(self)

Python prototype (for reference):

interpolationFlags() -> retval

panoConfidenceThresh(self)

Python prototype (for reference):

panoConfidenceThresh() -> retval

registrationResol(self)

Python prototype (for reference):

registrationResol() -> retval

seamEstimationResol(self)

Python prototype (for reference):

seamEstimationResol() -> retval

setCompositingResol(self, resol_mpx)

Positional Arguments

resol_mpx: double

Python prototype (for reference):

setCompositingResol(resol_mpx) -> None

setInterpolationFlags(self, interp_flags)

Positional Arguments

interp_flags: InterpolationFlags

Python prototype (for reference):

setInterpolationFlags(interp_flags) -> None

setPanoConfidenceThresh(self, conf_thresh)

Positional Arguments

conf_thresh: double

Python prototype (for reference):

setPanoConfidenceThresh(conf_thresh) -> None

setRegistrationResol(self, resol_mpx)

Positional Arguments

resol_mpx: double

Python prototype (for reference):

setRegistrationResol(resol_mpx) -> None

setSeamEstimationResol(self, resol_mpx)

Positional Arguments

resol_mpx: double

Python prototype (for reference):

setSeamEstimationResol(resol_mpx) -> None

setWaveCorrection(self, flag)

Positional Arguments

flag: bool

Python prototype (for reference):

setWaveCorrection(flag) -> None

stitch(self, images)

Positional Arguments

images: [Evision.Mat]

Return

pano: Evision.Mat.

Has overloading in C++

Python prototype (for reference):

stitch(images[, pano]) -> retval, pano

stitch(self, images, opts)

Variant 1:

These functions try to stitch the given images.

Positional Arguments

images: [Evision.Mat].

Input images.

masks: [Evision.Mat].

Masks for each input image specifying where to look for keypoints (optional).

Return

pano: Evision.Mat.

Final pano.

@return Status code.

Python prototype (for reference):

stitch(images, masks[, pano]) -> retval, pano

Variant 2:

Positional Arguments

images: [Evision.Mat]

Return

pano: Evision.Mat.

Has overloading in C++

Python prototype (for reference):

stitch(images[, pano]) -> retval, pano

stitch(self, images, masks, opts)

These functions try to stitch the given images.

Positional Arguments

images: [Evision.Mat].

Input images.

masks: [Evision.Mat].

Masks for each input image specifying where to look for keypoints (optional).

Return

pano: Evision.Mat.

Final pano.

@return Status code.

Python prototype (for reference):

stitch(images, masks[, pano]) -> retval, pano

waveCorrection(self)

Python prototype (for reference):

waveCorrection() -> retval

workScale(self)

Python prototype (for reference):

workScale() -> retval

Functions

composePanorama!(self)

Raising version of composePanorama/1.

composePanorama!(self, images)

Raising version of composePanorama/2.

composePanorama!(self, images, opts)

Raising version of composePanorama/3.

compositingResol!(self)

Raising version of compositingResol/1.

create!()

Raising version of create/0.

create!(opts)

Raising version of create/1.

estimateTransform!(self, images)

Raising version of estimateTransform/2.

estimateTransform!(self, images, opts)

Raising version of estimateTransform/3.

interpolationFlags!(self)

Raising version of interpolationFlags/1.

panoConfidenceThresh!(self)

Raising version of panoConfidenceThresh/1.

registrationResol!(self)

Raising version of registrationResol/1.

seamEstimationResol!(self)

Raising version of seamEstimationResol/1.

setCompositingResol!(self, resol_mpx)

Raising version of setCompositingResol/2.

setInterpolationFlags!(self, interp_flags)

Raising version of setInterpolationFlags/2.

setPanoConfidenceThresh!(self, conf_thresh)

Raising version of setPanoConfidenceThresh/2.

setRegistrationResol!(self, resol_mpx)

Raising version of setRegistrationResol/2.

setSeamEstimationResol!(self, resol_mpx)

Raising version of setSeamEstimationResol/2.

setWaveCorrection!(self, flag)

Raising version of setWaveCorrection/2.

stitch!(self, images)

Raising version of stitch/2.

stitch!(self, images, masks)

Raising version of stitch/3.

stitch!(self, images, masks, opts)

Raising version of stitch/4.

waveCorrection!(self)

Raising version of waveCorrection/1.

workScale!(self)

Raising version of workScale/1.