Evision.VideoIORegistry (Evision v0.1.7-dev)

cv.videoio_registry

getBackendName(api)

Returns backend API name or "UnknownVideoAPI(xxx)"

Python prototype (for reference):

getBackendName(api) -> retval

getBackends()

Returns list of all available backends

Python prototype (for reference):

getBackends() -> retval

getCameraBackendPluginVersion(api)

Returns description and ABI/API version of videoio plugin's camera interface

Python prototype (for reference):

getCameraBackendPluginVersion(api) -> retval, version_ABI, version_API

getCameraBackends()

Returns list of available backends which works via cv::VideoCapture(int index)

Python prototype (for reference):

getCameraBackends() -> retval

getStreamBackendPluginVersion(api)

Returns description and ABI/API version of videoio plugin's stream capture interface

Python prototype (for reference):

getStreamBackendPluginVersion(api) -> retval, version_ABI, version_API

getStreamBackends()

Returns list of available backends which works via cv::VideoCapture(filename)

Python prototype (for reference):

getStreamBackends() -> retval

getWriterBackendPluginVersion(api)

Returns description and ABI/API version of videoio plugin's writer interface

Python prototype (for reference):

getWriterBackendPluginVersion(api) -> retval, version_ABI, version_API

getWriterBackends()

Returns list of available backends which works via cv::VideoWriter()

Python prototype (for reference):

getWriterBackends() -> retval

hasBackend(api)

Returns true if backend is available

Python prototype (for reference):

hasBackend(api) -> retval

isBackendBuiltIn(api)

Returns true if backend is built in (false if backend is used as plugin)

Python prototype (for reference):

isBackendBuiltIn(api) -> retval

Functions

getBackendName!(api)

Raising version of getBackendName/1.

getBackends!()

Raising version of getBackends/0.

getCameraBackendPluginVersion!(api)

Raising version of getCameraBackendPluginVersion/1.

getCameraBackends!()

Raising version of getCameraBackends/0.

getStreamBackendPluginVersion!(api)

Raising version of getStreamBackendPluginVersion/1.

getStreamBackends!()

Raising version of getStreamBackends/0.

getWriterBackendPluginVersion!(api)

Raising version of getWriterBackendPluginVersion/1.

getWriterBackends!()

Raising version of getWriterBackends/0.

hasBackend!(api)

Raising version of hasBackend/1.

isBackendBuiltIn!(api)

Raising version of isBackendBuiltIn/1.