OpenCV.AsyncArray (Evision v0.1.0-dev)

cv

asyncarray()

get(self)

get(self, timeoutNs)

get(self, timeoutNs, opts)

Retrieving the result with timeout

dst.

timeoutNs. @returns true if result is ready, false if the timeout has expired Note: Result or stored exception can be fetched only once.

dst : UMat

Python prototype (for reference):

get(timeoutNs[, dst]) -> retval, dst

release(self)

valid(self)

wait_for(self, timeoutNs)