OpenCV.CUDA (Evision v0.1.0-dev)

cv.cuda

createContinuous(rows, cols, type)

createContinuous(rows, cols, type, opts)

Creates a continuous matrix.

rows.

cols.

type.

arr. \f$\texttt{rows} \times \texttt{cols}\f$ ). Matrix is called continuous if its elements are stored continuously, that is, without gaps at the end of each row.

arr : UMat

Python prototype (for reference):

createContinuous(rows, cols, type[, arr]) -> arr

ensureSizeIsEnough(rows, cols, type)

ensureSizeIsEnough(rows, cols, type, opts)

Ensures that the size of a matrix is big enough and the matrix has a proper type.

rows.

cols.

type.

arr. The function does not reallocate memory if the matrix has proper attributes already.

arr : UMat

Python prototype (for reference):

ensureSizeIsEnough(rows, cols, type[, arr]) -> arr

getCudaEnabledDeviceCount()

getDevice()

printCudaDeviceInfo(device)

printShortCudaDeviceInfo(device)

registerPageLocked(m)

resetDevice()

setBufferPoolConfig(deviceId, stackSize, stackCount)

setBufferPoolUsage(on)

setDevice(device)

unregisterPageLocked(m)