OpenCV.CUDA.GpuMat (Evision v0.1.0-dev)

cv.cuda

adjustROI(self, dtop, dbottom, dleft, dright)

assignTo(self, m)

assignTo(self, m, opts)

type : int

Python prototype (for reference):

assignTo(m[, type]) -> None

channels(self)

clone(self)

col(self, x)

colRange(self, r)

colRange(self, startcol, endcol)

convertTo(self, rtype, stream)

convertTo(self, rtype, alpha, stream)

convertTo(self, rtype, alpha, beta, stream)

convertTo(self, rtype, alpha, beta, stream, opts)

dst : UMat

Python prototype (for reference):

convertTo(rtype, alpha, beta, stream[, dst]) -> dst

copyTo(self, mask, stream)

copyTo(self, mask, stream, opts)

dst : UMat

Python prototype (for reference):

copyTo(mask, stream[, dst]) -> dst

create(self, size, type)

create(self, rows, cols, type)

cudaPtr(self)

defaultAllocator()

depth(self)

download(self, stream)

download(self, stream, opts)

Performs data download from GpuMat (Non-Blocking call) This function copies data from device memory to host memory. As being a non-blocking call, this function may return even if the copy operation is not finished. The copy operation may be overlapped with operations in other non-default streams if \p stream is not the default stream and \p dst is HostMem allocated with HostMem::PAGE_LOCKED option.

dst : UMat

Python prototype (for reference):

download(stream[, dst]) -> dst

elemSize(self)

elemSize1(self)

empty(self)

gpuMat(arr)

gpuMat(size, type)

gpuMat(size, type, s)

gpuMat(rows, cols, type, s)

gpuMat(rows, cols, type, s, opts)

allocator : GpuMat_Allocator*

Python prototype (for reference):

GpuMat(rows, cols, type, s[, allocator]) -> <cuda_GpuMat object>

isContinuous(self)

locateROI(self, wholeSize, ofs)

reshape(self, cn)

reshape(self, cn, opts)

rows : int

Python prototype (for reference):

reshape(cn[, rows]) -> retval

row(self, y)

rowRange(self, r)

rowRange(self, startrow, endrow)

setDefaultAllocator(allocator)

setTo(self, s, mask, stream)

size(self)

step1(self)

swap(self, mat)

type(self)

updateContinuityFlag(self)

upload(self, arr, stream)

Functions

get_step(self)