OpenCV.CascadeClassifier (Evision v0.1.0-dev)

cv

cascadeclassifier()

cascadeclassifier(filename)

convert(oldcascade, newcascade)

detectmultiscale(self, image)

detectmultiscale(self, image, opts)

Detects objects of different sizes in the input image. The detected objects are returned as a list of rectangles.

image.

objects. rectangles may be partially outside the original image.

scaleFactor.

minNeighbors. to retain it.

flags. cvHaarDetectObjects. It is not used for a new cascade.

minSize.

maxSize. The function is parallelized with the TBB library. Note:

(Python) A face detection example using cascade classifiers can be found at opencv_source_code/samples/python/facedetect.py

scaleFactor : double

minNeighbors : int

flags : int

minSize : Size

maxSize : Size

Python prototype (for reference):

detectMultiScale(image[, scaleFactor[, minNeighbors[, flags[, minSize[, maxSize]]]]]) -> objects

detectmultiscale2(self, image)

detectmultiscale2(self, image, opts)

Has overloading in C++

image.

objects. rectangles may be partially outside the original image.

numDetections. of detections is the number of neighboring positively classified rectangles that were joined together to form the object.

scaleFactor.

minNeighbors. to retain it.

flags. cvHaarDetectObjects. It is not used for a new cascade.

minSize.

maxSize.

scaleFactor : double

minNeighbors : int

flags : int

minSize : Size

maxSize : Size

Python prototype (for reference):

detectMultiScale2(image[, scaleFactor[, minNeighbors[, flags[, minSize[, maxSize]]]]]) -> objects, numDetections

detectmultiscale3(self, image)

detectmultiscale3(self, image, opts)

Has overloading in C++

This function allows you to retrieve the final stage decision certainty of classification. For this, one needs to set outputRejectLevels on true and provide the rejectLevels and levelWeights parameter. For each resulting detection, levelWeights will then contain the certainty of classification at the final stage. This value can then be used to separate strong from weaker classifications. A code sample on how to use it efficiently can be found below:

Mat img;
vector<double> weights;
vector<int> levels;
vector<Rect> detections;
CascadeClassifier model("/path/to/your/model.xml");
model.detectMultiScale(img, detections, levels, weights, 1.1, 3, 0, Size(), Size(), true);
cerr << "Detection " << detections[0] << " with weight " << weights[0] << endl;

scaleFactor : double

minNeighbors : int

flags : int

minSize : Size

maxSize : Size

outputRejectLevels : bool

Python prototype (for reference):

detectMultiScale3(image[, scaleFactor[, minNeighbors[, flags[, minSize[, maxSize[, outputRejectLevels]]]]]]) -> objects, rejectLevels, levelWeights

empty(self)

getfeaturetype(self)

getoriginalwindowsize(self)

isoldformatcascade(self)

load(self, filename)

read(self, node)