OpenCV.DNN.KeypointsModel (Evision v0.1.0-dev)

Functions

estimate(self, frame)

estimate(self, frame, opts)

Given the @p input frame, create input blob, run net

**.

thresh. @returns a vector holding the x and y coordinates of each detected keypoint

thresh : float

Python prototype (for reference):

estimate(frame[, thresh]) -> retval

keypointsModel(model)

keypointsModel(model, opts)

Create keypoints model from network represented in one of the supported formats. An order of @p model and @p config arguments does not matter.

model.

config.

config : String

Python prototype (for reference):

KeypointsModel(model[, config]) -> <dnn_KeypointsModel object>