OpenCV.DNN.Net (Evision v0.1.0-dev)

Functions

connect(self, outPin, inpPin)

dump(self)

dumpToFile(self, path)

empty(self)

enableFusion(self, fusion)

forward(self, opt \\ [])

Runs forward pass to compute outputs of layers listed in @p outBlobNames.

outputBlobs.

outBlobNames.

outputBlobs : vector_UMat

Python prototype (for reference):

forward(outBlobNames[, outputBlobs]) -> outputBlobs

forwardAndRetrieve(self, outBlobNames)

forwardAsync(self)

getFLOPS(self, netInputShapes)

getFLOPS(self, layerId, netInputShapes)

getInputDetails(self)

getLayer(self, layerId)

getLayerId(self, layer)

getLayerNames(self)

getLayerShapes(self, netInputShapes, layerId)

getLayerTypes(self)

getLayersCount(self, layerType)

getLayersShapes(self, netInputShapes)

getMemoryConsumption(self, netInputShape)

getMemoryConsumption(self, layerId, netInputShapes)

getOutputDetails(self)

getParam(self, layer)

getParam(self, layer, opts)

Returns parameter blob of the layer.

layer.

numParam. @see Layer::blobs

numParam : int

Python prototype (for reference):

getParam(layer[, numParam]) -> retval

getPerfProfile(self)

getUnconnectedOutLayers(self)

getUnconnectedOutLayersNames(self)

net()

quantize(self, calibData, inputsDtype, outputsDtype)

readFromModelOptimizer(bufferModelConfig, bufferWeights)

setHalideScheduler(self, scheduler)

setInput(self, blob)

setInput(self, blob, opts)

Sets the new input value for the network

blob.

name.

scalefactor.

mean. @see connect(String, String) to know format of the descriptor. If scale or mean values are specified, a final input blob is computed as: \f[input(n,c,h,w) = scalefactor \times (blob(n,c,h,w) - mean_c)\f]

name : String

scalefactor : double

mean : Scalar

Python prototype (for reference):

setInput(blob[, name[, scalefactor[, mean]]]) -> None

setInputShape(self, inputName, shape)

setInputsNames(self, inputBlobNames)

setParam(self, layer, numParam, blob)

setPreferableBackend(self, backendId)

setPreferableTarget(self, targetId)