OpenCV.DnnLayer (Evision v0.1.0-dev)

Functions

blobs(self)

finalize(self, inputs)

finalize(self, inputs, opts)

Computes and sets internal parameters according to inputs, outputs and blobs.

**.

outputs. If this method is called after network has allocated all memory for input and output blobs and before inferencing.

outputs : vector_UMat

Python prototype (for reference):

finalize(inputs[, outputs]) -> outputs

name(self)

outputnametoindex(self, outputName)

preferabletarget(self)

run(self, inputs, internals)

run(self, inputs, internals, opts)

Allocates layer and computes output. @deprecated This method will be removed in the future release.

outputs : vector_Mat

Python prototype (for reference):

run(inputs, internals[, outputs]) -> outputs, internals

set_blobs(self, opts)

type(self)