OpenCV.DnnModel (Evision v0.1.0-dev)

Functions

dnn_model(model)

dnn_model(model, opts)

Create model from deep learning network represented in one of the supported formats. An order of @p model and @p config arguments does not matter.

model.

config.

config : String

Python prototype (for reference):

Model(model[, config]) -> <dnn_Model object>

predict(self, frame)

predict(self, frame, opts)

Given the @p input frame, create input blob, run net and return the output @p blobs.

**.

outs.

outs : vector_UMat

Python prototype (for reference):

predict(frame[, outs]) -> outs

setinputcrop(self, crop)

setinputmean(self, mean)

setinputparams(self)

setinputscale(self, scale)

setinputsize(self, size)

setinputsize(self, width, height)

setinputswaprb(self, swapRB)

setpreferablebackend(self, backendId)

setpreferabletarget(self, targetId)