OpenCV.FaceRecognizerSF (Evision v0.1.0-dev)

cv

alignCrop(self, src_img, face_box)

alignCrop(self, src_img, face_box, opts)

Aligning image to put face on the standard position

src_img.

face_box.

aligned_img.

aligned_img : UMat

Python prototype (for reference):

alignCrop(src_img, face_box[, aligned_img]) -> aligned_img

create(model, config)

create(model, config, opts)

Creates an instance of this class with given parameters

model.

config.

backend_id.

target_id.

backend_id : int

target_id : int

Python prototype (for reference):

create(model, config[, backend_id[, target_id]]) -> retval

feature(self, aligned_img)

feature(self, aligned_img, opts)

Extracting face feature from aligned image

aligned_img.

face_feature.

face_feature : UMat

Python prototype (for reference):

feature(aligned_img[, face_feature]) -> face_feature

match(self, face_feature1, face_feature2)

match(self, face_feature1, face_feature2, opts)

Calculating the distance between two face features

face_feature1.

face_feature2.

dis_type.

dis_type : int

Python prototype (for reference):

match(_face_feature1, _face_feature2[, dis_type]) -> retval