OpenCV.Gui (Evision v0.1.0-dev)

OpenCV High-level Graphical User Interface

cv.highgui

destroy_all_windows()

close all windows

destroy_window(winname)

close a named window

imshow(winname, mat)

Show a mat in a named window

winname. The name of the window.

mat. The image.

Example

{:ok, mat} = OpenCV.imread("example.jpg")
OpenCV.imshow("OpenCV", mat)
# the following line may be necessary on your system
# will try to improve this later
OpenCV.waitkey(0)

waitkey(delay)

wait for user keyboard event for a delay amount of time (ms)