OpenCV.ML.ANNMLP (Evision v0.1.0-dev)

cv.ml

calcError(self, data, test)

calcError(self, data, test, opts)

Computes error on the training or test dataset

data.

test. computed over the training subset of the data. Please note that if you loaded a completely different dataset to evaluate already trained classifier, you will probably want not to set the test subset at all with TrainData::setTrainTestSplitRatio and specify test=false, so that the error is computed for the whole new set. Yes, this sounds a bit confusing.

resp. The method uses StatModel::predict to compute the error. For regression models the error is computed as RMS, for classifiers - as a percent of missclassified samples (0%-100%).

resp : UMat

Python prototype (for reference):

calcError(data, test[, resp]) -> retval, resp

create()

empty(self)

getAnnealCoolingRatio(self)

getAnnealFinalT(self)

getAnnealInitialT(self)

getAnnealItePerStep(self)

getBackpropMomentumScale(self)

getBackpropWeightScale(self)

getLayerSizes(self)

getRpropDW0(self)

getRpropDWMax(self)

getRpropDWMin(self)

getRpropDWMinus(self)

getRpropDWPlus(self)

getTermCriteria(self)

getTrainMethod(self)

getVarCount(self)

getWeights(self, layerIdx)

isClassifier(self)

isTrained(self)

load(filepath)

predict(self, samples)

predict(self, samples, opts)

Predicts response(s) for the provided sample(s)

samples.

results.

flags.

results : UMat

flags : int

Python prototype (for reference):

predict(samples[, results[, flags]]) -> retval, results

setActivationFunction(self, type)

setActivationFunction(self, type, opts)

Initialize the activation function for each neuron. Currently the default and the only fully supported activation function is ANN_MLP::SIGMOID_SYM.

type.

param1.

param2.

param1 : double

param2 : double

Python prototype (for reference):

setActivationFunction(type[, param1[, param2]]) -> None

setAnnealCoolingRatio(self, val)

setAnnealFinalT(self, val)

setAnnealInitialT(self, val)

setAnnealItePerStep(self, val)

setBackpropMomentumScale(self, val)

setBackpropWeightScale(self, val)

setLayerSizes(self, layer_sizes)

setRpropDW0(self, val)

setRpropDWMax(self, val)

setRpropDWMin(self, val)

setRpropDWMinus(self, val)

setRpropDWPlus(self, val)

setTermCriteria(self, val)

setTrainMethod(self, method)

setTrainMethod(self, method, opts)

Sets training method and common parameters.

method.

param1.

param2.

param1 : double

param2 : double

Python prototype (for reference):

setTrainMethod(method[, param1[, param2]]) -> None

train(self, trainData)

train(self, trainData, opts)

train(self, samples, layout, responses)