OpenCV.ML.NormalBayesClassifier (Evision v0.1.0-dev)

cv.ml

calcError(self, data, test)

calcError(self, data, test, opts)

Computes error on the training or test dataset

data.

test. computed over the training subset of the data. Please note that if you loaded a completely different dataset to evaluate already trained classifier, you will probably want not to set the test subset at all with TrainData::setTrainTestSplitRatio and specify test=false, so that the error is computed for the whole new set. Yes, this sounds a bit confusing.

resp. The method uses StatModel::predict to compute the error. For regression models the error is computed as RMS, for classifiers - as a percent of missclassified samples (0%-100%).

resp : UMat

Python prototype (for reference):

calcError(data, test[, resp]) -> retval, resp

create()

empty(self)

getVarCount(self)

isClassifier(self)

isTrained(self)

load(filepath)

load(filepath, opts)

Loads and creates a serialized NormalBayesClassifier from a file Use NormalBayesClassifier::save to serialize and store an NormalBayesClassifier to disk. Load the NormalBayesClassifier from this file again, by calling this function with the path to the file. Optionally specify the node for the file containing the classifier

filepath.

nodeName.

nodeName : String

Python prototype (for reference):

load(filepath[, nodeName]) -> retval

predict(self, samples)

predict(self, samples, opts)

Predicts response(s) for the provided sample(s)

samples.

results.

flags.

results : UMat

flags : int

Python prototype (for reference):

predict(samples[, results[, flags]]) -> retval, results

predictProb(self, inputs)

predictProb(self, inputs, opts)

Predicts the response for sample(s). The method estimates the most probable classes for input vectors. Input vectors (one or more) are stored as rows of the matrix inputs. In case of multiple input vectors, there should be one output vector outputs. The predicted class for a single input vector is returned by the method. The vector outputProbs contains the output probabilities corresponding to each element of result.

outputs : UMat

outputProbs : UMat

flags : int

Python prototype (for reference):

predictProb(inputs[, outputs[, outputProbs[, flags]]]) -> retval, outputs, outputProbs

train(self, trainData)

train(self, trainData, opts)

train(self, samples, layout, responses)