OpenCV.ML.RTrees (Evision v0.1.0-dev)

cv.ml

create()

getActiveVarCount(self)

getCalculateVarImportance(self)

getOOBError(self)

getTermCriteria(self)

getVarImportance(self)

getVotes(self, samples, flags)

getVotes(self, samples, flags, opts)

Returns the result of each individual tree in the forest. In case the model is a regression problem, the method will return each of the trees' results for each of the sample cases. If the model is a classifier, it will return a Mat with samples + 1 rows, where the first row gives the class number and the following rows return the votes each class had for each sample.

samples.

results.

flags.

results : UMat

Python prototype (for reference):

getVotes(samples, flags[, results]) -> results

load(filepath)

load(filepath, opts)

Loads and creates a serialized RTree from a file Use RTree::save to serialize and store an RTree to disk. Load the RTree from this file again, by calling this function with the path to the file. Optionally specify the node for the file containing the classifier

filepath.

nodeName.

nodeName : String

Python prototype (for reference):

load(filepath[, nodeName]) -> retval

setActiveVarCount(self, val)

setCalculateVarImportance(self, val)

setTermCriteria(self, val)