OpenCV.ML.SVM (Evision v0.1.0-dev)

cv.ml

calcError(self, data, test)

calcError(self, data, test, opts)

Computes error on the training or test dataset

data.

test. computed over the training subset of the data. Please note that if you loaded a completely different dataset to evaluate already trained classifier, you will probably want not to set the test subset at all with TrainData::setTrainTestSplitRatio and specify test=false, so that the error is computed for the whole new set. Yes, this sounds a bit confusing.

resp. The method uses StatModel::predict to compute the error. For regression models the error is computed as RMS, for classifiers - as a percent of missclassified samples (0%-100%).

resp : UMat

Python prototype (for reference):

calcError(data, test[, resp]) -> retval, resp

create()

empty(self)

getC(self)

getClassWeights(self)

getCoef0(self)

getDecisionFunction(self, i)

getDecisionFunction(self, i, opts)

Retrieves the decision function

i. 2-class classification, then there will be just one decision function and the index should always be 0. Otherwise, in the case of N-class classification, there will be \f$N(N-1)/2\f$ decision functions.

alpha. In the case of linear %SVM all the alpha's will be 1's.

svidx. support vectors (which can be retrieved by SVM::getSupportVectors). In the case of linear %SVM each decision function consists of a single "compressed" support vector. The method returns rho parameter of the decision function, a scalar subtracted from the weighted sum of kernel responses.

alpha : UMat

svidx : UMat

Python prototype (for reference):

getDecisionFunction(i[, alpha[, svidx]]) -> retval, alpha, svidx

getDefaultGridPtr(param_id)

getDegree(self)

getGamma(self)

getKernelType(self)

getNu(self)

getP(self)

getSupportVectors(self)

getTermCriteria(self)

getType(self)

getUncompressedSupportVectors(self)

getVarCount(self)

isClassifier(self)

isTrained(self)

load(filepath)

predict(self, samples)

predict(self, samples, opts)

Predicts response(s) for the provided sample(s)

samples.

results.

flags.

results : UMat

flags : int

Python prototype (for reference):

predict(samples[, results[, flags]]) -> retval, results

setC(self, val)

setClassWeights(self, val)

setCoef0(self, val)

setDegree(self, val)

setGamma(self, val)

setKernel(self, kernelType)

setNu(self, val)

setP(self, val)

setTermCriteria(self, val)

setType(self, val)

train(self, trainData)

train(self, trainData, opts)

train(self, samples, layout, responses)

trainAuto(self, samples, layout, responses)

trainAuto(self, samples, layout, responses, opts)

Trains an %SVM with optimal parameters

samples.

layout.

responses.

kFold. subset is used to test the model, the others form the train set. So, the %SVM algorithm is

cgrid.

gammaGrid.

pGrid.

nuGrid.

coeffGrid.

degreeGrid.

balanced. balanced cross-validation subsets that is proportions between classes in subsets are close to such proportion in the whole train dataset. The method trains the %SVM model automatically by choosing the optimal parameters C, gamma, p, nu, coef0, degree. Parameters are considered optimal when the cross-validation estimate of the test set error is minimal. This function only makes use of SVM::getDefaultGrid for parameter optimization and thus only offers rudimentary parameter options. This function works for the classification (SVM::C_SVC or SVM::NU_SVC) as well as for the regression (SVM::EPS_SVR or SVM::NU_SVR). If it is SVM::ONE_CLASS, no optimization is made and the usual %SVM with parameters specified in params is executed.

kFold : int

cgrid : Ptr<ParamGrid>

gammaGrid : Ptr<ParamGrid>

pGrid : Ptr<ParamGrid>

nuGrid : Ptr<ParamGrid>

coeffGrid : Ptr<ParamGrid>

degreeGrid : Ptr<ParamGrid>

balanced : bool

Python prototype (for reference):

trainAuto(samples, layout, responses[, kFold[, Cgrid[, gammaGrid[, pGrid[, nuGrid[, coeffGrid[, degreeGrid[, balanced]]]]]]]]) -> retval