OpenCV.ML.TrainData (Evision v0.1.0-dev)

cv.ml

create(samples, layout, responses)

create(samples, layout, responses, opts)

Creates training data from in-memory arrays.

samples.

layout.

responses. single row or as a single column. The matrix should have type CV_32F or CV_32S (in the former case the responses are considered as ordered by default; in the latter case - as categorical)

varIdx. (CV_32S) containing 0-based variable indices or byte vector (CV_8U) containing a mask of active variables.

sampleIdx. vector (CV_32S) containing 0-based sample indices or byte vector (CV_8U) containing a mask of training samples.

sampleWeights.

varType. <number_of_variables_in_responses>`, containing types of each input and output variable. See ml::VariableTypes.

varIdx : UMat

sampleIdx : UMat

sampleWeights : UMat

varType : UMat

Python prototype (for reference):

create(samples, layout, responses[, varIdx[, sampleIdx[, sampleWeights[, varType]]]]) -> retval

getCatCount(self, vi)

getCatMap(self)

getCatOfs(self)

getClassLabels(self)

getDefaultSubstValues(self)

getLayout(self)

getMissing(self)

getNAllVars(self)

getNSamples(self)

getNTestSamples(self)

getNTrainSamples(self)

getNVars(self)

getNames(self, names)

getNormCatResponses(self)

getResponseType(self)

getResponses(self)

getSample(self, varIdx, sidx, buf)

getSampleWeights(self)

getSamples(self)

getSubMatrix(matrix, idx, layout)

getSubVector(vec, idx)

getTestNormCatResponses(self)

getTestResponses(self)

getTestSampleIdx(self)

getTestSampleWeights(self)

getTestSamples(self)

getTrainNormCatResponses(self)

getTrainResponses(self)

getTrainSampleIdx(self)

getTrainSampleWeights(self)

getTrainSamples(self)

getValues(self, vi, sidx, values)

getVarIdx(self)

getVarSymbolFlags(self)

getVarType(self)

setTrainTestSplit(self, count)

setTrainTestSplit(self, count, opts)

Splits the training data into the training and test parts @sa TrainData::setTrainTestSplitRatio

shuffle : bool

Python prototype (for reference):

setTrainTestSplit(count[, shuffle]) -> None

setTrainTestSplitRatio(self, ratio)

setTrainTestSplitRatio(self, ratio, opts)

Splits the training data into the training and test parts The function selects a subset of specified relative size and then returns it as the training set. If the function is not called, all the data is used for training. Please, note that for each of TrainData::getTrain* there is corresponding TrainData::getTest*, so that the test subset can be retrieved and processed as well. @sa TrainData::setTrainTestSplit

shuffle : bool

Python prototype (for reference):

setTrainTestSplitRatio(ratio[, shuffle]) -> None

shuffleTrainTest(self)