OpenCV.MlEM (Evision v0.1.0-dev)

cv.ml

create()

getclustersnumber(self)

getcovariancematrixtype(self)

getcovs(self)

getmeans(self)

gettermcriteria(self)

getweights(self)

load(filepath)

load(filepath, opts)

Loads and creates a serialized EM from a file Use EM::save to serialize and store an EM to disk. Load the EM from this file again, by calling this function with the path to the file. Optionally specify the node for the file containing the classifier

filepath.

nodeName.

nodeName : String

Python prototype (for reference):

load(filepath[, nodeName]) -> retval

predict(self, samples)

predict(self, samples, opts)

Returns posterior probabilities for the provided samples

samples.

results. posterior probabilities for each sample from the input

flags.

results : UMat

flags : int

Python prototype (for reference):

predict(samples[, results[, flags]]) -> retval, results

predict2(self, sample)

predict2(self, sample, opts)

Returns a likelihood logarithm value and an index of the most probable mixture component for the given sample.

sample. \f$1 \times dims\f$ or \f$dims \times 1\f$ size.

probs. given the sample. It has \f$1 \times nclusters\f$ size and CV_64FC1 type. The method returns a two-element double vector. Zero element is a likelihood logarithm value for the sample. First element is an index of the most probable mixture component for the given sample.

probs : UMat

Python prototype (for reference):

predict2(sample[, probs]) -> retval, probs

setclustersnumber(self, val)

setcovariancematrixtype(self, val)

settermcriteria(self, val)

traine(self, samples, means0)

traine(self, samples, means0, opts)

Estimate the Gaussian mixture parameters from a samples set. This variation starts with Expectation step. You need to provide initial means \f$a_k\f$ of mixture components. Optionally you can pass initial weights \f$\pi_k\f$ and covariance matrices \f$S_k\f$ of mixture components.

samples. one-channel matrix, each row of which is a sample. If the matrix does not have CV_64F type it will be converted to the inner matrix of such type for the further computing.

means0. \f$nclusters \times dims\f$ size. If the matrix does not have CV_64F type it will be converted to the inner matrix of such type for the further computing.

covs0. covariance matrices is a one-channel matrix of \f$dims \times dims\f$ size. If the matrices do not have CV_64F type they will be converted to the inner matrices of such type for the further computing.

weights0. floating-point matrix with \f$1 \times nclusters\f$ or \f$nclusters \times 1\f$ size.

logLikelihoods. each sample. It has \f$nsamples \times 1\f$ size and CV_64FC1 type.

labels. \f$\texttt{labels}_i=\texttt{arg max}_k(p_{i,k}), i=1..N\f$ (indices of the most probable mixture component for each sample). It has \f$nsamples \times 1\f$ size and CV_32SC1 type.

probs. mixture component given the each sample. It has \f$nsamples \times nclusters\f$ size and CV_64FC1 type.

covs0 : UMat

weights0 : UMat

logLikelihoods : UMat

labels : UMat

probs : UMat

Python prototype (for reference):

trainE(samples, means0[, covs0[, weights0[, logLikelihoods[, labels[, probs]]]]]) -> retval, logLikelihoods, labels, probs

trainem(self, samples)

trainem(self, samples, opts)

Estimate the Gaussian mixture parameters from a samples set. This variation starts with Expectation step. Initial values of the model parameters will be estimated by the k-means algorithm. Unlike many of the ML models, %EM is an unsupervised learning algorithm and it does not take responses (class labels or function values) as input. Instead, it computes the Maximum Likelihood Estimate of the Gaussian mixture parameters from an input sample set, stores all the parameters inside the structure: \f$p_{i,k}\f$ in probs, \f$a_k\f$ in means , \f$S_k\f$ in covs[k], \f$\pi_k\f$ in weights , and optionally computes the output "class label" for each sample: \f$\texttt{labels}_i=\texttt{arg max}_k(p_{i,k}), i=1..N\f$ (indices of the most probable mixture component for each sample). The trained model can be used further for prediction, just like any other classifier. The trained model is similar to the NormalBayesClassifier.

samples. one-channel matrix, each row of which is a sample. If the matrix does not have CV_64F type it will be converted to the inner matrix of such type for the further computing.

logLikelihoods. each sample. It has \f$nsamples \times 1\f$ size and CV_64FC1 type.

labels. \f$\texttt{labels}_i=\texttt{arg max}_k(p_{i,k}), i=1..N\f$ (indices of the most probable mixture component for each sample). It has \f$nsamples \times 1\f$ size and CV_32SC1 type.

probs. mixture component given the each sample. It has \f$nsamples \times nclusters\f$ size and CV_64FC1 type.

logLikelihoods : UMat

labels : UMat

probs : UMat

Python prototype (for reference):

trainEM(samples[, logLikelihoods[, labels[, probs]]]) -> retval, logLikelihoods, labels, probs

trainm(self, samples, probs0)

trainm(self, samples, probs0, opts)

Estimate the Gaussian mixture parameters from a samples set. This variation starts with Maximization step. You need to provide initial probabilities \f$p_{i,k}\f$ to use this option.

samples. one-channel matrix, each row of which is a sample. If the matrix does not have CV_64F type it will be converted to the inner matrix of such type for the further computing.

probs0.

logLikelihoods. each sample. It has \f$nsamples \times 1\f$ size and CV_64FC1 type.

labels. \f$\texttt{labels}_i=\texttt{arg max}_k(p_{i,k}), i=1..N\f$ (indices of the most probable mixture component for each sample). It has \f$nsamples \times 1\f$ size and CV_32SC1 type.

probs. mixture component given the each sample. It has \f$nsamples \times nclusters\f$ size and CV_64FC1 type.

logLikelihoods : UMat

labels : UMat

probs : UMat

Python prototype (for reference):

trainM(samples, probs0[, logLikelihoods[, labels[, probs]]]) -> retval, logLikelihoods, labels, probs