OpenCV.MlLogisticRegression (Evision v0.1.0-dev)

cv.ml

create()

get_learnt_thetas(self)

getiterations(self)

getlearningrate(self)

getminibatchsize(self)

getregularization(self)

gettermcriteria(self)

gettrainmethod(self)

load(filepath)

load(filepath, opts)

Loads and creates a serialized LogisticRegression from a file Use LogisticRegression::save to serialize and store an LogisticRegression to disk. Load the LogisticRegression from this file again, by calling this function with the path to the file. Optionally specify the node for the file containing the classifier

filepath.

nodeName.

nodeName : String

Python prototype (for reference):

load(filepath[, nodeName]) -> retval

predict(self, samples)

predict(self, samples, opts)

Predicts responses for input samples and returns a float type.

samples. contains variables (features) of one object being classified. Should have data type CV_32F.

results.

flags.

results : UMat

flags : int

Python prototype (for reference):

predict(samples[, results[, flags]]) -> retval, results

setiterations(self, val)

setlearningrate(self, val)

setminibatchsize(self, val)

setregularization(self, val)

settermcriteria(self, val)

settrainmethod(self, val)