OpenCV.MlSVM (Evision v0.1.0-dev)

cv.ml

create()

getc(self)

getclassweights(self)

getcoef0(self)

getdecisionfunction(self, i)

getdecisionfunction(self, i, opts)

Retrieves the decision function

i. 2-class classification, then there will be just one decision function and the index should always be 0. Otherwise, in the case of N-class classification, there will be \f$N(N-1)/2\f$ decision functions.

alpha. In the case of linear %SVM all the alpha's will be 1's.

svidx. support vectors (which can be retrieved by SVM::getSupportVectors). In the case of linear %SVM each decision function consists of a single "compressed" support vector. The method returns rho parameter of the decision function, a scalar subtracted from the weighted sum of kernel responses.

alpha : UMat

svidx : UMat

Python prototype (for reference):

getDecisionFunction(i[, alpha[, svidx]]) -> retval, alpha, svidx

getdefaultgridptr(param_id)

getdegree(self)

getgamma(self)

getkerneltype(self)

getnu(self)

getp(self)

getsupportvectors(self)

gettermcriteria(self)

gettype(self)

getuncompressedsupportvectors(self)

load(filepath)

setc(self, val)

setclassweights(self, val)

setcoef0(self, val)

setdegree(self, val)

setgamma(self, val)

setkernel(self, kernelType)

setnu(self, val)

setp(self, val)

settermcriteria(self, val)

settype(self, val)

trainauto(self, samples, layout, responses)

trainauto(self, samples, layout, responses, opts)

Trains an %SVM with optimal parameters

samples.

layout.

responses.

kFold. subset is used to test the model, the others form the train set. So, the %SVM algorithm is

cgrid.

gammaGrid.

pGrid.

nuGrid.

coeffGrid.

degreeGrid.

balanced. balanced cross-validation subsets that is proportions between classes in subsets are close to such proportion in the whole train dataset. The method trains the %SVM model automatically by choosing the optimal parameters C, gamma, p, nu, coef0, degree. Parameters are considered optimal when the cross-validation estimate of the test set error is minimal. This function only makes use of SVM::getDefaultGrid for parameter optimization and thus only offers rudimentary parameter options. This function works for the classification (SVM::C_SVC or SVM::NU_SVC) as well as for the regression (SVM::EPS_SVR or SVM::NU_SVR). If it is SVM::ONE_CLASS, no optimization is made and the usual %SVM with parameters specified in params is executed.

kFold : int

cgrid : Ptr<ParamGrid>

gammaGrid : Ptr<ParamGrid>

pGrid : Ptr<ParamGrid>

nuGrid : Ptr<ParamGrid>

coeffGrid : Ptr<ParamGrid>

degreeGrid : Ptr<ParamGrid>

balanced : bool

Python prototype (for reference):

trainAuto(samples, layout, responses[, kFold[, Cgrid[, gammaGrid[, pGrid[, nuGrid[, coeffGrid[, degreeGrid[, balanced]]]]]]]]) -> retval