OpenCV.MlSVMSGD (Evision v0.1.0-dev)

cv.ml

create()

getinitialstepsize(self)

getmarginregularization(self)

getmargintype(self)

getshift(self)

getstepdecreasingpower(self)

getsvmsgdtype(self)

gettermcriteria(self)

getweights(self)

load(filepath)

load(filepath, opts)

Loads and creates a serialized SVMSGD from a file Use SVMSGD::save to serialize and store an SVMSGD to disk. Load the SVMSGD from this file again, by calling this function with the path to the file. Optionally specify the node for the file containing the classifier

filepath.

nodeName.

nodeName : String

Python prototype (for reference):

load(filepath[, nodeName]) -> retval

setinitialstepsize(self, initialStepSize)

setmarginregularization(self, marginRegularization)

setmargintype(self, marginType)

setoptimalparameters(self)

setstepdecreasingpower(self, stepDecreasingPower)

setsvmsgdtype(self, svmsgdType)

settermcriteria(self, val)