OpenCV.MlStatModel (Evision v0.1.0-dev)

cv.ml

calcerror(self, data, test)

calcerror(self, data, test, opts)

Computes error on the training or test dataset

data.

test. computed over the training subset of the data. Please note that if you loaded a completely different dataset to evaluate already trained classifier, you will probably want not to set the test subset at all with TrainData::setTrainTestSplitRatio and specify test=false, so that the error is computed for the whole new set. Yes, this sounds a bit confusing.

resp. The method uses StatModel::predict to compute the error. For regression models the error is computed as RMS, for classifiers - as a percent of missclassified samples (0%-100%).

resp : UMat

Python prototype (for reference):

calcError(data, test[, resp]) -> retval, resp

empty(self)

getvarcount(self)

isclassifier(self)

istrained(self)

predict(self, samples)

predict(self, samples, opts)

Predicts response(s) for the provided sample(s)

samples.

results.

flags.

results : UMat

flags : int

Python prototype (for reference):

predict(samples[, results[, flags]]) -> retval, results

train(self, trainData)

train(self, trainData, opts)

train(self, samples, layout, responses)