OpenCV.ORB (Evision v0.1.0-dev)

cv

create()

create(opts)

The ORB constructor

nfeatures.

scaleFactor. pyramid, where each next level has 4x less pixels than the previous, but such a big scale factor will degrade feature matching scores dramatically. On the other hand, too close to 1 scale factor will mean that to cover certain scale range you will need more pyramid levels and so the speed will suffer.

nlevels. input_image_linear_size/pow(scaleFactor, nlevels - firstLevel).

edgeThreshold. roughly match the patchSize parameter.

firstLevel. with upscaled source image.

wTA_K. default value 2 means the BRIEF where we take a random point pair and compare their brightnesses, so we get 0/1 response. Other possible values are 3 and 4. For example, 3 means that we take 3 random points (of course, those point coordinates are random, but they are generated from the pre-defined seed, so each element of BRIEF descriptor is computed deterministically from the pixel rectangle), find point of maximum brightness and output index of the winner (0, 1 or 2). Such output will occupy 2 bits, and therefore it will need a special variant of Hamming distance, denoted as NORM_HAMMING2 (2 bits per bin). When WTA_K=4, we take 4 random points to compute each bin (that will also occupy 2 bits with possible values 0, 1, 2 or 3).

scoreType. (the score is written to KeyPoint::score and is used to retain best nfeatures features); FAST_SCORE is alternative value of the parameter that produces slightly less stable keypoints, but it is a little faster to compute.

patchSize. pyramid layers the perceived image area covered by a feature will be larger.

fastThreshold.

nfeatures : int

scaleFactor : float

nlevels : int

edgeThreshold : int

firstLevel : int

wTA_K : int

scoreType : ORB_ScoreType

patchSize : int

fastThreshold : int

Python prototype (for reference):

create([, nfeatures[, scaleFactor[, nlevels[, edgeThreshold[, firstLevel[, WTA_K[, scoreType[, patchSize[, fastThreshold]]]]]]]]]) -> retval

getdefaultname(self)

getedgethreshold(self)

getfastthreshold(self)

getfirstlevel(self)

getmaxfeatures(self)

getnlevels(self)

getpatchsize(self)

getscalefactor(self)

getscoretype(self)

getwta_k(self)

setedgethreshold(self, edgeThreshold)

setfastthreshold(self, fastThreshold)

setfirstlevel(self, firstLevel)

setmaxfeatures(self, maxFeatures)

setnlevels(self, nlevels)

setpatchsize(self, patchSize)

setscalefactor(self, scaleFactor)

setscoretype(self, scoreType)

setwta_k(self, wta_k)