OpenCV.QRCodeDetector (Evision v0.1.0-dev)

cv

decode(self, img, points)

decode(self, img, points, opts)

Decodes QR code in image once it's found by the detect() method. Returns UTF8-encoded output string or empty string if the code cannot be decoded.

img.

points.

straight_qrcode.

straight_qrcode : UMat

Python prototype (for reference):

decode(img, points[, straight_qrcode]) -> retval, straight_qrcode

decodecurved(self, img, points)

decodecurved(self, img, points, opts)

Decodes QR code on a curved surface in image once it's found by the detect() method. Returns UTF8-encoded output string or empty string if the code cannot be decoded.

img.

points.

straight_qrcode.

straight_qrcode : UMat

Python prototype (for reference):

decodeCurved(img, points[, straight_qrcode]) -> retval, straight_qrcode

decodemulti(self, img, points)

decodemulti(self, img, points, opts)

Decodes QR codes in image once it's found by the detect() method.

img.

decoded_info.

points.

straight_qrcode.

straight_qrcode : vector_UMat

Python prototype (for reference):

decodeMulti(img, points[, straight_qrcode]) -> retval, decoded_info, straight_qrcode

detect(self, img)

detect(self, img, opts)

Detects QR code in image and returns the quadrangle containing the code.

img.

points.

points : UMat

Python prototype (for reference):

detect(img[, points]) -> retval, points

detectanddecode(self, img)

detectanddecode(self, img, opts)

Both detects and decodes QR code

img.

points.

straight_qrcode.

points : UMat

straight_qrcode : UMat

Python prototype (for reference):

detectAndDecode(img[, points[, straight_qrcode]]) -> retval, points, straight_qrcode

detectanddecodecurved(self, img)

detectanddecodecurved(self, img, opts)

Both detects and decodes QR code on a curved surface

img.

points.

straight_qrcode.

points : UMat

straight_qrcode : UMat

Python prototype (for reference):

detectAndDecodeCurved(img[, points[, straight_qrcode]]) -> retval, points, straight_qrcode

detectanddecodemulti(self, img)

detectanddecodemulti(self, img, opts)

Both detects and decodes QR codes

img.

decoded_info.

points.

straight_qrcode.

points : UMat

straight_qrcode : vector_UMat

Python prototype (for reference):

detectAndDecodeMulti(img[, points[, straight_qrcode]]) -> retval, decoded_info, points, straight_qrcode

detectmulti(self, img)

detectmulti(self, img, opts)

Detects QR codes in image and returns the vector of the quadrangles containing the codes.

img.

points.

points : UMat

Python prototype (for reference):

detectMulti(img[, points]) -> retval, points

qrcodedetector()

setepsx(self, epsX)

setepsy(self, epsY)