OpenCV.Segmentation.IntelligentScissorsMB (Evision v0.1.0-dev)

cv.segmentation

applyImage(self, image)

applyImageFeatures(self, non_edge, gradient_direction, gradient_magnitude)

applyImageFeatures(self, non_edge, gradient_direction, gradient_magnitude, opts)

Specify custom features of imput image Customized advanced variant of applyImage() call.

non_edge.

gradient_direction.

gradient_magnitude.

image.

image : UMat

Python prototype (for reference):

applyImageFeatures(non_edge, gradient_direction, gradient_magnitude[, image]) -> retval

buildMap(self, sourcePt)

getContour(self, targetPt)

getContour(self, targetPt, opts)

Extracts optimal contour for the given target point on the image Note: buildMap() must be called before this call

targetPt.

contour.

backward.

contour : UMat

backward : bool

Python prototype (for reference):

getContour(targetPt[, contour[, backward]]) -> contour

intelligentScissorsMB()

setEdgeFeatureCannyParameters(self, threshold1, threshold2)

setEdgeFeatureCannyParameters(self, threshold1, threshold2, opts)

Switch edge feature extractor to use Canny edge detector Note: "Laplacian Zero-Crossing" feature extractor is used by default (following to original article) @sa Canny

apertureSize : int

l2gradient : bool

Python prototype (for reference):

setEdgeFeatureCannyParameters(threshold1, threshold2[, apertureSize[, L2gradient]]) -> retval

setEdgeFeatureZeroCrossingParameters(self)

setGradientMagnitudeMaxLimit(self)

setWeights(self, weight_non_edge, weight_gradient_direction, weight_gradient_magnitude)