OpenCV.StereoBM (Evision v0.1.0-dev)

cv

create()

create(opts)

Creates StereoBM object

numDisparities. disparity from 0 (default minimum disparity) to numDisparities. The search range can then be shifted by changing the minimum disparity.

blockSize. (as the block is centered at the current pixel). Larger block size implies smoother, though less accurate disparity map. Smaller block size gives more detailed disparity map, but there is higher chance for algorithm to find a wrong correspondence. The function create StereoBM object. You can then call StereoBM::compute() to compute disparity for a specific stereo pair.

numDisparities : int

blockSize : int

Python prototype (for reference):

create([, numDisparities[, blockSize]]) -> retval

getPreFilterCap(self)

getPreFilterSize(self)

getPreFilterType(self)

getROI1(self)

getROI2(self)

getSmallerBlockSize(self)

getTextureThreshold(self)

getUniquenessRatio(self)

setPreFilterCap(self, preFilterCap)

setPreFilterSize(self, preFilterSize)

setPreFilterType(self, preFilterType)

setROI1(self, roi1)

setROI2(self, roi2)

setSmallerBlockSize(self, blockSize)

setTextureThreshold(self, textureThreshold)

setUniquenessRatio(self, uniquenessRatio)