OpenCV.StereoSGBM (Evision v0.1.0-dev)

cv

create()

create(opts)

Creates StereoSGBM object

minDisparity. rectification algorithms can shift images, so this parameter needs to be adjusted accordingly.

numDisparities. zero. In the current implementation, this parameter must be divisible by 16.

blockSize. somewhere in the 3..11 range.

p1.

p2. the smoother the disparity is. P1 is the penalty on the disparity change by plus or minus 1 between neighbor pixels. P2 is the penalty on the disparity change by more than 1 between neighbor pixels. The algorithm requires P2 > P1 . See stereo_match.cpp sample where some reasonably good P1 and P2 values are shown (like 8*number_of_image_channels*blockSize*blockSize and 32*number_of_image_channels*blockSize*blockSize , respectively).

disp12MaxDiff. disparity check. Set it to a non-positive value to disable the check.

preFilterCap. computes x-derivative at each pixel and clips its value by [-preFilterCap, preFilterCap] interval. The result values are passed to the Birchfield-Tomasi pixel cost function.

uniquenessRatio. value should "win" the second best value to consider the found match correct. Normally, a value within the 5-15 range is good enough.

speckleWindowSize. and invalidate. Set it to 0 to disable speckle filtering. Otherwise, set it somewhere in the 50-200 range.

speckleRange. filtering, set the parameter to a positive value, it will be implicitly multiplied by 16. Normally, 1 or 2 is good enough.

mode. algorithm. It will consume O(W*H*numDisparities) bytes, which is large for 640x480 stereo and huge for HD-size pictures. By default, it is set to false . The first constructor initializes StereoSGBM with all the default parameters. So, you only have to set StereoSGBM::numDisparities at minimum. The second constructor enables you to set each parameter to a custom value.

minDisparity : int

numDisparities : int

blockSize : int

p1 : int

p2 : int

disp12MaxDiff : int

preFilterCap : int

uniquenessRatio : int

speckleWindowSize : int

speckleRange : int

mode : int

Python prototype (for reference):

create([, minDisparity[, numDisparities[, blockSize[, P1[, P2[, disp12MaxDiff[, preFilterCap[, uniquenessRatio[, speckleWindowSize[, speckleRange[, mode]]]]]]]]]]]) -> retval

getmode(self)

getp1(self)

getp2(self)

getprefiltercap(self)

getuniquenessratio(self)

setmode(self, mode)

setp1(self, p1)

setp2(self, p2)

setprefiltercap(self, preFilterCap)

setuniquenessratio(self, uniquenessRatio)