Evision.AffineFeature (Evision v0.1.9)

cv

compute(self, images, keypoints)

Variant 1:

Positional Arguments

images: [Evision.Mat].

Image set.

Return

keypoints: [vector_KeyPoint].

Input collection of keypoints. Keypoints for which a descriptor cannot be computed are removed. Sometimes new keypoints can be added, for example: SIFT duplicates keypoint with several dominant orientations (for each orientation).

descriptors: [Evision.Mat].

Computed descriptors. In the second variant of the method descriptors[i] are descriptors computed for a keypoints[i]. Row j is the keypoints (or keypoints[i]) is the descriptor for keypoint j-th keypoint.

Has overloading in C++

Python prototype (for reference):

compute(images, keypoints[, descriptors]) -> keypoints, descriptors

Variant 2:

Computes the descriptors for a set of keypoints detected in an image (first variant) or image set (second variant).

Positional Arguments

image: Evision.Mat.

Image.

Return

keypoints: [KeyPoint].

Input collection of keypoints. Keypoints for which a descriptor cannot be computed are removed. Sometimes new keypoints can be added, for example: SIFT duplicates keypoint with several dominant orientations (for each orientation).

descriptors: Evision.Mat.

Computed descriptors. In the second variant of the method descriptors[i] are descriptors computed for a keypoints[i]. Row j is the keypoints (or keypoints[i]) is the descriptor for keypoint j-th keypoint.

Python prototype (for reference):

compute(image, keypoints[, descriptors]) -> keypoints, descriptors

compute(self, images, keypoints, opts)

Variant 1:

Positional Arguments

images: [Evision.Mat].

Image set.

Return

keypoints: [vector_KeyPoint].

Input collection of keypoints. Keypoints for which a descriptor cannot be computed are removed. Sometimes new keypoints can be added, for example: SIFT duplicates keypoint with several dominant orientations (for each orientation).

descriptors: [Evision.Mat].

Computed descriptors. In the second variant of the method descriptors[i] are descriptors computed for a keypoints[i]. Row j is the keypoints (or keypoints[i]) is the descriptor for keypoint j-th keypoint.

Has overloading in C++

Python prototype (for reference):

compute(images, keypoints[, descriptors]) -> keypoints, descriptors

Variant 2:

Computes the descriptors for a set of keypoints detected in an image (first variant) or image set (second variant).

Positional Arguments

image: Evision.Mat.

Image.

Return

keypoints: [KeyPoint].

Input collection of keypoints. Keypoints for which a descriptor cannot be computed are removed. Sometimes new keypoints can be added, for example: SIFT duplicates keypoint with several dominant orientations (for each orientation).

descriptors: Evision.Mat.

Computed descriptors. In the second variant of the method descriptors[i] are descriptors computed for a keypoints[i]. Row j is the keypoints (or keypoints[i]) is the descriptor for keypoint j-th keypoint.

Python prototype (for reference):

compute(image, keypoints[, descriptors]) -> keypoints, descriptors

create(backend)

Positional Arguments

backend: Ptr<Feature2D>.

The detector/extractor you want to use as backend.

Keyword Arguments

maxTilt: int.

The highest power index of tilt factor. 5 is used in the paper as tilt sampling range n.

minTilt: int.

The lowest power index of tilt factor. 0 is used in the paper.

tiltStep: float.

Tilt sampling step \f$\delta_t\f$ in Algorithm 1 in the paper.

rotateStepBase: float.

Rotation sampling step factor b in Algorithm 1 in the paper.

Python prototype (for reference):

create(backend[, maxTilt[, minTilt[, tiltStep[, rotateStepBase]]]]) -> retval

create(backend, opts)

Positional Arguments

backend: Ptr<Feature2D>.

The detector/extractor you want to use as backend.

Keyword Arguments

maxTilt: int.

The highest power index of tilt factor. 5 is used in the paper as tilt sampling range n.

minTilt: int.

The lowest power index of tilt factor. 0 is used in the paper.

tiltStep: float.

Tilt sampling step \f$\delta_t\f$ in Algorithm 1 in the paper.

rotateStepBase: float.

Rotation sampling step factor b in Algorithm 1 in the paper.

Python prototype (for reference):

create(backend[, maxTilt[, minTilt[, tiltStep[, rotateStepBase]]]]) -> retval

defaultNorm(self)

Python prototype (for reference):

defaultNorm() -> retval

descriptorSize(self)

Python prototype (for reference):

descriptorSize() -> retval

descriptorType(self)

Python prototype (for reference):

descriptorType() -> retval

detect(self, images)

Variant 1:

Positional Arguments

images: [Evision.Mat].

Image set.

Keyword Arguments

masks: [Evision.Mat].

Masks for each input image specifying where to look for keypoints (optional). masks[i] is a mask for images[i].

Return

keypoints: [vector_KeyPoint].

The detected keypoints. In the second variant of the method keypoints[i] is a set of keypoints detected in images[i] .

Has overloading in C++

Python prototype (for reference):

detect(images[, masks]) -> keypoints

Variant 2:

Detects keypoints in an image (first variant) or image set (second variant).

Positional Arguments

image: Evision.Mat.

Image.

Keyword Arguments

mask: Evision.Mat.

Mask specifying where to look for keypoints (optional). It must be a 8-bit integer matrix with non-zero values in the region of interest.

Return

keypoints: [KeyPoint].

The detected keypoints. In the second variant of the method keypoints[i] is a set of keypoints detected in images[i] .

Python prototype (for reference):

detect(image[, mask]) -> keypoints

detect(self, images, opts)

Variant 1:

Positional Arguments

images: [Evision.Mat].

Image set.

Keyword Arguments

masks: [Evision.Mat].

Masks for each input image specifying where to look for keypoints (optional). masks[i] is a mask for images[i].

Return

keypoints: [vector_KeyPoint].

The detected keypoints. In the second variant of the method keypoints[i] is a set of keypoints detected in images[i] .

Has overloading in C++

Python prototype (for reference):

detect(images[, masks]) -> keypoints

Variant 2:

Detects keypoints in an image (first variant) or image set (second variant).

Positional Arguments

image: Evision.Mat.

Image.

Keyword Arguments

mask: Evision.Mat.

Mask specifying where to look for keypoints (optional). It must be a 8-bit integer matrix with non-zero values in the region of interest.

Return

keypoints: [KeyPoint].

The detected keypoints. In the second variant of the method keypoints[i] is a set of keypoints detected in images[i] .

Python prototype (for reference):

detect(image[, mask]) -> keypoints

detectAndCompute(self, image, mask)

Positional Arguments

image: Evision.Mat

mask: Evision.Mat

Keyword Arguments

useProvidedKeypoints: bool.

Return

keypoints: [KeyPoint]

descriptors: Evision.Mat.

Detects keypoints and computes the descriptors

Python prototype (for reference):

detectAndCompute(image, mask[, descriptors[, useProvidedKeypoints]]) -> keypoints, descriptors

detectAndCompute(self, image, mask, opts)

Positional Arguments

image: Evision.Mat

mask: Evision.Mat

Keyword Arguments

useProvidedKeypoints: bool.

Return

keypoints: [KeyPoint]

descriptors: Evision.Mat.

Detects keypoints and computes the descriptors

Python prototype (for reference):

detectAndCompute(image, mask[, descriptors[, useProvidedKeypoints]]) -> keypoints, descriptors

empty(self)

Python prototype (for reference):

empty() -> retval

getDefaultName(self)

Python prototype (for reference):

getDefaultName() -> retval

getViewParams(self, tilts, rolls)

Positional Arguments

tilts: [float]

rolls: [float]

Python prototype (for reference):

getViewParams(tilts, rolls) -> None

read(self, fileName)

Variant 1:

Positional Arguments

fileName: String

Python prototype (for reference):

read(fileName) -> None

Variant 2:

Positional Arguments

arg1: FileNode

Python prototype (for reference):

read(arg1) -> None

setViewParams(self, tilts, rolls)

Positional Arguments

tilts: [float]

rolls: [float]

Python prototype (for reference):

setViewParams(tilts, rolls) -> None

write(self, fs)

Variant 1:

Positional Arguments

fs: Ptr<FileStorage>

Keyword Arguments

name: String.

Python prototype (for reference):

write(fs[, name]) -> None

Variant 2:

Positional Arguments

fileName: String

Python prototype (for reference):

write(fileName) -> None

write(self, fs, opts)

Positional Arguments

fs: Ptr<FileStorage>

Keyword Arguments

name: String.

Python prototype (for reference):

write(fs[, name]) -> None

Functions

compute!(self, images, keypoints)

Raising version of compute/3.

compute!(self, images, keypoints, opts)

Raising version of compute/4.

create!(backend)

Raising version of create/1.

create!(backend, opts)

Raising version of create/2.

defaultNorm!(self)

Raising version of defaultNorm/1.

descriptorSize!(self)

Raising version of descriptorSize/1.

descriptorType!(self)

Raising version of descriptorType/1.

detect!(self, images)

Raising version of detect/2.

detect!(self, images, opts)

Raising version of detect/3.

detectAndCompute!(self, image, mask)

Raising version of detectAndCompute/3.

detectAndCompute!(self, image, mask, opts)

Raising version of detectAndCompute/4.

empty!(self)

Raising version of empty/1.

getDefaultName!(self)

Raising version of getDefaultName/1.

getViewParams!(self, tilts, rolls)

Raising version of getViewParams/3.

read!(self, fileName)

Raising version of read/2.

setViewParams!(self, tilts, rolls)

Raising version of setViewParams/3.

write!(self, fs)

Raising version of write/2.

write!(self, fs, opts)

Raising version of write/3.