Evision.AlignMTB (Evision v0.1.7-dev)

cv

calculateShift(self, img0, img1)

Calculates shift between two images, i. e. how to shift the second image to correspond it with the first.

Python prototype (for reference):

calculateShift(img0, img1) -> retval

computeBitmaps(self, img)

Positional Arguments

img: UMat.

Keyword Arguments

tb: UMat.

eb: UMat.

Computes median threshold and exclude bitmaps of given image.

Python prototype (for reference):

computeBitmaps(img[, tb[, eb]]) -> tb, eb

computeBitmaps(self, img, opts)

Positional Arguments

img: UMat.

Keyword Arguments

tb: UMat.

eb: UMat.

Computes median threshold and exclude bitmaps of given image.

Python prototype (for reference):

computeBitmaps(img[, tb[, eb]]) -> tb, eb

getCut(self)

Python prototype (for reference):

getCut() -> retval

getExcludeRange(self)

Python prototype (for reference):

getExcludeRange() -> retval

getMaxBits(self)

Python prototype (for reference):

getMaxBits() -> retval

process(self, src, dst)

Short version of process, that doesn't take extra arguments.

Python prototype (for reference):

process(src, dst) -> None

process(self, src, dst, times, response)

Python prototype (for reference):

process(src, dst, times, response) -> None

setCut(self, value)

Python prototype (for reference):

setCut(value) -> None

setExcludeRange(self, exclude_range)

Python prototype (for reference):

setExcludeRange(exclude_range) -> None

setMaxBits(self, max_bits)

Python prototype (for reference):

setMaxBits(max_bits) -> None

shiftMat(self, src, shift)

Positional Arguments

src: UMat.

shift: Point.

Keyword Arguments

dst: UMat.

Helper function, that shift Mat filling new regions with zeros.

Python prototype (for reference):

shiftMat(src, shift[, dst]) -> dst

shiftMat(self, src, shift, opts)

Positional Arguments

src: UMat.

shift: Point.

Keyword Arguments

dst: UMat.

Helper function, that shift Mat filling new regions with zeros.

Python prototype (for reference):

shiftMat(src, shift[, dst]) -> dst

Functions

calculateShift!(self, img0, img1)

Raising version of calculateShift/3.

computeBitmaps!(self, img)

Raising version of computeBitmaps/2.

computeBitmaps!(self, img, opts)

Raising version of computeBitmaps/3.

getCut!(self)

Raising version of getCut/1.

getExcludeRange!(self)

Raising version of getExcludeRange/1.

getMaxBits!(self)

Raising version of getMaxBits/1.

process!(self, src, dst)

Raising version of process/3.

process!(self, src, dst, times, response)

Raising version of process/5.

setCut!(self, value)

Raising version of setCut/2.

setExcludeRange!(self, exclude_range)

Raising version of setExcludeRange/2.

setMaxBits!(self, max_bits)

Raising version of setMaxBits/2.

shiftMat!(self, src, shift)

Raising version of shiftMat/3.

shiftMat!(self, src, shift, opts)

Raising version of shiftMat/4.