Evision.BFMatcher (Evision v0.1.5-dev)

cv

add(self, descriptors)

bfMatcher()

bfMatcher(opts)

Brute-force matcher constructor (obsolete). Please use BFMatcher.create()

normType : int

crossCheck : bool

Python prototype (for reference):

BFMatcher([, normType[, crossCheck]]) -> <BFMatcher object>

clear(self)

clone(self)

create()

create(opts)

Brute-force matcher create method.

normType. preferable choices for SIFT and SURF descriptors, NORM_HAMMING should be used with ORB, BRISK and BRIEF, NORM_HAMMING2 should be used with ORB when WTA_K==3 or 4 (see ORB::ORB constructor description).

crossCheck. nearest neighbors for each query descriptor. If crossCheck==true, then the knnMatch() method with k=1 will only return pairs (i,j) such that for i-th query descriptor the j-th descriptor in the matcher's collection is the nearest and vice versa, i.e. the BFMatcher will only return consistent pairs. Such technique usually produces best results with minimal number of outliers when there are enough matches. This is alternative to the ratio test, used by D. Lowe in SIFT paper.

normType : int

crossCheck : bool

Python prototype (for reference):

create([, normType[, crossCheck]]) -> retval

empty(self)

getTrainDescriptors(self)

isMaskSupported(self)

knnMatch(self, queryDescriptors, k)

knnMatch(self, queryDescriptors, k, opts)

knnMatch(self, queryDescriptors, trainDescriptors, k, opts)

Finds the k best matches for each descriptor from a query set.

queryDescriptors.

trainDescriptors. collection stored in the class object.

mask. descriptors.

matches.

k. less than k possible matches in total.

compactResult. false, the matches vector has the same size as queryDescriptors rows. If compactResult is true, the matches vector does not contain matches for fully masked-out query descriptors. These extended variants of DescriptorMatcher::match methods find several best matches for each query descriptor. The matches are returned in the distance increasing order. See DescriptorMatcher::match for the details about query and train descriptors.

mask : UMat

compactResult : bool

Python prototype (for reference):

knnMatch(queryDescriptors, trainDescriptors, k[, mask[, compactResult]]) -> matches

match(self, queryDescriptors)

match(self, queryDescriptors, opts)

match(self, queryDescriptors, trainDescriptors, opts)

Finds the best match for each descriptor from a query set.

queryDescriptors.

trainDescriptors. collection stored in the class object.

matches. descriptor. So, matches size may be smaller than the query descriptors count.

mask. descriptors. In the first variant of this method, the train descriptors are passed as an input argument. In the second variant of the method, train descriptors collection that was set by DescriptorMatcher::add is used. Optional mask (or masks) can be passed to specify which query and training descriptors can be matched. Namely, queryDescriptors[i] can be matched with trainDescriptors[j] only if mask.at\<uchar>(i,j) is non-zero.

mask : UMat

Python prototype (for reference):

match(queryDescriptors, trainDescriptors[, mask]) -> matches

radiusMatch(self, queryDescriptors, maxDistance)

radiusMatch(self, queryDescriptors, maxDistance, opts)

radiusMatch(self, queryDescriptors, trainDescriptors, maxDistance, opts)

For each query descriptor, finds the training descriptors not farther than the specified distance.

queryDescriptors.

trainDescriptors. collection stored in the class object.

matches.

compactResult. false, the matches vector has the same size as queryDescriptors rows. If compactResult is true, the matches vector does not contain matches for fully masked-out query descriptors.

maxDistance. metric distance (e.g. Hamming distance), not the distance between coordinates (which is measured in Pixels)!

mask. descriptors. For each query descriptor, the methods find such training descriptors that the distance between the query descriptor and the training descriptor is equal or smaller than maxDistance. Found matches are returned in the distance increasing order.

mask : UMat

compactResult : bool

Python prototype (for reference):

radiusMatch(queryDescriptors, trainDescriptors, maxDistance[, mask[, compactResult]]) -> matches

read(self, arg1)

train(self)

write(self, fs)

write(self, fs, opts)

name : String

Python prototype (for reference):

write(fs[, name]) -> None

Functions

add!(self, descriptors)

Raising version of add/2.

bfMatcher!()

Raising version of bfMatcher/0.

bfMatcher!(opts)

Raising version of bfMatcher/1.

clear!(self)

Raising version of clear/1.

clone!(self)

Raising version of clone/1.

create!()

Raising version of create/0.

create!(opts)

Raising version of create/1.

empty!(self)

Raising version of empty/1.

getTrainDescriptors!(self)

Raising version of getTrainDescriptors/1.

isMaskSupported!(self)

Raising version of isMaskSupported/1.

knnMatch!(self, queryDescriptors, k)

Raising version of knnMatch/3.

knnMatch!(self, queryDescriptors, trainDescriptors, k)

Raising version of knnMatch/4.

knnMatch!(self, queryDescriptors, trainDescriptors, k, opts)

Raising version of knnMatch/5.

match!(self, queryDescriptors)

Raising version of match/2.

match!(self, queryDescriptors, trainDescriptors)

Raising version of match/3.

match!(self, queryDescriptors, trainDescriptors, opts)

Raising version of match/4.

radiusMatch!(self, queryDescriptors, maxDistance)

Raising version of radiusMatch/3.

radiusMatch!(self, queryDescriptors, trainDescriptors, maxDistance)

Raising version of radiusMatch/4.

radiusMatch!(self, queryDescriptors, trainDescriptors, maxDistance, opts)

Raising version of radiusMatch/5.

read!(self, arg1)

Raising version of read/2.

train!(self)

Raising version of train/1.

write!(self, fs)

Raising version of write/2.

write!(self, fs, opts)

Raising version of write/3.