Evision.BRISK (Evision v0.1.9)

cv

compute(self, images, keypoints)

Variant 1:

Positional Arguments

images: [Evision.Mat].

Image set.

Return

keypoints: [vector_KeyPoint].

Input collection of keypoints. Keypoints for which a descriptor cannot be computed are removed. Sometimes new keypoints can be added, for example: SIFT duplicates keypoint with several dominant orientations (for each orientation).

descriptors: [Evision.Mat].

Computed descriptors. In the second variant of the method descriptors[i] are descriptors computed for a keypoints[i]. Row j is the keypoints (or keypoints[i]) is the descriptor for keypoint j-th keypoint.

Has overloading in C++

Python prototype (for reference):

compute(images, keypoints[, descriptors]) -> keypoints, descriptors

Variant 2:

Computes the descriptors for a set of keypoints detected in an image (first variant) or image set (second variant).

Positional Arguments

image: Evision.Mat.

Image.

Return

keypoints: [KeyPoint].

Input collection of keypoints. Keypoints for which a descriptor cannot be computed are removed. Sometimes new keypoints can be added, for example: SIFT duplicates keypoint with several dominant orientations (for each orientation).

descriptors: Evision.Mat.

Computed descriptors. In the second variant of the method descriptors[i] are descriptors computed for a keypoints[i]. Row j is the keypoints (or keypoints[i]) is the descriptor for keypoint j-th keypoint.

Python prototype (for reference):

compute(image, keypoints[, descriptors]) -> keypoints, descriptors

compute(self, images, keypoints, opts)

Variant 1:

Positional Arguments

images: [Evision.Mat].

Image set.

Return

keypoints: [vector_KeyPoint].

Input collection of keypoints. Keypoints for which a descriptor cannot be computed are removed. Sometimes new keypoints can be added, for example: SIFT duplicates keypoint with several dominant orientations (for each orientation).

descriptors: [Evision.Mat].

Computed descriptors. In the second variant of the method descriptors[i] are descriptors computed for a keypoints[i]. Row j is the keypoints (or keypoints[i]) is the descriptor for keypoint j-th keypoint.

Has overloading in C++

Python prototype (for reference):

compute(images, keypoints[, descriptors]) -> keypoints, descriptors

Variant 2:

Computes the descriptors for a set of keypoints detected in an image (first variant) or image set (second variant).

Positional Arguments

image: Evision.Mat.

Image.

Return

keypoints: [KeyPoint].

Input collection of keypoints. Keypoints for which a descriptor cannot be computed are removed. Sometimes new keypoints can be added, for example: SIFT duplicates keypoint with several dominant orientations (for each orientation).

descriptors: Evision.Mat.

Computed descriptors. In the second variant of the method descriptors[i] are descriptors computed for a keypoints[i]. Row j is the keypoints (or keypoints[i]) is the descriptor for keypoint j-th keypoint.

Python prototype (for reference):

compute(image, keypoints[, descriptors]) -> keypoints, descriptors

create()

The BRISK constructor

Keyword Arguments

thresh: int.

AGAST detection threshold score.

octaves: int.

detection octaves. Use 0 to do single scale.

patternScale: float.

apply this scale to the pattern used for sampling the neighbourhood of a keypoint.

Python prototype (for reference):

create([, thresh[, octaves[, patternScale]]]) -> retval

create(opts)

The BRISK constructor

Keyword Arguments

thresh: int.

AGAST detection threshold score.

octaves: int.

detection octaves. Use 0 to do single scale.

patternScale: float.

apply this scale to the pattern used for sampling the neighbourhood of a keypoint.

Python prototype (for reference):

create([, thresh[, octaves[, patternScale]]]) -> retval

create(radiusList, numberList)

The BRISK constructor for a custom pattern

Positional Arguments

radiusList: [float].

defines the radii (in pixels) where the samples around a keypoint are taken (for keypoint scale 1).

numberList: [int].

defines the number of sampling points on the sampling circle. Must be the same size as radiusList..

Keyword Arguments

dMax: float.

threshold for the short pairings used for descriptor formation (in pixels for keypoint scale 1).

dMin: float.

threshold for the long pairings used for orientation determination (in pixels for keypoint scale 1).

indexChange: [int].

index remapping of the bits.

Python prototype (for reference):

create(radiusList, numberList[, dMax[, dMin[, indexChange]]]) -> retval

create(radiusList, numberList, opts)

The BRISK constructor for a custom pattern

Positional Arguments

radiusList: [float].

defines the radii (in pixels) where the samples around a keypoint are taken (for keypoint scale 1).

numberList: [int].

defines the number of sampling points on the sampling circle. Must be the same size as radiusList..

Keyword Arguments

dMax: float.

threshold for the short pairings used for descriptor formation (in pixels for keypoint scale 1).

dMin: float.

threshold for the long pairings used for orientation determination (in pixels for keypoint scale 1).

indexChange: [int].

index remapping of the bits.

Python prototype (for reference):

create(radiusList, numberList[, dMax[, dMin[, indexChange]]]) -> retval

create(thresh, octaves, radiusList, numberList)

The BRISK constructor for a custom pattern, detection threshold and octaves

Positional Arguments

thresh: int.

AGAST detection threshold score.

octaves: int.

detection octaves. Use 0 to do single scale.

radiusList: [float].

defines the radii (in pixels) where the samples around a keypoint are taken (for keypoint scale 1).

numberList: [int].

defines the number of sampling points on the sampling circle. Must be the same size as radiusList..

Keyword Arguments

dMax: float.

threshold for the short pairings used for descriptor formation (in pixels for keypoint scale 1).

dMin: float.

threshold for the long pairings used for orientation determination (in pixels for keypoint scale 1).

indexChange: [int].

index remapping of the bits.

Python prototype (for reference):

create(thresh, octaves, radiusList, numberList[, dMax[, dMin[, indexChange]]]) -> retval

create(thresh, octaves, radiusList, numberList, opts)

The BRISK constructor for a custom pattern, detection threshold and octaves

Positional Arguments

thresh: int.

AGAST detection threshold score.

octaves: int.

detection octaves. Use 0 to do single scale.

radiusList: [float].

defines the radii (in pixels) where the samples around a keypoint are taken (for keypoint scale 1).

numberList: [int].

defines the number of sampling points on the sampling circle. Must be the same size as radiusList..

Keyword Arguments

dMax: float.

threshold for the short pairings used for descriptor formation (in pixels for keypoint scale 1).

dMin: float.

threshold for the long pairings used for orientation determination (in pixels for keypoint scale 1).

indexChange: [int].

index remapping of the bits.

Python prototype (for reference):

create(thresh, octaves, radiusList, numberList[, dMax[, dMin[, indexChange]]]) -> retval

defaultNorm(self)

Python prototype (for reference):

defaultNorm() -> retval

descriptorSize(self)

Python prototype (for reference):

descriptorSize() -> retval

descriptorType(self)

Python prototype (for reference):

descriptorType() -> retval

detect(self, images)

Variant 1:

Positional Arguments

images: [Evision.Mat].

Image set.

Keyword Arguments

masks: [Evision.Mat].

Masks for each input image specifying where to look for keypoints (optional). masks[i] is a mask for images[i].

Return

keypoints: [vector_KeyPoint].

The detected keypoints. In the second variant of the method keypoints[i] is a set of keypoints detected in images[i] .

Has overloading in C++

Python prototype (for reference):

detect(images[, masks]) -> keypoints

Variant 2:

Detects keypoints in an image (first variant) or image set (second variant).

Positional Arguments

image: Evision.Mat.

Image.

Keyword Arguments

mask: Evision.Mat.

Mask specifying where to look for keypoints (optional). It must be a 8-bit integer matrix with non-zero values in the region of interest.

Return

keypoints: [KeyPoint].

The detected keypoints. In the second variant of the method keypoints[i] is a set of keypoints detected in images[i] .

Python prototype (for reference):

detect(image[, mask]) -> keypoints

detect(self, images, opts)

Variant 1:

Positional Arguments

images: [Evision.Mat].

Image set.

Keyword Arguments

masks: [Evision.Mat].

Masks for each input image specifying where to look for keypoints (optional). masks[i] is a mask for images[i].

Return

keypoints: [vector_KeyPoint].

The detected keypoints. In the second variant of the method keypoints[i] is a set of keypoints detected in images[i] .

Has overloading in C++

Python prototype (for reference):

detect(images[, masks]) -> keypoints

Variant 2:

Detects keypoints in an image (first variant) or image set (second variant).

Positional Arguments

image: Evision.Mat.

Image.

Keyword Arguments

mask: Evision.Mat.

Mask specifying where to look for keypoints (optional). It must be a 8-bit integer matrix with non-zero values in the region of interest.

Return

keypoints: [KeyPoint].

The detected keypoints. In the second variant of the method keypoints[i] is a set of keypoints detected in images[i] .

Python prototype (for reference):

detect(image[, mask]) -> keypoints

detectAndCompute(self, image, mask)

Positional Arguments

image: Evision.Mat

mask: Evision.Mat

Keyword Arguments

useProvidedKeypoints: bool.

Return

keypoints: [KeyPoint]

descriptors: Evision.Mat.

Detects keypoints and computes the descriptors

Python prototype (for reference):

detectAndCompute(image, mask[, descriptors[, useProvidedKeypoints]]) -> keypoints, descriptors

detectAndCompute(self, image, mask, opts)

Positional Arguments

image: Evision.Mat

mask: Evision.Mat

Keyword Arguments

useProvidedKeypoints: bool.

Return

keypoints: [KeyPoint]

descriptors: Evision.Mat.

Detects keypoints and computes the descriptors

Python prototype (for reference):

detectAndCompute(image, mask[, descriptors[, useProvidedKeypoints]]) -> keypoints, descriptors

empty(self)

Python prototype (for reference):

empty() -> retval

getDefaultName(self)

Python prototype (for reference):

getDefaultName() -> retval

getOctaves(self)

Python prototype (for reference):

getOctaves() -> retval

getThreshold(self)

Python prototype (for reference):

getThreshold() -> retval

read(self, fileName)

Variant 1:

Positional Arguments

fileName: String

Python prototype (for reference):

read(fileName) -> None

Variant 2:

Positional Arguments

arg1: FileNode

Python prototype (for reference):

read(arg1) -> None

setOctaves(self, octaves)

Set detection octaves.

Positional Arguments

octaves: int.

detection octaves. Use 0 to do single scale.

Python prototype (for reference):

setOctaves(octaves) -> None

setThreshold(self, threshold)

Set detection threshold.

Positional Arguments

threshold: int.

AGAST detection threshold score.

Python prototype (for reference):

setThreshold(threshold) -> None

write(self, fs)

Variant 1:

Positional Arguments

fs: Ptr<FileStorage>

Keyword Arguments

name: String.

Python prototype (for reference):

write(fs[, name]) -> None

Variant 2:

Positional Arguments

fileName: String

Python prototype (for reference):

write(fileName) -> None

write(self, fs, opts)

Positional Arguments

fs: Ptr<FileStorage>

Keyword Arguments

name: String.

Python prototype (for reference):

write(fs[, name]) -> None

Functions

compute!(self, images, keypoints)

Raising version of compute/3.

compute!(self, images, keypoints, opts)

Raising version of compute/4.

create!()

Raising version of create/0.

create!(opts)

Raising version of create/1.

create!(radiusList, numberList)

Raising version of create/2.

create!(radiusList, numberList, opts)

Raising version of create/3.

create!(thresh, octaves, radiusList, numberList)

Raising version of create/4.

create!(thresh, octaves, radiusList, numberList, opts)

Raising version of create/5.

defaultNorm!(self)

Raising version of defaultNorm/1.

descriptorSize!(self)

Raising version of descriptorSize/1.

descriptorType!(self)

Raising version of descriptorType/1.

detect!(self, images)

Raising version of detect/2.

detect!(self, images, opts)

Raising version of detect/3.

detectAndCompute!(self, image, mask)

Raising version of detectAndCompute/3.

detectAndCompute!(self, image, mask, opts)

Raising version of detectAndCompute/4.

empty!(self)

Raising version of empty/1.

getDefaultName!(self)

Raising version of getDefaultName/1.

getOctaves!(self)

Raising version of getOctaves/1.

getThreshold!(self)

Raising version of getThreshold/1.

read!(self, fileName)

Raising version of read/2.

setOctaves!(self, octaves)

Raising version of setOctaves/2.

setThreshold!(self, threshold)

Raising version of setThreshold/2.

write!(self, fs)

Raising version of write/2.

write!(self, fs, opts)

Raising version of write/3.