Evision.CUDA (Evision v0.1.7-dev)

cv.cuda

createContinuous(rows, cols, type)

Positional Arguments

rows: int.

cols: int.

type: int.

Keyword Arguments

arr: UMat.

Creates a continuous matrix.

\f$\texttt{rows} \times \texttt{cols}\f$ ).

Matrix is called continuous if its elements are stored continuously, that is, without gaps at the end of each row.

Python prototype (for reference):

createContinuous(rows, cols, type[, arr]) -> arr

createContinuous(rows, cols, type, opts)

Positional Arguments

rows: int.

cols: int.

type: int.

Keyword Arguments

arr: UMat.

Creates a continuous matrix.

\f$\texttt{rows} \times \texttt{cols}\f$ ).

Matrix is called continuous if its elements are stored continuously, that is, without gaps at the end of each row.

Python prototype (for reference):

createContinuous(rows, cols, type[, arr]) -> arr

ensureSizeIsEnough(rows, cols, type)

Positional Arguments

rows: int.

cols: int.

type: int.

Keyword Arguments

arr: UMat.

Ensures that the size of a matrix is big enough and the matrix has a proper type. The function does not reallocate memory if the matrix has proper attributes already.

Python prototype (for reference):

ensureSizeIsEnough(rows, cols, type[, arr]) -> arr

ensureSizeIsEnough(rows, cols, type, opts)

Positional Arguments

rows: int.

cols: int.

type: int.

Keyword Arguments

arr: UMat.

Ensures that the size of a matrix is big enough and the matrix has a proper type. The function does not reallocate memory if the matrix has proper attributes already.

Python prototype (for reference):

ensureSizeIsEnough(rows, cols, type[, arr]) -> arr

getCudaEnabledDeviceCount()

Returns the number of installed CUDA-enabled devices. Use this function before any other CUDA functions calls. If OpenCV is compiled without CUDA support, this function returns 0. If the CUDA driver is not installed, or is incompatible, this function returns -1.

Python prototype (for reference):

getCudaEnabledDeviceCount() -> retval

getDevice()

Returns the current device index set by cuda::setDevice or initialized by default.

Python prototype (for reference):

getDevice() -> retval

printCudaDeviceInfo(device)

Python prototype (for reference):

printCudaDeviceInfo(device) -> None

printShortCudaDeviceInfo(device)

Python prototype (for reference):

printShortCudaDeviceInfo(device) -> None

registerPageLocked(m)

Page-locks the memory of matrix and maps it for the device(s).

Python prototype (for reference):

registerPageLocked(m) -> None

resetDevice()

Explicitly destroys and cleans up all resources associated with the current device in the current process. Any subsequent API call to this device will reinitialize the device.

Python prototype (for reference):

resetDevice() -> None

setBufferPoolConfig(deviceId, stackSize, stackCount)

Python prototype (for reference):

setBufferPoolConfig(deviceId, stackSize, stackCount) -> None

setBufferPoolUsage(on)

Python prototype (for reference):

setBufferPoolUsage(on) -> None

setDevice(device)

Sets a device and initializes it for the current thread. If the call of this function is omitted, a default device is initialized at the fist CUDA usage.

Python prototype (for reference):

setDevice(device) -> None

unregisterPageLocked(m)

Unmaps the memory of matrix and makes it pageable again.

Python prototype (for reference):

unregisterPageLocked(m) -> None

Functions

createContinuous!(rows, cols, type)

Raising version of createContinuous/3.

createContinuous!(rows, cols, type, opts)

Raising version of createContinuous/4.

ensureSizeIsEnough!(rows, cols, type)

Raising version of ensureSizeIsEnough/3.

ensureSizeIsEnough!(rows, cols, type, opts)

Raising version of ensureSizeIsEnough/4.

getCudaEnabledDeviceCount!()

Raising version of getCudaEnabledDeviceCount/0.

getDevice!()

Raising version of getDevice/0.

printCudaDeviceInfo!(device)

Raising version of printCudaDeviceInfo/1.

printShortCudaDeviceInfo!(device)

Raising version of printShortCudaDeviceInfo/1.

registerPageLocked!(m)

Raising version of registerPageLocked/1.

resetDevice!()

Raising version of resetDevice/0.

setBufferPoolConfig!(deviceId, stackSize, stackCount)

Raising version of setBufferPoolConfig/3.

setBufferPoolUsage!(on)

Raising version of setBufferPoolUsage/1.

setDevice!(device)

Raising version of setDevice/1.

unregisterPageLocked!(m)

Raising version of unregisterPageLocked/1.