Evision.CUDA.GpuMat (Evision v0.1.5-dev)

cv.cuda

adjustROI(self, dtop, dbottom, dleft, dright)

assignTo(self, m)

assignTo(self, m, opts)

type : int

Python prototype (for reference):

assignTo(m[, type]) -> None

channels(self)

clone(self)

col(self, x)

colRange(self, r)

colRange(self, startcol, endcol)

convertTo(self, rtype)

convertTo(self, rtype, opts)

convertTo(self, rtype, alpha, opts)

convertTo(self, rtype, alpha, stream, opts)

convertTo(self, rtype, alpha, beta, stream, opts)

dst : UMat

Python prototype (for reference):

convertTo(rtype, alpha, beta, stream[, dst]) -> dst

copyTo(self)

copyTo(self, mask)

copyTo(self, mask, opts)

copyTo(self, mask, stream, opts)

dst : UMat

Python prototype (for reference):

copyTo(mask, stream[, dst]) -> dst

create(self, size, type)

create(self, rows, cols, type)

cudaPtr(self)

defaultAllocator()

depth(self)

download(self)

download(self, stream)

download(self, stream, opts)

Performs data download from GpuMat (Non-Blocking call) This function copies data from device memory to host memory. As being a non-blocking call, this function may return even if the copy operation is not finished. The copy operation may be overlapped with operations in other non-default streams if \p stream is not the default stream and \p dst is HostMem allocated with HostMem::PAGE_LOCKED option.

dst : UMat

Python prototype (for reference):

download(stream[, dst]) -> dst

elemSize1(self)

elemSize(self)

empty(self)

gpuMat()

gpuMat(opts)

gpuMat(arr, opts)

gpuMat(size, type, opts)

gpuMat(size, type, s, opts)

gpuMat(rows, cols, type, s, opts)

allocator : GpuMat_Allocator*

Python prototype (for reference):

GpuMat(rows, cols, type, s[, allocator]) -> <cuda_GpuMat object>

isContinuous(self)

locateROI(self, wholeSize, ofs)

release(self)

reshape(self, cn)

reshape(self, cn, opts)

rows : int

Python prototype (for reference):

reshape(cn[, rows]) -> retval

row(self, y)

rowRange(self, r)

rowRange(self, startrow, endrow)

setDefaultAllocator(allocator)

setTo(self, s)

setTo(self, s, mask)

setTo(self, s, mask, stream)

size(self)

step1(self)

swap(self, mat)

type(self)

updateContinuityFlag(self)

upload(self, arr)

upload(self, arr, stream)

Functions

adjustROI!(self, dtop, dbottom, dleft, dright)

Raising version of adjustROI/5.

assignTo!(self, m)

Raising version of assignTo/2.

assignTo!(self, m, opts)

Raising version of assignTo/3.

channels!(self)

Raising version of channels/1.

clone!(self)

Raising version of clone/1.

col!(self, x)

Raising version of col/2.

colRange!(self, r)

Raising version of colRange/2.

colRange!(self, startcol, endcol)

Raising version of colRange/3.

convertTo!(self, rtype)

Raising version of convertTo/2.

convertTo!(self, rtype, alpha)

Raising version of convertTo/3.

convertTo!(self, rtype, alpha, stream)

Raising version of convertTo/4.

convertTo!(self, rtype, alpha, beta, stream)

Raising version of convertTo/5.

convertTo!(self, rtype, alpha, beta, stream, opts)

Raising version of convertTo/6.

copyTo!(self)

Raising version of copyTo/1.

copyTo!(self, mask)

Raising version of copyTo/2.

copyTo!(self, mask, stream)

Raising version of copyTo/3.

copyTo!(self, mask, stream, opts)

Raising version of copyTo/4.

create!(self, size, type)

Raising version of create/3.

create!(self, rows, cols, type)

Raising version of create/4.

cudaPtr!(self)

Raising version of cudaPtr/1.

defaultAllocator!()

Raising version of defaultAllocator/0.

depth!(self)

Raising version of depth/1.

download!(self)

Raising version of download/1.

download!(self, stream)

Raising version of download/2.

download!(self, stream, opts)

Raising version of download/3.

elemSize1!(self)

Raising version of elemSize1/1.

elemSize!(self)

Raising version of elemSize/1.

empty!(self)

Raising version of empty/1.

get_step!(self)

Raising version of get_step/1.

get_step(self)

gpuMat!()

Raising version of gpuMat/0.

gpuMat!(arr)

Raising version of gpuMat/1.

gpuMat!(m, roi)

Raising version of gpuMat/2.

gpuMat!(m, rowRange, colRange)

Raising version of gpuMat/3.

gpuMat!(rows, cols, type, s)

Raising version of gpuMat/4.

gpuMat!(rows, cols, type, s, opts)

Raising version of gpuMat/5.

isContinuous!(self)

Raising version of isContinuous/1.

locateROI!(self, wholeSize, ofs)

Raising version of locateROI/3.

release!(self)

Raising version of release/1.

reshape!(self, cn)

Raising version of reshape/2.

reshape!(self, cn, opts)

Raising version of reshape/3.

row!(self, y)

Raising version of row/2.

rowRange!(self, r)

Raising version of rowRange/2.

rowRange!(self, startrow, endrow)

Raising version of rowRange/3.

setDefaultAllocator!(allocator)

Raising version of setDefaultAllocator/1.

setTo!(self, s)

Raising version of setTo/2.

setTo!(self, s, mask)

Raising version of setTo/3.

setTo!(self, s, mask, stream)

Raising version of setTo/4.

size!(self)

Raising version of size/1.

step1!(self)

Raising version of step1/1.

swap!(self, mat)

Raising version of swap/2.

type!(self)

Raising version of type/1.

updateContinuityFlag!(self)

Raising version of updateContinuityFlag/1.

upload!(self, arr)

Raising version of upload/2.

upload!(self, arr, stream)

Raising version of upload/3.