Evision.CUDA.GpuMat (Evision v0.1.9)

cv.cuda

adjustROI(self, dtop, dbottom, dleft, dright)

Positional Arguments

dtop: int

dbottom: int

dleft: int

dright: int

Python prototype (for reference):

adjustROI(dtop, dbottom, dleft, dright) -> retval

assignTo(self, m)

Positional Arguments

m: GpuMat

Keyword Arguments

type: int.

Python prototype (for reference):

assignTo(m[, type]) -> None

assignTo(self, m, opts)

Positional Arguments

m: GpuMat

Keyword Arguments

type: int.

Python prototype (for reference):

assignTo(m[, type]) -> None

channels(self)

Python prototype (for reference):

channels() -> retval

clone(self)

Python prototype (for reference):

clone() -> retval

col(self, x)

Positional Arguments

x: int

Python prototype (for reference):

col(x) -> retval

colRange(self, r)

Positional Arguments

r: Range

Python prototype (for reference):

colRange(r) -> retval

colRange(self, startcol, endcol)

Positional Arguments

startcol: int

endcol: int

Python prototype (for reference):

colRange(startcol, endcol) -> retval

convertTo(self, rtype)

Positional Arguments

rtype: int

Return

dst: Evision.Mat.

Python prototype (for reference):

convertTo(rtype[, dst]) -> dst

convertTo(self, rtype, opts)

Variant 1:

Positional Arguments

rtype: int

alpha: double

Keyword Arguments

beta: double.

Return

dst: Evision.Mat.

Python prototype (for reference):

convertTo(rtype, alpha[, dst[, beta]]) -> dst

Variant 2:

Positional Arguments

rtype: int

stream: Stream

Return

dst: Evision.Mat.

Python prototype (for reference):

convertTo(rtype, stream[, dst]) -> dst

Variant 3:

Positional Arguments

rtype: int

Return

dst: Evision.Mat.

Python prototype (for reference):

convertTo(rtype[, dst]) -> dst

convertTo(self, rtype, alpha, opts)

Variant 1:

Positional Arguments

rtype: int

alpha: double

stream: Stream

Return

dst: Evision.Mat.

Python prototype (for reference):

convertTo(rtype, alpha, stream[, dst]) -> dst

Variant 2:

Positional Arguments

rtype: int

alpha: double

Keyword Arguments

beta: double.

Return

dst: Evision.Mat.

Python prototype (for reference):

convertTo(rtype, alpha[, dst[, beta]]) -> dst

Variant 3:

Positional Arguments

rtype: int

stream: Stream

Return

dst: Evision.Mat.

Python prototype (for reference):

convertTo(rtype, stream[, dst]) -> dst

convertTo(self, rtype, alpha, stream, opts)

Variant 1:

Positional Arguments

rtype: int

alpha: double

beta: double

stream: Stream

Return

dst: Evision.Mat.

Python prototype (for reference):

convertTo(rtype, alpha, beta, stream[, dst]) -> dst

Variant 2:

Positional Arguments

rtype: int

alpha: double

stream: Stream

Return

dst: Evision.Mat.

Python prototype (for reference):

convertTo(rtype, alpha, stream[, dst]) -> dst

convertTo(self, rtype, alpha, beta, stream, opts)

Positional Arguments

rtype: int

alpha: double

beta: double

stream: Stream

Return

dst: Evision.Mat.

Python prototype (for reference):

convertTo(rtype, alpha, beta, stream[, dst]) -> dst

copyTo(self)

##### Return

dst: Evision.Mat.

Python prototype (for reference):

copyTo([, dst]) -> dst

copyTo(self, mask)

Positional Arguments

mask: Evision.Mat

Return

dst: Evision.Mat.

Python prototype (for reference):

copyTo(mask[, dst]) -> dst

copyTo(self, mask, opts)

Variant 1:

Positional Arguments

mask: Evision.Mat

stream: Stream

Return

dst: Evision.Mat.

Python prototype (for reference):

copyTo(mask, stream[, dst]) -> dst

Variant 2:

Positional Arguments

mask: Evision.Mat

Return

dst: Evision.Mat.

Python prototype (for reference):

copyTo(mask[, dst]) -> dst

copyTo(self, mask, stream, opts)

Positional Arguments

mask: Evision.Mat

stream: Stream

Return

dst: Evision.Mat.

Python prototype (for reference):

copyTo(mask, stream[, dst]) -> dst

create(self, size, type)

Positional Arguments

size: Size

type: int

Python prototype (for reference):

create(size, type) -> None

create(self, rows, cols, type)

Positional Arguments

rows: int

cols: int

type: int

Python prototype (for reference):

create(rows, cols, type) -> None

cuda_GpuMat()

##### Keyword Arguments

allocator: GpuMat_Allocator*.

Python prototype (for reference):

GpuMat([, allocator]) -> <cuda_GpuMat object>

cuda_GpuMat(opts)

Variant 1:

Positional Arguments

arr: Evision.Mat

Keyword Arguments

allocator: GpuMat_Allocator*.

Python prototype (for reference):

GpuMat(arr[, allocator]) -> <cuda_GpuMat object>

Variant 2:

##### Keyword Arguments

allocator: GpuMat_Allocator*.

Python prototype (for reference):

GpuMat([, allocator]) -> <cuda_GpuMat object>

cuda_GpuMat(arr, opts)

Variant 1:

Positional Arguments

m: GpuMat

roi: Rect

Python prototype (for reference):

GpuMat(m, roi) -> <cuda_GpuMat object>

Variant 2:

Positional Arguments

size: Size

type: int

Keyword Arguments

allocator: GpuMat_Allocator*.

Python prototype (for reference):

GpuMat(size, type[, allocator]) -> <cuda_GpuMat object>

Variant 3:

Positional Arguments

arr: Evision.Mat

Keyword Arguments

allocator: GpuMat_Allocator*.

Python prototype (for reference):

GpuMat(arr[, allocator]) -> <cuda_GpuMat object>

cuda_GpuMat(size, type, opts)

Variant 1:

Positional Arguments

m: GpuMat

rowRange: Range

colRange: Range

Python prototype (for reference):

GpuMat(m, rowRange, colRange) -> <cuda_GpuMat object>

Variant 2:

Positional Arguments

size: Size

type: int

s: Scalar

Keyword Arguments

allocator: GpuMat_Allocator*.

Python prototype (for reference):

GpuMat(size, type, s[, allocator]) -> <cuda_GpuMat object>

Variant 3:

Positional Arguments

rows: int

cols: int

type: int

Keyword Arguments

allocator: GpuMat_Allocator*.

Python prototype (for reference):

GpuMat(rows, cols, type[, allocator]) -> <cuda_GpuMat object>

Variant 4:

Positional Arguments

size: Size

type: int

Keyword Arguments

allocator: GpuMat_Allocator*.

Python prototype (for reference):

GpuMat(size, type[, allocator]) -> <cuda_GpuMat object>

cuda_GpuMat(size, type, s, opts)

Variant 1:

Positional Arguments

rows: int

cols: int

type: int

s: Scalar

Keyword Arguments

allocator: GpuMat_Allocator*.

Python prototype (for reference):

GpuMat(rows, cols, type, s[, allocator]) -> <cuda_GpuMat object>

Variant 2:

Positional Arguments

size: Size

type: int

s: Scalar

Keyword Arguments

allocator: GpuMat_Allocator*.

Python prototype (for reference):

GpuMat(size, type, s[, allocator]) -> <cuda_GpuMat object>

Variant 3:

Positional Arguments

rows: int

cols: int

type: int

Keyword Arguments

allocator: GpuMat_Allocator*.

Python prototype (for reference):

GpuMat(rows, cols, type[, allocator]) -> <cuda_GpuMat object>

cuda_GpuMat(rows, cols, type, s, opts)

Positional Arguments

rows: int

cols: int

type: int

s: Scalar

Keyword Arguments

allocator: GpuMat_Allocator*.

Python prototype (for reference):

GpuMat(rows, cols, type, s[, allocator]) -> <cuda_GpuMat object>

cudaPtr(self)

Python prototype (for reference):

cudaPtr() -> retval

defaultAllocator()

Python prototype (for reference):

defaultAllocator() -> retval

depth(self)

Python prototype (for reference):

depth() -> retval

download(self)

Performs data download from GpuMat (Blocking call)

Return

dst: Evision.Mat.

This function copies data from device memory to host memory. As being a blocking call, it is guaranteed that the copy operation is finished when this function returns.

Python prototype (for reference):

download([, dst]) -> dst

download(self, stream)

Performs data download from GpuMat (Non-Blocking call)

Positional Arguments

stream: Stream

Return

dst: Evision.Mat.

This function copies data from device memory to host memory. As being a non-blocking call, this function may return even if the copy operation is not finished. The copy operation may be overlapped with operations in other non-default streams if \p stream is not the default stream and \p dst is HostMem allocated with HostMem::PAGE_LOCKED option.

Python prototype (for reference):

download(stream[, dst]) -> dst

download(self, stream, opts)

Performs data download from GpuMat (Non-Blocking call)

Positional Arguments

stream: Stream

Return

dst: Evision.Mat.

This function copies data from device memory to host memory. As being a non-blocking call, this function may return even if the copy operation is not finished. The copy operation may be overlapped with operations in other non-default streams if \p stream is not the default stream and \p dst is HostMem allocated with HostMem::PAGE_LOCKED option.

Python prototype (for reference):

download(stream[, dst]) -> dst

elemSize1(self)

Python prototype (for reference):

elemSize1() -> retval

elemSize(self)

Python prototype (for reference):

elemSize() -> retval

empty(self)

Python prototype (for reference):

empty() -> retval

isContinuous(self)

Python prototype (for reference):

isContinuous() -> retval

locateROI(self, wholeSize, ofs)

Positional Arguments

wholeSize: Size

ofs: Point

Python prototype (for reference):

locateROI(wholeSize, ofs) -> None

release(self)

Python prototype (for reference):

release() -> None

reshape(self, cn)

Positional Arguments

cn: int

Keyword Arguments

rows: int.

Python prototype (for reference):

reshape(cn[, rows]) -> retval

reshape(self, cn, opts)

Positional Arguments

cn: int

Keyword Arguments

rows: int.

Python prototype (for reference):

reshape(cn[, rows]) -> retval

row(self, y)

Positional Arguments

y: int

Python prototype (for reference):

row(y) -> retval

rowRange(self, r)

Positional Arguments

r: Range

Python prototype (for reference):

rowRange(r) -> retval

rowRange(self, startrow, endrow)

Positional Arguments

startrow: int

endrow: int

Python prototype (for reference):

rowRange(startrow, endrow) -> retval

setDefaultAllocator(allocator)

Positional Arguments

allocator: GpuMat_Allocator*

Python prototype (for reference):

setDefaultAllocator(allocator) -> None

setTo(self, s)

Positional Arguments

s: Scalar

Python prototype (for reference):

setTo(s) -> retval

setTo(self, s, mask)

Positional Arguments

s: Scalar

mask: Evision.Mat

Python prototype (for reference):

setTo(s, mask) -> retval

setTo(self, s, mask, stream)

Positional Arguments

s: Scalar

mask: Evision.Mat

stream: Stream

Python prototype (for reference):

setTo(s, mask, stream) -> retval

size(self)

Python prototype (for reference):

size() -> retval

step1(self)

Python prototype (for reference):

step1() -> retval

swap(self, mat)

Positional Arguments

mat: GpuMat

Python prototype (for reference):

swap(mat) -> None

type(self)

Python prototype (for reference):

type() -> retval

updateContinuityFlag(self)

Python prototype (for reference):

updateContinuityFlag() -> None

upload(self, arr)

Performs data upload to GpuMat (Blocking call)

Positional Arguments

arr: Evision.Mat

This function copies data from host memory to device memory. As being a blocking call, it is guaranteed that the copy operation is finished when this function returns.

Python prototype (for reference):

upload(arr) -> None

upload(self, arr, stream)

Performs data upload to GpuMat (Non-Blocking call)

Positional Arguments

arr: Evision.Mat

stream: Stream

This function copies data from host memory to device memory. As being a non-blocking call, this function may return even if the copy operation is not finished. The copy operation may be overlapped with operations in other non-default streams if \p stream is not the default stream and \p dst is HostMem allocated with HostMem::PAGE_LOCKED option.

Python prototype (for reference):

upload(arr, stream) -> None

Functions

adjustROI!(self, dtop, dbottom, dleft, dright)

Raising version of adjustROI/5.

assignTo!(self, m)

Raising version of assignTo/2.

assignTo!(self, m, opts)

Raising version of assignTo/3.

channels!(self)

Raising version of channels/1.

clone!(self)

Raising version of clone/1.

col!(self, x)

Raising version of col/2.

colRange!(self, r)

Raising version of colRange/2.

colRange!(self, startcol, endcol)

Raising version of colRange/3.

convertTo!(self, rtype)

Raising version of convertTo/2.

convertTo!(self, rtype, alpha)

Raising version of convertTo/3.

convertTo!(self, rtype, alpha, stream)

Raising version of convertTo/4.

convertTo!(self, rtype, alpha, beta, stream)

Raising version of convertTo/5.

convertTo!(self, rtype, alpha, beta, stream, opts)

Raising version of convertTo/6.

copyTo!(self)

Raising version of copyTo/1.

copyTo!(self, mask)

Raising version of copyTo/2.

copyTo!(self, mask, stream)

Raising version of copyTo/3.

copyTo!(self, mask, stream, opts)

Raising version of copyTo/4.

create!(self, size, type)

Raising version of create/3.

create!(self, rows, cols, type)

Raising version of create/4.

cuda_GpuMat!()

Raising version of cuda_GpuMat/0.

cuda_GpuMat!(arr)

Raising version of cuda_GpuMat/1.

cuda_GpuMat!(m, roi)

Raising version of cuda_GpuMat/2.

cuda_GpuMat!(m, rowRange, colRange)

Raising version of cuda_GpuMat/3.

cuda_GpuMat!(rows, cols, type, s)

Raising version of cuda_GpuMat/4.

cuda_GpuMat!(rows, cols, type, s, opts)

Raising version of cuda_GpuMat/5.

cudaPtr!(self)

Raising version of cudaPtr/1.

defaultAllocator!()

Raising version of defaultAllocator/0.

depth!(self)

Raising version of depth/1.

download!(self)

Raising version of download/1.

download!(self, stream)

Raising version of download/2.

download!(self, stream, opts)

Raising version of download/3.

elemSize1!(self)

Raising version of elemSize1/1.

elemSize!(self)

Raising version of elemSize/1.

empty!(self)

Raising version of empty/1.

get_step!(self)

Raising version of get_step/1.

get_step(self)

isContinuous!(self)

Raising version of isContinuous/1.

locateROI!(self, wholeSize, ofs)

Raising version of locateROI/3.

release!(self)

Raising version of release/1.

reshape!(self, cn)

Raising version of reshape/2.

reshape!(self, cn, opts)

Raising version of reshape/3.

row!(self, y)

Raising version of row/2.

rowRange!(self, r)

Raising version of rowRange/2.

rowRange!(self, startrow, endrow)

Raising version of rowRange/3.

setDefaultAllocator!(allocator)

Raising version of setDefaultAllocator/1.

setTo!(self, s)

Raising version of setTo/2.

setTo!(self, s, mask)

Raising version of setTo/3.

setTo!(self, s, mask, stream)

Raising version of setTo/4.

size!(self)

Raising version of size/1.

step1!(self)

Raising version of step1/1.

swap!(self, mat)

Raising version of swap/2.

type!(self)

Raising version of type/1.

updateContinuityFlag!(self)

Raising version of updateContinuityFlag/1.

upload!(self, arr)

Raising version of upload/2.

upload!(self, arr, stream)

Raising version of upload/3.