Evision.DNN (Evision v0.1.7-dev)

Functions

blobFromImage!(image)

Raising version of blobFromImage/1.

blobFromImage!(image, opts)

Raising version of blobFromImage/2.

blobFromImage(image)

Positional Arguments

image: UMat.

Keyword Arguments

scalefactor: double.

size: Size.

mean: Scalar.

swapRB: bool.

crop: bool.

ddepth: int.

Creates 4-dimensional blob from image. Optionally resizes and crops @p image from center,

subtract @p mean values, scales values by @p scalefactor, swap Blue and Red channels.
  to be in (mean-R, mean-G, mean-B) order if @p image has BGR ordering and @p swapRB is true.
  in 3-channel image is necessary.

@details if @p crop is true, input image is resized so one side after resize is equal to corresponding

dimension in @p size and another one is equal or larger. Then, crop from the center is performed.
If @p crop is false, direct resize without cropping and preserving aspect ratio is performed.

@returns 4-dimensional Mat with NCHW dimensions order.

Python prototype (for reference):

blobFromImage(image[, scalefactor[, size[, mean[, swapRB[, crop[, ddepth]]]]]]) -> retval

blobFromImage(image, opts)

Positional Arguments

image: UMat.

Keyword Arguments

scalefactor: double.

size: Size.

mean: Scalar.

swapRB: bool.

crop: bool.

ddepth: int.

Creates 4-dimensional blob from image. Optionally resizes and crops @p image from center,

subtract @p mean values, scales values by @p scalefactor, swap Blue and Red channels.
  to be in (mean-R, mean-G, mean-B) order if @p image has BGR ordering and @p swapRB is true.
  in 3-channel image is necessary.

@details if @p crop is true, input image is resized so one side after resize is equal to corresponding

dimension in @p size and another one is equal or larger. Then, crop from the center is performed.
If @p crop is false, direct resize without cropping and preserving aspect ratio is performed.

@returns 4-dimensional Mat with NCHW dimensions order.

Python prototype (for reference):

blobFromImage(image[, scalefactor[, size[, mean[, swapRB[, crop[, ddepth]]]]]]) -> retval

blobFromImages!(images)

Raising version of blobFromImages/1.

blobFromImages!(images, opts)

Raising version of blobFromImages/2.

blobFromImages(images)

Positional Arguments

images: vector_UMat.

Keyword Arguments

scalefactor: double.

size: Size.

mean: Scalar.

swapRB: bool.

crop: bool.

ddepth: int.

Creates 4-dimensional blob from series of images. Optionally resizes and

crops @p images from center, subtract @p mean values, scales values by @p scalefactor,
swap Blue and Red channels.
  to be in (mean-R, mean-G, mean-B) order if @p image has BGR ordering and @p swapRB is true.
  in 3-channel image is necessary.

@details if @p crop is true, input image is resized so one side after resize is equal to corresponding

dimension in @p size and another one is equal or larger. Then, crop from the center is performed.
If @p crop is false, direct resize without cropping and preserving aspect ratio is performed.

@returns 4-dimensional Mat with NCHW dimensions order.

Python prototype (for reference):

blobFromImages(images[, scalefactor[, size[, mean[, swapRB[, crop[, ddepth]]]]]]) -> retval

blobFromImages(images, opts)

Positional Arguments

images: vector_UMat.

Keyword Arguments

scalefactor: double.

size: Size.

mean: Scalar.

swapRB: bool.

crop: bool.

ddepth: int.

Creates 4-dimensional blob from series of images. Optionally resizes and

crops @p images from center, subtract @p mean values, scales values by @p scalefactor,
swap Blue and Red channels.
  to be in (mean-R, mean-G, mean-B) order if @p image has BGR ordering and @p swapRB is true.
  in 3-channel image is necessary.

@details if @p crop is true, input image is resized so one side after resize is equal to corresponding

dimension in @p size and another one is equal or larger. Then, crop from the center is performed.
If @p crop is false, direct resize without cropping and preserving aspect ratio is performed.

@returns 4-dimensional Mat with NCHW dimensions order.

Python prototype (for reference):

blobFromImages(images[, scalefactor[, size[, mean[, swapRB[, crop[, ddepth]]]]]]) -> retval

getAvailableTargets!(be)

Raising version of getAvailableTargets/1.

getAvailableTargets(be)

Python prototype (for reference):

getAvailableTargets(be) -> retval

imagesFromBlob!(blob_)

Raising version of imagesFromBlob/1.

imagesFromBlob!(blob_, opts)

Raising version of imagesFromBlob/2.

imagesFromBlob(blob_)

Positional Arguments

blob_: Mat.

Keyword Arguments

images_: vector_UMat.

Parse a 4D blob and output the images it contains as 2D arrays through a simpler data structure

(std::vector<cv::Mat>).
  which you would like to extract the images.
  (CV_32F). They are non normalized neither mean added. The number of returned images equals the first dimension
  of the blob (batch size). Every image has a number of channels equals to the second dimension of the blob (depth).

Python prototype (for reference):

imagesFromBlob(blob_[, images_]) -> images_

imagesFromBlob(blob_, opts)

Positional Arguments

blob_: Mat.

Keyword Arguments

images_: vector_UMat.

Parse a 4D blob and output the images it contains as 2D arrays through a simpler data structure

(std::vector<cv::Mat>).
  which you would like to extract the images.
  (CV_32F). They are non normalized neither mean added. The number of returned images equals the first dimension
  of the blob (batch size). Every image has a number of channels equals to the second dimension of the blob (depth).

Python prototype (for reference):

imagesFromBlob(blob_[, images_]) -> images_

nmsBoxes!(bboxes, scores, score_threshold, nms_threshold)

Raising version of nmsBoxes/4.

nmsBoxes!(bboxes, scores, score_threshold, nms_threshold, opts)

Raising version of nmsBoxes/5.

nmsBoxes(bboxes, scores, score_threshold, nms_threshold)

Positional Arguments

bboxes: vector_Rect2d.

scores: vector_float.

score_threshold: float.

nms_threshold: float.

Keyword Arguments

eta: float.

top_k: int.

Performs non maximum suppression given boxes and corresponding scores.

Python prototype (for reference):

NMSBoxes(bboxes, scores, score_threshold, nms_threshold[, eta[, top_k]]) -> indices

nmsBoxes(bboxes, scores, score_threshold, nms_threshold, opts)

Positional Arguments

bboxes: vector_Rect2d.

scores: vector_float.

score_threshold: float.

nms_threshold: float.

Keyword Arguments

eta: float.

top_k: int.

Performs non maximum suppression given boxes and corresponding scores.

Python prototype (for reference):

NMSBoxes(bboxes, scores, score_threshold, nms_threshold[, eta[, top_k]]) -> indices

nmsBoxesRotated!(bboxes, scores, score_threshold, nms_threshold)

Raising version of nmsBoxesRotated/4.

nmsBoxesRotated!(bboxes, scores, score_threshold, nms_threshold, opts)

Raising version of nmsBoxesRotated/5.

nmsBoxesRotated(bboxes, scores, score_threshold, nms_threshold)

Positional Arguments

bboxes: vector_RotatedRect

scores: vector_float

score_threshold: float

nms_threshold: float

Keyword Arguments

eta: float.

top_k: int.

Python prototype (for reference):

NMSBoxesRotated(bboxes, scores, score_threshold, nms_threshold[, eta[, top_k]]) -> indices

nmsBoxesRotated(bboxes, scores, score_threshold, nms_threshold, opts)

Positional Arguments

bboxes: vector_RotatedRect

scores: vector_float

score_threshold: float

nms_threshold: float

Keyword Arguments

eta: float.

top_k: int.

Python prototype (for reference):

NMSBoxesRotated(bboxes, scores, score_threshold, nms_threshold[, eta[, top_k]]) -> indices

readNet!(model)

Raising version of readNet/1.

readNet!(framework, bufferModel)

Raising version of readNet/2.

readNet!(framework, bufferModel, opts)

Raising version of readNet/3.

readNet(model)

Positional Arguments

model: String.

Keyword Arguments

config: String.

framework: String.

Read deep learning network represented in one of the supported formats.

                  extensions are expected for models from different frameworks:
                  * `*.caffemodel` (Caffe, http://caffe.berkeleyvision.org/)
                  * `*.pb` (TensorFlow, https://www.tensorflow.org/)
                  * `*.t7` | `*.net` (Torch, http://torch.ch/)
                  * `*.weights` (Darknet, https://pjreddie.com/darknet/)
                  * `*.bin` (DLDT, https://software.intel.com/openvino-toolkit)
                  * `*.onnx` (ONNX, https://onnx.ai/)
                   file with the following extensions:
                  * `*.prototxt` (Caffe, http://caffe.berkeleyvision.org/)
                  * `*.pbtxt` (TensorFlow, https://www.tensorflow.org/)
                  * `*.cfg` (Darknet, https://pjreddie.com/darknet/)
                  * `*.xml` (DLDT, https://software.intel.com/openvino-toolkit)

@returns Net object. This function automatically detects an origin framework of trained model and calls an appropriate function such @ref readNetFromCaffe, @ref readNetFromTensorflow, @ref readNetFromTorch or @ref readNetFromDarknet. An order of @p model and @p config arguments does not matter.

Python prototype (for reference):

readNet(model[, config[, framework]]) -> retval

readNet(model, opts)

Variaint 1:

Positional Arguments

framework: String.

bufferModel: vector_uchar.

Keyword Arguments

bufferConfig: vector_uchar.

Read deep learning network represented in one of the supported formats. @details This is an overloaded member function, provided for convenience.

        It differs from the above function only in what argument(s) it accepts.

@returns Net object.

Python prototype (for reference):

readNet(framework, bufferModel[, bufferConfig]) -> retval

Variaint 2:

Positional Arguments

model: String.

Keyword Arguments

config: String.

framework: String.

Read deep learning network represented in one of the supported formats.

                  extensions are expected for models from different frameworks:
                  * `*.caffemodel` (Caffe, http://caffe.berkeleyvision.org/)
                  * `*.pb` (TensorFlow, https://www.tensorflow.org/)
                  * `*.t7` | `*.net` (Torch, http://torch.ch/)
                  * `*.weights` (Darknet, https://pjreddie.com/darknet/)
                  * `*.bin` (DLDT, https://software.intel.com/openvino-toolkit)
                  * `*.onnx` (ONNX, https://onnx.ai/)
                   file with the following extensions:
                  * `*.prototxt` (Caffe, http://caffe.berkeleyvision.org/)
                  * `*.pbtxt` (TensorFlow, https://www.tensorflow.org/)
                  * `*.cfg` (Darknet, https://pjreddie.com/darknet/)
                  * `*.xml` (DLDT, https://software.intel.com/openvino-toolkit)

@returns Net object. This function automatically detects an origin framework of trained model and calls an appropriate function such @ref readNetFromCaffe, @ref readNetFromTensorflow, @ref readNetFromTorch or @ref readNetFromDarknet. An order of @p model and @p config arguments does not matter.

Python prototype (for reference):

readNet(model[, config[, framework]]) -> retval

readNet(framework, bufferModel, opts)

Positional Arguments

framework: String.

bufferModel: vector_uchar.

Keyword Arguments

bufferConfig: vector_uchar.

Read deep learning network represented in one of the supported formats. @details This is an overloaded member function, provided for convenience.

        It differs from the above function only in what argument(s) it accepts.

@returns Net object.

Python prototype (for reference):

readNet(framework, bufferModel[, bufferConfig]) -> retval

readNetFromCaffe!(bufferProto)

Raising version of readNetFromCaffe/1.

readNetFromCaffe!(bufferProto, opts)

Raising version of readNetFromCaffe/2.

readNetFromCaffe(bufferProto)

Variaint 1:

Positional Arguments

bufferProto: vector_uchar.

Keyword Arguments

bufferModel: vector_uchar.

Reads a network model stored in Caffe model in memory. @returns Net object.

Python prototype (for reference):

readNetFromCaffe(bufferProto[, bufferModel]) -> retval

Variaint 2:

Positional Arguments

prototxt: String.

Keyword Arguments

caffeModel: String.

Reads a network model stored in <a href="http://caffe.berkeleyvision.org">Caffe</a> framework's format. @returns Net object.

Python prototype (for reference):

readNetFromCaffe(prototxt[, caffeModel]) -> retval

readNetFromCaffe(bufferProto, opts)

Variaint 1:

Positional Arguments

bufferProto: vector_uchar.

Keyword Arguments

bufferModel: vector_uchar.

Reads a network model stored in Caffe model in memory. @returns Net object.

Python prototype (for reference):

readNetFromCaffe(bufferProto[, bufferModel]) -> retval

Variaint 2:

Positional Arguments

prototxt: String.

Keyword Arguments

caffeModel: String.

Reads a network model stored in <a href="http://caffe.berkeleyvision.org">Caffe</a> framework's format. @returns Net object.

Python prototype (for reference):

readNetFromCaffe(prototxt[, caffeModel]) -> retval

readNetFromDarknet!(bufferCfg)

Raising version of readNetFromDarknet/1.

readNetFromDarknet!(bufferCfg, opts)

Raising version of readNetFromDarknet/2.

readNetFromDarknet(bufferCfg)

Variaint 1:

Positional Arguments

bufferCfg: vector_uchar.

Keyword Arguments

bufferModel: vector_uchar.

Reads a network model stored in <a href="https://pjreddie.com/darknet/">Darknet</a> model files. @returns Net object.

Python prototype (for reference):

readNetFromDarknet(bufferCfg[, bufferModel]) -> retval

Variaint 2:

Positional Arguments

cfgFile: String.

Keyword Arguments

darknetModel: String.

Reads a network model stored in <a href="https://pjreddie.com/darknet/">Darknet</a> model files. @returns Network object that ready to do forward, throw an exception in failure cases. @returns Net object.

Python prototype (for reference):

readNetFromDarknet(cfgFile[, darknetModel]) -> retval

readNetFromDarknet(bufferCfg, opts)

Variaint 1:

Positional Arguments

bufferCfg: vector_uchar.

Keyword Arguments

bufferModel: vector_uchar.

Reads a network model stored in <a href="https://pjreddie.com/darknet/">Darknet</a> model files. @returns Net object.

Python prototype (for reference):

readNetFromDarknet(bufferCfg[, bufferModel]) -> retval

Variaint 2:

Positional Arguments

cfgFile: String.

Keyword Arguments

darknetModel: String.

Reads a network model stored in <a href="https://pjreddie.com/darknet/">Darknet</a> model files. @returns Network object that ready to do forward, throw an exception in failure cases. @returns Net object.

Python prototype (for reference):

readNetFromDarknet(cfgFile[, darknetModel]) -> retval

readNetFromModelOptimizer!(bufferModelConfig, bufferWeights)

Raising version of readNetFromModelOptimizer/2.

readNetFromModelOptimizer(bufferModelConfig, bufferWeights)

Variaint 1:

Load a network from Intel's Model Optimizer intermediate representation. @returns Net object.

Networks imported from Intel's Model Optimizer are launched in Intel's Inference Engine
backend.

Python prototype (for reference):

readNetFromModelOptimizer(bufferModelConfig, bufferWeights) -> retval

Variaint 2:

Load a network from Intel's Model Optimizer intermediate representation. @returns Net object.

Networks imported from Intel's Model Optimizer are launched in Intel's Inference Engine
backend.

Python prototype (for reference):

readNetFromModelOptimizer(xml, bin) -> retval

readNetFromONNX!(buffer)

Raising version of readNetFromONNX/1.

readNetFromONNX(buffer)

Variaint 1:

Reads a network model from <a href="https://onnx.ai/">ONNX</a>

       in-memory buffer.

@returns Network object that ready to do forward, throw an exception

      in failure cases.

Python prototype (for reference):

readNetFromONNX(buffer) -> retval

Variaint 2:

Reads a network model <a href="https://onnx.ai/">ONNX</a>. @returns Network object that ready to do forward, throw an exception in failure cases.

Python prototype (for reference):

readNetFromONNX(onnxFile) -> retval

readNetFromTensorflow!(bufferModel)

Raising version of readNetFromTensorflow/1.

readNetFromTensorflow!(bufferModel, opts)

Raising version of readNetFromTensorflow/2.

readNetFromTensorflow(bufferModel)

Variaint 1:

Positional Arguments

bufferModel: vector_uchar.

Keyword Arguments

bufferConfig: vector_uchar.

Reads a network model stored in <a href="https://www.tensorflow.org/">TensorFlow</a> framework's format. @returns Net object.

Python prototype (for reference):

readNetFromTensorflow(bufferModel[, bufferConfig]) -> retval

Variaint 2:

Positional Arguments

model: String.

Keyword Arguments

config: String.

Reads a network model stored in <a href="https://www.tensorflow.org/">TensorFlow</a> framework's format.

               Resulting Net object is built by text graph using weights from a binary one that
               let us make it more flexible.

@returns Net object.

Python prototype (for reference):

readNetFromTensorflow(model[, config]) -> retval

readNetFromTensorflow(bufferModel, opts)

Variaint 1:

Positional Arguments

bufferModel: vector_uchar.

Keyword Arguments

bufferConfig: vector_uchar.

Reads a network model stored in <a href="https://www.tensorflow.org/">TensorFlow</a> framework's format. @returns Net object.

Python prototype (for reference):

readNetFromTensorflow(bufferModel[, bufferConfig]) -> retval

Variaint 2:

Positional Arguments

model: String.

Keyword Arguments

config: String.

Reads a network model stored in <a href="https://www.tensorflow.org/">TensorFlow</a> framework's format.

               Resulting Net object is built by text graph using weights from a binary one that
               let us make it more flexible.

@returns Net object.

Python prototype (for reference):

readNetFromTensorflow(model[, config]) -> retval

readNetFromTorch!(model)

Raising version of readNetFromTorch/1.

readNetFromTorch!(model, opts)

Raising version of readNetFromTorch/2.

readNetFromTorch(model)

Positional Arguments

model: String.

Keyword Arguments

isBinary: bool.

evaluate: bool.

Reads a network model stored in <a href="http://torch.ch">Torch7</a> framework's format. @returns Net object. Note: Ascii mode of Torch serializer is more preferable, because binary mode extensively use long type of C language,

which has various bit-length on different systems.

The loading file must contain serialized <a href="https://github.com/torch/nn/blob/master/doc/module.md">nn.Module</a> object with importing network. Try to eliminate a custom objects from serialazing data to avoid importing errors. List of supported layers (i.e. object instances derived from Torch nn.Module class):

nn.Sequential

nn.Parallel

nn.Concat

nn.Linear

nn.SpatialConvolution

nn.SpatialMaxPooling, nn.SpatialAveragePooling

nn.ReLU, nn.TanH, nn.Sigmoid

nn.Reshape

nn.SoftMax, nn.LogSoftMax Also some equivalents of these classes from cunn, cudnn, and fbcunn may be successfully imported.

Python prototype (for reference):

readNetFromTorch(model[, isBinary[, evaluate]]) -> retval

readNetFromTorch(model, opts)

Positional Arguments

model: String.

Keyword Arguments

isBinary: bool.

evaluate: bool.

Reads a network model stored in <a href="http://torch.ch">Torch7</a> framework's format. @returns Net object. Note: Ascii mode of Torch serializer is more preferable, because binary mode extensively use long type of C language,

which has various bit-length on different systems.

The loading file must contain serialized <a href="https://github.com/torch/nn/blob/master/doc/module.md">nn.Module</a> object with importing network. Try to eliminate a custom objects from serialazing data to avoid importing errors. List of supported layers (i.e. object instances derived from Torch nn.Module class):

nn.Sequential

nn.Parallel

nn.Concat

nn.Linear

nn.SpatialConvolution

nn.SpatialMaxPooling, nn.SpatialAveragePooling

nn.ReLU, nn.TanH, nn.Sigmoid

nn.Reshape

nn.SoftMax, nn.LogSoftMax Also some equivalents of these classes from cunn, cudnn, and fbcunn may be successfully imported.

Python prototype (for reference):

readNetFromTorch(model[, isBinary[, evaluate]]) -> retval

readTensorFromONNX!(path)

Raising version of readTensorFromONNX/1.

readTensorFromONNX(path)

Creates blob from .pb file. @returns Mat.

Python prototype (for reference):

readTensorFromONNX(path) -> retval

readTorchBlob!(filename)

Raising version of readTorchBlob/1.

readTorchBlob!(filename, opts)

Raising version of readTorchBlob/2.

readTorchBlob(filename)

Positional Arguments

filename: String

Keyword Arguments

isBinary: bool.

Loads blob which was serialized as torch.Tensor object of Torch7 framework. @warning This function has the same limitations as readNetFromTorch().

Python prototype (for reference):

readTorchBlob(filename[, isBinary]) -> retval

readTorchBlob(filename, opts)

Positional Arguments

filename: String

Keyword Arguments

isBinary: bool.

Loads blob which was serialized as torch.Tensor object of Torch7 framework. @warning This function has the same limitations as readNetFromTorch().

Python prototype (for reference):

readTorchBlob(filename[, isBinary]) -> retval

shrinkCaffeModel!(src, dst)

Raising version of shrinkCaffeModel/2.

shrinkCaffeModel!(src, dst, opts)

Raising version of shrinkCaffeModel/3.

shrinkCaffeModel(src, dst)

Positional Arguments

src: String.

dst: String.

Keyword Arguments

layersTypes: vector_String.

Convert all weights of Caffe network to half precision floating point.

            precision floating point weights (usually has `.caffemodel` extension).
                    By default, converts only Convolutional and Fully-Connected layers'
                    weights.

Note: Shrinked model has no origin float32 weights so it can't be used

     in origin Caffe framework anymore. However the structure of data
     is taken from NVidia's Caffe fork: https://github.com/NVIDIA/caffe.
     So the resulting model may be used there.

Python prototype (for reference):

shrinkCaffeModel(src, dst[, layersTypes]) -> None

shrinkCaffeModel(src, dst, opts)

Positional Arguments

src: String.

dst: String.

Keyword Arguments

layersTypes: vector_String.

Convert all weights of Caffe network to half precision floating point.

            precision floating point weights (usually has `.caffemodel` extension).
                    By default, converts only Convolutional and Fully-Connected layers'
                    weights.

Note: Shrinked model has no origin float32 weights so it can't be used

     in origin Caffe framework anymore. However the structure of data
     is taken from NVidia's Caffe fork: https://github.com/NVIDIA/caffe.
     So the resulting model may be used there.

Python prototype (for reference):

shrinkCaffeModel(src, dst[, layersTypes]) -> None

softNMSBoxes!(bboxes, scores, score_threshold, nms_threshold)

Raising version of softNMSBoxes/4.

softNMSBoxes!(bboxes, scores, score_threshold, nms_threshold, opts)

Raising version of softNMSBoxes/5.

softNMSBoxes(bboxes, scores, score_threshold, nms_threshold)

Positional Arguments

bboxes: vector_Rect.

scores: vector_float.

score_threshold: float.

nms_threshold: float.

Keyword Arguments

top_k: size_t.

sigma: float.

method: SoftNMSMethod.

Performs soft non maximum suppression given boxes and corresponding scores. Reference: https://arxiv.org/abs/1704.04503 @see SoftNMSMethod

Python prototype (for reference):

softNMSBoxes(bboxes, scores, score_threshold, nms_threshold[, top_k[, sigma[, method]]]) -> updated_scores, indices

softNMSBoxes(bboxes, scores, score_threshold, nms_threshold, opts)

Positional Arguments

bboxes: vector_Rect.

scores: vector_float.

score_threshold: float.

nms_threshold: float.

Keyword Arguments

top_k: size_t.

sigma: float.

method: SoftNMSMethod.

Performs soft non maximum suppression given boxes and corresponding scores. Reference: https://arxiv.org/abs/1704.04503 @see SoftNMSMethod

Python prototype (for reference):

softNMSBoxes(bboxes, scores, score_threshold, nms_threshold[, top_k[, sigma[, method]]]) -> updated_scores, indices

writeTextGraph!(model, output)

Raising version of writeTextGraph/2.

writeTextGraph(model, output)

Create a text representation for a binary network stored in protocol buffer format. Note: To reduce output file size, trained weights are not included.

Python prototype (for reference):

writeTextGraph(model, output) -> None