Evision.DNN.DetectionModel (Evision v0.1.7-dev)

Functions

detect!(self, frame)

Raising version of detect/2.

detect!(self, frame, opts)

Raising version of detect/3.

detect(self, frame)

Positional Arguments

frame: UMat

Keyword Arguments

confThreshold: float.

nmsThreshold: float.

Given the @p input frame, create input blob, run net and return result detections.

Python prototype (for reference):

detect(frame[, confThreshold[, nmsThreshold]]) -> classIds, confidences, boxes

detect(self, frame, opts)

Positional Arguments

frame: UMat

Keyword Arguments

confThreshold: float.

nmsThreshold: float.

Given the @p input frame, create input blob, run net and return result detections.

Python prototype (for reference):

detect(frame[, confThreshold[, nmsThreshold]]) -> classIds, confidences, boxes

dnn_DetectionModel!(network)

Raising version of dnn_DetectionModel/1.

dnn_DetectionModel!(model, opts)

Raising version of dnn_DetectionModel/2.

dnn_DetectionModel(network)

Variaint 1:

Create model from deep learning network.

Python prototype (for reference):

DetectionModel(network) -> <dnn_DetectionModel object>

Variaint 2:

Positional Arguments

model: String.

Keyword Arguments

config: String.

Create detection model from network represented in one of the supported formats. An order of @p model and @p config arguments does not matter.

Python prototype (for reference):

DetectionModel(model[, config]) -> <dnn_DetectionModel object>

dnn_DetectionModel(model, opts)

Positional Arguments

model: String.

Keyword Arguments

config: String.

Create detection model from network represented in one of the supported formats. An order of @p model and @p config arguments does not matter.

Python prototype (for reference):

DetectionModel(model[, config]) -> <dnn_DetectionModel object>

getNmsAcrossClasses!(self)

Raising version of getNmsAcrossClasses/1.

getNmsAcrossClasses(self)

Getter for nmsAcrossClasses. This variable defaults to false, such that when non max suppression is used during the detect() function, it will do so only per-class

Python prototype (for reference):

getNmsAcrossClasses() -> retval

setNmsAcrossClasses!(self, value)

Raising version of setNmsAcrossClasses/2.

setNmsAcrossClasses(self, value)

nmsAcrossClasses defaults to false, such that when non max suppression is used during the detect() function, it will do so per-class. This function allows you to toggle this behaviour.

Python prototype (for reference):

setNmsAcrossClasses(value) -> retval