Evision.DNN.Net (Evision v0.1.7-dev)

Functions

connect!(self, outPin, inpPin)

Raising version of connect/3.

connect(self, outPin, inpPin)

Connects output of the first layer to input of the second layer. Descriptors have the following template <DFN>&lt;layer_name&gt;[.input_number]</DFN>:

the first part of the template <DFN>layer_name</DFN> is string name of the added layer. If this part is empty then the network input pseudo layer will be used;

the second optional part of the template <DFN>input_number</DFN> is either number of the layer input, either label one. If this part is omitted then the first layer input will be used. @see setNetInputs(), Layer::inputNameToIndex(), Layer::outputNameToIndex()

Python prototype (for reference):

connect(outPin, inpPin) -> None

dnn_Net!()

Raising version of dnn_Net/0.

dnn_Net()

Python prototype (for reference):

Net() -> <dnn_Net object>

dump!(self)

Raising version of dump/1.

dump(self)

Dump net to String @returns String with structure, hyperparameters, backend, target and fusion

Call method after setInput(). To see correct backend, target and fusion run after forward().

Python prototype (for reference):

dump() -> retval

dumpToFile!(self, path)

Raising version of dumpToFile/2.

dumpToFile(self, path)

Dump net structure, hyperparameters, backend, target and fusion to dot file @see dump()

Python prototype (for reference):

dumpToFile(path) -> None

empty!(self)

Raising version of empty/1.

empty(self)

Returns true if there are no layers in the network.

Python prototype (for reference):

empty() -> retval

enableFusion!(self, fusion)

Raising version of enableFusion/2.

enableFusion(self, fusion)

Enables or disables layer fusion in the network.

Python prototype (for reference):

enableFusion(fusion) -> None

forward!(self, opts)

Raising version of forward/2.

forward(self, opts \\ [])

Positional Arguments

outBlobNames: vector_String.

Keyword Arguments

outputBlobs: vector_UMat.

Runs forward pass to compute outputs of layers listed in @p outBlobNames.

Python prototype (for reference):

forward(outBlobNames[, outputBlobs]) -> outputBlobs

forwardAndRetrieve!(self, outBlobNames)

Raising version of forwardAndRetrieve/2.

forwardAndRetrieve(self, outBlobNames)

Runs forward pass to compute outputs of layers listed in @p outBlobNames.

Python prototype (for reference):

forwardAndRetrieve(outBlobNames) -> outputBlobs

forwardAsync!(self)

Raising version of forwardAsync/1.

forwardAsync(self)

Positional Arguments

Keyword Arguments

outputName: String.

Runs forward pass to compute output of layer with name @p outputName. @details By default runs forward pass for the whole network.

This is an asynchronous version of forward(const String&).
dnn::DNN_BACKEND_INFERENCE_ENGINE backend is required.

Python prototype (for reference):

forwardAsync([, outputName]) -> retval

getFLOPS!(self, netInputShape)

Raising version of getFLOPS/2.

getFLOPS!(self, layerId, netInputShape)

Raising version of getFLOPS/3.

getFLOPS(self, netInputShape)

Variaint 1:

Has overloading in C++

Python prototype (for reference):

getFLOPS(netInputShape) -> retval

Variaint 2:

Computes FLOP for whole loaded model with specified input shapes. @returns computed FLOP.

Python prototype (for reference):

getFLOPS(netInputShapes) -> retval

getFLOPS(self, layerId, netInputShape)

Variaint 1:

Has overloading in C++

Python prototype (for reference):

getFLOPS(layerId, netInputShape) -> retval

Variaint 2:

Has overloading in C++

Python prototype (for reference):

getFLOPS(layerId, netInputShapes) -> retval

getInputDetails!(self)

Raising version of getInputDetails/1.

getInputDetails(self)

Returns input scale and zeropoint for a quantized Net.

Python prototype (for reference):

getInputDetails() -> scales, zeropoints

getLayer!(self, layerName)

Raising version of getLayer/2.

getLayer(self, layerName)

Variaint 1:

Has overloading in C++

@deprecated Use int getLayerId(const String &layer)

Python prototype (for reference):

getLayer(layerName) -> retval

Variaint 2:

Returns pointer to layer with specified id or name which the network use.

Python prototype (for reference):

getLayer(layerId) -> retval

Variaint 3:

Has overloading in C++

@deprecated to be removed

Python prototype (for reference):

getLayer(layerId) -> retval

getLayerId!(self, layer)

Raising version of getLayerId/2.

getLayerId(self, layer)

Converts string name of the layer to the integer identifier. @returns id of the layer, or -1 if the layer wasn't found.

Python prototype (for reference):

getLayerId(layer) -> retval

getLayerNames!(self)

Raising version of getLayerNames/1.

getLayerNames(self)

Python prototype (for reference):

getLayerNames() -> retval

getLayerShapes!(self, opts)

Raising version of getLayerShapes/2.

getLayerShapes(self, opts \\ [])

Has overloading in C++

Python prototype (for reference):

getLayerShapes(netInputShapes, layerId) -> inLayerShapes, outLayerShapes

getLayersCount!(self, layerType)

Raising version of getLayersCount/2.

getLayersCount(self, layerType)

Returns count of layers of specified type. @returns count of layers

Python prototype (for reference):

getLayersCount(layerType) -> retval

getLayersShapes!(self, opts)

Raising version of getLayersShapes/2.

getLayersShapes(self, opts \\ [])

Has overloading in C++

Python prototype (for reference):

getLayersShapes(netInputShape) -> layersIds, inLayersShapes, outLayersShapes

getLayerTypes!(self)

Raising version of getLayerTypes/1.

getLayerTypes(self)

Returns list of types for layer used in model.

Python prototype (for reference):

getLayerTypes() -> layersTypes

getMemoryConsumption!(self, netInputShape)

Raising version of getMemoryConsumption/2.

getMemoryConsumption!(self, layerId, netInputShape)

Raising version of getMemoryConsumption/3.

getMemoryConsumption(self, netInputShape)

Has overloading in C++

Python prototype (for reference):

getMemoryConsumption(netInputShape) -> weights, blobs

getMemoryConsumption(self, layerId, netInputShape)

Variaint 1:

Has overloading in C++

Python prototype (for reference):

getMemoryConsumption(layerId, netInputShape) -> weights, blobs

Variaint 2:

Has overloading in C++

Python prototype (for reference):

getMemoryConsumption(layerId, netInputShapes) -> weights, blobs

getOutputDetails!(self)

Raising version of getOutputDetails/1.

getOutputDetails(self)

Returns output scale and zeropoint for a quantized Net.

Python prototype (for reference):

getOutputDetails() -> scales, zeropoints

getParam!(self, layerName)

Raising version of getParam/2.

getParam!(self, layerName, opts)

Raising version of getParam/3.

getParam(self, layerName)

Variaint 1:

Positional Arguments

layerName: String

Keyword Arguments

numParam: int.

Python prototype (for reference):

getParam(layerName[, numParam]) -> retval

Variaint 2:

Positional Arguments

layer: int.

Keyword Arguments

numParam: int.

Returns parameter blob of the layer. @see Layer::blobs

Python prototype (for reference):

getParam(layer[, numParam]) -> retval

getParam(self, layerName, opts)

Variaint 1:

Positional Arguments

layerName: String

Keyword Arguments

numParam: int.

Python prototype (for reference):

getParam(layerName[, numParam]) -> retval

Variaint 2:

Positional Arguments

layer: int.

Keyword Arguments

numParam: int.

Returns parameter blob of the layer. @see Layer::blobs

Python prototype (for reference):

getParam(layer[, numParam]) -> retval

getPerfProfile!(self)

Raising version of getPerfProfile/1.

getPerfProfile(self)

Returns overall time for inference and timings (in ticks) for layers. Indexes in returned vector correspond to layers ids. Some layers can be fused with others, in this case zero ticks count will be return for that skipped layers. Supported by DNN_BACKEND_OPENCV on DNN_TARGET_CPU only. @return overall ticks for model inference.

Python prototype (for reference):

getPerfProfile() -> retval, timings

getUnconnectedOutLayers!(self)

Raising version of getUnconnectedOutLayers/1.

getUnconnectedOutLayers(self)

Returns indexes of layers with unconnected outputs. FIXIT: Rework API to registerOutput() approach, deprecate this call

Python prototype (for reference):

getUnconnectedOutLayers() -> retval

getUnconnectedOutLayersNames!(self)

Raising version of getUnconnectedOutLayersNames/1.

getUnconnectedOutLayersNames(self)

Returns names of layers with unconnected outputs. FIXIT: Rework API to registerOutput() approach, deprecate this call

Python prototype (for reference):

getUnconnectedOutLayersNames() -> retval

quantize!(self, calibData, inputsDtype, outputsDtype)

Raising version of quantize/4.

quantize(self, calibData, inputsDtype, outputsDtype)

Returns a quantized Net from a floating-point Net.

Python prototype (for reference):

quantize(calibData, inputsDtype, outputsDtype) -> retval

readFromModelOptimizer!(bufferModelConfig, bufferWeights)

Raising version of readFromModelOptimizer/2.

readFromModelOptimizer(bufferModelConfig, bufferWeights)

Variaint 1:

Create a network from Intel's Model Optimizer in-memory buffers with intermediate representation (IR). @returns Net object.

Python prototype (for reference):

readFromModelOptimizer(bufferModelConfig, bufferWeights) -> retval

Variaint 2:

Create a network from Intel's Model Optimizer intermediate representation (IR).

  Networks imported from Intel's Model Optimizer are launched in Intel's Inference Engine
  backend.

Python prototype (for reference):

readFromModelOptimizer(xml, bin) -> retval

setHalideScheduler!(self, scheduler)

Raising version of setHalideScheduler/2.

setHalideScheduler(self, scheduler)

Compile Halide layers. @see setPreferableBackend Schedule layers that support Halide backend. Then compile them for specific target. For layers that not represented in scheduling file or if no manual scheduling used at all, automatic scheduling will be applied.

Python prototype (for reference):

setHalideScheduler(scheduler) -> None

setInput!(self, blob)

Raising version of setInput/2.

setInput!(self, blob, opts)

Raising version of setInput/3.

setInput(self, blob)

Positional Arguments

blob: UMat.

Keyword Arguments

name: String.

scalefactor: double.

mean: Scalar.

Sets the new input value for the network @see connect(String, String) to know format of the descriptor.

If scale or mean values are specified, a final input blob is computed
as:

\f[input(n,c,h,w) = scalefactor \times (blob(n,c,h,w) - mean_c)\f]

Python prototype (for reference):

setInput(blob[, name[, scalefactor[, mean]]]) -> None

setInput(self, blob, opts)

Positional Arguments

blob: UMat.

Keyword Arguments

name: String.

scalefactor: double.

mean: Scalar.

Sets the new input value for the network @see connect(String, String) to know format of the descriptor.

If scale or mean values are specified, a final input blob is computed
as:

\f[input(n,c,h,w) = scalefactor \times (blob(n,c,h,w) - mean_c)\f]

Python prototype (for reference):

setInput(blob[, name[, scalefactor[, mean]]]) -> None

setInputShape!(self, inputName, shape)

Raising version of setInputShape/3.

setInputShape(self, inputName, shape)

Specify shape of network input.

Python prototype (for reference):

setInputShape(inputName, shape) -> None

setInputsNames!(self, inputBlobNames)

Raising version of setInputsNames/2.

setInputsNames(self, inputBlobNames)

Sets outputs names of the network input pseudo layer. Each net always has special own the network input pseudo layer with id=0. This layer stores the user blobs only and don't make any computations. In fact, this layer provides the only way to pass user data into the network. As any other layer, this layer can label its outputs and this function provides an easy way to do this.

Python prototype (for reference):

setInputsNames(inputBlobNames) -> None

setParam!(self, layerName, numParam, blob)

Raising version of setParam/4.

setParam(self, layerName, numParam, blob)

Variaint 1:

Python prototype (for reference):

setParam(layerName, numParam, blob) -> None

Variaint 2:

Sets the new value for the learned param of the layer. @see Layer::blobs Note: If shape of the new blob differs from the previous shape,

then the following forward pass may fail.

Python prototype (for reference):

setParam(layer, numParam, blob) -> None

setPreferableBackend!(self, backendId)

Raising version of setPreferableBackend/2.

setPreferableBackend(self, backendId)

Ask network to use specific computation backend where it supported. @see Backend If OpenCV is compiled with Intel's Inference Engine library, DNN_BACKEND_DEFAULT means DNN_BACKEND_INFERENCE_ENGINE. Otherwise it equals to DNN_BACKEND_OPENCV.

Python prototype (for reference):

setPreferableBackend(backendId) -> None

setPreferableTarget!(self, targetId)

Raising version of setPreferableTarget/2.

setPreferableTarget(self, targetId)

Ask network to make computations on specific target device. @see Target List of supported combinations backend / target: | | DNN_BACKEND_OPENCV | DNN_BACKEND_INFERENCE_ENGINE | DNN_BACKEND_HALIDE | DNN_BACKEND_CUDA | |------------------------|--------------------|------------------------------|--------------------|-------------------| | DNN_TARGET_CPU | + | + | + | | | DNN_TARGET_OPENCL | + | + | + | | | DNN_TARGET_OPENCL_FP16 | + | + | | | | DNN_TARGET_MYRIAD | | + | | | | DNN_TARGET_FPGA | | + | | | | DNN_TARGET_CUDA | | | | + | | DNN_TARGET_CUDA_FP16 | | | | + | | DNN_TARGET_HDDL | | + | | |

Python prototype (for reference):

setPreferableTarget(targetId) -> None