Evision.DNN.TextRecognitionModel (Evision v0.1.7-dev)

Functions

dnn_TextRecognitionModel!(model)

Raising version of dnn_TextRecognitionModel/1.

dnn_TextRecognitionModel!(model, opts)

Raising version of dnn_TextRecognitionModel/2.

dnn_TextRecognitionModel(model)

Positional Arguments

model: string.

Keyword Arguments

config: string.

Create text recognition model from network represented in one of the supported formats Call setDecodeType() and setVocabulary() after constructor to initialize the decoding method

Python prototype (for reference):

TextRecognitionModel(model[, config]) -> <dnn_TextRecognitionModel object>

dnn_TextRecognitionModel(model, opts)

Positional Arguments

model: string.

Keyword Arguments

config: string.

Create text recognition model from network represented in one of the supported formats Call setDecodeType() and setVocabulary() after constructor to initialize the decoding method

Python prototype (for reference):

TextRecognitionModel(model[, config]) -> <dnn_TextRecognitionModel object>

getDecodeType!(self)

Raising version of getDecodeType/1.

getDecodeType(self)

Get the decoding method @return the decoding method

Python prototype (for reference):

getDecodeType() -> retval

getVocabulary!(self)

Raising version of getVocabulary/1.

getVocabulary(self)

Get the vocabulary for recognition. @return vocabulary the associated vocabulary

Python prototype (for reference):

getVocabulary() -> retval

recognize!(self, frame)

Raising version of recognize/2.

recognize!(self, frame, roiRects)

Raising version of recognize/3.

recognize(self, frame)

Given the @p input frame, create input blob, run net and return recognition result @return The text recognition result

Python prototype (for reference):

recognize(frame) -> retval

recognize(self, frame, roiRects)

Given the @p input frame, create input blob, run net and return recognition result

Python prototype (for reference):

recognize(frame, roiRects) -> results

setDecodeOptsCTCPrefixBeamSearch!(self, beamSize)

Raising version of setDecodeOptsCTCPrefixBeamSearch/2.

setDecodeOptsCTCPrefixBeamSearch!(self, beamSize, opts)

Raising version of setDecodeOptsCTCPrefixBeamSearch/3.

setDecodeOptsCTCPrefixBeamSearch(self, beamSize)

Positional Arguments

beamSize: int.

Keyword Arguments

vocPruneSize: int.

Set the decoding method options for "CTC-prefix-beam-search" decode usage

 only take top @p vocPruneSize tokens in each search step, @p vocPruneSize <= 0 stands for disable this prune.

Python prototype (for reference):

setDecodeOptsCTCPrefixBeamSearch(beamSize[, vocPruneSize]) -> retval

setDecodeOptsCTCPrefixBeamSearch(self, beamSize, opts)

Positional Arguments

beamSize: int.

Keyword Arguments

vocPruneSize: int.

Set the decoding method options for "CTC-prefix-beam-search" decode usage

 only take top @p vocPruneSize tokens in each search step, @p vocPruneSize <= 0 stands for disable this prune.

Python prototype (for reference):

setDecodeOptsCTCPrefixBeamSearch(beamSize[, vocPruneSize]) -> retval

setDecodeType!(self, decodeType)

Raising version of setDecodeType/2.

setDecodeType(self, decodeType)

Set the decoding method of translating the network output into string

    - `"CTC-greedy"` greedy decoding for the output of CTC-based methods
    - `"CTC-prefix-beam-search"` Prefix beam search decoding for the output of CTC-based methods

Python prototype (for reference):

setDecodeType(decodeType) -> retval

setVocabulary!(self, vocabulary)

Raising version of setVocabulary/2.

setVocabulary(self, vocabulary)

Set the vocabulary for recognition.

Python prototype (for reference):

setVocabulary(vocabulary) -> retval