Evision.FileStorage (Evision v0.1.7-dev)

cv

endWriteStruct(self)

Finishes writing nested structure (should pair startWriteStruct())

Python prototype (for reference):

endWriteStruct() -> None

fileStorage()

The constructors. The full constructor opens the file. Alternatively you can use the default constructor and then call FileStorage::open.

Python prototype (for reference):

FileStorage() -> <FileStorage object>

fileStorage(filename, flags)

Positional Arguments

filename: String

flags: int

Keyword Arguments

encoding: String.

Has overloading in C++

@copydoc open()

Python prototype (for reference):

FileStorage(filename, flags[, encoding]) -> <FileStorage object>

fileStorage(filename, flags, opts)

Positional Arguments

filename: String

flags: int

Keyword Arguments

encoding: String.

Has overloading in C++

@copydoc open()

Python prototype (for reference):

FileStorage(filename, flags[, encoding]) -> <FileStorage object>

getFirstTopLevelNode(self)

Returns the first element of the top-level mapping. @returns The first element of the top-level mapping.

Python prototype (for reference):

getFirstTopLevelNode() -> retval

getFormat(self)

Returns the current format. @returns The current format, see FileStorage::Mode

Python prototype (for reference):

getFormat() -> retval

getNode(self, nodename)

Has overloading in C++

Python prototype (for reference):

getNode(nodename) -> retval

isOpened(self)

Checks whether the file is opened. @returns true if the object is associated with the current file and false otherwise. It is a good practice to call this method after you tried to open a file.

Python prototype (for reference):

isOpened() -> retval

open(self, filename, flags)

Positional Arguments

filename: String.

flags: int.

Keyword Arguments

encoding: String.

Opens a file. See description of parameters in FileStorage::FileStorage. The method calls FileStorage::release before opening the file.

Extension of the file (.xml, .yml/.yaml or .json) determines its format (XML, YAML or JSON
respectively). Also you can append .gz to work with compressed files, for example myHugeMatrix.xml.gz. If both
FileStorage::WRITE and FileStorage::MEMORY flags are specified, source is used just to specify
the output file format (e.g. mydata.xml, .yml etc.). A file name can also contain parameters.
You can use this format, "*?base64" (e.g. "file.json?base64" (case sensitive)), as an alternative to
FileStorage::BASE64 flag.
you should use 8-bit encoding instead of it.

Python prototype (for reference):

open(filename, flags[, encoding]) -> retval

open(self, filename, flags, opts)

Positional Arguments

filename: String.

flags: int.

Keyword Arguments

encoding: String.

Opens a file. See description of parameters in FileStorage::FileStorage. The method calls FileStorage::release before opening the file.

Extension of the file (.xml, .yml/.yaml or .json) determines its format (XML, YAML or JSON
respectively). Also you can append .gz to work with compressed files, for example myHugeMatrix.xml.gz. If both
FileStorage::WRITE and FileStorage::MEMORY flags are specified, source is used just to specify
the output file format (e.g. mydata.xml, .yml etc.). A file name can also contain parameters.
You can use this format, "*?base64" (e.g. "file.json?base64" (case sensitive)), as an alternative to
FileStorage::BASE64 flag.
you should use 8-bit encoding instead of it.

Python prototype (for reference):

open(filename, flags[, encoding]) -> retval

release(self)

Closes the file and releases all the memory buffers. Call this method after all I/O operations with the storage are finished.

Python prototype (for reference):

release() -> None

releaseAndGetString(self)

Closes the file and releases all the memory buffers. Call this method after all I/O operations with the storage are finished. If the storage was opened for writing data and FileStorage::WRITE was specified

Python prototype (for reference):

releaseAndGetString() -> retval

root(self)

Positional Arguments

Keyword Arguments

streamidx: int.

Returns the top-level mapping

However, YAML supports multiple streams and so there can be several.

@returns The top-level mapping.

Python prototype (for reference):

root([, streamidx]) -> retval

startWriteStruct(self, name, flags)

Positional Arguments

name: String.

flags: int.

Keyword Arguments

typeName: String.

Starts to write a nested structure (sequence or a mapping).

I.e. if the format has a specification for storing type information, this parameter is used.

Python prototype (for reference):

startWriteStruct(name, flags[, typeName]) -> None

startWriteStruct(self, name, flags, opts)

Positional Arguments

name: String.

flags: int.

Keyword Arguments

typeName: String.

Starts to write a nested structure (sequence or a mapping).

I.e. if the format has a specification for storing type information, this parameter is used.

Python prototype (for reference):

startWriteStruct(name, flags[, typeName]) -> None

write(self, name, val)

Variaint 1:

Python prototype (for reference):

write(name, val) -> None

Variaint 2:

Python prototype (for reference):

write(name, val) -> None

Variaint 3:

Python prototype (for reference):

write(name, val) -> None

Variaint 4:

Python prototype (for reference):

write(name, val) -> None

Variaint 5:

Simplified writing API to use with bindings.

Python prototype (for reference):

write(name, val) -> None

writeComment(self, comment)

Positional Arguments

comment: String.

Keyword Arguments

append: bool.

Writes a comment. The function writes a comment into file storage. The comments are skipped when the storage is read.

Else if the comment is multi-line, or if it does not fit at the end of the current
line, the comment starts a new line.

Python prototype (for reference):

writeComment(comment[, append]) -> None

writeComment(self, comment, opts)

Positional Arguments

comment: String.

Keyword Arguments

append: bool.

Writes a comment. The function writes a comment into file storage. The comments are skipped when the storage is read.

Else if the comment is multi-line, or if it does not fit at the end of the current
line, the comment starts a new line.

Python prototype (for reference):

writeComment(comment[, append]) -> None

Functions

endWriteStruct!(self)

Raising version of endWriteStruct/1.

fileStorage!()

Raising version of fileStorage/0.

fileStorage!(filename, flags)

Raising version of fileStorage/2.

fileStorage!(filename, flags, opts)

Raising version of fileStorage/3.

getFirstTopLevelNode!(self)

Raising version of getFirstTopLevelNode/1.

getFormat!(self)

Raising version of getFormat/1.

getNode!(self, nodename)

Raising version of getNode/2.

isOpened!(self)

Raising version of isOpened/1.

open!(self, filename, flags)

Raising version of open/3.

open!(self, filename, flags, opts)

Raising version of open/4.

release!(self)

Raising version of release/1.

releaseAndGetString!(self)

Raising version of releaseAndGetString/1.

root!(self)

Raising version of root/1.

startWriteStruct!(self, name, flags)

Raising version of startWriteStruct/3.

startWriteStruct!(self, name, flags, opts)

Raising version of startWriteStruct/4.

write!(self, name, val)

Raising version of write/3.

writeComment!(self, comment)

Raising version of writeComment/2.

writeComment!(self, comment, opts)

Raising version of writeComment/3.