Evision.FlannBasedMatcher (Evision v0.1.7-dev)

cv

add(self, descriptors)

Adds descriptors to train a CPU(trainDescCollectionis) or GPU(utrainDescCollectionis) descriptor collection. If the collection is not empty, the new descriptors are added to existing train descriptors.

train image.

Python prototype (for reference):

add(descriptors) -> None

clear(self)

Clears the train descriptor collections.

Python prototype (for reference):

clear() -> None

clone(self)

Positional Arguments

Keyword Arguments

emptyTrainData: bool.

Clones the matcher.

that is, copies both parameters and train data. If emptyTrainData is true, the method creates an
object copy with the current parameters but with empty train data.

Python prototype (for reference):

clone([, emptyTrainData]) -> retval

create()

Python prototype (for reference):

create() -> retval

empty(self)

Returns true if there are no train descriptors in the both collections.

Python prototype (for reference):

empty() -> retval

flannBasedMatcher()

Positional Arguments

Keyword Arguments

indexParams: Ptr<flann::IndexParams>.

searchParams: Ptr<flann::SearchParams>.

Python prototype (for reference):

FlannBasedMatcher([, indexParams[, searchParams]]) -> <FlannBasedMatcher object>

flannBasedMatcher(opts)

Positional Arguments

Keyword Arguments

indexParams: Ptr<flann::IndexParams>.

searchParams: Ptr<flann::SearchParams>.

Python prototype (for reference):

FlannBasedMatcher([, indexParams[, searchParams]]) -> <FlannBasedMatcher object>

getTrainDescriptors(self)

Returns a constant link to the train descriptor collection trainDescCollection .

Python prototype (for reference):

getTrainDescriptors() -> retval

isMaskSupported(self)

Returns true if the descriptor matcher supports masking permissible matches.

Python prototype (for reference):

isMaskSupported() -> retval

knnMatch(self, queryDescriptors, k)

Positional Arguments

queryDescriptors: UMat.

k: int.

Keyword Arguments

masks: vector_UMat.

compactResult: bool.

Has overloading in C++

less than k possible matches in total.
descriptors and stored train descriptors from the i-th image trainDescCollection[i].
false, the matches vector has the same size as queryDescriptors rows. If compactResult is true,
the matches vector does not contain matches for fully masked-out query descriptors.

Python prototype (for reference):

knnMatch(queryDescriptors, k[, masks[, compactResult]]) -> matches

knnMatch(self, queryDescriptors, k, opts)

Variaint 1:

Positional Arguments

queryDescriptors: UMat.

trainDescriptors: UMat.

k: int.

Keyword Arguments

mask: UMat.

compactResult: bool.

Finds the k best matches for each descriptor from a query set.

collection stored in the class object.
descriptors.
less than k possible matches in total.
false, the matches vector has the same size as queryDescriptors rows. If compactResult is true,
the matches vector does not contain matches for fully masked-out query descriptors.

These extended variants of DescriptorMatcher::match methods find several best matches for each query descriptor. The matches are returned in the distance increasing order. See DescriptorMatcher::match for the details about query and train descriptors.

Python prototype (for reference):

knnMatch(queryDescriptors, trainDescriptors, k[, mask[, compactResult]]) -> matches

Variaint 2:

Positional Arguments

queryDescriptors: UMat.

k: int.

Keyword Arguments

masks: vector_UMat.

compactResult: bool.

Has overloading in C++

less than k possible matches in total.
descriptors and stored train descriptors from the i-th image trainDescCollection[i].
false, the matches vector has the same size as queryDescriptors rows. If compactResult is true,
the matches vector does not contain matches for fully masked-out query descriptors.

Python prototype (for reference):

knnMatch(queryDescriptors, k[, masks[, compactResult]]) -> matches

knnMatch(self, queryDescriptors, trainDescriptors, k, opts)

Positional Arguments

queryDescriptors: UMat.

trainDescriptors: UMat.

k: int.

Keyword Arguments

mask: UMat.

compactResult: bool.

Finds the k best matches for each descriptor from a query set.

collection stored in the class object.
descriptors.
less than k possible matches in total.
false, the matches vector has the same size as queryDescriptors rows. If compactResult is true,
the matches vector does not contain matches for fully masked-out query descriptors.

These extended variants of DescriptorMatcher::match methods find several best matches for each query descriptor. The matches are returned in the distance increasing order. See DescriptorMatcher::match for the details about query and train descriptors.

Python prototype (for reference):

knnMatch(queryDescriptors, trainDescriptors, k[, mask[, compactResult]]) -> matches

match(self, queryDescriptors)

Positional Arguments

queryDescriptors: UMat.

Keyword Arguments

masks: vector_UMat.

Has overloading in C++

descriptor. So, matches size may be smaller than the query descriptors count.
descriptors and stored train descriptors from the i-th image trainDescCollection[i].

Python prototype (for reference):

match(queryDescriptors[, masks]) -> matches

match(self, queryDescriptors, opts)

Variaint 1:

Positional Arguments

queryDescriptors: UMat.

trainDescriptors: UMat.

Keyword Arguments

mask: UMat.

Finds the best match for each descriptor from a query set.

collection stored in the class object.
descriptor. So, matches size may be smaller than the query descriptors count.
descriptors.

In the first variant of this method, the train descriptors are passed as an input argument. In the second variant of the method, train descriptors collection that was set by DescriptorMatcher::add is used. Optional mask (or masks) can be passed to specify which query and training descriptors can be matched. Namely, queryDescriptors[i] can be matched with trainDescriptors[j] only if mask.at\<uchar>(i,j) is non-zero.

Python prototype (for reference):

match(queryDescriptors, trainDescriptors[, mask]) -> matches

Variaint 2:

Positional Arguments

queryDescriptors: UMat.

Keyword Arguments

masks: vector_UMat.

Has overloading in C++

descriptor. So, matches size may be smaller than the query descriptors count.
descriptors and stored train descriptors from the i-th image trainDescCollection[i].

Python prototype (for reference):

match(queryDescriptors[, masks]) -> matches

match(self, queryDescriptors, trainDescriptors, opts)

Positional Arguments

queryDescriptors: UMat.

trainDescriptors: UMat.

Keyword Arguments

mask: UMat.

Finds the best match for each descriptor from a query set.

collection stored in the class object.
descriptor. So, matches size may be smaller than the query descriptors count.
descriptors.

In the first variant of this method, the train descriptors are passed as an input argument. In the second variant of the method, train descriptors collection that was set by DescriptorMatcher::add is used. Optional mask (or masks) can be passed to specify which query and training descriptors can be matched. Namely, queryDescriptors[i] can be matched with trainDescriptors[j] only if mask.at\<uchar>(i,j) is non-zero.

Python prototype (for reference):

match(queryDescriptors, trainDescriptors[, mask]) -> matches

radiusMatch(self, queryDescriptors, maxDistance)

Positional Arguments

queryDescriptors: UMat.

maxDistance: float.

Keyword Arguments

masks: vector_UMat.

compactResult: bool.

Has overloading in C++

metric distance (e.g. Hamming distance), not the distance between coordinates (which is measured
in Pixels)!
descriptors and stored train descriptors from the i-th image trainDescCollection[i].
false, the matches vector has the same size as queryDescriptors rows. If compactResult is true,
the matches vector does not contain matches for fully masked-out query descriptors.

Python prototype (for reference):

radiusMatch(queryDescriptors, maxDistance[, masks[, compactResult]]) -> matches

radiusMatch(self, queryDescriptors, maxDistance, opts)

Variaint 1:

Positional Arguments

queryDescriptors: UMat.

trainDescriptors: UMat.

maxDistance: float.

Keyword Arguments

mask: UMat.

compactResult: bool.

For each query descriptor, finds the training descriptors not farther than the specified distance.

collection stored in the class object.
false, the matches vector has the same size as queryDescriptors rows. If compactResult is true,
the matches vector does not contain matches for fully masked-out query descriptors.
metric distance (e.g. Hamming distance), not the distance between coordinates (which is measured
in Pixels)!
descriptors.

For each query descriptor, the methods find such training descriptors that the distance between the query descriptor and the training descriptor is equal or smaller than maxDistance. Found matches are returned in the distance increasing order.

Python prototype (for reference):

radiusMatch(queryDescriptors, trainDescriptors, maxDistance[, mask[, compactResult]]) -> matches

Variaint 2:

Positional Arguments

queryDescriptors: UMat.

maxDistance: float.

Keyword Arguments

masks: vector_UMat.

compactResult: bool.

Has overloading in C++

metric distance (e.g. Hamming distance), not the distance between coordinates (which is measured
in Pixels)!
descriptors and stored train descriptors from the i-th image trainDescCollection[i].
false, the matches vector has the same size as queryDescriptors rows. If compactResult is true,
the matches vector does not contain matches for fully masked-out query descriptors.

Python prototype (for reference):

radiusMatch(queryDescriptors, maxDistance[, masks[, compactResult]]) -> matches

radiusMatch(self, queryDescriptors, trainDescriptors, maxDistance, opts)

Positional Arguments

queryDescriptors: UMat.

trainDescriptors: UMat.

maxDistance: float.

Keyword Arguments

mask: UMat.

compactResult: bool.

For each query descriptor, finds the training descriptors not farther than the specified distance.

collection stored in the class object.
false, the matches vector has the same size as queryDescriptors rows. If compactResult is true,
the matches vector does not contain matches for fully masked-out query descriptors.
metric distance (e.g. Hamming distance), not the distance between coordinates (which is measured
in Pixels)!
descriptors.

For each query descriptor, the methods find such training descriptors that the distance between the query descriptor and the training descriptor is equal or smaller than maxDistance. Found matches are returned in the distance increasing order.

Python prototype (for reference):

radiusMatch(queryDescriptors, trainDescriptors, maxDistance[, mask[, compactResult]]) -> matches

read(self, arg1)

Variaint 1:

Python prototype (for reference):

read(arg1) -> None

Variaint 2:

Python prototype (for reference):

read(fileName) -> None

train(self)

Trains a descriptor matcher Trains a descriptor matcher (for example, the flann index). In all methods to match, the method train() is run every time before matching. Some descriptor matchers (for example, BruteForceMatcher) have an empty implementation of this method. Other matchers really train their inner structures (for example, FlannBasedMatcher trains flann::Index ).

Python prototype (for reference):

train() -> None

write(self, fs)

Variaint 1:

Positional Arguments

fs: Ptr<FileStorage>

Keyword Arguments

name: String.

Python prototype (for reference):

write(fs[, name]) -> None

Variaint 2:

Python prototype (for reference):

write(fileName) -> None

write(self, fs, opts)

Positional Arguments

fs: Ptr<FileStorage>

Keyword Arguments

name: String.

Python prototype (for reference):

write(fs[, name]) -> None

Functions

add!(self, descriptors)

Raising version of add/2.

clear!(self)

Raising version of clear/1.

clone!(self)

Raising version of clone/1.

create!()

Raising version of create/0.

empty!(self)

Raising version of empty/1.

flannBasedMatcher!()

Raising version of flannBasedMatcher/0.

flannBasedMatcher!(opts)

Raising version of flannBasedMatcher/1.

getTrainDescriptors!(self)

Raising version of getTrainDescriptors/1.

isMaskSupported!(self)

Raising version of isMaskSupported/1.

knnMatch!(self, queryDescriptors, k)

Raising version of knnMatch/3.

knnMatch!(self, queryDescriptors, trainDescriptors, k)

Raising version of knnMatch/4.

knnMatch!(self, queryDescriptors, trainDescriptors, k, opts)

Raising version of knnMatch/5.

match!(self, queryDescriptors)

Raising version of match/2.

match!(self, queryDescriptors, trainDescriptors)

Raising version of match/3.

match!(self, queryDescriptors, trainDescriptors, opts)

Raising version of match/4.

radiusMatch!(self, queryDescriptors, maxDistance)

Raising version of radiusMatch/3.

radiusMatch!(self, queryDescriptors, trainDescriptors, maxDistance)

Raising version of radiusMatch/4.

radiusMatch!(self, queryDescriptors, trainDescriptors, maxDistance, opts)

Raising version of radiusMatch/5.

read!(self, arg1)

Raising version of read/2.

train!(self)

Raising version of train/1.

write!(self, fs)

Raising version of write/2.

write!(self, fs, opts)

Raising version of write/3.