Evision.HOGDescriptor (Evision v0.1.5-dev)

cv

checkDetectorSize(self)

compute(self, img)

compute(self, img, opts)

Computes HOG descriptors of given image.

img.

descriptors.

winStride.

padding : Padding

locations.

winStride : Size

padding : Size

locations : vector_Point

Python prototype (for reference):

compute(img[, winStride[, padding[, locations]]]) -> descriptors

computeGradient(self, img, grad, angleOfs)

computeGradient(self, img, grad, angleOfs, opts)

Computes gradients and quantized gradient orientations.

img.

grad.

angleOfs.

paddingTL.

paddingBR.

paddingTL : Size

paddingBR : Size

Python prototype (for reference):

computeGradient(img, grad, angleOfs[, paddingTL[, paddingBR]]) -> grad, angleOfs

detect(self, img)

detect(self, img, opts)

Performs object detection without a multi-scale window.

img.

foundLocations.

weights.

hitThreshold. Usually it is 0 and should be specified in the detector coefficients (as the last free coefficient). But if the free coefficient is omitted (which is allowed), you can specify it manually here.

winStride.

padding : Padding

searchLocations.

hitThreshold : double

winStride : Size

padding : Size

searchLocations : vector_Point

Python prototype (for reference):

detect(img[, hitThreshold[, winStride[, padding[, searchLocations]]]]) -> foundLocations, weights

detectMultiScale(self, img)

detectMultiScale(self, img, opts)

Detects objects of different sizes in the input image. The detected objects are returned as a list of rectangles.

img.

foundLocations.

foundWeights.

hitThreshold. Usually it is 0 and should be specified in the detector coefficients (as the last free coefficient). But if the free coefficient is omitted (which is allowed), you can specify it manually here.

winStride.

padding : Padding

scale.

groupThreshold. by many rectangles. 0 means not to perform grouping.

useMeanshiftGrouping.

hitThreshold : double

winStride : Size

padding : Size

scale : double

groupThreshold : double

useMeanshiftGrouping : bool

Python prototype (for reference):

detectMultiScale(img[, hitThreshold[, winStride[, padding[, scale[, groupThreshold[, useMeanshiftGrouping]]]]]]) -> foundLocations, foundWeights

getDaimlerPeopleDetector()

getDefaultPeopleDetector()

getDescriptorSize(self)

getWinSigma(self)

hogDescriptor()

hogDescriptor(filename)

hogDescriptor(winSize, blockSize, blockStride, cellSize, nbins)

hogDescriptor(winSize, blockSize, blockStride, cellSize, nbins, opts)

Has overloading in C++

winSize.

blockSize.

blockStride.

cellSize.

nbins.

derivAperture.

winSigma.

histogramNormType.

l2HysThreshold.

gammaCorrection.

nlevels.

signedGradient.

derivAperture : int

winSigma : double

histogramNormType : HOGDescriptor_HistogramNormType

l2HysThreshold : double

gammaCorrection : bool

nlevels : int

signedGradient : bool

Python prototype (for reference):

HOGDescriptor(_winSize, _blockSize, _blockStride, _cellSize, _nbins[, _derivAperture[, _winSigma[, _histogramNormType[, _L2HysThreshold[, _gammaCorrection[, _nlevels[, _signedGradient]]]]]]]) -> <HOGDescriptor object>

load(self, filename)

load(self, filename, opts)

loads HOGDescriptor parameters and coefficients for the linear SVM classifier from a file

filename.

objname.

objname : String

Python prototype (for reference):

load(filename[, objname]) -> retval

save(self, filename)

save(self, filename, opts)

saves HOGDescriptor parameters and coefficients for the linear SVM classifier to a file

filename.

objname.

objname : String

Python prototype (for reference):

save(filename[, objname]) -> None

setSVMDetector(self, svmdetector)

Functions

checkDetectorSize!(self)

Raising version of checkDetectorSize/1.

compute!(self, img)

Raising version of compute/2.

compute!(self, img, opts)

Raising version of compute/3.

computeGradient!(self, img, grad, angleOfs)

Raising version of computeGradient/4.

computeGradient!(self, img, grad, angleOfs, opts)

Raising version of computeGradient/5.

detect!(self, img)

Raising version of detect/2.

detect!(self, img, opts)

Raising version of detect/3.

detectMultiScale!(self, img)

Raising version of detectMultiScale/2.

detectMultiScale!(self, img, opts)

Raising version of detectMultiScale/3.

get_blockSize!(self)

Raising version of get_blockSize/1.

get_blockSize(self)

get_blockStride!(self)

Raising version of get_blockStride/1.

get_blockStride(self)

get_cellSize!(self)

Raising version of get_cellSize/1.

get_cellSize(self)

get_derivAperture!(self)

Raising version of get_derivAperture/1.

get_derivAperture(self)

get_gammaCorrection!(self)

Raising version of get_gammaCorrection/1.

get_gammaCorrection(self)

get_histogramNormType!(self)

Raising version of get_histogramNormType/1.

get_histogramNormType(self)

get_L2HysThreshold!(self)

Raising version of get_L2HysThreshold/1.

get_L2HysThreshold(self)

get_nbins!(self)

Raising version of get_nbins/1.

get_nbins(self)

get_nlevels!(self)

Raising version of get_nlevels/1.

get_nlevels(self)

get_signedGradient!(self)

Raising version of get_signedGradient/1.

get_signedGradient(self)

get_svmDetector!(self)

Raising version of get_svmDetector/1.

get_svmDetector(self)

get_winSigma!(self)

Raising version of get_winSigma/1.

get_winSigma(self)

get_winSize!(self)

Raising version of get_winSize/1.

get_winSize(self)

getDaimlerPeopleDetector!()

Raising version of getDaimlerPeopleDetector/0.

getDefaultPeopleDetector!()

Raising version of getDefaultPeopleDetector/0.

getDescriptorSize!(self)

Raising version of getDescriptorSize/1.

getWinSigma!(self)

Raising version of getWinSigma/1.

hogDescriptor!()

Raising version of hogDescriptor/0.

hogDescriptor!(filename)

Raising version of hogDescriptor/1.

hogDescriptor!(winSize, blockSize, blockStride, cellSize, nbins)

Raising version of hogDescriptor/5.

hogDescriptor!(winSize, blockSize, blockStride, cellSize, nbins, opts)

Raising version of hogDescriptor/6.

load!(self, filename)

Raising version of load/2.

load!(self, filename, opts)

Raising version of load/3.

save!(self, filename)

Raising version of save/2.

save!(self, filename, opts)

Raising version of save/3.

setSVMDetector!(self, svmdetector)

Raising version of setSVMDetector/2.