Evision.HOGDescriptor (Evision v0.1.7-dev)

cv

checkDetectorSize(self)

Checks if detector size equal to descriptor size.

Python prototype (for reference):

checkDetectorSize() -> retval

compute(self, img)

Positional Arguments

img: UMat.

Keyword Arguments

winStride: Size.

padding: Size.

locations: vector_Point.

Computes HOG descriptors of given image.

Python prototype (for reference):

compute(img[, winStride[, padding[, locations]]]) -> descriptors

compute(self, img, opts)

Positional Arguments

img: UMat.

Keyword Arguments

winStride: Size.

padding: Size.

locations: vector_Point.

Computes HOG descriptors of given image.

Python prototype (for reference):

compute(img[, winStride[, padding[, locations]]]) -> descriptors

computeGradient(self, img, grad, angleOfs)

Positional Arguments

img: UMat.

grad: UMat.

angleOfs: UMat.

Keyword Arguments

paddingTL: Size.

paddingBR: Size.

Computes gradients and quantized gradient orientations.

Python prototype (for reference):

computeGradient(img, grad, angleOfs[, paddingTL[, paddingBR]]) -> grad, angleOfs

computeGradient(self, img, grad, angleOfs, opts)

Positional Arguments

img: UMat.

grad: UMat.

angleOfs: UMat.

Keyword Arguments

paddingTL: Size.

paddingBR: Size.

Computes gradients and quantized gradient orientations.

Python prototype (for reference):

computeGradient(img, grad, angleOfs[, paddingTL[, paddingBR]]) -> grad, angleOfs

detect(self, img)

Positional Arguments

img: UMat.

Keyword Arguments

hitThreshold: double.

winStride: Size.

padding: Size.

searchLocations: vector_Point.

Performs object detection without a multi-scale window.

Usually it is 0 and should be specified in the detector coefficients (as the last free coefficient).
But if the free coefficient is omitted (which is allowed), you can specify it manually here.

Python prototype (for reference):

detect(img[, hitThreshold[, winStride[, padding[, searchLocations]]]]) -> foundLocations, weights

detect(self, img, opts)

Positional Arguments

img: UMat.

Keyword Arguments

hitThreshold: double.

winStride: Size.

padding: Size.

searchLocations: vector_Point.

Performs object detection without a multi-scale window.

Usually it is 0 and should be specified in the detector coefficients (as the last free coefficient).
But if the free coefficient is omitted (which is allowed), you can specify it manually here.

Python prototype (for reference):

detect(img[, hitThreshold[, winStride[, padding[, searchLocations]]]]) -> foundLocations, weights

detectMultiScale(self, img)

Positional Arguments

img: UMat.

Keyword Arguments

hitThreshold: double.

winStride: Size.

padding: Size.

scale: double.

groupThreshold: double.

useMeanshiftGrouping: bool.

Detects objects of different sizes in the input image. The detected objects are returned as a list of rectangles.

Usually it is 0 and should be specified in the detector coefficients (as the last free coefficient).
But if the free coefficient is omitted (which is allowed), you can specify it manually here.
by many rectangles. 0 means not to perform grouping.

Python prototype (for reference):

detectMultiScale(img[, hitThreshold[, winStride[, padding[, scale[, groupThreshold[, useMeanshiftGrouping]]]]]]) -> foundLocations, foundWeights

detectMultiScale(self, img, opts)

Positional Arguments

img: UMat.

Keyword Arguments

hitThreshold: double.

winStride: Size.

padding: Size.

scale: double.

groupThreshold: double.

useMeanshiftGrouping: bool.

Detects objects of different sizes in the input image. The detected objects are returned as a list of rectangles.

Usually it is 0 and should be specified in the detector coefficients (as the last free coefficient).
But if the free coefficient is omitted (which is allowed), you can specify it manually here.
by many rectangles. 0 means not to perform grouping.

Python prototype (for reference):

detectMultiScale(img[, hitThreshold[, winStride[, padding[, scale[, groupThreshold[, useMeanshiftGrouping]]]]]]) -> foundLocations, foundWeights

getDaimlerPeopleDetector()

Returns coefficients of the classifier trained for people detection (for 48x96 windows).

Python prototype (for reference):

getDaimlerPeopleDetector() -> retval

getDefaultPeopleDetector()

Returns coefficients of the classifier trained for people detection (for 64x128 windows).

Python prototype (for reference):

getDefaultPeopleDetector() -> retval

getDescriptorSize(self)

Returns the number of coefficients required for the classification.

Python prototype (for reference):

getDescriptorSize() -> retval

getWinSigma(self)

Returns winSigma value

Python prototype (for reference):

getWinSigma() -> retval

hogDescriptor()

Creates the HOG descriptor and detector with default parameters. aqual to HOGDescriptor(Size(64,128), Size(16,16), Size(8,8), Size(8,8), 9 )

Python prototype (for reference):

HOGDescriptor() -> <HOGDescriptor object>

hogDescriptor(filename)

Has overloading in C++

Creates the HOG descriptor and detector and loads HOGDescriptor parameters and coefficients for the linear SVM classifier from a file.

Python prototype (for reference):

HOGDescriptor(filename) -> <HOGDescriptor object>

hogDescriptor(winSize, blockSize, blockStride, cellSize, nbins)

Positional Arguments

winSize: Size.

blockSize: Size.

blockStride: Size.

cellSize: Size.

nbins: int.

Keyword Arguments

derivAperture: int.

winSigma: double.

histogramNormType: HOGDescriptor_HistogramNormType.

l2HysThreshold: double.

gammaCorrection: bool.

nlevels: int.

signedGradient: bool.

Has overloading in C++

Python prototype (for reference):

HOGDescriptor(_winSize, _blockSize, _blockStride, _cellSize, _nbins[, _derivAperture[, _winSigma[, _histogramNormType[, _L2HysThreshold[, _gammaCorrection[, _nlevels[, _signedGradient]]]]]]]) -> <HOGDescriptor object>

hogDescriptor(winSize, blockSize, blockStride, cellSize, nbins, opts)

Positional Arguments

winSize: Size.

blockSize: Size.

blockStride: Size.

cellSize: Size.

nbins: int.

Keyword Arguments

derivAperture: int.

winSigma: double.

histogramNormType: HOGDescriptor_HistogramNormType.

l2HysThreshold: double.

gammaCorrection: bool.

nlevels: int.

signedGradient: bool.

Has overloading in C++

Python prototype (for reference):

HOGDescriptor(_winSize, _blockSize, _blockStride, _cellSize, _nbins[, _derivAperture[, _winSigma[, _histogramNormType[, _L2HysThreshold[, _gammaCorrection[, _nlevels[, _signedGradient]]]]]]]) -> <HOGDescriptor object>

load(self, filename)

Positional Arguments

filename: String.

Keyword Arguments

objname: String.

loads HOGDescriptor parameters and coefficients for the linear SVM classifier from a file

Python prototype (for reference):

load(filename[, objname]) -> retval

load(self, filename, opts)

Positional Arguments

filename: String.

Keyword Arguments

objname: String.

loads HOGDescriptor parameters and coefficients for the linear SVM classifier from a file

Python prototype (for reference):

load(filename[, objname]) -> retval

save(self, filename)

Positional Arguments

filename: String.

Keyword Arguments

objname: String.

saves HOGDescriptor parameters and coefficients for the linear SVM classifier to a file

Python prototype (for reference):

save(filename[, objname]) -> None

save(self, filename, opts)

Positional Arguments

filename: String.

Keyword Arguments

objname: String.

saves HOGDescriptor parameters and coefficients for the linear SVM classifier to a file

Python prototype (for reference):

save(filename[, objname]) -> None

setSVMDetector(self, svmdetector)

Sets coefficients for the linear SVM classifier.

Python prototype (for reference):

setSVMDetector(svmdetector) -> None

Functions

checkDetectorSize!(self)

Raising version of checkDetectorSize/1.

compute!(self, img)

Raising version of compute/2.

compute!(self, img, opts)

Raising version of compute/3.

computeGradient!(self, img, grad, angleOfs)

Raising version of computeGradient/4.

computeGradient!(self, img, grad, angleOfs, opts)

Raising version of computeGradient/5.

detect!(self, img)

Raising version of detect/2.

detect!(self, img, opts)

Raising version of detect/3.

detectMultiScale!(self, img)

Raising version of detectMultiScale/2.

detectMultiScale!(self, img, opts)

Raising version of detectMultiScale/3.

get_blockSize!(self)

Raising version of get_blockSize/1.

get_blockSize(self)

get_blockStride!(self)

Raising version of get_blockStride/1.

get_blockStride(self)

get_cellSize!(self)

Raising version of get_cellSize/1.

get_cellSize(self)

get_derivAperture!(self)

Raising version of get_derivAperture/1.

get_derivAperture(self)

get_gammaCorrection!(self)

Raising version of get_gammaCorrection/1.

get_gammaCorrection(self)

get_histogramNormType!(self)

Raising version of get_histogramNormType/1.

get_histogramNormType(self)

get_L2HysThreshold!(self)

Raising version of get_L2HysThreshold/1.

get_L2HysThreshold(self)

get_nbins!(self)

Raising version of get_nbins/1.

get_nbins(self)

get_nlevels!(self)

Raising version of get_nlevels/1.

get_nlevels(self)

get_signedGradient!(self)

Raising version of get_signedGradient/1.

get_signedGradient(self)

get_svmDetector!(self)

Raising version of get_svmDetector/1.

get_svmDetector(self)

get_winSigma!(self)

Raising version of get_winSigma/1.

get_winSigma(self)

get_winSize!(self)

Raising version of get_winSize/1.

get_winSize(self)

getDaimlerPeopleDetector!()

Raising version of getDaimlerPeopleDetector/0.

getDefaultPeopleDetector!()

Raising version of getDefaultPeopleDetector/0.

getDescriptorSize!(self)

Raising version of getDescriptorSize/1.

getWinSigma!(self)

Raising version of getWinSigma/1.

hogDescriptor!()

Raising version of hogDescriptor/0.

hogDescriptor!(filename)

Raising version of hogDescriptor/1.

hogDescriptor!(winSize, blockSize, blockStride, cellSize, nbins)

Raising version of hogDescriptor/5.

hogDescriptor!(winSize, blockSize, blockStride, cellSize, nbins, opts)

Raising version of hogDescriptor/6.

load!(self, filename)

Raising version of load/2.

load!(self, filename, opts)

Raising version of load/3.

save!(self, filename)

Raising version of save/2.

save!(self, filename, opts)

Raising version of save/3.

setSVMDetector!(self, svmdetector)

Raising version of setSVMDetector/2.