Evision.ML.ANNMLP (Evision v0.1.9)

cv

clear(self)

Clears the algorithm state

Python prototype (for reference):

clear() -> None

getDefaultName(self)

Returns the algorithm string identifier. This string is used as top level xml/yml node tag when the object is saved to a file or string.

Python prototype (for reference):

getDefaultName() -> retval

read(self, fn_)

Reads algorithm parameters from a file storage

Positional Arguments

fn_: FileNode

Python prototype (for reference):

read(fn_) -> None

save(self, filename)

Positional Arguments

filename: String

Saves the algorithm to a file. In order to make this method work, the derived class must implement Algorithm::write(FileStorage& fs).

Python prototype (for reference):

save(filename) -> None

write(self, fs)

simplified API for language bindings

Positional Arguments

fs: Ptr<FileStorage>

Keyword Arguments

name: String.

Has overloading in C++

Python prototype (for reference):

write(fs[, name]) -> None

write(self, fs, opts)

simplified API for language bindings

Positional Arguments

fs: Ptr<FileStorage>

Keyword Arguments

name: String.

Has overloading in C++

Python prototype (for reference):

write(fs[, name]) -> None

cv.ml

calcError(self, data, test)

Computes error on the training or test dataset

Positional Arguments

data: Ptr<TrainData>.

the training data

test: bool.

if true, the error is computed over the test subset of the data, otherwise it's computed over the training subset of the data. Please note that if you loaded a completely different dataset to evaluate already trained classifier, you will probably want not to set the test subset at all with TrainData::setTrainTestSplitRatio and specify test=false, so that the error is computed for the whole new set. Yes, this sounds a bit confusing.

Return

resp: Evision.Mat.

the optional output responses.

The method uses StatModel::predict to compute the error. For regression models the error is computed as RMS, for classifiers - as a percent of missclassified samples (0%-100%).

Python prototype (for reference):

calcError(data, test[, resp]) -> retval, resp

calcError(self, data, test, opts)

Computes error on the training or test dataset

Positional Arguments

data: Ptr<TrainData>.

the training data

test: bool.

if true, the error is computed over the test subset of the data, otherwise it's computed over the training subset of the data. Please note that if you loaded a completely different dataset to evaluate already trained classifier, you will probably want not to set the test subset at all with TrainData::setTrainTestSplitRatio and specify test=false, so that the error is computed for the whole new set. Yes, this sounds a bit confusing.

Return

resp: Evision.Mat.

the optional output responses.

The method uses StatModel::predict to compute the error. For regression models the error is computed as RMS, for classifiers - as a percent of missclassified samples (0%-100%).

Python prototype (for reference):

calcError(data, test[, resp]) -> retval, resp

create()

Creates empty model

Use StatModel::train to train the model, Algorithm::load\<ANN_MLP>(filename) to load the pre-trained model. Note that the train method has optional flags: ANN_MLP::TrainFlags.

Python prototype (for reference):

create() -> retval

empty(self)

Python prototype (for reference):

empty() -> retval

getAnnealCoolingRatio(self)

@see setAnnealCoolingRatio

Python prototype (for reference):

getAnnealCoolingRatio() -> retval

getAnnealFinalT(self)

@see setAnnealFinalT

Python prototype (for reference):

getAnnealFinalT() -> retval

getAnnealInitialT(self)

@see setAnnealInitialT

Python prototype (for reference):

getAnnealInitialT() -> retval

getAnnealItePerStep(self)

@see setAnnealItePerStep

Python prototype (for reference):

getAnnealItePerStep() -> retval

getBackpropMomentumScale(self)

@see setBackpropMomentumScale

Python prototype (for reference):

getBackpropMomentumScale() -> retval

getBackpropWeightScale(self)

@see setBackpropWeightScale

Python prototype (for reference):

getBackpropWeightScale() -> retval

getLayerSizes(self)

Integer vector specifying the number of neurons in each layer including the input and output layers. The very first element specifies the number of elements in the input layer. The last element - number of elements in the output layer. @sa setLayerSizes

Python prototype (for reference):

getLayerSizes() -> retval

getRpropDW0(self)

@see setRpropDW0

Python prototype (for reference):

getRpropDW0() -> retval

getRpropDWMax(self)

@see setRpropDWMax

Python prototype (for reference):

getRpropDWMax() -> retval

getRpropDWMin(self)

@see setRpropDWMin

Python prototype (for reference):

getRpropDWMin() -> retval

getRpropDWMinus(self)

@see setRpropDWMinus

Python prototype (for reference):

getRpropDWMinus() -> retval

getRpropDWPlus(self)

@see setRpropDWPlus

Python prototype (for reference):

getRpropDWPlus() -> retval

getTermCriteria(self)

@see setTermCriteria

Python prototype (for reference):

getTermCriteria() -> retval

getTrainMethod(self)

Returns current training method

Python prototype (for reference):

getTrainMethod() -> retval

getVarCount(self)

Returns the number of variables in training samples

Python prototype (for reference):

getVarCount() -> retval

getWeights(self, layerIdx)

Positional Arguments

layerIdx: int

Python prototype (for reference):

getWeights(layerIdx) -> retval

isClassifier(self)

Returns true if the model is classifier

Python prototype (for reference):

isClassifier() -> retval

isTrained(self)

Returns true if the model is trained

Python prototype (for reference):

isTrained() -> retval

load(filepath)

Loads and creates a serialized ANN from a file

Positional Arguments

filepath: String.

path to serialized ANN

Use ANN::save to serialize and store an ANN to disk. Load the ANN from this file again, by calling this function with the path to the file.

Python prototype (for reference):

load(filepath) -> retval

predict(self, samples)

Predicts response(s) for the provided sample(s)

Positional Arguments

samples: Evision.Mat.

The input samples, floating-point matrix

Keyword Arguments

flags: int.

The optional flags, model-dependent. See cv::ml::StatModel::Flags.

Return

results: Evision.Mat.

The optional output matrix of results.

Python prototype (for reference):

predict(samples[, results[, flags]]) -> retval, results

predict(self, samples, opts)

Predicts response(s) for the provided sample(s)

Positional Arguments

samples: Evision.Mat.

The input samples, floating-point matrix

Keyword Arguments

flags: int.

The optional flags, model-dependent. See cv::ml::StatModel::Flags.

Return

results: Evision.Mat.

The optional output matrix of results.

Python prototype (for reference):

predict(samples[, results[, flags]]) -> retval, results

setActivationFunction(self, type)

Positional Arguments

type: int.

The type of activation function. See ANN_MLP::ActivationFunctions.

Keyword Arguments

param1: double.

The first parameter of the activation function, \f$\alpha\f$. Default value is 0.

param2: double.

The second parameter of the activation function, \f$\beta\f$. Default value is 0.

Initialize the activation function for each neuron. Currently the default and the only fully supported activation function is ANN_MLP::SIGMOID_SYM.

Python prototype (for reference):

setActivationFunction(type[, param1[, param2]]) -> None

setActivationFunction(self, type, opts)

Positional Arguments

type: int.

The type of activation function. See ANN_MLP::ActivationFunctions.

Keyword Arguments

param1: double.

The first parameter of the activation function, \f$\alpha\f$. Default value is 0.

param2: double.

The second parameter of the activation function, \f$\beta\f$. Default value is 0.

Initialize the activation function for each neuron. Currently the default and the only fully supported activation function is ANN_MLP::SIGMOID_SYM.

Python prototype (for reference):

setActivationFunction(type[, param1[, param2]]) -> None

setAnnealCoolingRatio(self, val)

Positional Arguments

val: double

@copybrief getAnnealCoolingRatio @see getAnnealCoolingRatio

Python prototype (for reference):

setAnnealCoolingRatio(val) -> None

setAnnealFinalT(self, val)

Positional Arguments

val: double

@copybrief getAnnealFinalT @see getAnnealFinalT

Python prototype (for reference):

setAnnealFinalT(val) -> None

setAnnealInitialT(self, val)

Positional Arguments

val: double

@copybrief getAnnealInitialT @see getAnnealInitialT

Python prototype (for reference):

setAnnealInitialT(val) -> None

setAnnealItePerStep(self, val)

Positional Arguments

val: int

@copybrief getAnnealItePerStep @see getAnnealItePerStep

Python prototype (for reference):

setAnnealItePerStep(val) -> None

setBackpropMomentumScale(self, val)

Positional Arguments

val: double

@copybrief getBackpropMomentumScale @see getBackpropMomentumScale

Python prototype (for reference):

setBackpropMomentumScale(val) -> None

setBackpropWeightScale(self, val)

Positional Arguments

val: double

@copybrief getBackpropWeightScale @see getBackpropWeightScale

Python prototype (for reference):

setBackpropWeightScale(val) -> None

setLayerSizes(self, layer_sizes)

Positional Arguments

layer_sizes: Evision.Mat

Integer vector specifying the number of neurons in each layer including the input and output layers. The very first element specifies the number of elements in the input layer. The last element - number of elements in the output layer. Default value is empty Mat. @sa getLayerSizes

Python prototype (for reference):

setLayerSizes(_layer_sizes) -> None

setRpropDW0(self, val)

Positional Arguments

val: double

@copybrief getRpropDW0 @see getRpropDW0

Python prototype (for reference):

setRpropDW0(val) -> None

setRpropDWMax(self, val)

Positional Arguments

val: double

@copybrief getRpropDWMax @see getRpropDWMax

Python prototype (for reference):

setRpropDWMax(val) -> None

setRpropDWMin(self, val)

Positional Arguments

val: double

@copybrief getRpropDWMin @see getRpropDWMin

Python prototype (for reference):

setRpropDWMin(val) -> None

setRpropDWMinus(self, val)

Positional Arguments

val: double

@copybrief getRpropDWMinus @see getRpropDWMinus

Python prototype (for reference):

setRpropDWMinus(val) -> None

setRpropDWPlus(self, val)

Positional Arguments

val: double

@copybrief getRpropDWPlus @see getRpropDWPlus

Python prototype (for reference):

setRpropDWPlus(val) -> None

setTermCriteria(self, val)

Positional Arguments

val: TermCriteria

@copybrief getTermCriteria @see getTermCriteria

Python prototype (for reference):

setTermCriteria(val) -> None

setTrainMethod(self, method)

Positional Arguments

method: int.

Default value is ANN_MLP::RPROP. See ANN_MLP::TrainingMethods.

Keyword Arguments

param1: double.

passed to setRpropDW0 for ANN_MLP::RPROP and to setBackpropWeightScale for ANN_MLP::BACKPROP and to initialT for ANN_MLP::ANNEAL.

param2: double.

passed to setRpropDWMin for ANN_MLP::RPROP and to setBackpropMomentumScale for ANN_MLP::BACKPROP and to finalT for ANN_MLP::ANNEAL.

Sets training method and common parameters.

Python prototype (for reference):

setTrainMethod(method[, param1[, param2]]) -> None

setTrainMethod(self, method, opts)

Positional Arguments

method: int.

Default value is ANN_MLP::RPROP. See ANN_MLP::TrainingMethods.

Keyword Arguments

param1: double.

passed to setRpropDW0 for ANN_MLP::RPROP and to setBackpropWeightScale for ANN_MLP::BACKPROP and to initialT for ANN_MLP::ANNEAL.

param2: double.

passed to setRpropDWMin for ANN_MLP::RPROP and to setBackpropMomentumScale for ANN_MLP::BACKPROP and to finalT for ANN_MLP::ANNEAL.

Sets training method and common parameters.

Python prototype (for reference):

setTrainMethod(method[, param1[, param2]]) -> None

train(self, trainData)

Trains the statistical model

Positional Arguments

trainData: Ptr<TrainData>.

training data that can be loaded from file using TrainData::loadFromCSV or created with TrainData::create.

Keyword Arguments

flags: int.

optional flags, depending on the model. Some of the models can be updated with the new training samples, not completely overwritten (such as NormalBayesClassifier or ANN_MLP).

Python prototype (for reference):

train(trainData[, flags]) -> retval

train(self, trainData, opts)

Trains the statistical model

Positional Arguments

trainData: Ptr<TrainData>.

training data that can be loaded from file using TrainData::loadFromCSV or created with TrainData::create.

Keyword Arguments

flags: int.

optional flags, depending on the model. Some of the models can be updated with the new training samples, not completely overwritten (such as NormalBayesClassifier or ANN_MLP).

Python prototype (for reference):

train(trainData[, flags]) -> retval

train(self, samples, layout, responses)

Trains the statistical model

Positional Arguments

samples: Evision.Mat.

training samples

layout: int.

See ml::SampleTypes.

responses: Evision.Mat.

vector of responses associated with the training samples.

Python prototype (for reference):

train(samples, layout, responses) -> retval

Functions

calcError!(self, data, test)

Raising version of calcError/3.

calcError!(self, data, test, opts)

Raising version of calcError/4.

clear!(self)

Raising version of clear/1.

create!()

Raising version of create/0.

empty!(self)

Raising version of empty/1.

getAnnealCoolingRatio!(self)

Raising version of getAnnealCoolingRatio/1.

getAnnealFinalT!(self)

Raising version of getAnnealFinalT/1.

getAnnealInitialT!(self)

Raising version of getAnnealInitialT/1.

getAnnealItePerStep!(self)

Raising version of getAnnealItePerStep/1.

getBackpropMomentumScale!(self)

Raising version of getBackpropMomentumScale/1.

getBackpropWeightScale!(self)

Raising version of getBackpropWeightScale/1.

getDefaultName!(self)

Raising version of getDefaultName/1.

getLayerSizes!(self)

Raising version of getLayerSizes/1.

getRpropDW0!(self)

Raising version of getRpropDW0/1.

getRpropDWMax!(self)

Raising version of getRpropDWMax/1.

getRpropDWMin!(self)

Raising version of getRpropDWMin/1.

getRpropDWMinus!(self)

Raising version of getRpropDWMinus/1.

getRpropDWPlus!(self)

Raising version of getRpropDWPlus/1.

getTermCriteria!(self)

Raising version of getTermCriteria/1.

getTrainMethod!(self)

Raising version of getTrainMethod/1.

getVarCount!(self)

Raising version of getVarCount/1.

getWeights!(self, layerIdx)

Raising version of getWeights/2.

isClassifier!(self)

Raising version of isClassifier/1.

isTrained!(self)

Raising version of isTrained/1.

load!(filepath)

Raising version of load/1.

predict!(self, samples)

Raising version of predict/2.

predict!(self, samples, opts)

Raising version of predict/3.

read!(self, fn_)

Raising version of read/2.

save!(self, filename)

Raising version of save/2.

setActivationFunction!(self, type)

Raising version of setActivationFunction/2.

setActivationFunction!(self, type, opts)

Raising version of setActivationFunction/3.

setAnnealCoolingRatio!(self, val)

Raising version of setAnnealCoolingRatio/2.

setAnnealFinalT!(self, val)

Raising version of setAnnealFinalT/2.

setAnnealInitialT!(self, val)

Raising version of setAnnealInitialT/2.

setAnnealItePerStep!(self, val)

Raising version of setAnnealItePerStep/2.

setBackpropMomentumScale!(self, val)

Raising version of setBackpropMomentumScale/2.

setBackpropWeightScale!(self, val)

Raising version of setBackpropWeightScale/2.

setLayerSizes!(self, layer_sizes)

Raising version of setLayerSizes/2.

setRpropDW0!(self, val)

Raising version of setRpropDW0/2.

setRpropDWMax!(self, val)

Raising version of setRpropDWMax/2.

setRpropDWMin!(self, val)

Raising version of setRpropDWMin/2.

setRpropDWMinus!(self, val)

Raising version of setRpropDWMinus/2.

setRpropDWPlus!(self, val)

Raising version of setRpropDWPlus/2.

setTermCriteria!(self, val)

Raising version of setTermCriteria/2.

setTrainMethod!(self, method)

Raising version of setTrainMethod/2.

setTrainMethod!(self, method, opts)

Raising version of setTrainMethod/3.

train!(self, trainData)

Raising version of train/2.

train!(self, trainData, opts)

Raising version of train/3.

train!(self, samples, layout, responses)

Raising version of train/4.

write!(self, fs)

Raising version of write/2.

write!(self, fs, opts)

Raising version of write/3.