Evision.ML.Boost (Evision v0.1.7-dev)

cv

clear(self)

Clears the algorithm state

Python prototype (for reference):

clear() -> None

getDefaultName(self)

Returns the algorithm string identifier. This string is used as top level xml/yml node tag when the object is saved to a file or string.

Python prototype (for reference):

getDefaultName() -> retval

read(self, fn_)

Reads algorithm parameters from a file storage

Python prototype (for reference):

read(fn_) -> None

save(self, filename)

Saves the algorithm to a file. In order to make this method work, the derived class must implement Algorithm::write(FileStorage& fs).

Python prototype (for reference):

save(filename) -> None

write(self, fs)

Positional Arguments

fs: Ptr<FileStorage>

Keyword Arguments

name: String.

simplified API for language bindings Has overloading in C++

Python prototype (for reference):

write(fs[, name]) -> None

write(self, fs, opts)

Positional Arguments

fs: Ptr<FileStorage>

Keyword Arguments

name: String.

simplified API for language bindings Has overloading in C++

Python prototype (for reference):

write(fs[, name]) -> None

cv.ml

calcError(self, data, test)

Positional Arguments

data: Ptr<TrainData>.

test: bool.

Keyword Arguments

resp: UMat.

Computes error on the training or test dataset

computed over the training subset of the data. Please note that if you loaded a completely
different dataset to evaluate already trained classifier, you will probably want not to set
the test subset at all with TrainData::setTrainTestSplitRatio and specify test=false, so
that the error is computed for the whole new set. Yes, this sounds a bit confusing.

The method uses StatModel::predict to compute the error. For regression models the error is computed as RMS, for classifiers - as a percent of missclassified samples (0%-100%).

Python prototype (for reference):

calcError(data, test[, resp]) -> retval, resp

calcError(self, data, test, opts)

Positional Arguments

data: Ptr<TrainData>.

test: bool.

Keyword Arguments

resp: UMat.

Computes error on the training or test dataset

computed over the training subset of the data. Please note that if you loaded a completely
different dataset to evaluate already trained classifier, you will probably want not to set
the test subset at all with TrainData::setTrainTestSplitRatio and specify test=false, so
that the error is computed for the whole new set. Yes, this sounds a bit confusing.

The method uses StatModel::predict to compute the error. For regression models the error is computed as RMS, for classifiers - as a percent of missclassified samples (0%-100%).

Python prototype (for reference):

calcError(data, test[, resp]) -> retval, resp

create()

Creates the empty model. Use StatModel::train to train the model, Algorithm::load\<Boost>(filename) to load the pre-trained model.

Python prototype (for reference):

create() -> retval

empty(self)

Python prototype (for reference):

empty() -> retval

getBoostType(self)

@see setBoostType

Python prototype (for reference):

getBoostType() -> retval

getCVFolds(self)

@see setCVFolds

Python prototype (for reference):

getCVFolds() -> retval

getMaxCategories(self)

@see setMaxCategories

Python prototype (for reference):

getMaxCategories() -> retval

getMaxDepth(self)

@see setMaxDepth

Python prototype (for reference):

getMaxDepth() -> retval

getMinSampleCount(self)

@see setMinSampleCount

Python prototype (for reference):

getMinSampleCount() -> retval

getPriors(self)

@see setPriors

Python prototype (for reference):

getPriors() -> retval

getRegressionAccuracy(self)

@see setRegressionAccuracy

Python prototype (for reference):

getRegressionAccuracy() -> retval

getTruncatePrunedTree(self)

@see setTruncatePrunedTree

Python prototype (for reference):

getTruncatePrunedTree() -> retval

getUse1SERule(self)

@see setUse1SERule

Python prototype (for reference):

getUse1SERule() -> retval

getUseSurrogates(self)

@see setUseSurrogates

Python prototype (for reference):

getUseSurrogates() -> retval

getVarCount(self)

Returns the number of variables in training samples

Python prototype (for reference):

getVarCount() -> retval

getWeakCount(self)

@see setWeakCount

Python prototype (for reference):

getWeakCount() -> retval

getWeightTrimRate(self)

@see setWeightTrimRate

Python prototype (for reference):

getWeightTrimRate() -> retval

isClassifier(self)

Returns true if the model is classifier

Python prototype (for reference):

isClassifier() -> retval

isTrained(self)

Returns true if the model is trained

Python prototype (for reference):

isTrained() -> retval

load(filepath)

Positional Arguments

filepath: String.

Keyword Arguments

nodeName: String.

Loads and creates a serialized Boost from a file Use Boost::save to serialize and store an RTree to disk. Load the Boost from this file again, by calling this function with the path to the file. Optionally specify the node for the file containing the classifier

Python prototype (for reference):

load(filepath[, nodeName]) -> retval

load(filepath, opts)

Positional Arguments

filepath: String.

Keyword Arguments

nodeName: String.

Loads and creates a serialized Boost from a file Use Boost::save to serialize and store an RTree to disk. Load the Boost from this file again, by calling this function with the path to the file. Optionally specify the node for the file containing the classifier

Python prototype (for reference):

load(filepath[, nodeName]) -> retval

predict(self, samples)

Positional Arguments

samples: UMat.

Keyword Arguments

results: UMat.

flags: int.

Predicts response(s) for the provided sample(s)

Python prototype (for reference):

predict(samples[, results[, flags]]) -> retval, results

predict(self, samples, opts)

Positional Arguments

samples: UMat.

Keyword Arguments

results: UMat.

flags: int.

Predicts response(s) for the provided sample(s)

Python prototype (for reference):

predict(samples[, results[, flags]]) -> retval, results

setBoostType(self, val)

@copybrief getBoostType @see getBoostType

Python prototype (for reference):

setBoostType(val) -> None

setCVFolds(self, val)

@copybrief getCVFolds @see getCVFolds

Python prototype (for reference):

setCVFolds(val) -> None

setMaxCategories(self, val)

@copybrief getMaxCategories @see getMaxCategories

Python prototype (for reference):

setMaxCategories(val) -> None

setMaxDepth(self, val)

@copybrief getMaxDepth @see getMaxDepth

Python prototype (for reference):

setMaxDepth(val) -> None

setMinSampleCount(self, val)

@copybrief getMinSampleCount @see getMinSampleCount

Python prototype (for reference):

setMinSampleCount(val) -> None

setPriors(self, val)

@copybrief getPriors @see getPriors

Python prototype (for reference):

setPriors(val) -> None

setRegressionAccuracy(self, val)

@copybrief getRegressionAccuracy @see getRegressionAccuracy

Python prototype (for reference):

setRegressionAccuracy(val) -> None

setTruncatePrunedTree(self, val)

@copybrief getTruncatePrunedTree @see getTruncatePrunedTree

Python prototype (for reference):

setTruncatePrunedTree(val) -> None

setUse1SERule(self, val)

@copybrief getUse1SERule @see getUse1SERule

Python prototype (for reference):

setUse1SERule(val) -> None

setUseSurrogates(self, val)

@copybrief getUseSurrogates @see getUseSurrogates

Python prototype (for reference):

setUseSurrogates(val) -> None

setWeakCount(self, val)

@copybrief getWeakCount @see getWeakCount

Python prototype (for reference):

setWeakCount(val) -> None

setWeightTrimRate(self, val)

@copybrief getWeightTrimRate @see getWeightTrimRate

Python prototype (for reference):

setWeightTrimRate(val) -> None

train(self, trainData)

Positional Arguments

trainData: Ptr<TrainData>.

Keyword Arguments

flags: int.

Trains the statistical model

created with TrainData::create.
new training samples, not completely overwritten (such as NormalBayesClassifier or ANN_MLP).

Python prototype (for reference):

train(trainData[, flags]) -> retval

train(self, trainData, opts)

Positional Arguments

trainData: Ptr<TrainData>.

Keyword Arguments

flags: int.

Trains the statistical model

created with TrainData::create.
new training samples, not completely overwritten (such as NormalBayesClassifier or ANN_MLP).

Python prototype (for reference):

train(trainData[, flags]) -> retval

train(self, samples, layout, responses)

Trains the statistical model

Python prototype (for reference):

train(samples, layout, responses) -> retval

Functions

calcError!(self, data, test)

Raising version of calcError/3.

calcError!(self, data, test, opts)

Raising version of calcError/4.

clear!(self)

Raising version of clear/1.

create!()

Raising version of create/0.

empty!(self)

Raising version of empty/1.

getBoostType!(self)

Raising version of getBoostType/1.

getCVFolds!(self)

Raising version of getCVFolds/1.

getDefaultName!(self)

Raising version of getDefaultName/1.

getMaxCategories!(self)

Raising version of getMaxCategories/1.

getMaxDepth!(self)

Raising version of getMaxDepth/1.

getMinSampleCount!(self)

Raising version of getMinSampleCount/1.

getPriors!(self)

Raising version of getPriors/1.

getRegressionAccuracy!(self)

Raising version of getRegressionAccuracy/1.

getTruncatePrunedTree!(self)

Raising version of getTruncatePrunedTree/1.

getUse1SERule!(self)

Raising version of getUse1SERule/1.

getUseSurrogates!(self)

Raising version of getUseSurrogates/1.

getVarCount!(self)

Raising version of getVarCount/1.

getWeakCount!(self)

Raising version of getWeakCount/1.

getWeightTrimRate!(self)

Raising version of getWeightTrimRate/1.

isClassifier!(self)

Raising version of isClassifier/1.

isTrained!(self)

Raising version of isTrained/1.

load!(filepath)

Raising version of load/1.

load!(filepath, opts)

Raising version of load/2.

predict!(self, samples)

Raising version of predict/2.

predict!(self, samples, opts)

Raising version of predict/3.

read!(self, fn_)

Raising version of read/2.

save!(self, filename)

Raising version of save/2.

setBoostType!(self, val)

Raising version of setBoostType/2.

setCVFolds!(self, val)

Raising version of setCVFolds/2.

setMaxCategories!(self, val)

Raising version of setMaxCategories/2.

setMaxDepth!(self, val)

Raising version of setMaxDepth/2.

setMinSampleCount!(self, val)

Raising version of setMinSampleCount/2.

setPriors!(self, val)

Raising version of setPriors/2.

setRegressionAccuracy!(self, val)

Raising version of setRegressionAccuracy/2.

setTruncatePrunedTree!(self, val)

Raising version of setTruncatePrunedTree/2.

setUse1SERule!(self, val)

Raising version of setUse1SERule/2.

setUseSurrogates!(self, val)

Raising version of setUseSurrogates/2.

setWeakCount!(self, val)

Raising version of setWeakCount/2.

setWeightTrimRate!(self, val)

Raising version of setWeightTrimRate/2.

train!(self, trainData)

Raising version of train/2.

train!(self, trainData, opts)

Raising version of train/3.

train!(self, samples, layout, responses)

Raising version of train/4.

write!(self, fs)

Raising version of write/2.

write!(self, fs, opts)

Raising version of write/3.