Evision.ML.RTrees (Evision v0.1.9)

cv

clear(self)

Clears the algorithm state

Python prototype (for reference):

clear() -> None

getDefaultName(self)

Returns the algorithm string identifier. This string is used as top level xml/yml node tag when the object is saved to a file or string.

Python prototype (for reference):

getDefaultName() -> retval

read(self, fn_)

Reads algorithm parameters from a file storage

Positional Arguments

fn_: FileNode

Python prototype (for reference):

read(fn_) -> None

save(self, filename)

Positional Arguments

filename: String

Saves the algorithm to a file. In order to make this method work, the derived class must implement Algorithm::write(FileStorage& fs).

Python prototype (for reference):

save(filename) -> None

write(self, fs)

simplified API for language bindings

Positional Arguments

fs: Ptr<FileStorage>

Keyword Arguments

name: String.

Has overloading in C++

Python prototype (for reference):

write(fs[, name]) -> None

write(self, fs, opts)

simplified API for language bindings

Positional Arguments

fs: Ptr<FileStorage>

Keyword Arguments

name: String.

Has overloading in C++

Python prototype (for reference):

write(fs[, name]) -> None

cv.ml

calcError(self, data, test)

Computes error on the training or test dataset

Positional Arguments

data: Ptr<TrainData>.

the training data

test: bool.

if true, the error is computed over the test subset of the data, otherwise it's computed over the training subset of the data. Please note that if you loaded a completely different dataset to evaluate already trained classifier, you will probably want not to set the test subset at all with TrainData::setTrainTestSplitRatio and specify test=false, so that the error is computed for the whole new set. Yes, this sounds a bit confusing.

Return

resp: Evision.Mat.

the optional output responses.

The method uses StatModel::predict to compute the error. For regression models the error is computed as RMS, for classifiers - as a percent of missclassified samples (0%-100%).

Python prototype (for reference):

calcError(data, test[, resp]) -> retval, resp

calcError(self, data, test, opts)

Computes error on the training or test dataset

Positional Arguments

data: Ptr<TrainData>.

the training data

test: bool.

if true, the error is computed over the test subset of the data, otherwise it's computed over the training subset of the data. Please note that if you loaded a completely different dataset to evaluate already trained classifier, you will probably want not to set the test subset at all with TrainData::setTrainTestSplitRatio and specify test=false, so that the error is computed for the whole new set. Yes, this sounds a bit confusing.

Return

resp: Evision.Mat.

the optional output responses.

The method uses StatModel::predict to compute the error. For regression models the error is computed as RMS, for classifiers - as a percent of missclassified samples (0%-100%).

Python prototype (for reference):

calcError(data, test[, resp]) -> retval, resp

create()

Creates the empty model. Use StatModel::train to train the model, StatModel::train to create and train the model, Algorithm::load to load the pre-trained model.

Python prototype (for reference):

create() -> retval

empty(self)

Python prototype (for reference):

empty() -> retval

getActiveVarCount(self)

@see setActiveVarCount

Python prototype (for reference):

getActiveVarCount() -> retval

getCalculateVarImportance(self)

@see setCalculateVarImportance

Python prototype (for reference):

getCalculateVarImportance() -> retval

getCVFolds(self)

@see setCVFolds

Python prototype (for reference):

getCVFolds() -> retval

getMaxCategories(self)

@see setMaxCategories

Python prototype (for reference):

getMaxCategories() -> retval

getMaxDepth(self)

@see setMaxDepth

Python prototype (for reference):

getMaxDepth() -> retval

getMinSampleCount(self)

@see setMinSampleCount

Python prototype (for reference):

getMinSampleCount() -> retval

getOOBError(self)

Returns the OOB error value, computed at the training stage when calcOOBError is set to true. If this flag was set to false, 0 is returned. The OOB error is also scaled by sample weighting.

Python prototype (for reference):

getOOBError() -> retval

getPriors(self)

@see setPriors

Python prototype (for reference):

getPriors() -> retval

getRegressionAccuracy(self)

@see setRegressionAccuracy

Python prototype (for reference):

getRegressionAccuracy() -> retval

getTermCriteria(self)

@see setTermCriteria

Python prototype (for reference):

getTermCriteria() -> retval

getTruncatePrunedTree(self)

@see setTruncatePrunedTree

Python prototype (for reference):

getTruncatePrunedTree() -> retval

getUse1SERule(self)

@see setUse1SERule

Python prototype (for reference):

getUse1SERule() -> retval

getUseSurrogates(self)

@see setUseSurrogates

Python prototype (for reference):

getUseSurrogates() -> retval

getVarCount(self)

Returns the number of variables in training samples

Python prototype (for reference):

getVarCount() -> retval

getVarImportance(self)

Returns the variable importance array. The method returns the variable importance vector, computed at the training stage when CalculateVarImportance is set to true. If this flag was set to false, the empty matrix is returned.

Python prototype (for reference):

getVarImportance() -> retval

getVotes(self, samples, flags)

Positional Arguments

samples: Evision.Mat.

Array containing the samples for which votes will be calculated.

flags: int.

Flags for defining the type of RTrees.

Return

results: Evision.Mat.

Array where the result of the calculation will be written.

Returns the result of each individual tree in the forest. In case the model is a regression problem, the method will return each of the trees' results for each of the sample cases. If the model is a classifier, it will return a Mat with samples + 1 rows, where the first row gives the class number and the following rows return the votes each class had for each sample.

Python prototype (for reference):

getVotes(samples, flags[, results]) -> results

getVotes(self, samples, flags, opts)

Positional Arguments

samples: Evision.Mat.

Array containing the samples for which votes will be calculated.

flags: int.

Flags for defining the type of RTrees.

Return

results: Evision.Mat.

Array where the result of the calculation will be written.

Returns the result of each individual tree in the forest. In case the model is a regression problem, the method will return each of the trees' results for each of the sample cases. If the model is a classifier, it will return a Mat with samples + 1 rows, where the first row gives the class number and the following rows return the votes each class had for each sample.

Python prototype (for reference):

getVotes(samples, flags[, results]) -> results

isClassifier(self)

Returns true if the model is classifier

Python prototype (for reference):

isClassifier() -> retval

isTrained(self)

Returns true if the model is trained

Python prototype (for reference):

isTrained() -> retval

load(filepath)

Loads and creates a serialized RTree from a file

Positional Arguments

filepath: String.

path to serialized RTree

Keyword Arguments

nodeName: String.

name of node containing the classifier

Use RTree::save to serialize and store an RTree to disk. Load the RTree from this file again, by calling this function with the path to the file. Optionally specify the node for the file containing the classifier

Python prototype (for reference):

load(filepath[, nodeName]) -> retval

load(filepath, opts)

Loads and creates a serialized RTree from a file

Positional Arguments

filepath: String.

path to serialized RTree

Keyword Arguments

nodeName: String.

name of node containing the classifier

Use RTree::save to serialize and store an RTree to disk. Load the RTree from this file again, by calling this function with the path to the file. Optionally specify the node for the file containing the classifier

Python prototype (for reference):

load(filepath[, nodeName]) -> retval

predict(self, samples)

Predicts response(s) for the provided sample(s)

Positional Arguments

samples: Evision.Mat.

The input samples, floating-point matrix

Keyword Arguments

flags: int.

The optional flags, model-dependent. See cv::ml::StatModel::Flags.

Return

results: Evision.Mat.

The optional output matrix of results.

Python prototype (for reference):

predict(samples[, results[, flags]]) -> retval, results

predict(self, samples, opts)

Predicts response(s) for the provided sample(s)

Positional Arguments

samples: Evision.Mat.

The input samples, floating-point matrix

Keyword Arguments

flags: int.

The optional flags, model-dependent. See cv::ml::StatModel::Flags.

Return

results: Evision.Mat.

The optional output matrix of results.

Python prototype (for reference):

predict(samples[, results[, flags]]) -> retval, results

setActiveVarCount(self, val)

Positional Arguments

val: int

@copybrief getActiveVarCount @see getActiveVarCount

Python prototype (for reference):

setActiveVarCount(val) -> None

setCalculateVarImportance(self, val)

Positional Arguments

val: bool

@copybrief getCalculateVarImportance @see getCalculateVarImportance

Python prototype (for reference):

setCalculateVarImportance(val) -> None

setCVFolds(self, val)

Positional Arguments

val: int

@copybrief getCVFolds @see getCVFolds

Python prototype (for reference):

setCVFolds(val) -> None

setMaxCategories(self, val)

Positional Arguments

val: int

@copybrief getMaxCategories @see getMaxCategories

Python prototype (for reference):

setMaxCategories(val) -> None

setMaxDepth(self, val)

Positional Arguments

val: int

@copybrief getMaxDepth @see getMaxDepth

Python prototype (for reference):

setMaxDepth(val) -> None

setMinSampleCount(self, val)

Positional Arguments

val: int

@copybrief getMinSampleCount @see getMinSampleCount

Python prototype (for reference):

setMinSampleCount(val) -> None

setPriors(self, val)

Positional Arguments

val: Evision.Mat

@copybrief getPriors @see getPriors

Python prototype (for reference):

setPriors(val) -> None

setRegressionAccuracy(self, val)

Positional Arguments

val: float

@copybrief getRegressionAccuracy @see getRegressionAccuracy

Python prototype (for reference):

setRegressionAccuracy(val) -> None

setTermCriteria(self, val)

Positional Arguments

val: TermCriteria

@copybrief getTermCriteria @see getTermCriteria

Python prototype (for reference):

setTermCriteria(val) -> None

setTruncatePrunedTree(self, val)

Positional Arguments

val: bool

@copybrief getTruncatePrunedTree @see getTruncatePrunedTree

Python prototype (for reference):

setTruncatePrunedTree(val) -> None

setUse1SERule(self, val)

Positional Arguments

val: bool

@copybrief getUse1SERule @see getUse1SERule

Python prototype (for reference):

setUse1SERule(val) -> None

setUseSurrogates(self, val)

Positional Arguments

val: bool

@copybrief getUseSurrogates @see getUseSurrogates

Python prototype (for reference):

setUseSurrogates(val) -> None

train(self, trainData)

Trains the statistical model

Positional Arguments

trainData: Ptr<TrainData>.

training data that can be loaded from file using TrainData::loadFromCSV or created with TrainData::create.

Keyword Arguments

flags: int.

optional flags, depending on the model. Some of the models can be updated with the new training samples, not completely overwritten (such as NormalBayesClassifier or ANN_MLP).

Python prototype (for reference):

train(trainData[, flags]) -> retval

train(self, trainData, opts)

Trains the statistical model

Positional Arguments

trainData: Ptr<TrainData>.

training data that can be loaded from file using TrainData::loadFromCSV or created with TrainData::create.

Keyword Arguments

flags: int.

optional flags, depending on the model. Some of the models can be updated with the new training samples, not completely overwritten (such as NormalBayesClassifier or ANN_MLP).

Python prototype (for reference):

train(trainData[, flags]) -> retval

train(self, samples, layout, responses)

Trains the statistical model

Positional Arguments

samples: Evision.Mat.

training samples

layout: int.

See ml::SampleTypes.

responses: Evision.Mat.

vector of responses associated with the training samples.

Python prototype (for reference):

train(samples, layout, responses) -> retval

Functions

calcError!(self, data, test)

Raising version of calcError/3.

calcError!(self, data, test, opts)

Raising version of calcError/4.

clear!(self)

Raising version of clear/1.

create!()

Raising version of create/0.

empty!(self)

Raising version of empty/1.

getActiveVarCount!(self)

Raising version of getActiveVarCount/1.

getCalculateVarImportance!(self)

Raising version of getCalculateVarImportance/1.

getCVFolds!(self)

Raising version of getCVFolds/1.

getDefaultName!(self)

Raising version of getDefaultName/1.

getMaxCategories!(self)

Raising version of getMaxCategories/1.

getMaxDepth!(self)

Raising version of getMaxDepth/1.

getMinSampleCount!(self)

Raising version of getMinSampleCount/1.

getOOBError!(self)

Raising version of getOOBError/1.

getPriors!(self)

Raising version of getPriors/1.

getRegressionAccuracy!(self)

Raising version of getRegressionAccuracy/1.

getTermCriteria!(self)

Raising version of getTermCriteria/1.

getTruncatePrunedTree!(self)

Raising version of getTruncatePrunedTree/1.

getUse1SERule!(self)

Raising version of getUse1SERule/1.

getUseSurrogates!(self)

Raising version of getUseSurrogates/1.

getVarCount!(self)

Raising version of getVarCount/1.

getVarImportance!(self)

Raising version of getVarImportance/1.

getVotes!(self, samples, flags)

Raising version of getVotes/3.

getVotes!(self, samples, flags, opts)

Raising version of getVotes/4.

isClassifier!(self)

Raising version of isClassifier/1.

isTrained!(self)

Raising version of isTrained/1.

load!(filepath)

Raising version of load/1.

load!(filepath, opts)

Raising version of load/2.

predict!(self, samples)

Raising version of predict/2.

predict!(self, samples, opts)

Raising version of predict/3.

read!(self, fn_)

Raising version of read/2.

save!(self, filename)

Raising version of save/2.

setActiveVarCount!(self, val)

Raising version of setActiveVarCount/2.

setCalculateVarImportance!(self, val)

Raising version of setCalculateVarImportance/2.

setCVFolds!(self, val)

Raising version of setCVFolds/2.

setMaxCategories!(self, val)

Raising version of setMaxCategories/2.

setMaxDepth!(self, val)

Raising version of setMaxDepth/2.

setMinSampleCount!(self, val)

Raising version of setMinSampleCount/2.

setPriors!(self, val)

Raising version of setPriors/2.

setRegressionAccuracy!(self, val)

Raising version of setRegressionAccuracy/2.

setTermCriteria!(self, val)

Raising version of setTermCriteria/2.

setTruncatePrunedTree!(self, val)

Raising version of setTruncatePrunedTree/2.

setUse1SERule!(self, val)

Raising version of setUse1SERule/2.

setUseSurrogates!(self, val)

Raising version of setUseSurrogates/2.

train!(self, trainData)

Raising version of train/2.

train!(self, trainData, opts)

Raising version of train/3.

train!(self, samples, layout, responses)

Raising version of train/4.

write!(self, fs)

Raising version of write/2.

write!(self, fs, opts)

Raising version of write/3.