Evision.ML.SVM (Evision v0.1.5-dev)

cv

clear(self)

getDefaultName(self)

read(self, fn_)

save(self, filename)

write(self, fs)

write(self, fs, opts)

simplified API for language bindings Has overloading in C++

name : String

Python prototype (for reference):

write(fs[, name]) -> None

cv.ml

calcError(self, data, test)

calcError(self, data, test, opts)

Computes error on the training or test dataset

data.

test. computed over the training subset of the data. Please note that if you loaded a completely different dataset to evaluate already trained classifier, you will probably want not to set the test subset at all with TrainData::setTrainTestSplitRatio and specify test=false, so that the error is computed for the whole new set. Yes, this sounds a bit confusing.

resp. The method uses StatModel::predict to compute the error. For regression models the error is computed as RMS, for classifiers - as a percent of missclassified samples (0%-100%).

resp : UMat

Python prototype (for reference):

calcError(data, test[, resp]) -> retval, resp

create()

empty(self)

getC(self)

getClassWeights(self)

getCoef0(self)

getDecisionFunction(self, i)

getDecisionFunction(self, i, opts)

Retrieves the decision function

i. 2-class classification, then there will be just one decision function and the index should always be 0. Otherwise, in the case of N-class classification, there will be \f$N(N-1)/2\f$ decision functions.

alpha. In the case of linear %SVM all the alpha's will be 1's.

svidx. support vectors (which can be retrieved by SVM::getSupportVectors). In the case of linear %SVM each decision function consists of a single "compressed" support vector. The method returns rho parameter of the decision function, a scalar subtracted from the weighted sum of kernel responses.

alpha : UMat

svidx : UMat

Python prototype (for reference):

getDecisionFunction(i[, alpha[, svidx]]) -> retval, alpha, svidx

getDefaultGridPtr(param_id)

getDegree(self)

getGamma(self)

getKernelType(self)

getNu(self)

getP(self)

getSupportVectors(self)

getTermCriteria(self)

getType(self)

getUncompressedSupportVectors(self)

getVarCount(self)

isClassifier(self)

isTrained(self)

load(filepath)

predict(self, samples)

predict(self, samples, opts)

Predicts response(s) for the provided sample(s)

samples.

results.

flags.

results : UMat

flags : int

Python prototype (for reference):

predict(samples[, results[, flags]]) -> retval, results

setC(self, val)

setClassWeights(self, val)

setCoef0(self, val)

setDegree(self, val)

setGamma(self, val)

setKernel(self, kernelType)

setNu(self, val)

setP(self, val)

setTermCriteria(self, val)

setType(self, val)

train(self, trainData)

train(self, trainData, opts)

train(self, samples, layout, responses)

trainAuto(self, samples, layout, responses)

trainAuto(self, samples, layout, responses, opts)

Trains an %SVM with optimal parameters

samples.

layout.

responses.

kFold. subset is used to test the model, the others form the train set. So, the %SVM algorithm is

cgrid.

gammaGrid.

pGrid.

nuGrid.

coeffGrid.

degreeGrid.

balanced. balanced cross-validation subsets that is proportions between classes in subsets are close to such proportion in the whole train dataset. The method trains the %SVM model automatically by choosing the optimal parameters C, gamma, p, nu, coef0, degree. Parameters are considered optimal when the cross-validation estimate of the test set error is minimal. This function only makes use of SVM::getDefaultGrid for parameter optimization and thus only offers rudimentary parameter options. This function works for the classification (SVM::C_SVC or SVM::NU_SVC) as well as for the regression (SVM::EPS_SVR or SVM::NU_SVR). If it is SVM::ONE_CLASS, no optimization is made and the usual %SVM with parameters specified in params is executed.

kFold : int

cgrid : Ptr<ParamGrid>

gammaGrid : Ptr<ParamGrid>

pGrid : Ptr<ParamGrid>

nuGrid : Ptr<ParamGrid>

coeffGrid : Ptr<ParamGrid>

degreeGrid : Ptr<ParamGrid>

balanced : bool

Python prototype (for reference):

trainAuto(samples, layout, responses[, kFold[, Cgrid[, gammaGrid[, pGrid[, nuGrid[, coeffGrid[, degreeGrid[, balanced]]]]]]]]) -> retval

Functions

calcError!(self, data, test)

Raising version of calcError/3.

calcError!(self, data, test, opts)

Raising version of calcError/4.

clear!(self)

Raising version of clear/1.

create!()

Raising version of create/0.

empty!(self)

Raising version of empty/1.

getC!(self)

Raising version of getC/1.

getClassWeights!(self)

Raising version of getClassWeights/1.

getCoef0!(self)

Raising version of getCoef0/1.

getDecisionFunction!(self, i)

Raising version of getDecisionFunction/2.

getDecisionFunction!(self, i, opts)

Raising version of getDecisionFunction/3.

getDefaultGridPtr!(param_id)

Raising version of getDefaultGridPtr/1.

getDefaultName!(self)

Raising version of getDefaultName/1.

getDegree!(self)

Raising version of getDegree/1.

getGamma!(self)

Raising version of getGamma/1.

getKernelType!(self)

Raising version of getKernelType/1.

getNu!(self)

Raising version of getNu/1.

getP!(self)

Raising version of getP/1.

getSupportVectors!(self)

Raising version of getSupportVectors/1.

getTermCriteria!(self)

Raising version of getTermCriteria/1.

getType!(self)

Raising version of getType/1.

getUncompressedSupportVectors!(self)

Raising version of getUncompressedSupportVectors/1.

getVarCount!(self)

Raising version of getVarCount/1.

isClassifier!(self)

Raising version of isClassifier/1.

isTrained!(self)

Raising version of isTrained/1.

load!(filepath)

Raising version of load/1.

predict!(self, samples)

Raising version of predict/2.

predict!(self, samples, opts)

Raising version of predict/3.

read!(self, fn_)

Raising version of read/2.

save!(self, filename)

Raising version of save/2.

setC!(self, val)

Raising version of setC/2.

setClassWeights!(self, val)

Raising version of setClassWeights/2.

setCoef0!(self, val)

Raising version of setCoef0/2.

setDegree!(self, val)

Raising version of setDegree/2.

setGamma!(self, val)

Raising version of setGamma/2.

setKernel!(self, kernelType)

Raising version of setKernel/2.

setNu!(self, val)

Raising version of setNu/2.

setP!(self, val)

Raising version of setP/2.

setTermCriteria!(self, val)

Raising version of setTermCriteria/2.

setType!(self, val)

Raising version of setType/2.

train!(self, trainData)

Raising version of train/2.

train!(self, trainData, opts)

Raising version of train/3.

train!(self, samples, layout, responses)

Raising version of train/4.

trainAuto!(self, samples, layout, responses)

Raising version of trainAuto/4.

trainAuto!(self, samples, layout, responses, opts)

Raising version of trainAuto/5.

write!(self, fs)

Raising version of write/2.

write!(self, fs, opts)

Raising version of write/3.