Evision.ML.SVMSGD (Evision v0.1.5-dev)

cv

clear(self)

getDefaultName(self)

read(self, fn_)

save(self, filename)

write(self, fs)

write(self, fs, opts)

simplified API for language bindings Has overloading in C++

name : String

Python prototype (for reference):

write(fs[, name]) -> None

cv.ml

calcError(self, data, test)

calcError(self, data, test, opts)

Computes error on the training or test dataset

data.

test. computed over the training subset of the data. Please note that if you loaded a completely different dataset to evaluate already trained classifier, you will probably want not to set the test subset at all with TrainData::setTrainTestSplitRatio and specify test=false, so that the error is computed for the whole new set. Yes, this sounds a bit confusing.

resp. The method uses StatModel::predict to compute the error. For regression models the error is computed as RMS, for classifiers - as a percent of missclassified samples (0%-100%).

resp : UMat

Python prototype (for reference):

calcError(data, test[, resp]) -> retval, resp

create()

empty(self)

getInitialStepSize(self)

getMarginRegularization(self)

getMarginType(self)

getShift(self)

getStepDecreasingPower(self)

getSvmsgdType(self)

getTermCriteria(self)

getVarCount(self)

getWeights(self)

isClassifier(self)

isTrained(self)

load(filepath)

load(filepath, opts)

Loads and creates a serialized SVMSGD from a file Use SVMSGD::save to serialize and store an SVMSGD to disk. Load the SVMSGD from this file again, by calling this function with the path to the file. Optionally specify the node for the file containing the classifier

filepath.

nodeName.

nodeName : String

Python prototype (for reference):

load(filepath[, nodeName]) -> retval

predict(self, samples)

predict(self, samples, opts)

Predicts response(s) for the provided sample(s)

samples.

results.

flags.

results : UMat

flags : int

Python prototype (for reference):

predict(samples[, results[, flags]]) -> retval, results

setInitialStepSize(self, initialStepSize)

setMarginRegularization(self, marginRegularization)

setMarginType(self, marginType)

setOptimalParameters(self)

setStepDecreasingPower(self, stepDecreasingPower)

setSvmsgdType(self, svmsgdType)

setTermCriteria(self, val)

train(self, trainData)

train(self, trainData, opts)

train(self, samples, layout, responses)

Functions

calcError!(self, data, test)

Raising version of calcError/3.

calcError!(self, data, test, opts)

Raising version of calcError/4.

clear!(self)

Raising version of clear/1.

create!()

Raising version of create/0.

empty!(self)

Raising version of empty/1.

getDefaultName!(self)

Raising version of getDefaultName/1.

getInitialStepSize!(self)

Raising version of getInitialStepSize/1.

getMarginRegularization!(self)

Raising version of getMarginRegularization/1.

getMarginType!(self)

Raising version of getMarginType/1.

getShift!(self)

Raising version of getShift/1.

getStepDecreasingPower!(self)

Raising version of getStepDecreasingPower/1.

getSvmsgdType!(self)

Raising version of getSvmsgdType/1.

getTermCriteria!(self)

Raising version of getTermCriteria/1.

getVarCount!(self)

Raising version of getVarCount/1.

getWeights!(self)

Raising version of getWeights/1.

isClassifier!(self)

Raising version of isClassifier/1.

isTrained!(self)

Raising version of isTrained/1.

load!(filepath)

Raising version of load/1.

load!(filepath, opts)

Raising version of load/2.

predict!(self, samples)

Raising version of predict/2.

predict!(self, samples, opts)

Raising version of predict/3.

read!(self, fn_)

Raising version of read/2.

save!(self, filename)

Raising version of save/2.

setInitialStepSize!(self, initialStepSize)

Raising version of setInitialStepSize/2.

setMarginRegularization!(self, marginRegularization)

Raising version of setMarginRegularization/2.

setMarginType!(self, marginType)

Raising version of setMarginType/2.

setOptimalParameters!(self)

Raising version of setOptimalParameters/1.

setStepDecreasingPower!(self, stepDecreasingPower)

Raising version of setStepDecreasingPower/2.

setSvmsgdType!(self, svmsgdType)

Raising version of setSvmsgdType/2.

setTermCriteria!(self, val)

Raising version of setTermCriteria/2.

train!(self, trainData)

Raising version of train/2.

train!(self, trainData, opts)

Raising version of train/3.

train!(self, samples, layout, responses)

Raising version of train/4.

write!(self, fs)

Raising version of write/2.

write!(self, fs, opts)

Raising version of write/3.