Evision.ML.TrainData (Evision v0.1.7-dev)

cv.ml

create(samples, layout, responses)

Positional Arguments

samples: UMat.

layout: int.

responses: UMat.

Keyword Arguments

varIdx: UMat.

sampleIdx: UMat.

sampleWeights: UMat.

varType: UMat.

Creates training data from in-memory arrays.

single row or as a single column. The matrix should have type CV_32F or CV_32S (in the
former case the responses are considered as ordered by default; in the latter case - as
categorical)
(CV_32S) containing 0-based variable indices or byte vector (CV_8U) containing a mask of
active variables.
vector (CV_32S) containing 0-based sample indices or byte vector (CV_8U) containing a mask
of training samples.
<number_of_variables_in_responses>`, containing types of each input and output variable. See
ml::VariableTypes.

Python prototype (for reference):

create(samples, layout, responses[, varIdx[, sampleIdx[, sampleWeights[, varType]]]]) -> retval

create(samples, layout, responses, opts)

Positional Arguments

samples: UMat.

layout: int.

responses: UMat.

Keyword Arguments

varIdx: UMat.

sampleIdx: UMat.

sampleWeights: UMat.

varType: UMat.

Creates training data from in-memory arrays.

single row or as a single column. The matrix should have type CV_32F or CV_32S (in the
former case the responses are considered as ordered by default; in the latter case - as
categorical)
(CV_32S) containing 0-based variable indices or byte vector (CV_8U) containing a mask of
active variables.
vector (CV_32S) containing 0-based sample indices or byte vector (CV_8U) containing a mask
of training samples.
<number_of_variables_in_responses>`, containing types of each input and output variable. See
ml::VariableTypes.

Python prototype (for reference):

create(samples, layout, responses[, varIdx[, sampleIdx[, sampleWeights[, varType]]]]) -> retval

getCatCount(self, vi)

Python prototype (for reference):

getCatCount(vi) -> retval

getCatMap(self)

Python prototype (for reference):

getCatMap() -> retval

getCatOfs(self)

Python prototype (for reference):

getCatOfs() -> retval

getClassLabels(self)

Returns the vector of class labels The function returns vector of unique labels occurred in the responses.

Python prototype (for reference):

getClassLabels() -> retval

getDefaultSubstValues(self)

Python prototype (for reference):

getDefaultSubstValues() -> retval

getLayout(self)

Python prototype (for reference):

getLayout() -> retval

getMissing(self)

Python prototype (for reference):

getMissing() -> retval

getNAllVars(self)

Python prototype (for reference):

getNAllVars() -> retval

getNames(self, names)

Returns vector of symbolic names captured in loadFromCSV()

Python prototype (for reference):

getNames(names) -> None

getNormCatResponses(self)

Python prototype (for reference):

getNormCatResponses() -> retval

getNSamples(self)

Python prototype (for reference):

getNSamples() -> retval

getNTestSamples(self)

Python prototype (for reference):

getNTestSamples() -> retval

getNTrainSamples(self)

Python prototype (for reference):

getNTrainSamples() -> retval

getNVars(self)

Python prototype (for reference):

getNVars() -> retval

getResponses(self)

Python prototype (for reference):

getResponses() -> retval

getResponseType(self)

Python prototype (for reference):

getResponseType() -> retval

getSample(self, varIdx, sidx, buf)

Python prototype (for reference):

getSample(varIdx, sidx, buf) -> None

getSamples(self)

Python prototype (for reference):

getSamples() -> retval

getSampleWeights(self)

Python prototype (for reference):

getSampleWeights() -> retval

getSubMatrix(matrix, idx, layout)

Extract from matrix rows/cols specified by passed indexes.

Python prototype (for reference):

getSubMatrix(matrix, idx, layout) -> retval

getSubVector(vec, idx)

Extract from 1D vector elements specified by passed indexes.

Python prototype (for reference):

getSubVector(vec, idx) -> retval

getTestNormCatResponses(self)

Python prototype (for reference):

getTestNormCatResponses() -> retval

getTestResponses(self)

Python prototype (for reference):

getTestResponses() -> retval

getTestSampleIdx(self)

Python prototype (for reference):

getTestSampleIdx() -> retval

getTestSamples(self)

Returns matrix of test samples

Python prototype (for reference):

getTestSamples() -> retval

getTestSampleWeights(self)

Python prototype (for reference):

getTestSampleWeights() -> retval

getTrainNormCatResponses(self)

Returns the vector of normalized categorical responses The function returns vector of responses. Each response is integer from 0 to <number of classes>-1. The actual label value can be retrieved then from the class label vector, see TrainData::getClassLabels.

Python prototype (for reference):

getTrainNormCatResponses() -> retval

getTrainResponses(self)

Returns the vector of responses The function returns ordered or the original categorical responses. Usually it's used in regression algorithms.

Python prototype (for reference):

getTrainResponses() -> retval

getTrainSampleIdx(self)

Python prototype (for reference):

getTrainSampleIdx() -> retval

getTrainSamples(self)

Positional Arguments

Keyword Arguments

layout: int.

compressSamples: bool.

compressVars: bool.

Returns matrix of train samples

transposed. See ml::SampleTypes.
sampleIdx)
the active variables.

In current implementation the function tries to avoid physical data copying and returns the matrix stored inside TrainData (unless the transposition or compression is needed).

Python prototype (for reference):

getTrainSamples([, layout[, compressSamples[, compressVars]]]) -> retval

getTrainSampleWeights(self)

Python prototype (for reference):

getTrainSampleWeights() -> retval

getValues(self, vi, sidx, values)

Python prototype (for reference):

getValues(vi, sidx, values) -> None

getVarIdx(self)

Python prototype (for reference):

getVarIdx() -> retval

getVarSymbolFlags(self)

Python prototype (for reference):

getVarSymbolFlags() -> retval

getVarType(self)

Python prototype (for reference):

getVarType() -> retval

setTrainTestSplit(self, count)

Positional Arguments

count: int

Keyword Arguments

shuffle: bool.

Splits the training data into the training and test parts @sa TrainData::setTrainTestSplitRatio

Python prototype (for reference):

setTrainTestSplit(count[, shuffle]) -> None

setTrainTestSplit(self, count, opts)

Positional Arguments

count: int

Keyword Arguments

shuffle: bool.

Splits the training data into the training and test parts @sa TrainData::setTrainTestSplitRatio

Python prototype (for reference):

setTrainTestSplit(count[, shuffle]) -> None

setTrainTestSplitRatio(self, ratio)

Positional Arguments

ratio: double

Keyword Arguments

shuffle: bool.

Splits the training data into the training and test parts The function selects a subset of specified relative size and then returns it as the training set. If the function is not called, all the data is used for training. Please, note that for each of TrainData::getTrain* there is corresponding TrainData::getTest*, so that the test subset can be retrieved and processed as well. @sa TrainData::setTrainTestSplit

Python prototype (for reference):

setTrainTestSplitRatio(ratio[, shuffle]) -> None

setTrainTestSplitRatio(self, ratio, opts)

Positional Arguments

ratio: double

Keyword Arguments

shuffle: bool.

Splits the training data into the training and test parts The function selects a subset of specified relative size and then returns it as the training set. If the function is not called, all the data is used for training. Please, note that for each of TrainData::getTrain* there is corresponding TrainData::getTest*, so that the test subset can be retrieved and processed as well. @sa TrainData::setTrainTestSplit

Python prototype (for reference):

setTrainTestSplitRatio(ratio[, shuffle]) -> None

shuffleTrainTest(self)

Python prototype (for reference):

shuffleTrainTest() -> None

Functions

create!(samples, layout, responses)

Raising version of create/3.

create!(samples, layout, responses, opts)

Raising version of create/4.

getCatCount!(self, vi)

Raising version of getCatCount/2.

getCatMap!(self)

Raising version of getCatMap/1.

getCatOfs!(self)

Raising version of getCatOfs/1.

getClassLabels!(self)

Raising version of getClassLabels/1.

getDefaultSubstValues!(self)

Raising version of getDefaultSubstValues/1.

getLayout!(self)

Raising version of getLayout/1.

getMissing!(self)

Raising version of getMissing/1.

getNAllVars!(self)

Raising version of getNAllVars/1.

getNames!(self, names)

Raising version of getNames/2.

getNormCatResponses!(self)

Raising version of getNormCatResponses/1.

getNSamples!(self)

Raising version of getNSamples/1.

getNTestSamples!(self)

Raising version of getNTestSamples/1.

getNTrainSamples!(self)

Raising version of getNTrainSamples/1.

getNVars!(self)

Raising version of getNVars/1.

getResponses!(self)

Raising version of getResponses/1.

getResponseType!(self)

Raising version of getResponseType/1.

getSample!(self, varIdx, sidx, buf)

Raising version of getSample/4.

getSamples!(self)

Raising version of getSamples/1.

getSampleWeights!(self)

Raising version of getSampleWeights/1.

getSubMatrix!(matrix, idx, layout)

Raising version of getSubMatrix/3.

getSubVector!(vec, idx)

Raising version of getSubVector/2.

getTestNormCatResponses!(self)

Raising version of getTestNormCatResponses/1.

getTestResponses!(self)

Raising version of getTestResponses/1.

getTestSampleIdx!(self)

Raising version of getTestSampleIdx/1.

getTestSamples!(self)

Raising version of getTestSamples/1.

getTestSampleWeights!(self)

Raising version of getTestSampleWeights/1.

getTrainNormCatResponses!(self)

Raising version of getTrainNormCatResponses/1.

getTrainResponses!(self)

Raising version of getTrainResponses/1.

getTrainSampleIdx!(self)

Raising version of getTrainSampleIdx/1.

getTrainSamples!(self)

Raising version of getTrainSamples/1.

getTrainSampleWeights!(self)

Raising version of getTrainSampleWeights/1.

getValues!(self, vi, sidx, values)

Raising version of getValues/4.

getVarIdx!(self)

Raising version of getVarIdx/1.

getVarSymbolFlags!(self)

Raising version of getVarSymbolFlags/1.

getVarType!(self)

Raising version of getVarType/1.

setTrainTestSplit!(self, count)

Raising version of setTrainTestSplit/2.

setTrainTestSplit!(self, count, opts)

Raising version of setTrainTestSplit/3.

setTrainTestSplitRatio!(self, ratio)

Raising version of setTrainTestSplitRatio/2.

setTrainTestSplitRatio!(self, ratio, opts)

Raising version of setTrainTestSplitRatio/3.

shuffleTrainTest!(self)

Raising version of shuffleTrainTest/1.