Evision.Mat (Evision v0.1.7-dev)

OpenCV Mat

Types

channels_from_binary()

@type channels_from_binary() :: 1 | 2 | 3 | 4

mat_type()

@type mat_type() ::
  {:u, 8}
  | {:u, 16}
  | {:s, 8}
  | {:s, 16}
  | {:s, 32}
  | {:f, 32}
  | {:f, 64}
  | {:f, 16}
  | :u8
  | :u16
  | :s8
  | :s16
  | :s32
  | :f32
  | :f64
  | :f16

Types for mat

t()

@type t() :: %{__struct__: atom()}

cv.Mat

abs(mat)

add(lhs, rhs)

add(lhs, rhs, type)

as_shape(mat, as_shape)

This method does not change the underlying data. It only changes the steps when accessing the matrix.

If intended to change the underlying data to the new shape, please use Evision.Mat.reshape/2.

as_type(mat, type)

at(mat, position)

bitwise_and(lhs, rhs)

bitwise_not(mat)

bitwise_or(lhs, rhs)

bitwise_xor(lhs, rhs)

ceil(mat)

channels(mat)

The method returns the number of matrix channels.

clip(mat, lower, upper)

clone(mat)

cmp(lhs, rhs, op)

depth(mat)

Returns the depth of a matrix element.

The method returns the identifier of the matrix element depth (the type of each individual channel). For example, for a 16-bit signed element array, the method returns CV_16S. A complete list of matrix types contains the following values:

CV_8U - 8-bit unsigned integers ( 0..255 )

CV_8S - 8-bit signed integers ( -128..127 )

CV_16U - 16-bit unsigned integers ( 0..65535 )

CV_16S - 16-bit signed integers ( -32768..32767 )

CV_32S - 32-bit signed integers ( -2147483648..2147483647 )

CV_32F - 32-bit floating-point numbers ( -FLT_MAX..FLT_MAX, INF, NAN )

CV_64F - 64-bit floating-point numbers ( -DBL_MAX..DBL_MAX, INF, NAN )

divide(lhs, rhs)

divide(lhs, rhs, type)

dot(mat_a, mat_b)

elemSize1(mat)

Returns the size of each matrix element channel in bytes.

The method returns the matrix element channel size in bytes, that is, it ignores the number of channels. For example, if the matrix type is CV_16SC3 , the method returns sizeof(short) or 2.

elemSize(mat)

Returns the matrix element size in bytes.

The method returns the matrix element size in bytes. For example, if the matrix type is CV_16SC3, the method returns 3*sizeof(short) or 6.

empty()

expm1(mat)

eye(n, type)

floor(mat)

@spec floor(reference()) :: {:ok, reference()} | {:error, String.t()}

from_binary(binary, type, rows, cols, channels)

@spec from_binary(
  binary(),
  mat_type(),
  pos_integer(),
  pos_integer(),
  channels_from_binary()
) ::
  {:ok, reference()} | {:error, String.t()}

Create Mat from binary (pixel) data

binary. The binary pixel data

type. type={t, l} is one of [{:u, 8}, {:s, 8}, {:u, 16}, {:s, 16}, {:s, 32}, {:f, 32}, {:f, 64}]

rows. Number of rows (i.e., the height of the image)

cols. Number of cols (i.e., the width of the image)

channels. Number of channels, only valid if in [1, 3, 4]

from_binary_by_shape(binary, type, shape)

isContinuous(mat)

isSubmatrix(mat)

last_dim_as_channel(mat)

This function would convert the input tensor with dims [height, width, dims] to a dims-channel image with dims [height, width].

Note that OpenCV has limitation on the number of channels. Currently the maximum number of channels is 512.

literal(list)

Create an Evision.Mat from list literals.

Example

Creating Evision.Mat from empty list literal ([]) is the same as calling Evision.Mat.empty().

iex> Evision.Mat.literal!([])
%Evision.Mat{
  channels: 1,
  dims: 0,
  type: {:u, 8},
  raw_type: 0,
  shape: {},
  ref: #Reference<0.1204050731.2031747092.46781>
}

By default, the shape of the Mat will stay as is.

iex> Evision.Mat.literal!([[[1,1,1],[2,2,2],[3,3,3]]], :u8)
%Evision.Mat{
  channels: 1,
  dims: 3,
  type: {:u, 8},
  raw_type: 0,
  shape: {1, 3, 3},
  ref: #Reference<0.512519210.691404819.106300>
}

Evision.Mat.literal/3 will return a vaild 2D image if the keyword argument, as_2d, is set to true and if the list literal can be represented as a 2D image.

iex> Evision.Mat.literal!([[[1,1,1],[2,2,2],[3,3,3]]], :u8, as_2d: true)
%Evision.Mat{
  channels: 3,
  dims: 2,
  type: {:u, 8},
  raw_type: 16,
  shape: {1, 3, 3},
  ref: #Reference<0.512519210.691404820.106293>
}

logical_and(lhs, rhs)

logical_or(lhs, rhs)

logical_xor(lhs, rhs)

matrix_multiply(lhs, rhs, out_type \\ nil)

multiply(lhs, rhs)

multiply(lhs, rhs, type)

negate(mat)

number(number, type)

ones(shape, type)

raw_type(mat)

Returns the type of a matrix.

As Evision.Mat.type/1 returns the type used by Nx, this method gives the raw value of cv::Mat.type()

reshape(mat, shape)

round(mat)

setTo(mat, value, mask)

shape(mat)

sign(mat)

size(mat)

Returns the cv::MatSize of the matrix.

The method returns a tuple {dims, p} where dims is the number of dimensions, and p is a list with dims elements.

subtract(lhs, rhs)

subtract(lhs, rhs, type)

to_binary(mat, limit \\ 0)

total(mat)

Returns the total number of array elements.

The method returns the number of array elements (a number of pixels if the array represents an image).

total(mat, start_dim, end_dim \\ 4_294_967_295)

Returns the total number of array elements.

The method returns the number of elements within a certain sub-array slice with start_dim <= dim < end_dim

type(mat)

This method returns the type-tuple used by Nx. To get the raw value of cv::Mat.type(), please use Evision.Mat.raw_type/1.

zeros(shape, type)

Functions

abs!(mat)

Raising version of abs/1.

add!(lhs, rhs)

Raising version of add/2.

add!(lhs, rhs, type)

Raising version of add/3.

arange!(from, to, step, type)

Raising version of arange/4.

arange!(from, to, step, type, shape)

Raising version of arange/5.

arange(from, to, step, type)

arange(from, to, step, type, shape)

as_shape!(mat, as_shape)

Raising version of as_shape/2.

as_type!(mat, type)

Raising version of as_type/2.

at!(mat, position)

Raising version of at/2.

bitwise_and!(lhs, rhs)

Raising version of bitwise_and/2.

bitwise_not!(mat)

Raising version of bitwise_not/1.

bitwise_or!(lhs, rhs)

Raising version of bitwise_or/2.

bitwise_xor!(lhs, rhs)

Raising version of bitwise_xor/2.

broadcast_to!(mat, to_shape)

Raising version of broadcast_to/2.

broadcast_to!(mat, to_shape, force_src_shape)

Raising version of broadcast_to/3.

broadcast_to(mat, to_shape)

broadcast_to(mat, to_shape, force_src_shape)

ceil!(mat)

Raising version of ceil/1.

channel_as_last_dim!(mat)

Raising version of channel_as_last_dim/1.

channel_as_last_dim(mat)

This function does the opposite as to Evision.Mat.last_dim_as_channel/1.

If the number of channels of the input Evision.Mat is greater than 1, then this function would convert the input Evision.Mat with dims dims=list(int()) to a 1-channel Evision.Mat with dims [dims | channels].

If the number of channels of the input Evision.Mat is equal to 1,

if dims == shape, then nothing happens

otherwise, a new Evision.Mat that has dims=[dims | channels] will be returned

channels!(mat)

Raising version of channels/1.

clip!(mat, lower, upper)

Raising version of clip/3.

clone!(mat)

Raising version of clone/1.

cmp!(lhs, rhs, op)

Raising version of cmp/3.

depth!(mat)

Raising version of depth/1.

divide!(lhs, rhs)

Raising version of divide/2.

divide!(lhs, rhs, type)

Raising version of divide/3.

dot!(mat_a, mat_b)

Raising version of dot/2.

elemSize1!(mat)

Raising version of elemSize1/1.

elemSize!(mat)

Raising version of elemSize/1.

empty!()

Raising version of empty/0.

expm1!(mat)

Raising version of expm1/1.

eye!(n, type)

Raising version of eye/2.

floor!(mat)

Raising version of floor/1.

from_binary!(binary, type, rows, cols, channels)

Raising version of from_binary/5.

from_binary_by_shape!(binary, type, shape)

Raising version of from_binary_by_shape/3.

full!(shape, number, type)

Raising version of full/3.

full(shape, number, type)

isContinuous!(mat)

Raising version of isContinuous/1.

isSubmatrix!(mat)

Raising version of isSubmatrix/1.

last_dim_as_channel!(mat)

Raising version of last_dim_as_channel/1.

literal!(list)

Raising version of literal/1.

literal!(literal, type)

Raising version of literal/2.

literal!(literal, type, opts)

Raising version of literal/3.

literal(literal, type, opts \\ [])

@spec literal(list(), mat_type(), Keyword.t()) ::
  {:ok,
   %Evision.Mat{
     channels: term(),
     dims: term(),
     raw_type: term(),
     ref: term(),
     shape: term(),
     type: term()
   }}
  | {:error, String.t()}

logical_and!(lhs, rhs)

Raising version of logical_and/2.

logical_or!(lhs, rhs)

Raising version of logical_or/2.

logical_xor!(lhs, rhs)

Raising version of logical_xor/2.

matrix_multiply!(lhs, rhs)

Raising version of matrix_multiply/2.

matrix_multiply!(lhs, rhs, out_type)

Raising version of matrix_multiply/3.

multiply!(lhs, rhs)

Raising version of multiply/2.

multiply!(lhs, rhs, type)

Raising version of multiply/3.

negate!(mat)

Raising version of negate/1.

number!(number, type)

Raising version of number/2.

ones!(shape, type)

Raising version of ones/2.

raw_type!(mat)

Raising version of raw_type/1.

reshape!(mat, shape)

Raising version of reshape/2.

round!(mat)

Raising version of round/1.

setTo!(mat, value, mask)

Raising version of setTo/3.

shape!(mat)

Raising version of shape/1.

sign!(mat)

Raising version of sign/1.

size!(mat)

Raising version of size/1.

squeeze!(mat)

Raising version of squeeze/1.

squeeze(mat)

subtract!(lhs, rhs)

Raising version of subtract/2.

subtract!(lhs, rhs, type)

Raising version of subtract/3.

to_batched!(mat, batch_size, opts)

Raising version of to_batched/3.

to_batched!(mat, batch_size, as_shape, opts)

Raising version of to_batched/4.

to_batched(mat, batch_size, opts)

to_batched(mat, batch_size, as_shape, opts)

to_binary!(mat)

Raising version of to_binary/1.

to_binary!(mat, limit)

Raising version of to_binary/2.

total!(mat)

Raising version of total/1.

total!(mat, start_dim)

Raising version of total/2.

total!(mat, start_dim, end_dim)

Raising version of total/3.

transpose!(mat)

Raising version of transpose/1.

transpose!(mat, axes)

Raising version of transpose/2.

transpose!(mat, axes, opts)

Raising version of transpose/3.

transpose(mat)

Transpose a matrix

Parameters

mat. The matrix. by default it reverses the order of the axes.

transpose(mat, axes, opts \\ [])

Transpose a matrix

Parameters

mat. The matrix.

axes. list of ints. It must be a list which contains a permutation of [0,1,..,N-1] where N is the number of axes of mat. The i’th axis of the returned array will correspond to the axis numbered axes[i] of the input.

opts. Keyword options.

as_shape. A tuple or list which overwrites the shape of the matrix (the total number of elements must be equal to the one as in its original shape). For example, a 4x4 matrix can be treated as a 2x2x2x2 matrix and transposed with axes=[2,1,3,0] in a single call.

When specified, it combines the reshape and transpose operation in a single NIF call.

type!(mat)

Raising version of type/1.

zeros!(shape, type)

Raising version of zeros/2.