Evision.ORB (Evision v0.1.5-dev)

cv

compute(self, images, keypoints)

compute(self, images, keypoints, opts)

Has overloading in C++

images.

keypoints. computed are removed. Sometimes new keypoints can be added, for example: SIFT duplicates keypoint with several dominant orientations (for each orientation).

descriptors. descriptors computed for a keypoints[i]. Row j is the keypoints (or keypoints[i]) is the descriptor for keypoint j-th keypoint.

descriptors : vector_UMat

Python prototype (for reference):

compute(images, keypoints[, descriptors]) -> keypoints, descriptors

create()

create(opts)

The ORB constructor

nfeatures.

scaleFactor. pyramid, where each next level has 4x less pixels than the previous, but such a big scale factor will degrade feature matching scores dramatically. On the other hand, too close to 1 scale factor will mean that to cover certain scale range you will need more pyramid levels and so the speed will suffer.

nlevels. input_image_linear_size/pow(scaleFactor, nlevels - firstLevel).

edgeThreshold. roughly match the patchSize parameter.

firstLevel. with upscaled source image.

wTA_K. default value 2 means the BRIEF where we take a random point pair and compare their brightnesses, so we get 0/1 response. Other possible values are 3 and 4. For example, 3 means that we take 3 random points (of course, those point coordinates are random, but they are generated from the pre-defined seed, so each element of BRIEF descriptor is computed deterministically from the pixel rectangle), find point of maximum brightness and output index of the winner (0, 1 or 2). Such output will occupy 2 bits, and therefore it will need a special variant of Hamming distance, denoted as NORM_HAMMING2 (2 bits per bin). When WTA_K=4, we take 4 random points to compute each bin (that will also occupy 2 bits with possible values 0, 1, 2 or 3).

scoreType. (the score is written to KeyPoint::score and is used to retain best nfeatures features); FAST_SCORE is alternative value of the parameter that produces slightly less stable keypoints, but it is a little faster to compute.

patchSize. pyramid layers the perceived image area covered by a feature will be larger.

fastThreshold.

nfeatures : int

scaleFactor : float

nlevels : int

edgeThreshold : int

firstLevel : int

wTA_K : int

scoreType : ORB_ScoreType

patchSize : int

fastThreshold : int

Python prototype (for reference):

create([, nfeatures[, scaleFactor[, nlevels[, edgeThreshold[, firstLevel[, WTA_K[, scoreType[, patchSize[, fastThreshold]]]]]]]]]) -> retval

defaultNorm(self)

descriptorSize(self)

descriptorType(self)

detect(self, images)

detect(self, images, opts)

Has overloading in C++

images.

keypoints. of keypoints detected in images[i] .

masks. masks[i] is a mask for images[i].

masks : vector_UMat

Python prototype (for reference):

detect(images[, masks]) -> keypoints

detectAndCompute(self, image, mask)

detectAndCompute(self, image, mask, opts)

Detects keypoints and computes the descriptors

descriptors : UMat

useProvidedKeypoints : bool

Python prototype (for reference):

detectAndCompute(image, mask[, descriptors[, useProvidedKeypoints]]) -> keypoints, descriptors

empty(self)

getDefaultName(self)

getEdgeThreshold(self)

getFastThreshold(self)

getFirstLevel(self)

getMaxFeatures(self)

getNLevels(self)

getPatchSize(self)

getScaleFactor(self)

getScoreType(self)

getWTA_K(self)

read(self, arg1)

setEdgeThreshold(self, edgeThreshold)

setFastThreshold(self, fastThreshold)

setFirstLevel(self, firstLevel)

setMaxFeatures(self, maxFeatures)

setNLevels(self, nlevels)

setPatchSize(self, patchSize)

setScaleFactor(self, scaleFactor)

setScoreType(self, scoreType)

setWTA_K(self, wta_k)

write(self, fs)

write(self, fs, opts)

name : String

Python prototype (for reference):

write(fs[, name]) -> None

Functions

compute!(self, images, keypoints)

Raising version of compute/3.

compute!(self, images, keypoints, opts)

Raising version of compute/4.

create!()

Raising version of create/0.

create!(opts)

Raising version of create/1.

defaultNorm!(self)

Raising version of defaultNorm/1.

descriptorSize!(self)

Raising version of descriptorSize/1.

descriptorType!(self)

Raising version of descriptorType/1.

detect!(self, images)

Raising version of detect/2.

detect!(self, images, opts)

Raising version of detect/3.

detectAndCompute!(self, image, mask)

Raising version of detectAndCompute/3.

detectAndCompute!(self, image, mask, opts)

Raising version of detectAndCompute/4.

empty!(self)

Raising version of empty/1.

getDefaultName!(self)

Raising version of getDefaultName/1.

getEdgeThreshold!(self)

Raising version of getEdgeThreshold/1.

getFastThreshold!(self)

Raising version of getFastThreshold/1.

getFirstLevel!(self)

Raising version of getFirstLevel/1.

getMaxFeatures!(self)

Raising version of getMaxFeatures/1.

getNLevels!(self)

Raising version of getNLevels/1.

getPatchSize!(self)

Raising version of getPatchSize/1.

getScaleFactor!(self)

Raising version of getScaleFactor/1.

getScoreType!(self)

Raising version of getScoreType/1.

getWTA_K!(self)

Raising version of getWTA_K/1.

read!(self, arg1)

Raising version of read/2.

setEdgeThreshold!(self, edgeThreshold)

Raising version of setEdgeThreshold/2.

setFastThreshold!(self, fastThreshold)

Raising version of setFastThreshold/2.

setFirstLevel!(self, firstLevel)

Raising version of setFirstLevel/2.

setMaxFeatures!(self, maxFeatures)

Raising version of setMaxFeatures/2.

setNLevels!(self, nlevels)

Raising version of setNLevels/2.

setPatchSize!(self, patchSize)

Raising version of setPatchSize/2.

setScaleFactor!(self, scaleFactor)

Raising version of setScaleFactor/2.

setScoreType!(self, scoreType)

Raising version of setScoreType/2.

setWTA_K!(self, wta_k)

Raising version of setWTA_K/2.

write!(self, fs)

Raising version of write/2.

write!(self, fs, opts)

Raising version of write/3.