Evision.PyRotationWarper (Evision v0.1.9)

cv

buildMaps(self, src_size, k, r)

Builds the projection maps according to the given camera data.

Positional Arguments

src_size: Size.

Source image size

k: Evision.Mat.

Camera intrinsic parameters

r: Evision.Mat.

Camera rotation matrix

Return

xmap: Evision.Mat.

Projection map for the x axis

ymap: Evision.Mat.

Projection map for the y axis

@return Projected image minimum bounding box

Python prototype (for reference):

buildMaps(src_size, K, R[, xmap[, ymap]]) -> retval, xmap, ymap

buildMaps(self, src_size, k, r, opts)

Builds the projection maps according to the given camera data.

Positional Arguments

src_size: Size.

Source image size

k: Evision.Mat.

Camera intrinsic parameters

r: Evision.Mat.

Camera rotation matrix

Return

xmap: Evision.Mat.

Projection map for the x axis

ymap: Evision.Mat.

Projection map for the y axis

@return Projected image minimum bounding box

Python prototype (for reference):

buildMaps(src_size, K, R[, xmap[, ymap]]) -> retval, xmap, ymap

getScale(self)

Python prototype (for reference):

getScale() -> retval

pyRotationWarper()

Python prototype (for reference):

PyRotationWarper() -> <PyRotationWarper object>

pyRotationWarper(type, scale)

Positional Arguments

type: String

scale: float

Python prototype (for reference):

PyRotationWarper(type, scale) -> <PyRotationWarper object>

setScale(self, arg1)

Positional Arguments

arg1: float

Python prototype (for reference):

setScale(arg1) -> None

warp(self, src, k, r, interp_mode, border_mode)

Projects the image.

Positional Arguments

src: Evision.Mat.

Source image

k: Evision.Mat.

Camera intrinsic parameters

r: Evision.Mat.

Camera rotation matrix

interp_mode: int.

Interpolation mode

border_mode: int.

Border extrapolation mode

Return

dst: Evision.Mat.

Projected image

@return Project image top-left corner

Python prototype (for reference):

warp(src, K, R, interp_mode, border_mode[, dst]) -> retval, dst

warp(self, src, k, r, interp_mode, border_mode, opts)

Projects the image.

Positional Arguments

src: Evision.Mat.

Source image

k: Evision.Mat.

Camera intrinsic parameters

r: Evision.Mat.

Camera rotation matrix

interp_mode: int.

Interpolation mode

border_mode: int.

Border extrapolation mode

Return

dst: Evision.Mat.

Projected image

@return Project image top-left corner

Python prototype (for reference):

warp(src, K, R, interp_mode, border_mode[, dst]) -> retval, dst

warpBackward(self, src, k, r, interp_mode, border_mode, dst_size)

Projects the image backward.

Positional Arguments

src: Evision.Mat.

Projected image

k: Evision.Mat.

Camera intrinsic parameters

r: Evision.Mat.

Camera rotation matrix

interp_mode: int.

Interpolation mode

border_mode: int.

Border extrapolation mode

dst_size: Size.

Backward-projected image size

Return

dst: Evision.Mat.

Backward-projected image

Python prototype (for reference):

warpBackward(src, K, R, interp_mode, border_mode, dst_size[, dst]) -> dst

warpBackward(self, src, k, r, interp_mode, border_mode, dst_size, opts)

Projects the image backward.

Positional Arguments

src: Evision.Mat.

Projected image

k: Evision.Mat.

Camera intrinsic parameters

r: Evision.Mat.

Camera rotation matrix

interp_mode: int.

Interpolation mode

border_mode: int.

Border extrapolation mode

dst_size: Size.

Backward-projected image size

Return

dst: Evision.Mat.

Backward-projected image

Python prototype (for reference):

warpBackward(src, K, R, interp_mode, border_mode, dst_size[, dst]) -> dst

warpPoint(self, pt, k, r)

Projects the image point.

Positional Arguments

pt: Point2f.

Source point

k: Evision.Mat.

Camera intrinsic parameters

r: Evision.Mat.

Camera rotation matrix

@return Projected point

Python prototype (for reference):

warpPoint(pt, K, R) -> retval

warpPointBackward(self, pt, k, r)

Positional Arguments

pt: Point2f

k: Evision.Mat

r: Evision.Mat

Python prototype (for reference):

warpPointBackward(pt, K, R) -> retval

warpRoi(self, src_size, k, r)

Positional Arguments

src_size: Size.

Source image bounding box

k: Evision.Mat.

Camera intrinsic parameters

r: Evision.Mat.

Camera rotation matrix

@return Projected image minimum bounding box

Python prototype (for reference):

warpRoi(src_size, K, R) -> retval

Functions

buildMaps!(self, src_size, k, r)

Raising version of buildMaps/4.

buildMaps!(self, src_size, k, r, opts)

Raising version of buildMaps/5.

getScale!(self)

Raising version of getScale/1.

pyRotationWarper!()

Raising version of pyRotationWarper/0.

pyRotationWarper!(type, scale)

Raising version of pyRotationWarper/2.

setScale!(self, arg1)

Raising version of setScale/2.

warp!(self, src, k, r, interp_mode, border_mode)

Raising version of warp/6.

warp!(self, src, k, r, interp_mode, border_mode, opts)

Raising version of warp/7.

warpBackward!(self, src, k, r, interp_mode, border_mode, dst_size)

Raising version of warpBackward/7.

warpBackward!(self, src, k, r, interp_mode, border_mode, dst_size, opts)

Raising version of warpBackward/8.

warpPoint!(self, pt, k, r)

Raising version of warpPoint/4.

warpPointBackward!(self, pt, k, r)

Raising version of warpPointBackward/4.

warpRoi!(self, src_size, k, r)

Raising version of warpRoi/4.