Evision.Segmentation.IntelligentScissorsMB (Evision v0.1.7-dev)

cv.segmentation

applyImage(self, image)

Specify input image and extract image features

Python prototype (for reference):

applyImage(image) -> retval

applyImageFeatures(self, non_edge, gradient_direction, gradient_magnitude)

Positional Arguments

non_edge: UMat.

gradient_direction: UMat.

gradient_magnitude: UMat.

Keyword Arguments

image: UMat.

Specify custom features of input image Customized advanced variant of applyImage() call.

Python prototype (for reference):

applyImageFeatures(non_edge, gradient_direction, gradient_magnitude[, image]) -> retval

applyImageFeatures(self, non_edge, gradient_direction, gradient_magnitude, opts)

Positional Arguments

non_edge: UMat.

gradient_direction: UMat.

gradient_magnitude: UMat.

Keyword Arguments

image: UMat.

Specify custom features of input image Customized advanced variant of applyImage() call.

Python prototype (for reference):

applyImageFeatures(non_edge, gradient_direction, gradient_magnitude[, image]) -> retval

buildMap(self, sourcePt)

Prepares a map of optimal paths for the given source point on the image Note: applyImage() / applyImageFeatures() must be called before this call

Python prototype (for reference):

buildMap(sourcePt) -> None

getContour(self, targetPt)

Positional Arguments

targetPt: Point.

Keyword Arguments

contour: UMat.

backward: bool.

Extracts optimal contour for the given target point on the image Note: buildMap() must be called before this call

Python prototype (for reference):

getContour(targetPt[, contour[, backward]]) -> contour

getContour(self, targetPt, opts)

Positional Arguments

targetPt: Point.

Keyword Arguments

contour: UMat.

backward: bool.

Extracts optimal contour for the given target point on the image Note: buildMap() must be called before this call

Python prototype (for reference):

getContour(targetPt[, contour[, backward]]) -> contour

segmentation_IntelligentScissorsMB()

Python prototype (for reference):

IntelligentScissorsMB() -> <segmentation_IntelligentScissorsMB object>

setEdgeFeatureCannyParameters(self, threshold1, threshold2)

Positional Arguments

threshold1: double

threshold2: double

Keyword Arguments

apertureSize: int.

l2gradient: bool.

Switch edge feature extractor to use Canny edge detector Note: "Laplacian Zero-Crossing" feature extractor is used by default (following to original article) @sa Canny

Python prototype (for reference):

setEdgeFeatureCannyParameters(threshold1, threshold2[, apertureSize[, L2gradient]]) -> retval

setEdgeFeatureCannyParameters(self, threshold1, threshold2, opts)

Positional Arguments

threshold1: double

threshold2: double

Keyword Arguments

apertureSize: int.

l2gradient: bool.

Switch edge feature extractor to use Canny edge detector Note: "Laplacian Zero-Crossing" feature extractor is used by default (following to original article) @sa Canny

Python prototype (for reference):

setEdgeFeatureCannyParameters(threshold1, threshold2[, apertureSize[, L2gradient]]) -> retval

setEdgeFeatureZeroCrossingParameters(self)

Positional Arguments

Keyword Arguments

gradient_magnitude_min_value: float.

Switch to "Laplacian Zero-Crossing" edge feature extractor and specify its parameters This feature extractor is used by default according to article. Implementation has additional filtering for regions with low-amplitude noise. This filtering is enabled through parameter of minimal gradient amplitude (use some small value 4, 8, 16). Note: Current implementation of this feature extractor is based on processing of grayscale images (color image is converted to grayscale image first). Note: Canny edge detector is a bit slower, but provides better results (especially on color images): use setEdgeFeatureCannyParameters().

Python prototype (for reference):

setEdgeFeatureZeroCrossingParameters([, gradient_magnitude_min_value]) -> retval

setGradientMagnitudeMaxLimit(self)

Positional Arguments

Keyword Arguments

gradient_magnitude_threshold_max: float.

Specify gradient magnitude max value threshold Zero limit value is used to disable gradient magnitude thresholding (default behavior, as described in original article). Otherwize pixels with gradient magnitude >= threshold have zero cost. Note: Thresholding should be used for images with irregular regions (to avoid stuck on parameters from high-contract areas, like embedded logos).

Python prototype (for reference):

setGradientMagnitudeMaxLimit([, gradient_magnitude_threshold_max]) -> retval

setWeights(self, weight_non_edge, weight_gradient_direction, weight_gradient_magnitude)

Specify weights of feature functions Consider keeping weights normalized (sum of weights equals to 1.0) Discrete dynamic programming (DP) goal is minimization of costs between pixels.

Python prototype (for reference):

setWeights(weight_non_edge, weight_gradient_direction, weight_gradient_magnitude) -> retval

Functions

applyImage!(self, image)

Raising version of applyImage/2.

applyImageFeatures!(self, non_edge, gradient_direction, gradient_magnitude)

Raising version of applyImageFeatures/4.

applyImageFeatures!(self, non_edge, gradient_direction, gradient_magnitude, opts)

Raising version of applyImageFeatures/5.

buildMap!(self, sourcePt)

Raising version of buildMap/2.

getContour!(self, targetPt)

Raising version of getContour/2.

getContour!(self, targetPt, opts)

Raising version of getContour/3.

segmentation_IntelligentScissorsMB!()

Raising version of segmentation_IntelligentScissorsMB/0.

setEdgeFeatureCannyParameters!(self, threshold1, threshold2)

Raising version of setEdgeFeatureCannyParameters/3.

setEdgeFeatureCannyParameters!(self, threshold1, threshold2, opts)

Raising version of setEdgeFeatureCannyParameters/4.

setEdgeFeatureZeroCrossingParameters!(self)

Raising version of setEdgeFeatureZeroCrossingParameters/1.

setGradientMagnitudeMaxLimit!(self)

Raising version of setGradientMagnitudeMaxLimit/1.

setWeights!(self, weight_non_edge, weight_gradient_direction, weight_gradient_magnitude)

Raising version of setWeights/4.