Evision.StereoBM (Evision v0.1.7-dev)

cv

compute(self, left, right)

Positional Arguments

left: UMat.

right: UMat.

Keyword Arguments

disparity: UMat.

Computes disparity map for the specified stereo pair

like StereoBM or StereoSGBM compute 16-bit fixed-point disparity map (where each disparity value
has 4 fractional bits), whereas other algorithms output 32-bit floating-point disparity map.

Python prototype (for reference):

compute(left, right[, disparity]) -> disparity

compute(self, left, right, opts)

Positional Arguments

left: UMat.

right: UMat.

Keyword Arguments

disparity: UMat.

Computes disparity map for the specified stereo pair

like StereoBM or StereoSGBM compute 16-bit fixed-point disparity map (where each disparity value
has 4 fractional bits), whereas other algorithms output 32-bit floating-point disparity map.

Python prototype (for reference):

compute(left, right[, disparity]) -> disparity

create()

Positional Arguments

Keyword Arguments

numDisparities: int.

blockSize: int.

Creates StereoBM object

disparity from 0 (default minimum disparity) to numDisparities. The search range can then be
shifted by changing the minimum disparity.
(as the block is centered at the current pixel). Larger block size implies smoother, though less
accurate disparity map. Smaller block size gives more detailed disparity map, but there is higher
chance for algorithm to find a wrong correspondence.

The function create StereoBM object. You can then call StereoBM::compute() to compute disparity for a specific stereo pair.

Python prototype (for reference):

create([, numDisparities[, blockSize]]) -> retval

create(opts)

Positional Arguments

Keyword Arguments

numDisparities: int.

blockSize: int.

Creates StereoBM object

disparity from 0 (default minimum disparity) to numDisparities. The search range can then be
shifted by changing the minimum disparity.
(as the block is centered at the current pixel). Larger block size implies smoother, though less
accurate disparity map. Smaller block size gives more detailed disparity map, but there is higher
chance for algorithm to find a wrong correspondence.

The function create StereoBM object. You can then call StereoBM::compute() to compute disparity for a specific stereo pair.

Python prototype (for reference):

create([, numDisparities[, blockSize]]) -> retval

getBlockSize(self)

Python prototype (for reference):

getBlockSize() -> retval

getDisp12MaxDiff(self)

Python prototype (for reference):

getDisp12MaxDiff() -> retval

getMinDisparity(self)

Python prototype (for reference):

getMinDisparity() -> retval

getNumDisparities(self)

Python prototype (for reference):

getNumDisparities() -> retval

getPreFilterCap(self)

Python prototype (for reference):

getPreFilterCap() -> retval

getPreFilterSize(self)

Python prototype (for reference):

getPreFilterSize() -> retval

getPreFilterType(self)

Python prototype (for reference):

getPreFilterType() -> retval

getROI1(self)

Python prototype (for reference):

getROI1() -> retval

getROI2(self)

Python prototype (for reference):

getROI2() -> retval

getSmallerBlockSize(self)

Python prototype (for reference):

getSmallerBlockSize() -> retval

getSpeckleRange(self)

Python prototype (for reference):

getSpeckleRange() -> retval

getSpeckleWindowSize(self)

Python prototype (for reference):

getSpeckleWindowSize() -> retval

getTextureThreshold(self)

Python prototype (for reference):

getTextureThreshold() -> retval

getUniquenessRatio(self)

Python prototype (for reference):

getUniquenessRatio() -> retval

setBlockSize(self, blockSize)

Python prototype (for reference):

setBlockSize(blockSize) -> None

setDisp12MaxDiff(self, disp12MaxDiff)

Python prototype (for reference):

setDisp12MaxDiff(disp12MaxDiff) -> None

setMinDisparity(self, minDisparity)

Python prototype (for reference):

setMinDisparity(minDisparity) -> None

setNumDisparities(self, numDisparities)

Python prototype (for reference):

setNumDisparities(numDisparities) -> None

setPreFilterCap(self, preFilterCap)

Python prototype (for reference):

setPreFilterCap(preFilterCap) -> None

setPreFilterSize(self, preFilterSize)

Python prototype (for reference):

setPreFilterSize(preFilterSize) -> None

setPreFilterType(self, preFilterType)

Python prototype (for reference):

setPreFilterType(preFilterType) -> None

setROI1(self, roi1)

Python prototype (for reference):

setROI1(roi1) -> None

setROI2(self, roi2)

Python prototype (for reference):

setROI2(roi2) -> None

setSmallerBlockSize(self, blockSize)

Python prototype (for reference):

setSmallerBlockSize(blockSize) -> None

setSpeckleRange(self, speckleRange)

Python prototype (for reference):

setSpeckleRange(speckleRange) -> None

setSpeckleWindowSize(self, speckleWindowSize)

Python prototype (for reference):

setSpeckleWindowSize(speckleWindowSize) -> None

setTextureThreshold(self, textureThreshold)

Python prototype (for reference):

setTextureThreshold(textureThreshold) -> None

setUniquenessRatio(self, uniquenessRatio)

Python prototype (for reference):

setUniquenessRatio(uniquenessRatio) -> None

Functions

compute!(self, left, right)

Raising version of compute/3.

compute!(self, left, right, opts)

Raising version of compute/4.

create!()

Raising version of create/0.

create!(opts)

Raising version of create/1.

getBlockSize!(self)

Raising version of getBlockSize/1.

getDisp12MaxDiff!(self)

Raising version of getDisp12MaxDiff/1.

getMinDisparity!(self)

Raising version of getMinDisparity/1.

getNumDisparities!(self)

Raising version of getNumDisparities/1.

getPreFilterCap!(self)

Raising version of getPreFilterCap/1.

getPreFilterSize!(self)

Raising version of getPreFilterSize/1.

getPreFilterType!(self)

Raising version of getPreFilterType/1.

getROI1!(self)

Raising version of getROI1/1.

getROI2!(self)

Raising version of getROI2/1.

getSmallerBlockSize!(self)

Raising version of getSmallerBlockSize/1.

getSpeckleRange!(self)

Raising version of getSpeckleRange/1.

getSpeckleWindowSize!(self)

Raising version of getSpeckleWindowSize/1.

getTextureThreshold!(self)

Raising version of getTextureThreshold/1.

getUniquenessRatio!(self)

Raising version of getUniquenessRatio/1.

setBlockSize!(self, blockSize)

Raising version of setBlockSize/2.

setDisp12MaxDiff!(self, disp12MaxDiff)

Raising version of setDisp12MaxDiff/2.

setMinDisparity!(self, minDisparity)

Raising version of setMinDisparity/2.

setNumDisparities!(self, numDisparities)

Raising version of setNumDisparities/2.

setPreFilterCap!(self, preFilterCap)

Raising version of setPreFilterCap/2.

setPreFilterSize!(self, preFilterSize)

Raising version of setPreFilterSize/2.

setPreFilterType!(self, preFilterType)

Raising version of setPreFilterType/2.

setROI1!(self, roi1)

Raising version of setROI1/2.

setROI2!(self, roi2)

Raising version of setROI2/2.

setSmallerBlockSize!(self, blockSize)

Raising version of setSmallerBlockSize/2.

setSpeckleRange!(self, speckleRange)

Raising version of setSpeckleRange/2.

setSpeckleWindowSize!(self, speckleWindowSize)

Raising version of setSpeckleWindowSize/2.

setTextureThreshold!(self, textureThreshold)

Raising version of setTextureThreshold/2.

setUniquenessRatio!(self, uniquenessRatio)

Raising version of setUniquenessRatio/2.