Evision.StereoSGBM (Evision v0.1.7-dev)

cv

compute(self, left, right)

Positional Arguments

left: UMat.

right: UMat.

Keyword Arguments

disparity: UMat.

Computes disparity map for the specified stereo pair

like StereoBM or StereoSGBM compute 16-bit fixed-point disparity map (where each disparity value
has 4 fractional bits), whereas other algorithms output 32-bit floating-point disparity map.

Python prototype (for reference):

compute(left, right[, disparity]) -> disparity

compute(self, left, right, opts)

Positional Arguments

left: UMat.

right: UMat.

Keyword Arguments

disparity: UMat.

Computes disparity map for the specified stereo pair

like StereoBM or StereoSGBM compute 16-bit fixed-point disparity map (where each disparity value
has 4 fractional bits), whereas other algorithms output 32-bit floating-point disparity map.

Python prototype (for reference):

compute(left, right[, disparity]) -> disparity

create()

Positional Arguments

Keyword Arguments

minDisparity: int.

numDisparities: int.

blockSize: int.

p1: int.

p2: int.

disp12MaxDiff: int.

preFilterCap: int.

uniquenessRatio: int.

speckleWindowSize: int.

speckleRange: int.

mode: int.

Creates StereoSGBM object

rectification algorithms can shift images, so this parameter needs to be adjusted accordingly.
zero. In the current implementation, this parameter must be divisible by 16.
somewhere in the 3..11 range.
the smoother the disparity is. P1 is the penalty on the disparity change by plus or minus 1
between neighbor pixels. P2 is the penalty on the disparity change by more than 1 between neighbor
pixels. The algorithm requires P2 \> P1 . See stereo_match.cpp sample where some reasonably good
P1 and P2 values are shown (like 8\*number_of_image_channels\*blockSize\*blockSize and
32\*number_of_image_channels\*blockSize\*blockSize , respectively).
disparity check. Set it to a non-positive value to disable the check.
computes x-derivative at each pixel and clips its value by [-preFilterCap, preFilterCap] interval.
The result values are passed to the Birchfield-Tomasi pixel cost function.
value should "win" the second best value to consider the found match correct. Normally, a value
within the 5-15 range is good enough.
and invalidate. Set it to 0 to disable speckle filtering. Otherwise, set it somewhere in the
50-200 range.
filtering, set the parameter to a positive value, it will be implicitly multiplied by 16.
Normally, 1 or 2 is good enough.
algorithm. It will consume O(W\*H\*numDisparities) bytes, which is large for 640x480 stereo and
huge for HD-size pictures. By default, it is set to false .

The first constructor initializes StereoSGBM with all the default parameters. So, you only have to set StereoSGBM::numDisparities at minimum. The second constructor enables you to set each parameter to a custom value.

Python prototype (for reference):

create([, minDisparity[, numDisparities[, blockSize[, P1[, P2[, disp12MaxDiff[, preFilterCap[, uniquenessRatio[, speckleWindowSize[, speckleRange[, mode]]]]]]]]]]]) -> retval

create(opts)

Positional Arguments

Keyword Arguments

minDisparity: int.

numDisparities: int.

blockSize: int.

p1: int.

p2: int.

disp12MaxDiff: int.

preFilterCap: int.

uniquenessRatio: int.

speckleWindowSize: int.

speckleRange: int.

mode: int.

Creates StereoSGBM object

rectification algorithms can shift images, so this parameter needs to be adjusted accordingly.
zero. In the current implementation, this parameter must be divisible by 16.
somewhere in the 3..11 range.
the smoother the disparity is. P1 is the penalty on the disparity change by plus or minus 1
between neighbor pixels. P2 is the penalty on the disparity change by more than 1 between neighbor
pixels. The algorithm requires P2 \> P1 . See stereo_match.cpp sample where some reasonably good
P1 and P2 values are shown (like 8\*number_of_image_channels\*blockSize\*blockSize and
32\*number_of_image_channels\*blockSize\*blockSize , respectively).
disparity check. Set it to a non-positive value to disable the check.
computes x-derivative at each pixel and clips its value by [-preFilterCap, preFilterCap] interval.
The result values are passed to the Birchfield-Tomasi pixel cost function.
value should "win" the second best value to consider the found match correct. Normally, a value
within the 5-15 range is good enough.
and invalidate. Set it to 0 to disable speckle filtering. Otherwise, set it somewhere in the
50-200 range.
filtering, set the parameter to a positive value, it will be implicitly multiplied by 16.
Normally, 1 or 2 is good enough.
algorithm. It will consume O(W\*H\*numDisparities) bytes, which is large for 640x480 stereo and
huge for HD-size pictures. By default, it is set to false .

The first constructor initializes StereoSGBM with all the default parameters. So, you only have to set StereoSGBM::numDisparities at minimum. The second constructor enables you to set each parameter to a custom value.

Python prototype (for reference):

create([, minDisparity[, numDisparities[, blockSize[, P1[, P2[, disp12MaxDiff[, preFilterCap[, uniquenessRatio[, speckleWindowSize[, speckleRange[, mode]]]]]]]]]]]) -> retval

getBlockSize(self)

Python prototype (for reference):

getBlockSize() -> retval

getDisp12MaxDiff(self)

Python prototype (for reference):

getDisp12MaxDiff() -> retval

getMinDisparity(self)

Python prototype (for reference):

getMinDisparity() -> retval

getMode(self)

Python prototype (for reference):

getMode() -> retval

getNumDisparities(self)

Python prototype (for reference):

getNumDisparities() -> retval

getP1(self)

Python prototype (for reference):

getP1() -> retval

getP2(self)

Python prototype (for reference):

getP2() -> retval

getPreFilterCap(self)

Python prototype (for reference):

getPreFilterCap() -> retval

getSpeckleRange(self)

Python prototype (for reference):

getSpeckleRange() -> retval

getSpeckleWindowSize(self)

Python prototype (for reference):

getSpeckleWindowSize() -> retval

getUniquenessRatio(self)

Python prototype (for reference):

getUniquenessRatio() -> retval

setBlockSize(self, blockSize)

Python prototype (for reference):

setBlockSize(blockSize) -> None

setDisp12MaxDiff(self, disp12MaxDiff)

Python prototype (for reference):

setDisp12MaxDiff(disp12MaxDiff) -> None

setMinDisparity(self, minDisparity)

Python prototype (for reference):

setMinDisparity(minDisparity) -> None

setMode(self, mode)

Python prototype (for reference):

setMode(mode) -> None

setNumDisparities(self, numDisparities)

Python prototype (for reference):

setNumDisparities(numDisparities) -> None

setP1(self, p1)

Python prototype (for reference):

setP1(P1) -> None

setP2(self, p2)

Python prototype (for reference):

setP2(P2) -> None

setPreFilterCap(self, preFilterCap)

Python prototype (for reference):

setPreFilterCap(preFilterCap) -> None

setSpeckleRange(self, speckleRange)

Python prototype (for reference):

setSpeckleRange(speckleRange) -> None

setSpeckleWindowSize(self, speckleWindowSize)

Python prototype (for reference):

setSpeckleWindowSize(speckleWindowSize) -> None

setUniquenessRatio(self, uniquenessRatio)

Python prototype (for reference):

setUniquenessRatio(uniquenessRatio) -> None

Functions

compute!(self, left, right)

Raising version of compute/3.

compute!(self, left, right, opts)

Raising version of compute/4.

create!()

Raising version of create/0.

create!(opts)

Raising version of create/1.

getBlockSize!(self)

Raising version of getBlockSize/1.

getDisp12MaxDiff!(self)

Raising version of getDisp12MaxDiff/1.

getMinDisparity!(self)

Raising version of getMinDisparity/1.

getMode!(self)

Raising version of getMode/1.

getNumDisparities!(self)

Raising version of getNumDisparities/1.

getP1!(self)

Raising version of getP1/1.

getP2!(self)

Raising version of getP2/1.

getPreFilterCap!(self)

Raising version of getPreFilterCap/1.

getSpeckleRange!(self)

Raising version of getSpeckleRange/1.

getSpeckleWindowSize!(self)

Raising version of getSpeckleWindowSize/1.

getUniquenessRatio!(self)

Raising version of getUniquenessRatio/1.

setBlockSize!(self, blockSize)

Raising version of setBlockSize/2.

setDisp12MaxDiff!(self, disp12MaxDiff)

Raising version of setDisp12MaxDiff/2.

setMinDisparity!(self, minDisparity)

Raising version of setMinDisparity/2.

setMode!(self, mode)

Raising version of setMode/2.

setNumDisparities!(self, numDisparities)

Raising version of setNumDisparities/2.

setP1!(self, p1)

Raising version of setP1/2.

setP2!(self, p2)

Raising version of setP2/2.

setPreFilterCap!(self, preFilterCap)

Raising version of setPreFilterCap/2.

setSpeckleRange!(self, speckleRange)

Raising version of setSpeckleRange/2.

setSpeckleWindowSize!(self, speckleWindowSize)

Raising version of setSpeckleWindowSize/2.

setUniquenessRatio!(self, uniquenessRatio)

Raising version of setUniquenessRatio/2.