Evision.Stitcher (Evision v0.1.7-dev)

cv

composePanorama(self)

Positional Arguments

Keyword Arguments

pano: UMat.

Has overloading in C++

Python prototype (for reference):

composePanorama([, pano]) -> retval, pano

composePanorama(self, images)

Positional Arguments

images: vector_UMat.

Keyword Arguments

pano: UMat.

These functions try to compose the given images (or images stored internally from the other function calls) into the final pano under the assumption that the image transformations were estimated before. Note: Use the functions only if you're aware of the stitching pipeline, otherwise use Stitcher::stitch. @return Status code.

Python prototype (for reference):

composePanorama(images[, pano]) -> retval, pano

composePanorama(self, images, opts)

Positional Arguments

images: vector_UMat.

Keyword Arguments

pano: UMat.

These functions try to compose the given images (or images stored internally from the other function calls) into the final pano under the assumption that the image transformations were estimated before. Note: Use the functions only if you're aware of the stitching pipeline, otherwise use Stitcher::stitch. @return Status code.

Python prototype (for reference):

composePanorama(images[, pano]) -> retval, pano

compositingResol(self)

Python prototype (for reference):

compositingResol() -> retval

create()

Positional Arguments

Keyword Arguments

mode: Mode.

Creates a Stitcher configured in one of the stitching modes.

to stitch and their transformation. Default parameters will be chosen for operation in given
scenario.

@return Stitcher class instance.

Python prototype (for reference):

create([, mode]) -> retval

create(opts)

Positional Arguments

Keyword Arguments

mode: Mode.

Creates a Stitcher configured in one of the stitching modes.

to stitch and their transformation. Default parameters will be chosen for operation in given
scenario.

@return Stitcher class instance.

Python prototype (for reference):

create([, mode]) -> retval

estimateTransform(self, images)

Positional Arguments

images: vector_UMat.

Keyword Arguments

masks: vector_UMat.

These functions try to match the given images and to estimate rotations of each camera. Note: Use the functions only if you're aware of the stitching pipeline, otherwise use Stitcher::stitch. @return Status code.

Python prototype (for reference):

estimateTransform(images[, masks]) -> retval

estimateTransform(self, images, opts)

Positional Arguments

images: vector_UMat.

Keyword Arguments

masks: vector_UMat.

These functions try to match the given images and to estimate rotations of each camera. Note: Use the functions only if you're aware of the stitching pipeline, otherwise use Stitcher::stitch. @return Status code.

Python prototype (for reference):

estimateTransform(images[, masks]) -> retval

interpolationFlags(self)

Python prototype (for reference):

interpolationFlags() -> retval

panoConfidenceThresh(self)

Python prototype (for reference):

panoConfidenceThresh() -> retval

registrationResol(self)

Python prototype (for reference):

registrationResol() -> retval

seamEstimationResol(self)

Python prototype (for reference):

seamEstimationResol() -> retval

setCompositingResol(self, resol_mpx)

Python prototype (for reference):

setCompositingResol(resol_mpx) -> None

setInterpolationFlags(self, interp_flags)

Python prototype (for reference):

setInterpolationFlags(interp_flags) -> None

setPanoConfidenceThresh(self, conf_thresh)

Python prototype (for reference):

setPanoConfidenceThresh(conf_thresh) -> None

setRegistrationResol(self, resol_mpx)

Python prototype (for reference):

setRegistrationResol(resol_mpx) -> None

setSeamEstimationResol(self, resol_mpx)

Python prototype (for reference):

setSeamEstimationResol(resol_mpx) -> None

setWaveCorrection(self, flag)

Python prototype (for reference):

setWaveCorrection(flag) -> None

stitch(self, images)

Positional Arguments

images: vector_UMat

Keyword Arguments

pano: UMat.

Has overloading in C++

Python prototype (for reference):

stitch(images[, pano]) -> retval, pano

stitch(self, images, opts)

Variaint 1:

Positional Arguments

images: vector_UMat.

masks: vector_UMat.

Keyword Arguments

pano: UMat.

These functions try to stitch the given images. @return Status code.

Python prototype (for reference):

stitch(images, masks[, pano]) -> retval, pano

Variaint 2:

Positional Arguments

images: vector_UMat

Keyword Arguments

pano: UMat.

Has overloading in C++

Python prototype (for reference):

stitch(images[, pano]) -> retval, pano

stitch(self, images, masks, opts)

Positional Arguments

images: vector_UMat.

masks: vector_UMat.

Keyword Arguments

pano: UMat.

These functions try to stitch the given images. @return Status code.

Python prototype (for reference):

stitch(images, masks[, pano]) -> retval, pano

waveCorrection(self)

Python prototype (for reference):

waveCorrection() -> retval

workScale(self)

Python prototype (for reference):

workScale() -> retval

Functions

composePanorama!(self)

Raising version of composePanorama/1.

composePanorama!(self, images)

Raising version of composePanorama/2.

composePanorama!(self, images, opts)

Raising version of composePanorama/3.

compositingResol!(self)

Raising version of compositingResol/1.

create!()

Raising version of create/0.

create!(opts)

Raising version of create/1.

estimateTransform!(self, images)

Raising version of estimateTransform/2.

estimateTransform!(self, images, opts)

Raising version of estimateTransform/3.

interpolationFlags!(self)

Raising version of interpolationFlags/1.

panoConfidenceThresh!(self)

Raising version of panoConfidenceThresh/1.

registrationResol!(self)

Raising version of registrationResol/1.

seamEstimationResol!(self)

Raising version of seamEstimationResol/1.

setCompositingResol!(self, resol_mpx)

Raising version of setCompositingResol/2.

setInterpolationFlags!(self, interp_flags)

Raising version of setInterpolationFlags/2.

setPanoConfidenceThresh!(self, conf_thresh)

Raising version of setPanoConfidenceThresh/2.

setRegistrationResol!(self, resol_mpx)

Raising version of setRegistrationResol/2.

setSeamEstimationResol!(self, resol_mpx)

Raising version of setSeamEstimationResol/2.

setWaveCorrection!(self, flag)

Raising version of setWaveCorrection/2.

stitch!(self, images)

Raising version of stitch/2.

stitch!(self, images, masks)

Raising version of stitch/3.

stitch!(self, images, masks, opts)

Raising version of stitch/4.

waveCorrection!(self)

Raising version of waveCorrection/1.

workScale!(self)

Raising version of workScale/1.