Evision.VideoCapture (Evision v0.1.7-dev)

cv

get(self, propId)

Returns the specified VideoCapture property

or one from @ref videoio_flags_others

@return Value for the specified property. Value 0 is returned when querying a property that is not supported by the backend used by the VideoCapture instance. Note: Reading / writing properties involves many layers. Some unexpected result might happens along this chain.

VideoCapture -> API Backend -> Operating System -> Device Driver -> Device Hardware

The returned value might be different from what really used by the device or it could be encoded using device dependent rules (eg. steps or percentage). Effective behaviour depends from device driver and API Backend

Python prototype (for reference):

get(propId) -> retval

getBackendName(self)

Returns used backend API name Note: Stream should be opened.

Python prototype (for reference):

getBackendName() -> retval

getExceptionMode(self)

Python prototype (for reference):

getExceptionMode() -> retval

grab(self)

Grabs the next frame from video file or capturing device. @return true (non-zero) in the case of success. The method/function grabs the next frame from video file or camera and returns true (non-zero) in the case of success. The primary use of the function is in multi-camera environments, especially when the cameras do not have hardware synchronization. That is, you call VideoCapture::grab() for each camera and after that call the slower method VideoCapture::retrieve() to decode and get frame from each camera. This way the overhead on demosaicing or motion jpeg decompression etc. is eliminated and the retrieved frames from different cameras will be closer in time. Also, when a connected camera is multi-head (for example, a stereo camera or a Kinect device), the correct way of retrieving data from it is to call VideoCapture::grab() first and then call VideoCapture::retrieve() one or more times with different values of the channel parameter. @ref tutorial_kinect_openni

Python prototype (for reference):

grab() -> retval

isOpened(self)

Returns true if video capturing has been initialized already. If the previous call to VideoCapture constructor or VideoCapture::open() succeeded, the method returns true.

Python prototype (for reference):

isOpened() -> retval

open(self, index)

Variaint 1:

Positional Arguments

index: int

Keyword Arguments

apiPreference: int.

Opens a camera for video capturing Has overloading in C++

Parameters are same as the constructor VideoCapture(int index, int apiPreference = CAP_ANY) @return true if the camera has been successfully opened. The method first calls VideoCapture::release to close the already opened file or camera.

Python prototype (for reference):

open(index[, apiPreference]) -> retval

Variaint 2:

Positional Arguments

filename: String

Keyword Arguments

apiPreference: int.

Opens a video file or a capturing device or an IP video stream for video capturing. Has overloading in C++

Parameters are same as the constructor VideoCapture(const String& filename, int apiPreference = CAP_ANY) @return true if the file has been successfully opened The method first calls VideoCapture::release to close the already opened file or camera.

Python prototype (for reference):

open(filename[, apiPreference]) -> retval

open(self, index, opts)

Variaint 1:

Positional Arguments

index: int

Keyword Arguments

apiPreference: int.

Opens a camera for video capturing Has overloading in C++

Parameters are same as the constructor VideoCapture(int index, int apiPreference = CAP_ANY) @return true if the camera has been successfully opened. The method first calls VideoCapture::release to close the already opened file or camera.

Python prototype (for reference):

open(index[, apiPreference]) -> retval

Variaint 2:

Positional Arguments

filename: String

Keyword Arguments

apiPreference: int.

Opens a video file or a capturing device or an IP video stream for video capturing. Has overloading in C++

Parameters are same as the constructor VideoCapture(const String& filename, int apiPreference = CAP_ANY) @return true if the file has been successfully opened The method first calls VideoCapture::release to close the already opened file or camera.

Python prototype (for reference):

open(filename[, apiPreference]) -> retval

open(self, index, apiPreference, params)

Variaint 1:

Opens a camera for video capturing with API Preference and parameters Has overloading in C++

The params parameter allows to specify extra parameters encoded as pairs (paramId_1, paramValue_1, paramId_2, paramValue_2, ...). See cv::VideoCaptureProperties @return true if the camera has been successfully opened. The method first calls VideoCapture::release to close the already opened file or camera.

Python prototype (for reference):

open(index, apiPreference, params) -> retval

Variaint 2:

Opens a video file or a capturing device or an IP video stream for video capturing with API Preference and parameters Has overloading in C++

The params parameter allows to specify extra parameters encoded as pairs (paramId_1, paramValue_1, paramId_2, paramValue_2, ...). See cv::VideoCaptureProperties @return true if the file has been successfully opened The method first calls VideoCapture::release to close the already opened file or camera.

Python prototype (for reference):

open(filename, apiPreference, params) -> retval

read(self)

Positional Arguments

Keyword Arguments

image: UMat.

Grabs, decodes and returns the next video frame. @return false if no frames has been grabbed The method/function combines VideoCapture::grab() and VideoCapture::retrieve() in one call. This is the most convenient method for reading video files or capturing data from decode and returns the just grabbed frame. If no frames has been grabbed (camera has been disconnected, or there are no more frames in video file), the method returns false and the function returns empty image (with %cv::Mat, test it with Mat::empty()). Note: In @ref videoio_c "C API", functions cvRetrieveFrame() and cv.RetrieveFrame() return image stored inside the video capturing structure. It is not allowed to modify or release the image! You can copy the frame using cvCloneImage and then do whatever you want with the copy.

Python prototype (for reference):

read([, image]) -> retval, image

release(self)

Closes video file or capturing device. The method is automatically called by subsequent VideoCapture::open and by VideoCapture destructor. The C function also deallocates memory and clears *capture pointer.

Python prototype (for reference):

release() -> None

retrieve(self)

Positional Arguments

Keyword Arguments

image: UMat.

flag: int.

Decodes and returns the grabbed video frame. @return false if no frames has been grabbed The method decodes and returns the just grabbed frame. If no frames has been grabbed (camera has been disconnected, or there are no more frames in video file), the method returns false and the function returns an empty image (with %cv::Mat, test it with Mat::empty()). @sa read() Note: In @ref videoio_c "C API", functions cvRetrieveFrame() and cv.RetrieveFrame() return image stored inside the video capturing structure. It is not allowed to modify or release the image! You can copy the frame using cvCloneImage and then do whatever you want with the copy.

Python prototype (for reference):

retrieve([, image[, flag]]) -> retval, image

set(self, propId, value)

Sets a property in the VideoCapture.

or one from @ref videoio_flags_others

@return true if the property is supported by backend used by the VideoCapture instance. Note: Even if it returns true this doesn't ensure that the property value has been accepted by the capture device. See note in VideoCapture::get()

Python prototype (for reference):

set(propId, value) -> retval

setExceptionMode(self, enable)

Switches exceptions mode methods raise exceptions if not successful instead of returning an error code

Python prototype (for reference):

setExceptionMode(enable) -> None

videoCapture()

Default constructor Note: In @ref videoio_c "C API", when you finished working with video, release CvCapture structure with cvReleaseCapture(), or use Ptr\<CvCapture> that calls cvReleaseCapture() automatically in the destructor.

Python prototype (for reference):

VideoCapture() -> <VideoCapture object>

videoCapture(index)

Variaint 1:

Positional Arguments

index: int.

Keyword Arguments

apiPreference: int.

Has overloading in C++

Opens a camera for video capturing

(to backward compatibility usage of camera_id + domain_offset (CAP_*) is valid when apiPreference is CAP_ANY)
implementation if multiple are available: e.g. cv::CAP_DSHOW or cv::CAP_MSMF or cv::CAP_V4L.

@sa cv::VideoCaptureAPIs

Python prototype (for reference):

VideoCapture(index[, apiPreference]) -> <VideoCapture object>

Variaint 2:

Positional Arguments

filename: String.

Keyword Arguments

apiPreference: int.

Has overloading in C++

Opens a video file or a capturing device or an IP video stream for video capturing with API Preference

- name of video file (eg. `video.avi`)
- or image sequence (eg. `img_%02d.jpg`, which will read samples like `img_00.jpg, img_01.jpg, img_02.jpg, ...`)
- or URL of video stream (eg. `protocol://host:port/script_name?script_params|auth`)
- or GStreamer pipeline string in gst-launch tool format in case if GStreamer is used as backend
Note that each video stream or IP camera feed has its own URL scheme. Please refer to the
documentation of source stream to know the right URL.
implementation if multiple are available: e.g. cv::CAP_FFMPEG or cv::CAP_IMAGES or cv::CAP_DSHOW.

@sa cv::VideoCaptureAPIs

Python prototype (for reference):

VideoCapture(filename[, apiPreference]) -> <VideoCapture object>

videoCapture(index, opts)

Variaint 1:

Positional Arguments

index: int.

Keyword Arguments

apiPreference: int.

Has overloading in C++

Opens a camera for video capturing

(to backward compatibility usage of camera_id + domain_offset (CAP_*) is valid when apiPreference is CAP_ANY)
implementation if multiple are available: e.g. cv::CAP_DSHOW or cv::CAP_MSMF or cv::CAP_V4L.

@sa cv::VideoCaptureAPIs

Python prototype (for reference):

VideoCapture(index[, apiPreference]) -> <VideoCapture object>

Variaint 2:

Positional Arguments

filename: String.

Keyword Arguments

apiPreference: int.

Has overloading in C++

Opens a video file or a capturing device or an IP video stream for video capturing with API Preference

- name of video file (eg. `video.avi`)
- or image sequence (eg. `img_%02d.jpg`, which will read samples like `img_00.jpg, img_01.jpg, img_02.jpg, ...`)
- or URL of video stream (eg. `protocol://host:port/script_name?script_params|auth`)
- or GStreamer pipeline string in gst-launch tool format in case if GStreamer is used as backend
Note that each video stream or IP camera feed has its own URL scheme. Please refer to the
documentation of source stream to know the right URL.
implementation if multiple are available: e.g. cv::CAP_FFMPEG or cv::CAP_IMAGES or cv::CAP_DSHOW.

@sa cv::VideoCaptureAPIs

Python prototype (for reference):

VideoCapture(filename[, apiPreference]) -> <VideoCapture object>

videoCapture(index, apiPreference, params)

Variaint 1:

Has overloading in C++

Opens a camera for video capturing with API Preference and parameters The params parameter allows to specify extra parameters encoded as pairs (paramId_1, paramValue_1, paramId_2, paramValue_2, ...). See cv::VideoCaptureProperties

Python prototype (for reference):

VideoCapture(index, apiPreference, params) -> <VideoCapture object>

Variaint 2:

Has overloading in C++

Opens a video file or a capturing device or an IP video stream for video capturing with API Preference and parameters The params parameter allows to specify extra parameters encoded as pairs (paramId_1, paramValue_1, paramId_2, paramValue_2, ...). See cv::VideoCaptureProperties

Python prototype (for reference):

VideoCapture(filename, apiPreference, params) -> <VideoCapture object>

Functions

get!(self, propId)

Raising version of get/2.

getBackendName!(self)

Raising version of getBackendName/1.

getExceptionMode!(self)

Raising version of getExceptionMode/1.

grab!(self)

Raising version of grab/1.

isOpened!(self)

Raising version of isOpened/1.

open!(self, index)

Raising version of open/2.

open!(self, index, opts)

Raising version of open/3.

open!(self, index, apiPreference, params)

Raising version of open/4.

read!(self)

Raising version of read/1.

release!(self)

Raising version of release/1.

retrieve!(self)

Raising version of retrieve/1.

set!(self, propId, value)

Raising version of set/3.

setExceptionMode!(self, enable)

Raising version of setExceptionMode/2.

videoCapture!()

Raising version of videoCapture/0.

videoCapture!(index)

Raising version of videoCapture/1.

videoCapture!(index, opts)

Raising version of videoCapture/2.

videoCapture!(index, apiPreference, params)

Raising version of videoCapture/3.