OpenCV.BackgroundSubtractorMOG2 (Evision v0.1.0-dev)

cv

apply(self, image)

apply(self, image, opts)

Computes a foreground mask.

image.

fgmask.

learningRate. learnt. Negative parameter value makes the algorithm to use some automatically chosen learning rate. 0 means that the background model is not updated at all, 1 means that the background model is completely reinitialized from the last frame.

fgmask : UMat

learningRate : double

Python prototype (for reference):

apply(image[, fgmask[, learningRate]]) -> fgmask

getBackgroundRatio(self)

getComplexityReductionThreshold(self)

getDetectShadows(self)

getHistory(self)

getNMixtures(self)

getShadowThreshold(self)

getShadowValue(self)

getVarInit(self)

getVarMax(self)

getVarMin(self)

getVarThreshold(self)

getVarThresholdGen(self)

setBackgroundRatio(self, ratio)

setComplexityReductionThreshold(self, ct)

setDetectShadows(self, detectShadows)

setHistory(self, history)

setNMixtures(self, nmixtures)

setShadowThreshold(self, threshold)

setShadowValue(self, value)

setVarInit(self, varInit)

setVarMax(self, varMax)

setVarMin(self, varMin)

setVarThreshold(self, varThreshold)

setVarThresholdGen(self, varThresholdGen)