OpenCV.CudaGpuMat (Evision v0.1.0-dev)

cv.cuda

adjustroi(self, dtop, dbottom, dleft, dright)

assignto(self, m)

assignto(self, m, opts)

type : int

Python prototype (for reference):

assignTo(m[, type]) -> None

channels(self)

clone(self)

col(self, x)

colrange(self, r)

colrange(self, startcol, endcol)

convertto(self, rtype, stream)

convertto(self, rtype, alpha, stream)

convertto(self, rtype, alpha, beta, stream)

convertto(self, rtype, alpha, beta, stream, opts)

dst : UMat

Python prototype (for reference):

convertTo(rtype, alpha, beta, stream[, dst]) -> dst

copyto(self, mask, stream)

copyto(self, mask, stream, opts)

dst : UMat

Python prototype (for reference):

copyTo(mask, stream[, dst]) -> dst

create(self, size, type)

create(self, rows, cols, type)

cuda_gpumat(arr)

cuda_gpumat(size, type)

cuda_gpumat(size, type, s)

cuda_gpumat(rows, cols, type, s)

cuda_gpumat(rows, cols, type, s, opts)

allocator : GpuMat_Allocator*

Python prototype (for reference):

GpuMat(rows, cols, type, s[, allocator]) -> <cuda_GpuMat object>

cudaptr(self)

defaultallocator()

depth(self)

download(self, stream)

download(self, stream, opts)

Performs data download from GpuMat (Non-Blocking call) This function copies data from device memory to host memory. As being a non-blocking call, this function may return even if the copy operation is not finished. The copy operation may be overlapped with operations in other non-default streams if \p stream is not the default stream and \p dst is HostMem allocated with HostMem::PAGE_LOCKED option.

dst : UMat

Python prototype (for reference):

download(stream[, dst]) -> dst

elemsize(self)

elemsize1(self)

empty(self)

iscontinuous(self)

locateroi(self, wholeSize, ofs)

reshape(self, cn)

reshape(self, cn, opts)

rows : int

Python prototype (for reference):

reshape(cn[, rows]) -> retval

row(self, y)

rowrange(self, r)

rowrange(self, startrow, endrow)

setdefaultallocator(allocator)

setto(self, s, mask, stream)

size(self)

step1(self)

swap(self, mat)

type(self)

updatecontinuityflag(self)

upload(self, arr, stream)

Functions

step(self)