OpenCV.DnnNet (Evision v0.1.0-dev)

Functions

connect(self, outPin, inpPin)

dnn_net()

dump(self)

dumptofile(self, path)

empty(self)

enablefusion(self, fusion)

forward(self, opt \\ [])

Runs forward pass to compute outputs of layers listed in @p outBlobNames.

outputBlobs.

outBlobNames.

outputBlobs : vector_UMat

Python prototype (for reference):

forward(outBlobNames[, outputBlobs]) -> outputBlobs

forwardandretrieve(self, outBlobNames)

forwardasync(self)

getflops(self, netInputShapes)

getflops(self, layerId, netInputShapes)

getinputdetails(self)

getlayer(self, layerId)

getlayerid(self, layer)

getlayernames(self)

getlayerscount(self, layerType)

getlayershapes(self, opt \\ [])

Has overloading in C++

Python prototype (for reference):

getLayerShapes(netInputShapes, layerId) -> inLayerShapes, outLayerShapes

getlayersshapes(self, opt \\ [])

Returns input and output shapes for all layers in loaded model;

preliminary inferencing isn't necessary.

netInputShapes.

layersIds.

inLayersShapes. order is the same as in layersIds

outLayersShapes. order is the same as in layersIds

Python prototype (for reference):

getLayersShapes(netInputShapes) -> layersIds, inLayersShapes, outLayersShapes

getlayertypes(self)

getmemoryconsumption(self, netInputShape)

getmemoryconsumption(self, layerId, netInputShapes)

getoutputdetails(self)

getparam(self, layer)

getparam(self, layer, opts)

Returns parameter blob of the layer.

layer.

numParam. @see Layer::blobs

numParam : int

Python prototype (for reference):

getParam(layer[, numParam]) -> retval

getperfprofile(self)

getunconnectedoutlayers(self)

getunconnectedoutlayersnames(self)

quantize(self, calibData, inputsDtype, outputsDtype)

readfrommodeloptimizer(bufferModelConfig, bufferWeights)

sethalidescheduler(self, scheduler)

setinput(self, blob)

setinput(self, blob, opts)

Sets the new input value for the network

blob.

name.

scalefactor.

mean. @see connect(String, String) to know format of the descriptor. If scale or mean values are specified, a final input blob is computed as: \f[input(n,c,h,w) = scalefactor \times (blob(n,c,h,w) - mean_c)\f]

name : String

scalefactor : double

mean : Scalar

Python prototype (for reference):

setInput(blob[, name[, scalefactor[, mean]]]) -> None

setinputshape(self, inputName, shape)

setinputsnames(self, inputBlobNames)

setparam(self, layer, numParam, blob)

setpreferablebackend(self, backendId)

setpreferabletarget(self, targetId)