OpenCV.FileStorage (Evision v0.1.0-dev)

cv

endWriteStruct(self)

fileStorage()

fileStorage(filename, flags)

fileStorage(filename, flags, opts)

Has overloading in C++

@copydoc open()

encoding : String

Python prototype (for reference):

FileStorage(filename, flags[, encoding]) -> <FileStorage object>

getFirstTopLevelNode(self)

getFormat(self)

getNode(self, nodename)

isOpened(self)

open(self, filename, flags)

open(self, filename, flags, opts)

Opens a file. See description of parameters in FileStorage::FileStorage. The method calls FileStorage::release before opening the file.

filename. Extension of the file (.xml, .yml/.yaml or .json) determines its format (XML, YAML or JSON respectively). Also you can append .gz to work with compressed files, for example myHugeMatrix.xml.gz. If both FileStorage::WRITE and FileStorage::MEMORY flags are specified, source is used just to specify the output file format (e.g. mydata.xml, .yml etc.). A file name can also contain parameters. You can use this format, "*?base64" (e.g. "file.json?base64" (case sensitive)), as an alternative to FileStorage::BASE64 flag.

flags.

encoding. you should use 8-bit encoding instead of it.

encoding : String

Python prototype (for reference):

open(filename, flags[, encoding]) -> retval

release(self)

releaseAndGetString(self)

root(self)

startWriteStruct(self, name, flags)

startWriteStruct(self, name, flags, opts)

Starts to write a nested structure (sequence or a mapping).

name.

flags.

typeName. I.e. if the format has a specification for storing type information, this parameter is used.

typeName : String

Python prototype (for reference):

startWriteStruct(name, flags[, typeName]) -> None

write(self, name, val)

writeComment(self, comment)

writeComment(self, comment, opts)

Writes a comment. The function writes a comment into file storage. The comments are skipped when the storage is read.

comment.

append. Else if the comment is multi-line, or if it does not fit at the end of the current line, the comment starts a new line.

append : bool

Python prototype (for reference):

writeComment(comment[, append]) -> None