OpenCV.HOGDescriptor (Evision v0.1.0-dev)

cv

checkdetectorsize(self)

compute(self, img)

compute(self, img, opts)

Computes HOG descriptors of given image.

img.

descriptors.

winStride.

padding : Padding

locations.

winStride : Size

padding : Size

locations : vector_Point

Python prototype (for reference):

compute(img[, winStride[, padding[, locations]]]) -> descriptors

computegradient(self, img, grad, angleOfs)

computegradient(self, img, grad, angleOfs, opts)

Computes gradients and quantized gradient orientations.

img.

grad.

angleOfs.

paddingTL.

paddingBR.

paddingTL : Size

paddingBR : Size

Python prototype (for reference):

computeGradient(img, grad, angleOfs[, paddingTL[, paddingBR]]) -> grad, angleOfs

detect(self, img)

detect(self, img, opts)

Performs object detection without a multi-scale window.

img.

foundLocations.

weights.

hitThreshold. Usually it is 0 and should be specified in the detector coefficients (as the last free coefficient). But if the free coefficient is omitted (which is allowed), you can specify it manually here.

winStride.

padding : Padding

searchLocations.

hitThreshold : double

winStride : Size

padding : Size

searchLocations : vector_Point

Python prototype (for reference):

detect(img[, hitThreshold[, winStride[, padding[, searchLocations]]]]) -> foundLocations, weights

detectmultiscale(self, img)

detectmultiscale(self, img, opts)

Detects objects of different sizes in the input image. The detected objects are returned as a list of rectangles.

img.

foundLocations.

foundWeights.

hitThreshold. Usually it is 0 and should be specified in the detector coefficients (as the last free coefficient). But if the free coefficient is omitted (which is allowed), you can specify it manually here.

winStride.

padding : Padding

scale.

finalThreshold.

useMeanshiftGrouping.

hitThreshold : double

winStride : Size

padding : Size

scale : double

finalThreshold : double

useMeanshiftGrouping : bool

Python prototype (for reference):

detectMultiScale(img[, hitThreshold[, winStride[, padding[, scale[, finalThreshold[, useMeanshiftGrouping]]]]]]) -> foundLocations, foundWeights

getdaimlerpeopledetector()

getdefaultpeopledetector()

getdescriptorsize(self)

getwinsigma(self)

hogdescriptor()

hogdescriptor(filename)

hogdescriptor(winSize, blockSize, blockStride, cellSize, nbins)

hogdescriptor(winSize, blockSize, blockStride, cellSize, nbins, opts)

Has overloading in C++

winSize.

blockSize.

blockStride.

cellSize.

nbins.

derivAperture.

winSigma.

histogramNormType.

l2HysThreshold.

gammaCorrection.

nlevels.

signedGradient.

derivAperture : int

winSigma : double

histogramNormType : HOGDescriptor_HistogramNormType

l2HysThreshold : double

gammaCorrection : bool

nlevels : int

signedGradient : bool

Python prototype (for reference):

HOGDescriptor(_winSize, _blockSize, _blockStride, _cellSize, _nbins[, _derivAperture[, _winSigma[, _histogramNormType[, _L2HysThreshold[, _gammaCorrection[, _nlevels[, _signedGradient]]]]]]]) -> <HOGDescriptor object>

load(self, filename)

load(self, filename, opts)

loads HOGDescriptor parameters and coefficients for the linear SVM classifier from a file.

filename.

objname.

objname : String

Python prototype (for reference):

load(filename[, objname]) -> retval

save(self, filename)

save(self, filename, opts)

saves HOGDescriptor parameters and coefficients for the linear SVM classifier to a file

filename.

objname.

objname : String

Python prototype (for reference):

save(filename[, objname]) -> None

setsvmdetector(self, svmdetector)

Functions

blocksize(self)

blockstride(self)

cellsize(self)

derivaperture(self)

gammacorrection(self)

histogramnormtype(self)

l2hysthreshold(self)

nbins(self)

nlevels(self)

signedgradient(self)

svmdetector(self)

winsigma(self)

winsize(self)