OpenCV.MlDTrees (Evision v0.1.0-dev)

cv.ml

create()

getcvfolds(self)

getmaxcategories(self)

getmaxdepth(self)

getminsamplecount(self)

getpriors(self)

getregressionaccuracy(self)

gettruncateprunedtree(self)

getuse1serule(self)

getusesurrogates(self)

load(filepath)

load(filepath, opts)

Loads and creates a serialized DTrees from a file Use DTree::save to serialize and store an DTree to disk. Load the DTree from this file again, by calling this function with the path to the file. Optionally specify the node for the file containing the classifier

filepath.

nodeName.

nodeName : String

Python prototype (for reference):

load(filepath[, nodeName]) -> retval

setcvfolds(self, val)

setmaxcategories(self, val)

setmaxdepth(self, val)

setminsamplecount(self, val)

setpriors(self, val)

setregressionaccuracy(self, val)

settruncateprunedtree(self, val)

setuse1serule(self, val)

setusesurrogates(self, val)