OpenCV.MlKNearest (Evision v0.1.0-dev)

cv.ml

create()

findnearest(self, samples, k)

findnearest(self, samples, k, opts)

Finds the neighbors and predicts responses for input vectors.

samples. <number_of_samples> * k size.

k.

results. sample. It is a single-precision floating-point vector with <number_of_samples> elements.

neighborResponses. precision floating-point matrix of <number_of_samples> * k size.

dist. is a single-precision floating-point matrix of <number_of_samples> * k size. For each input vector (a row of the matrix samples), the method finds the k nearest neighbors. In case of regression, the predicted result is a mean value of the particular vector's neighbor responses. In case of classification, the class is determined by voting. For each input vector, the neighbors are sorted by their distances to the vector. In case of C++ interface you can use output pointers to empty matrices and the function will allocate memory itself. If only a single input vector is passed, all output matrices are optional and the predicted value is returned by the method. The function is parallelized with the TBB library.

results : UMat

neighborResponses : UMat

dist : UMat

Python prototype (for reference):

findNearest(samples, k[, results[, neighborResponses[, dist]]]) -> retval, results, neighborResponses, dist

getalgorithmtype(self)

getdefaultk(self)

getemax(self)

getisclassifier(self)

load(filepath)

setalgorithmtype(self, val)

setdefaultk(self, val)

setemax(self, val)

setisclassifier(self, val)