OpenCV.Stitcher (Evision v0.1.0-dev)

cv

composePanorama(self)

composePanorama(self, images)

composePanorama(self, images, opts)

These functions try to compose the given images (or images stored internally from the other function calls) into the final pano under the assumption that the image transformations were estimated before. Note: Use the functions only if you're aware of the stitching pipeline, otherwise use Stitcher::stitch.

images.

pano. @return Status code.

pano : UMat

Python prototype (for reference):

composePanorama(images[, pano]) -> retval, pano

compositingResol(self)

create()

create(opts)

Creates a Stitcher configured in one of the stitching modes.

mode. to stitch and their transformation. Default parameters will be chosen for operation in given scenario. @return Stitcher class instance.

mode : Mode

Python prototype (for reference):

create([, mode]) -> retval

estimateTransform(self, images)

estimateTransform(self, images, opts)

These functions try to match the given images and to estimate rotations of each camera. Note: Use the functions only if you're aware of the stitching pipeline, otherwise use Stitcher::stitch.

images.

masks. @return Status code.

masks : vector_UMat

Python prototype (for reference):

estimateTransform(images[, masks]) -> retval

interpolationFlags(self)

panoConfidenceThresh(self)

registrationResol(self)

seamEstimationResol(self)

setCompositingResol(self, resol_mpx)

setInterpolationFlags(self, interp_flags)

setPanoConfidenceThresh(self, conf_thresh)

setRegistrationResol(self, resol_mpx)

setSeamEstimationResol(self, resol_mpx)

setWaveCorrection(self, flag)

stitch(self, images)

stitch(self, images, masks)

stitch(self, images, masks, opts)

These functions try to stitch the given images.

images.

masks.

pano. @return Status code.

pano : UMat

Python prototype (for reference):

stitch(images, masks[, pano]) -> retval, pano

waveCorrection(self)

workScale(self)